% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpabridget.R
\name{AbridgedPacketCodec}
\alias{AbridgedPacketCodec}
\title{AbridgedPacketCodec}
\description{
An R6 class that implements the Abridged Packet Codec for encoding and decoding packets.
}
\section{Super class}{
\code{\link[telegramR:PacketCodec]{telegramR::PacketCodec}} -> \code{AbridgedPacketCodec}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tag}}{A raw byte representing the tag for the codec.}

\item{\code{obfuscate_tag}}{A raw vector used for obfuscation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbridgedPacketCodec-encode_packet}{\code{AbridgedPacketCodec$encode_packet()}}
\item \href{#method-AbridgedPacketCodec-read_packet}{\code{AbridgedPacketCodec$read_packet()}}
\item \href{#method-AbridgedPacketCodec-clone}{\code{AbridgedPacketCodec$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="PacketCodec" data-id="initialize"><a href='../../telegramR/html/PacketCodec.html#method-PacketCodec-initialize'><code>telegramR::PacketCodec$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbridgedPacketCodec-encode_packet"></a>}}
\if{latex}{\out{\hypertarget{method-AbridgedPacketCodec-encode_packet}{}}}
\subsection{Method \code{encode_packet()}}{
Encodes a packet by adding a length prefix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbridgedPacketCodec$encode_packet(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A raw vector representing the data to be encoded.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A raw vector containing the encoded packet with the length prefix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbridgedPacketCodec-read_packet"></a>}}
\if{latex}{\out{\hypertarget{method-AbridgedPacketCodec-read_packet}{}}}
\subsection{Method \code{read_packet()}}{
Reads and decodes a packet from a reader.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbridgedPacketCodec$read_packet(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{An object with a `readexactly` method to read a specific number of bytes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A promise that resolves to the decoded packet.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbridgedPacketCodec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbridgedPacketCodec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbridgedPacketCodec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
