% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{AcceptCallRequest}
\alias{AcceptCallRequest}
\title{AcceptCallRequest}
\description{
Telegram API type AcceptCallRequest
}
\details{
AcceptCallRequest R6 class

Represents AcceptCallRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{peer}}{Field.}

\item{\code{g_b}}{Field.}

\item{\code{protocol}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AcceptCallRequest-new}{\code{AcceptCallRequest$new()}}
\item \href{#method-AcceptCallRequest-to_list}{\code{AcceptCallRequest$to_list()}}
\item \href{#method-AcceptCallRequest-bytes}{\code{AcceptCallRequest$bytes()}}
\item \href{#method-AcceptCallRequest-from_reader}{\code{AcceptCallRequest$from_reader()}}
\item \href{#method-AcceptCallRequest-clone}{\code{AcceptCallRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptCallRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptCallRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an AcceptCallRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptCallRequest$new(peer, g_b, protocol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{Input phone call object.}

\item{\code{g_b}}{Raw bytes.}

\item{\code{protocol}}{Protocol object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptCallRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptCallRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptCallRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptCallRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptCallRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptCallRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptCallRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptCallRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptCallRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptCallRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptCallRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptCallRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
