% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{AcceptTermsOfServiceRequest}
\alias{AcceptTermsOfServiceRequest}
\title{AcceptTermsOfServiceRequest R6 class}
\description{
Accept terms of service (help.TermsOfService). Serializes a DataJSON id.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{id}}{TLObject-like TypeDataJSON (environment with to_bytes()/to_list() or raw)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AcceptTermsOfServiceRequest-new}{\code{AcceptTermsOfServiceRequest$new()}}
\item \href{#method-AcceptTermsOfServiceRequest-to_list}{\code{AcceptTermsOfServiceRequest$to_list()}}
\item \href{#method-AcceptTermsOfServiceRequest-to_bytes}{\code{AcceptTermsOfServiceRequest$to_bytes()}}
\item \href{#method-AcceptTermsOfServiceRequest-clone}{\code{AcceptTermsOfServiceRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptTermsOfServiceRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptTermsOfServiceRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptTermsOfServiceRequest$new(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{TLObject-like or raw
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptTermsOfServiceRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptTermsOfServiceRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptTermsOfServiceRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `_` discriminator and id as list (if possible)
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptTermsOfServiceRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptTermsOfServiceRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptTermsOfServiceRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptTermsOfServiceRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptTermsOfServiceRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptTermsOfServiceRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
