% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_messages.R
\name{AcceptUrlAuthRequest}
\alias{AcceptUrlAuthRequest}
\title{AcceptUrlAuthRequest}
\description{
Represents a request to accept URL auth. This class inherits from TLRequest.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{AcceptUrlAuthRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{The constructor ID.}

\item{\code{SUBCLASS_OF_ID}}{The subclass ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AcceptUrlAuthRequest-new}{\code{AcceptUrlAuthRequest$new()}}
\item \href{#method-AcceptUrlAuthRequest-resolve}{\code{AcceptUrlAuthRequest$resolve()}}
\item \href{#method-AcceptUrlAuthRequest-toDict}{\code{AcceptUrlAuthRequest$toDict()}}
\item \href{#method-AcceptUrlAuthRequest-bytes}{\code{AcceptUrlAuthRequest$bytes()}}
\item \href{#method-AcceptUrlAuthRequest-clone}{\code{AcceptUrlAuthRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptUrlAuthRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptUrlAuthRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the AcceptUrlAuthRequest object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptUrlAuthRequest$new(
  writeAllowed = NULL,
  peer = NULL,
  msgId = NULL,
  buttonId = NULL,
  url = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{writeAllowed}}{Whether write is allowed (optional).}

\item{\code{peer}}{The input peer (optional).}

\item{\code{msgId}}{The message ID (optional).}

\item{\code{buttonId}}{The button ID (optional).}

\item{\code{url}}{The URL (optional).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptUrlAuthRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptUrlAuthRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve the request using client and utils.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptUrlAuthRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The client object.}

\item{\code{utils}}{The utils object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptUrlAuthRequest-toDict"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptUrlAuthRequest-toDict}{}}}
\subsection{Method \code{toDict()}}{
Convert the object to a dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptUrlAuthRequest$toDict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptUrlAuthRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptUrlAuthRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize the object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptUrlAuthRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector of bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptUrlAuthRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptUrlAuthRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptUrlAuthRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
