% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AccessPointRule}
\alias{AccessPointRule}
\title{AccessPointRule}
\description{
Telegram API type AccessPointRule
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{AccessPointRule}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{phone_prefix_rules}}{A string representing phone prefix rules.}

\item{\code{dc_id}}{An integer representing the data center ID.}

\item{\code{ips}}{A list of TypeIpPort objects representing IPs.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AccessPointRule-new}{\code{AccessPointRule$new()}}
\item \href{#method-AccessPointRule-to_dict}{\code{AccessPointRule$to_dict()}}
\item \href{#method-AccessPointRule-bytes}{\code{AccessPointRule$bytes()}}
\item \href{#method-AccessPointRule-clone}{\code{AccessPointRule$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessPointRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-AccessPointRule-new}{}}}
\subsection{Method \code{new()}}{
Initializes the AccessPointRule with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessPointRule$new(phone_prefix_rules, dc_id, ips)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phone_prefix_rules}}{A string representing phone prefix rules.}

\item{\code{dc_id}}{An integer representing the data center ID.}

\item{\code{ips}}{A list of TypeIpPort objects representing IPs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessPointRule-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-AccessPointRule-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the AccessPointRule object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessPointRule$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the AccessPointRule object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessPointRule-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AccessPointRule-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the AccessPointRule object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessPointRule$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the AccessPointRule object.
Reads an AccessPointRule object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessPointRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AccessPointRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessPointRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
