% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{AccountMethods}
\alias{AccountMethods}
\title{Account Methods}
\description{
This file contains the `AccountMethods` class and the `TakeoutClient` class, which are used to manage account-related operations in the Telegram API.
}
\details{
The `AccountMethods` class provides methods for managing account-related operations, including takeout sessions.
The `TakeoutClient` class is used to handle the takeout session and its associated requests.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client}}{The client instance backing these methods.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AccountMethods-new}{\code{AccountMethods$new()}}
\item \href{#method-AccountMethods-clone}{\code{AccountMethods$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccountMethods-new"></a>}}
\if{latex}{\out{\hypertarget{method-AccountMethods-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AccountMethods wrapper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccountMethods$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The TelegramClient instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccountMethods-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AccountMethods-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccountMethods$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
