% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa.R
\name{add_key}
\alias{add_key}
\title{Adds a new public key to the server keys.}
\usage{
add_key(pub, old = FALSE)
}
\arguments{
\item{pub}{The public key in PEM format.}

\item{old}{Logical indicating if the key is old.}
}
\description{
Adds a new public key to the server keys.
}
