% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_user_to_chat.R
\name{add_user_to_chat}
\alias{add_user_to_chat}
\title{Add a user to a chat or channel}
\usage{
add_user_to_chat(client, chat, user, fwd_limit = 10)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{chat}{Chat/channel (username, link, or entity).}

\item{user}{User (username, link, or entity).}

\item{fwd_limit}{Forward limit for basic chats. Default is 10.}
}
\value{
The API response.
}
\description{
Add a user to a chat or channel
}
