% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{AES}
\alias{AES}
\title{AES}
\description{
Provides an interface to encrypt and decrypt text using the AES IGE mode.

Telegram API type AES
}
\details{
R6 Class Representing AES Encryption and Decryption


A pure R implementation of AES IGE mode encryption and decryption.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AES-decrypt_ige}{\code{AES$decrypt_ige()}}
\item \href{#method-AES-encrypt_ige}{\code{AES$encrypt_ige()}}
\item \href{#method-AES-clone}{\code{AES$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AES-decrypt_ige"></a>}}
\if{latex}{\out{\hypertarget{method-AES-decrypt_ige}{}}}
\subsection{Method \code{decrypt_ige()}}{
Decrypts the given text in 16-byte blocks using the provided key and IV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AES$decrypt_ige(cipher_text, key, iv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cipher_text}}{A raw vector representing the encrypted text.}

\item{\code{key}}{A raw vector representing the encryption key.}

\item{\code{iv}}{A raw vector representing the 32-byte initialization vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A raw vector representing the decrypted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AES-encrypt_ige"></a>}}
\if{latex}{\out{\hypertarget{method-AES-encrypt_ige}{}}}
\subsection{Method \code{encrypt_ige()}}{
Encrypts the given text in 16-byte blocks using the provided key and IV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AES$encrypt_ige(plain_text, key, iv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plain_text}}{A raw vector representing the text to encrypt.}

\item{\code{key}}{A raw vector representing the encryption key.}

\item{\code{iv}}{A raw vector representing the 32-byte initialization vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A raw vector representing the encrypted text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AES-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AES-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AES$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
