% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesctr.R, R/tcpmtproxy.R
\name{AESModeCTR}
\alias{AESModeCTR}
\title{AES CTR Mode R6 Class}
\description{
A wrapper around AES CTR mode with custom IV using the `digest` library.

A simple placeholder class for AES CTR mode encryption.
}
\details{
This class provides methods to encrypt and decrypt data using AES CTR mode.
It initializes with a key and a 16-byte initialization vector (IV).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{key}}{The encryption key as a raw vector.}

\item{\code{iv}}{The initialization vector (IV) as a 16-byte raw vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AESModeCTR-new}{\code{AESModeCTR$new()}}
\item \href{#method-AESModeCTR-encrypt}{\code{AESModeCTR$encrypt()}}
\item \href{#method-AESModeCTR-clone}{\code{AESModeCTR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AESModeCTR-new"></a>}}
\if{latex}{\out{\hypertarget{method-AESModeCTR-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AESModeCTR$new(key, iv)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AESModeCTR-encrypt"></a>}}
\if{latex}{\out{\hypertarget{method-AESModeCTR-encrypt}{}}}
\subsection{Method \code{encrypt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AESModeCTR$encrypt(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AESModeCTR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AESModeCTR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AESModeCTR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{key}}{The encryption key.}

\item{\code{iv}}{The initialization vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AESModeCTR-new}{\code{AESModeCTR$new()}}
\item \href{#method-AESModeCTR-encrypt}{\code{AESModeCTR$encrypt()}}
\item \href{#method-AESModeCTR-clone}{\code{AESModeCTR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AESModeCTR-new"></a>}}
\if{latex}{\out{\hypertarget{method-AESModeCTR-new}{}}}
\subsection{Method \code{new()}}{
Initialize AESModeCTR.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AESModeCTR$new(key, iv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A raw vector representing the encryption key.}

\item{\code{iv}}{A raw vector representing the initialization vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AESModeCTR-encrypt"></a>}}
\if{latex}{\out{\hypertarget{method-AESModeCTR-encrypt}{}}}
\subsection{Method \code{encrypt()}}{
Encrypt data using AES CTR mode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AESModeCTR$encrypt(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A raw vector to encrypt.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A raw vector containing the (simulated) encrypted data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AESModeCTR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AESModeCTR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AESModeCTR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
