% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{AlreadyInConversationError}
\alias{AlreadyInConversationError}
\title{AlreadyInConversationError}
\description{
Telegram API type AlreadyInConversationError
}
\details{
AlreadyInConversationError
}
\section{Super class}{
\code{\link[telegramR:Exception]{telegramR::Exception}} -> \code{AlreadyInConversationError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AlreadyInConversationError-new}{\code{AlreadyInConversationError$new()}}
\item \href{#method-AlreadyInConversationError-clone}{\code{AlreadyInConversationError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlreadyInConversationError-new"></a>}}
\if{latex}{\out{\hypertarget{method-AlreadyInConversationError-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the AlreadyInConversationError class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlreadyInConversationError$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{\code{character} The error message.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlreadyInConversationError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AlreadyInConversationError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlreadyInConversationError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
