% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{async_open_connection}
\alias{async_open_connection}
\title{Helper function for asynchronous open connection.}
\usage{
async_open_connection(
  host = NULL,
  port = NULL,
  ssl = NULL,
  local_addr = NULL,
  sock = NULL,
  timeout = NULL
)
}
\arguments{
\item{host}{Hostname.}

\item{port}{Port number.}

\item{ssl}{Optional SSL configuration.}

\item{local_addr}{Optional local address.}

\item{sock}{Optional socket.}

\item{timeout}{Optional timeout.}
}
\value{
A promise resolving with a list containing reader and writer.
}
\description{
Helper function for asynchronous open connection.
}
