% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AsyncClassWrapper}
\alias{AsyncClassWrapper}
\title{AsyncClassWrapper}
\description{
An R6 class that wraps an object and provides asynchronous method wrapping.
This class mimics the behavior of a Python class that wraps objects to handle asynchronous calls.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{wrapped}}{The wrapped object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AsyncClassWrapper-new}{\code{AsyncClassWrapper$new()}}
\item \href{#method-AsyncClassWrapper-get_attr}{\code{AsyncClassWrapper$get_attr()}}
\item \href{#method-AsyncClassWrapper-clone}{\code{AsyncClassWrapper$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsyncClassWrapper-new"></a>}}
\if{latex}{\out{\hypertarget{method-AsyncClassWrapper-new}{}}}
\subsection{Method \code{new()}}{
Initialize the AsyncClassWrapper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncClassWrapper$new(wrapped)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wrapped}}{The object to wrap.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsyncClassWrapper-get_attr"></a>}}
\if{latex}{\out{\hypertarget{method-AsyncClassWrapper-get_attr}{}}}
\subsection{Method \code{get_attr()}}{
Get an attribute from the wrapped object.
If the attribute is a function, return a wrapper function that checks if the result is awaitable
and awaits it if necessary (assuming the 'coro' package for async support).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncClassWrapper$get_attr(item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{The name of the attribute as a string.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The attribute value or a wrapped function.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsyncClassWrapper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AsyncClassWrapper-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncClassWrapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
