% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{AsyncQueue}
\alias{AsyncQueue}
\title{AsyncQueue Class}
\description{
Implements an asynchronous queue.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AsyncQueue-new}{\code{AsyncQueue$new()}}
\item \href{#method-AsyncQueue-put}{\code{AsyncQueue$put()}}
\item \href{#method-AsyncQueue-get}{\code{AsyncQueue$get()}}
\item \href{#method-AsyncQueue-clone}{\code{AsyncQueue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsyncQueue-new"></a>}}
\if{latex}{\out{\hypertarget{method-AsyncQueue-new}{}}}
\subsection{Method \code{new()}}{
Create a new AsyncQueue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncQueue$new(maxsize = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxsize}}{Maximum size of the queue.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsyncQueue-put"></a>}}
\if{latex}{\out{\hypertarget{method-AsyncQueue-put}{}}}
\subsection{Method \code{put()}}{
Add an item to the queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncQueue$put(item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{The item to add.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A promise that resolves when the item is added.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsyncQueue-get"></a>}}
\if{latex}{\out{\hypertarget{method-AsyncQueue-get}{}}}
\subsection{Method \code{get()}}{
Retrieve an item from the queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncQueue$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A promise that resolves with the item.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsyncQueue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AsyncQueue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncQueue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
