% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AttachMenuBot}
\alias{AttachMenuBot}
\title{AttachMenuBot}
\description{
Telegram API type AttachMenuBot
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{AttachMenuBot}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttachMenuBot-new}{\code{AttachMenuBot$new()}}
\item \href{#method-AttachMenuBot-to_dict}{\code{AttachMenuBot$to_dict()}}
\item \href{#method-AttachMenuBot-bytes}{\code{AttachMenuBot$bytes()}}
\item \href{#method-AttachMenuBot-clone}{\code{AttachMenuBot$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBot-new"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBot-new}{}}}
\subsection{Method \code{new()}}{
Initializes the AttachMenuBot with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBot$new(
  bot_id,
  short_name,
  icons,
  inactive = NULL,
  has_settings = NULL,
  request_write_access = NULL,
  show_in_attach_menu = NULL,
  show_in_side_menu = NULL,
  side_menu_disclaimer_needed = NULL,
  peer_types = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bot_id}}{An integer representing the bot ID.}

\item{\code{short_name}}{A string representing the short name.}

\item{\code{icons}}{A list of TypeAttachMenuBotIcon objects.}

\item{\code{inactive}}{Optional boolean indicating if inactive.}

\item{\code{has_settings}}{Optional boolean indicating if has settings.}

\item{\code{request_write_access}}{Optional boolean indicating if requests write access.}

\item{\code{show_in_attach_menu}}{Optional boolean indicating if shows in attach menu.}

\item{\code{show_in_side_menu}}{Optional boolean indicating if shows in side menu.}

\item{\code{side_menu_disclaimer_needed}}{Optional boolean indicating if side menu disclaimer needed.}

\item{\code{peer_types}}{Optional list of TypeAttachMenuPeerType objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBot-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBot-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the AttachMenuBot object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBot$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the AttachMenuBot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBot-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBot-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the AttachMenuBot object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBot$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the AttachMenuBot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
