% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AttachMenuBotIcon}
\alias{AttachMenuBotIcon}
\title{AttachMenuBotIcon}
\description{
Telegram API type AttachMenuBotIcon
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{AttachMenuBotIcon}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{name}}{Field.}

\item{\code{icon}}{Field.}

\item{\code{colors}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttachMenuBotIcon-new}{\code{AttachMenuBotIcon$new()}}
\item \href{#method-AttachMenuBotIcon-to_dict}{\code{AttachMenuBotIcon$to_dict()}}
\item \href{#method-AttachMenuBotIcon-bytes}{\code{AttachMenuBotIcon$bytes()}}
\item \href{#method-AttachMenuBotIcon-clone}{\code{AttachMenuBotIcon$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBotIcon-new"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBotIcon-new}{}}}
\subsection{Method \code{new()}}{
Initializes the AttachMenuBotIcon with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBotIcon$new(name, icon, colors = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A string representing the icon name.}

\item{\code{icon}}{A TypeDocument object representing the icon.}

\item{\code{colors}}{Optional list of TypeAttachMenuBotIconColor objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBotIcon-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBotIcon-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the AttachMenuBotIcon object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBotIcon$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the AttachMenuBotIcon object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBotIcon-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBotIcon-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the AttachMenuBotIcon object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBotIcon$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the AttachMenuBotIcon object.
Reads an AttachMenuBotIcon object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBotIcon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBotIcon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBotIcon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
