% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AttachMenuBots}
\alias{AttachMenuBots}
\title{AttachMenuBots}
\description{
Telegram API type AttachMenuBots
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{AttachMenuBots}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttachMenuBots-new}{\code{AttachMenuBots$new()}}
\item \href{#method-AttachMenuBots-to_dict}{\code{AttachMenuBots$to_dict()}}
\item \href{#method-AttachMenuBots-bytes}{\code{AttachMenuBots$bytes()}}
\item \href{#method-AttachMenuBots-clone}{\code{AttachMenuBots$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBots-new"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBots-new}{}}}
\subsection{Method \code{new()}}{
Initializes the AttachMenuBots with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBots$new(hash, bots, users)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{An integer representing the hash value.}

\item{\code{bots}}{A list of TypeAttachMenuBot objects.}

\item{\code{users}}{A list of TypeUser objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBots-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBots-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the AttachMenuBots object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBots$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the AttachMenuBots object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBots-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBots-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the AttachMenuBots object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBots$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the AttachMenuBots object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuBots-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuBots-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuBots$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
