% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AttachMenuPeerTypeBroadcast}
\alias{AttachMenuPeerTypeBroadcast}
\title{AttachMenuPeerTypeBroadcast}
\description{
Telegram API type AttachMenuPeerTypeBroadcast
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{AttachMenuPeerTypeBroadcast}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttachMenuPeerTypeBroadcast-new}{\code{AttachMenuPeerTypeBroadcast$new()}}
\item \href{#method-AttachMenuPeerTypeBroadcast-to_dict}{\code{AttachMenuPeerTypeBroadcast$to_dict()}}
\item \href{#method-AttachMenuPeerTypeBroadcast-bytes}{\code{AttachMenuPeerTypeBroadcast$bytes()}}
\item \href{#method-AttachMenuPeerTypeBroadcast-clone}{\code{AttachMenuPeerTypeBroadcast$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuPeerTypeBroadcast-new"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuPeerTypeBroadcast-new}{}}}
\subsection{Method \code{new()}}{
Initializes the AttachMenuPeerTypeBroadcast object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuPeerTypeBroadcast$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuPeerTypeBroadcast-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuPeerTypeBroadcast-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the AttachMenuPeerTypeBroadcast object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuPeerTypeBroadcast$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the AttachMenuPeerTypeBroadcast object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuPeerTypeBroadcast-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuPeerTypeBroadcast-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the AttachMenuPeerTypeBroadcast object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuPeerTypeBroadcast$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the AttachMenuPeerTypeBroadcast object.
Reads an AttachMenuPeerTypeBroadcast object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuPeerTypeBroadcast-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuPeerTypeBroadcast-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuPeerTypeBroadcast$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
