% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AttachMenuPeerTypeSameBotPM}
\alias{AttachMenuPeerTypeSameBotPM}
\title{AttachMenuPeerTypeSameBotPM}
\description{
Telegram API type AttachMenuPeerTypeSameBotPM
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{AttachMenuPeerTypeSameBotPM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttachMenuPeerTypeSameBotPM-new}{\code{AttachMenuPeerTypeSameBotPM$new()}}
\item \href{#method-AttachMenuPeerTypeSameBotPM-to_dict}{\code{AttachMenuPeerTypeSameBotPM$to_dict()}}
\item \href{#method-AttachMenuPeerTypeSameBotPM-bytes}{\code{AttachMenuPeerTypeSameBotPM$bytes()}}
\item \href{#method-AttachMenuPeerTypeSameBotPM-clone}{\code{AttachMenuPeerTypeSameBotPM$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuPeerTypeSameBotPM-new"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuPeerTypeSameBotPM-new}{}}}
\subsection{Method \code{new()}}{
Initializes the AttachMenuPeerTypeSameBotPM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuPeerTypeSameBotPM$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuPeerTypeSameBotPM-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuPeerTypeSameBotPM-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the AttachMenuPeerTypeSameBotPM object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuPeerTypeSameBotPM$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the AttachMenuPeerTypeSameBotPM object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuPeerTypeSameBotPM-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuPeerTypeSameBotPM-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the AttachMenuPeerTypeSameBotPM object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuPeerTypeSameBotPM$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the AttachMenuPeerTypeSameBotPM object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttachMenuPeerTypeSameBotPM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttachMenuPeerTypeSameBotPM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttachMenuPeerTypeSameBotPM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
