% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{AuthKeyNotFound}
\alias{AuthKeyNotFound}
\title{AuthKeyNotFound}
\description{
The server claims it doesn't know about the authorization key (session file)
currently being used. This might be because it either has never seen this
authorization key, or it used to know about the authorization key but has
forgotten it, either temporarily or permanently (possibly due to server errors).
}
\section{Super class}{
\code{\link[telegramR:Exception]{telegramR::Exception}} -> \code{AuthKeyNotFound}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthKeyNotFound-new}{\code{AuthKeyNotFound$new()}}
\item \href{#method-AuthKeyNotFound-clone}{\code{AuthKeyNotFound$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthKeyNotFound-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthKeyNotFound-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the AuthKeyNotFound class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthKeyNotFound$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{\code{character} The error message.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthKeyNotFound-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthKeyNotFound-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthKeyNotFound$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
