% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AutoDownloadSettings}
\alias{AutoDownloadSettings}
\title{AutoDownloadSettings}
\description{
Telegram API type AutoDownloadSettings
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{AutoDownloadSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutoDownloadSettings-new}{\code{AutoDownloadSettings$new()}}
\item \href{#method-AutoDownloadSettings-to_dict}{\code{AutoDownloadSettings$to_dict()}}
\item \href{#method-AutoDownloadSettings-bytes}{\code{AutoDownloadSettings$bytes()}}
\item \href{#method-AutoDownloadSettings-clone}{\code{AutoDownloadSettings$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoDownloadSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutoDownloadSettings-new}{}}}
\subsection{Method \code{new()}}{
Initializes the AutoDownloadSettings with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoDownloadSettings$new(
  photo_size_max,
  video_size_max,
  file_size_max,
  video_upload_maxbitrate,
  small_queue_active_operations_max,
  large_queue_active_operations_max,
  disabled = NULL,
  video_preload_large = NULL,
  audio_preload_next = NULL,
  phonecalls_lessdata = NULL,
  stories_preload = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{photo_size_max}}{An integer representing the maximum photo size.}

\item{\code{video_size_max}}{An integer representing the maximum video size.}

\item{\code{file_size_max}}{An integer representing the maximum file size.}

\item{\code{video_upload_maxbitrate}}{An integer representing the maximum video upload bitrate.}

\item{\code{small_queue_active_operations_max}}{An integer representing the maximum small queue active operations.}

\item{\code{large_queue_active_operations_max}}{An integer representing the maximum large queue active operations.}

\item{\code{disabled}}{Optional boolean indicating if disabled.}

\item{\code{video_preload_large}}{Optional boolean indicating if video preload large.}

\item{\code{audio_preload_next}}{Optional boolean indicating if audio preload next.}

\item{\code{phonecalls_lessdata}}{Optional boolean indicating if phonecalls use less data.}

\item{\code{stories_preload}}{Optional boolean indicating if stories preload.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoDownloadSettings-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-AutoDownloadSettings-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the AutoDownloadSettings object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoDownloadSettings$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the AutoDownloadSettings object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoDownloadSettings-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AutoDownloadSettings-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the AutoDownloadSettings object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoDownloadSettings$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the AutoDownloadSettings object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoDownloadSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutoDownloadSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoDownloadSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
