% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AutoSaveSettings}
\alias{AutoSaveSettings}
\title{AutoSaveSettings}
\description{
Telegram API type AutoSaveSettings
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{AutoSaveSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutoSaveSettings-new}{\code{AutoSaveSettings$new()}}
\item \href{#method-AutoSaveSettings-to_dict}{\code{AutoSaveSettings$to_dict()}}
\item \href{#method-AutoSaveSettings-bytes}{\code{AutoSaveSettings$bytes()}}
\item \href{#method-AutoSaveSettings-clone}{\code{AutoSaveSettings$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoSaveSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutoSaveSettings-new}{}}}
\subsection{Method \code{new()}}{
Initializes the AutoSaveSettings with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoSaveSettings$new(photos = NULL, videos = NULL, video_max_size = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{photos}}{Optional boolean indicating if photos should be auto-saved.}

\item{\code{videos}}{Optional boolean indicating if videos should be auto-saved.}

\item{\code{video_max_size}}{Optional integer representing the maximum video size.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoSaveSettings-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-AutoSaveSettings-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the AutoSaveSettings object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoSaveSettings$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the AutoSaveSettings object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoSaveSettings-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AutoSaveSettings-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the AutoSaveSettings object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoSaveSettings$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the AutoSaveSettings object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoSaveSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutoSaveSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoSaveSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
