% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AvailableEffect}
\alias{AvailableEffect}
\title{AvailableEffect}
\description{
Telegram API type AvailableEffect
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{AvailableEffect}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{id}}{Field.}

\item{\code{emoticon}}{Field.}

\item{\code{effect_sticker_id}}{Field.}

\item{\code{premium_required}}{Field.}

\item{\code{static_icon_id}}{Field.}

\item{\code{effect_animation_id}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AvailableEffect-new}{\code{AvailableEffect$new()}}
\item \href{#method-AvailableEffect-to_dict}{\code{AvailableEffect$to_dict()}}
\item \href{#method-AvailableEffect-bytes}{\code{AvailableEffect$bytes()}}
\item \href{#method-AvailableEffect-clone}{\code{AvailableEffect$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableEffect-new"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableEffect-new}{}}}
\subsection{Method \code{new()}}{
Initializes the AvailableEffect with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableEffect$new(
  id,
  emoticon,
  effect_sticker_id,
  premium_required = NULL,
  static_icon_id = NULL,
  effect_animation_id = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{An integer representing the effect ID.}

\item{\code{emoticon}}{A string representing the emoticon.}

\item{\code{effect_sticker_id}}{An integer representing the effect sticker ID.}

\item{\code{premium_required}}{Optional boolean indicating if premium is required.}

\item{\code{static_icon_id}}{Optional integer representing the static icon ID.}

\item{\code{effect_animation_id}}{Optional integer representing the effect animation ID.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableEffect-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableEffect-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the AvailableEffect object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableEffect$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the AvailableEffect object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableEffect-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableEffect-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the AvailableEffect object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableEffect$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the AvailableEffect object.
Reads an AvailableEffect object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableEffect-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableEffect-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableEffect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
