% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AvailableReaction}
\alias{AvailableReaction}
\title{AvailableReaction}
\description{
Telegram API type AvailableReaction
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{AvailableReaction}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{reaction}}{Field.}

\item{\code{title}}{Field.}

\item{\code{static_icon}}{Field.}

\item{\code{appear_animation}}{Field.}

\item{\code{select_animation}}{Field.}

\item{\code{activate_animation}}{Field.}

\item{\code{effect_animation}}{Field.}

\item{\code{inactive}}{Field.}

\item{\code{premium}}{Field.}

\item{\code{around_animation}}{Field.}

\item{\code{center_icon}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AvailableReaction-new}{\code{AvailableReaction$new()}}
\item \href{#method-AvailableReaction-to_dict}{\code{AvailableReaction$to_dict()}}
\item \href{#method-AvailableReaction-bytes}{\code{AvailableReaction$bytes()}}
\item \href{#method-AvailableReaction-clone}{\code{AvailableReaction$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableReaction-new"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableReaction-new}{}}}
\subsection{Method \code{new()}}{
Initializes the AvailableReaction with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableReaction$new(
  reaction,
  title,
  static_icon,
  appear_animation,
  select_animation,
  activate_animation,
  effect_animation,
  inactive = NULL,
  premium = NULL,
  around_animation = NULL,
  center_icon = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reaction}}{A string representing the reaction.}

\item{\code{title}}{A string representing the title.}

\item{\code{static_icon}}{A TLObject representing the static icon.}

\item{\code{appear_animation}}{A TLObject representing the appear animation.}

\item{\code{select_animation}}{A TLObject representing the select animation.}

\item{\code{activate_animation}}{A TLObject representing the activate animation.}

\item{\code{effect_animation}}{A TLObject representing the effect animation.}

\item{\code{inactive}}{A logical indicating if inactive. Optional.}

\item{\code{premium}}{A logical indicating if premium. Optional.}

\item{\code{around_animation}}{A TLObject representing the around animation. Optional.}

\item{\code{center_icon}}{A TLObject representing the center icon. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableReaction-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableReaction-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the AvailableReaction object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableReaction$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the AvailableReaction object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableReaction-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableReaction-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the AvailableReaction object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableReaction$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the AvailableReaction object.
Reads an AvailableReaction object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableReaction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableReaction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableReaction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
