% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{BadMessageError}
\alias{BadMessageError}
\title{BadMessageError}
\description{
Occurs when handling a bad_message_notification.

Telegram API type BadMessageError
}
\details{
BadMessageError
}
\section{Super class}{
\code{\link[telegramR:Exception]{telegramR::Exception}} -> \code{BadMessageError}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{request}}{\code{TLRequest} The request that caused the error.}

\item{\code{code}}{\code{numeric} The error code.}

\item{\code{ErrorMessages}}{\code{list} A list of error messages corresponding to the error codes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BadMessageError-new}{\code{BadMessageError$new()}}
\item \href{#method-BadMessageError-clone}{\code{BadMessageError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BadMessageError-new"></a>}}
\if{latex}{\out{\hypertarget{method-BadMessageError-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the BadMessageError class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BadMessageError$new(request, code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request}}{\code{TLRequest} The request that caused the error.}

\item{\code{code}}{\code{numeric} The error code.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BadMessageError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BadMessageError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BadMessageError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
