% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BadMsgNotification}
\alias{BadMsgNotification}
\title{BadMsgNotification}
\description{
Telegram API type BadMsgNotification
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BadMsgNotification}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{bad_msg_id}}{Field.}

\item{\code{bad_msg_seqno}}{Field.}

\item{\code{error_code}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BadMsgNotification-new}{\code{BadMsgNotification$new()}}
\item \href{#method-BadMsgNotification-to_dict}{\code{BadMsgNotification$to_dict()}}
\item \href{#method-BadMsgNotification-bytes}{\code{BadMsgNotification$bytes()}}
\item \href{#method-BadMsgNotification-clone}{\code{BadMsgNotification$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BadMsgNotification-new"></a>}}
\if{latex}{\out{\hypertarget{method-BadMsgNotification-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BadMsgNotification with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BadMsgNotification$new(bad_msg_id, bad_msg_seqno, error_code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bad_msg_id}}{An integer representing the bad message ID.}

\item{\code{bad_msg_seqno}}{An integer representing the bad message sequence number.}

\item{\code{error_code}}{An integer representing the error code.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BadMsgNotification-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BadMsgNotification-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BadMsgNotification object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BadMsgNotification$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BadMsgNotification object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BadMsgNotification-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BadMsgNotification-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BadMsgNotification object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BadMsgNotification$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BadMsgNotification object.
Reads a BadMsgNotification object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BadMsgNotification-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BadMsgNotification-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BadMsgNotification$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
