% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BaseThemeDay}
\alias{BaseThemeDay}
\title{BaseThemeDay}
\description{
Telegram API type BaseThemeDay
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BaseThemeDay}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseThemeDay-new}{\code{BaseThemeDay$new()}}
\item \href{#method-BaseThemeDay-to_dict}{\code{BaseThemeDay$to_dict()}}
\item \href{#method-BaseThemeDay-bytes}{\code{BaseThemeDay$bytes()}}
\item \href{#method-BaseThemeDay-clone}{\code{BaseThemeDay$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseThemeDay-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseThemeDay-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BaseThemeDay with no parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseThemeDay$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseThemeDay-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BaseThemeDay-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BaseThemeDay object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseThemeDay$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BaseThemeDay object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseThemeDay-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BaseThemeDay-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BaseThemeDay object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseThemeDay$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BaseThemeDay object.
Reads a BaseThemeDay object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseThemeDay-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseThemeDay-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseThemeDay$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
