% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BaseThemeNight}
\alias{BaseThemeNight}
\title{BaseThemeNight}
\description{
Telegram API type BaseThemeNight
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BaseThemeNight}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseThemeNight-new}{\code{BaseThemeNight$new()}}
\item \href{#method-BaseThemeNight-to_dict}{\code{BaseThemeNight$to_dict()}}
\item \href{#method-BaseThemeNight-bytes}{\code{BaseThemeNight$bytes()}}
\item \href{#method-BaseThemeNight-clone}{\code{BaseThemeNight$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseThemeNight-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseThemeNight-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BaseThemeNight with no parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseThemeNight$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseThemeNight-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BaseThemeNight-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BaseThemeNight object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseThemeNight$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BaseThemeNight object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseThemeNight-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BaseThemeNight-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BaseThemeNight object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseThemeNight$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BaseThemeNight object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseThemeNight-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseThemeNight-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseThemeNight$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
