% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{big_num_for_hash}
\alias{big_num_for_hash}
\title{Big Num for Hash}
\usage{
big_num_for_hash(g)
}
\arguments{
\item{g}{An integer.}
}
\value{
A raw vector.
}
\description{
Converts an integer to 256 bytes big-endian.
}
