% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BindAuthKeyInner}
\alias{BindAuthKeyInner}
\title{BindAuthKeyInner}
\description{
Telegram API type BindAuthKeyInner
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BindAuthKeyInner}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BindAuthKeyInner-new}{\code{BindAuthKeyInner$new()}}
\item \href{#method-BindAuthKeyInner-to_dict}{\code{BindAuthKeyInner$to_dict()}}
\item \href{#method-BindAuthKeyInner-bytes}{\code{BindAuthKeyInner$bytes()}}
\item \href{#method-BindAuthKeyInner-clone}{\code{BindAuthKeyInner$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BindAuthKeyInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-BindAuthKeyInner-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BindAuthKeyInner with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BindAuthKeyInner$new(
  nonce,
  temp_auth_key_id,
  perm_auth_key_id,
  temp_session_id,
  expires_at = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nonce}}{An integer representing the nonce.}

\item{\code{temp_auth_key_id}}{An integer representing the temporary auth key ID.}

\item{\code{perm_auth_key_id}}{An integer representing the permanent auth key ID.}

\item{\code{temp_session_id}}{An integer representing the temporary session ID.}

\item{\code{expires_at}}{A datetime object representing when this expires. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BindAuthKeyInner-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BindAuthKeyInner-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BindAuthKeyInner object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BindAuthKeyInner$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BindAuthKeyInner object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BindAuthKeyInner-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BindAuthKeyInner-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BindAuthKeyInner object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BindAuthKeyInner$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BindAuthKeyInner object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BindAuthKeyInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BindAuthKeyInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BindAuthKeyInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
