% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Birthday}
\alias{Birthday}
\title{Birthday}
\description{
Telegram API type Birthday
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{Birthday}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Birthday-new}{\code{Birthday$new()}}
\item \href{#method-Birthday-to_dict}{\code{Birthday$to_dict()}}
\item \href{#method-Birthday-bytes}{\code{Birthday$bytes()}}
\item \href{#method-Birthday-clone}{\code{Birthday$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Birthday-new"></a>}}
\if{latex}{\out{\hypertarget{method-Birthday-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Birthday with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Birthday$new(day, month, year = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{day}}{An integer representing the day.}

\item{\code{month}}{An integer representing the month.}

\item{\code{year}}{An integer representing the year. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Birthday-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-Birthday-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the Birthday object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Birthday$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the Birthday object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Birthday-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-Birthday-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the Birthday object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Birthday$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the Birthday object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Birthday-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Birthday-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Birthday$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
