% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bitwLength}
\alias{bitwLength}
\title{Bitwise Length}
\usage{
bitwLength(x)
}
\arguments{
\item{x}{An integer value.}
}
\value{
The bitwise length of the integer.
}
\description{
Calculates the bitwise length of an integer.
}
