% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Boost}
\alias{Boost}
\title{Boost}
\description{
Telegram API type Boost
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{Boost}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{id}}{Field.}

\item{\code{date}}{Field.}

\item{\code{expires}}{Field.}

\item{\code{gift}}{Field.}

\item{\code{giveaway}}{Field.}

\item{\code{unclaimed}}{Field.}

\item{\code{user_id}}{Field.}

\item{\code{giveaway_msg_id}}{Field.}

\item{\code{used_gift_slug}}{Field.}

\item{\code{multiplier}}{Field.}

\item{\code{stars}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Boost-new}{\code{Boost$new()}}
\item \href{#method-Boost-to_dict}{\code{Boost$to_dict()}}
\item \href{#method-Boost-bytes}{\code{Boost$bytes()}}
\item \href{#method-Boost-clone}{\code{Boost$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boost-new"></a>}}
\if{latex}{\out{\hypertarget{method-Boost-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Boost with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boost$new(
  id,
  date = NULL,
  expires = NULL,
  gift = NULL,
  giveaway = NULL,
  unclaimed = NULL,
  user_id = NULL,
  giveaway_msg_id = NULL,
  used_gift_slug = NULL,
  multiplier = NULL,
  stars = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{A string representing the boost ID.}

\item{\code{date}}{A datetime object representing the boost date. Optional.}

\item{\code{expires}}{A datetime object representing when the boost expires. Optional.}

\item{\code{gift}}{A logical indicating if this is a gift boost. Optional.}

\item{\code{giveaway}}{A logical indicating if this is a giveaway boost. Optional.}

\item{\code{unclaimed}}{A logical indicating if the boost is unclaimed. Optional.}

\item{\code{user_id}}{An integer representing the user ID. Optional.}

\item{\code{giveaway_msg_id}}{An integer representing the giveaway message ID. Optional.}

\item{\code{used_gift_slug}}{A string representing the used gift slug. Optional.}

\item{\code{multiplier}}{An integer representing the multiplier. Optional.}

\item{\code{stars}}{An integer representing the stars. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boost-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-Boost-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the Boost object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boost$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the Boost object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boost-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-Boost-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the Boost object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boost$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the Boost object.
Reads a Boost object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Boost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Boost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Boost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
