% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BotApp}
\alias{BotApp}
\title{BotApp}
\description{
Telegram API type BotApp
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BotApp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotApp-new}{\code{BotApp$new()}}
\item \href{#method-BotApp-to_dict}{\code{BotApp$to_dict()}}
\item \href{#method-BotApp-bytes}{\code{BotApp$bytes()}}
\item \href{#method-BotApp-clone}{\code{BotApp$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotApp-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotApp-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BotApp with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotApp$new(
  id,
  access_hash,
  short_name,
  title,
  description,
  photo,
  hash,
  document = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{An integer representing the bot app ID.}

\item{\code{access_hash}}{An integer representing the access hash.}

\item{\code{short_name}}{A string representing the short name.}

\item{\code{title}}{A string representing the title.}

\item{\code{description}}{A string representing the description.}

\item{\code{photo}}{A TLObject representing the photo.}

\item{\code{hash}}{An integer representing the hash.}

\item{\code{document}}{A TLObject representing the document. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotApp-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotApp-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BotApp object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotApp$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BotApp object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotApp-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotApp-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BotApp object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotApp$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BotApp object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotApp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotApp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotApp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
