% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BotAppSettings}
\alias{BotAppSettings}
\title{BotAppSettings}
\description{
Telegram API type BotAppSettings
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BotAppSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotAppSettings-new}{\code{BotAppSettings$new()}}
\item \href{#method-BotAppSettings-to_dict}{\code{BotAppSettings$to_dict()}}
\item \href{#method-BotAppSettings-bytes}{\code{BotAppSettings$bytes()}}
\item \href{#method-BotAppSettings-clone}{\code{BotAppSettings$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotAppSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotAppSettings-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BotAppSettings with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotAppSettings$new(
  placeholder_path = NULL,
  background_color = NULL,
  background_dark_color = NULL,
  header_color = NULL,
  header_dark_color = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{placeholder_path}}{A raw vector representing placeholder path bytes. Optional.}

\item{\code{background_color}}{An integer representing the background color. Optional.}

\item{\code{background_dark_color}}{An integer representing the dark background color. Optional.}

\item{\code{header_color}}{An integer representing the header color. Optional.}

\item{\code{header_dark_color}}{An integer representing the dark header color. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotAppSettings-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotAppSettings-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BotAppSettings object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotAppSettings$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BotAppSettings object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotAppSettings-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotAppSettings-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BotAppSettings object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotAppSettings$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BotAppSettings object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotAppSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotAppSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotAppSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
