% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_payments.R
\name{BotCancelStarsSubscriptionRequest}
\alias{BotCancelStarsSubscriptionRequest}
\title{BotCancelStarsSubscriptionRequest}
\description{
Telegram API type BotCancelStarsSubscriptionRequest
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{BotCancelStarsSubscriptionRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{user_id}}{Field.}

\item{\code{charge_id}}{Field.}

\item{\code{restore}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotCancelStarsSubscriptionRequest-new}{\code{BotCancelStarsSubscriptionRequest$new()}}
\item \href{#method-BotCancelStarsSubscriptionRequest-resolve}{\code{BotCancelStarsSubscriptionRequest$resolve()}}
\item \href{#method-BotCancelStarsSubscriptionRequest-to_dict}{\code{BotCancelStarsSubscriptionRequest$to_dict()}}
\item \href{#method-BotCancelStarsSubscriptionRequest-bytes}{\code{BotCancelStarsSubscriptionRequest$bytes()}}
\item \href{#method-BotCancelStarsSubscriptionRequest-from_reader}{\code{BotCancelStarsSubscriptionRequest$from_reader()}}
\item \href{#method-BotCancelStarsSubscriptionRequest-clone}{\code{BotCancelStarsSubscriptionRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCancelStarsSubscriptionRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotCancelStarsSubscriptionRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the BotCancelStarsSubscriptionRequest object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCancelStarsSubscriptionRequest$new(user_id, charge_id, restore = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{The input user (TypeInputUser).}

\item{\code{charge_id}}{The charge ID (string).}

\item{\code{restore}}{Optional boolean flag to restore the subscription.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCancelStarsSubscriptionRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-BotCancelStarsSubscriptionRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve the user_id using client and utils.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCancelStarsSubscriptionRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The client object.}

\item{\code{utils}}{The utils object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCancelStarsSubscriptionRequest-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotCancelStarsSubscriptionRequest-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Convert the object to a dictionary-like list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCancelStarsSubscriptionRequest$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object in dictionary form.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCancelStarsSubscriptionRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotCancelStarsSubscriptionRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize the object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCancelStarsSubscriptionRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector of bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCancelStarsSubscriptionRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-BotCancelStarsSubscriptionRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Deserialize from a reader.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCancelStarsSubscriptionRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{The reader object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of BotCancelStarsSubscriptionRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCancelStarsSubscriptionRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotCancelStarsSubscriptionRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCancelStarsSubscriptionRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
