% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BotCommand}
\alias{BotCommand}
\title{BotCommand}
\description{
Telegram API type BotCommand
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BotCommand}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotCommand-new}{\code{BotCommand$new()}}
\item \href{#method-BotCommand-to_dict}{\code{BotCommand$to_dict()}}
\item \href{#method-BotCommand-bytes}{\code{BotCommand$bytes()}}
\item \href{#method-BotCommand-clone}{\code{BotCommand$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommand-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommand-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BotCommand with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommand$new(command, description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{command}}{A string representing the command.}

\item{\code{description}}{A string representing the description.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommand-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommand-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BotCommand object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommand$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BotCommand object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommand-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommand-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BotCommand object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommand$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BotCommand object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommand-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
