% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BotCommandScopeChatAdmins}
\alias{BotCommandScopeChatAdmins}
\title{BotCommandScopeChatAdmins}
\description{
Telegram API type BotCommandScopeChatAdmins
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BotCommandScopeChatAdmins}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotCommandScopeChatAdmins-new}{\code{BotCommandScopeChatAdmins$new()}}
\item \href{#method-BotCommandScopeChatAdmins-to_dict}{\code{BotCommandScopeChatAdmins$to_dict()}}
\item \href{#method-BotCommandScopeChatAdmins-bytes}{\code{BotCommandScopeChatAdmins$bytes()}}
\item \href{#method-BotCommandScopeChatAdmins-clone}{\code{BotCommandScopeChatAdmins$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommandScopeChatAdmins-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommandScopeChatAdmins-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BotCommandScopeChatAdmins with no parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommandScopeChatAdmins$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommandScopeChatAdmins-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommandScopeChatAdmins-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BotCommandScopeChatAdmins object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommandScopeChatAdmins$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BotCommandScopeChatAdmins object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommandScopeChatAdmins-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommandScopeChatAdmins-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BotCommandScopeChatAdmins object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommandScopeChatAdmins$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BotCommandScopeChatAdmins object.
Reads a BotCommandScopeChatAdmins object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommandScopeChatAdmins-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommandScopeChatAdmins-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommandScopeChatAdmins$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
