% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BotCommandScopePeerAdmins}
\alias{BotCommandScopePeerAdmins}
\title{BotCommandScopePeerAdmins}
\description{
Telegram API type BotCommandScopePeerAdmins
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BotCommandScopePeerAdmins}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotCommandScopePeerAdmins-new}{\code{BotCommandScopePeerAdmins$new()}}
\item \href{#method-BotCommandScopePeerAdmins-to_dict}{\code{BotCommandScopePeerAdmins$to_dict()}}
\item \href{#method-BotCommandScopePeerAdmins-bytes}{\code{BotCommandScopePeerAdmins$bytes()}}
\item \href{#method-BotCommandScopePeerAdmins-clone}{\code{BotCommandScopePeerAdmins$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommandScopePeerAdmins-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommandScopePeerAdmins-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BotCommandScopePeerAdmins with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommandScopePeerAdmins$new(peer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{A TLObject representing the peer.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommandScopePeerAdmins-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommandScopePeerAdmins-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BotCommandScopePeerAdmins object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommandScopePeerAdmins$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BotCommandScopePeerAdmins object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommandScopePeerAdmins-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommandScopePeerAdmins-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BotCommandScopePeerAdmins object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommandScopePeerAdmins$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BotCommandScopePeerAdmins object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotCommandScopePeerAdmins-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotCommandScopePeerAdmins-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotCommandScopePeerAdmins$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
