% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BotInfo}
\alias{BotInfo}
\title{BotInfo}
\description{
Telegram API type BotInfo
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BotInfo}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{has_preview_medias}}{Field.}

\item{\code{user_id}}{Field.}

\item{\code{description}}{Field.}

\item{\code{description_photo}}{Field.}

\item{\code{description_document}}{Field.}

\item{\code{commands}}{Field.}

\item{\code{menu_button}}{Field.}

\item{\code{privacy_policy_url}}{Field.}

\item{\code{app_settings}}{Field.}

\item{\code{verifier_settings}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotInfo-new}{\code{BotInfo$new()}}
\item \href{#method-BotInfo-to_dict}{\code{BotInfo$to_dict()}}
\item \href{#method-BotInfo-bytes}{\code{BotInfo$bytes()}}
\item \href{#method-BotInfo-clone}{\code{BotInfo$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotInfo-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BotInfo with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInfo$new(
  has_preview_medias = NULL,
  user_id = NULL,
  description = NULL,
  description_photo = NULL,
  description_document = NULL,
  commands = NULL,
  menu_button = NULL,
  privacy_policy_url = NULL,
  app_settings = NULL,
  verifier_settings = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{has_preview_medias}}{A boolean indicating if has preview medias. Optional.}

\item{\code{user_id}}{An integer representing the user ID. Optional.}

\item{\code{description}}{A string representing the description. Optional.}

\item{\code{description_photo}}{A TLObject representing the description photo. Optional.}

\item{\code{description_document}}{A TLObject representing the description document. Optional.}

\item{\code{commands}}{A list of TLObjects representing commands. Optional.}

\item{\code{menu_button}}{A TLObject representing the menu button. Optional.}

\item{\code{privacy_policy_url}}{A string representing the privacy policy URL. Optional.}

\item{\code{app_settings}}{A TLObject representing the app settings. Optional.}

\item{\code{verifier_settings}}{A TLObject representing the verifier settings. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInfo-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotInfo-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BotInfo object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInfo$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BotInfo object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInfo-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotInfo-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BotInfo object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInfo$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BotInfo object.
Reads a BotInfo object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
