% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BotInlineMediaResult}
\alias{BotInlineMediaResult}
\title{BotInlineMediaResult}
\description{
Telegram API type BotInlineMediaResult
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BotInlineMediaResult}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotInlineMediaResult-new}{\code{BotInlineMediaResult$new()}}
\item \href{#method-BotInlineMediaResult-to_dict}{\code{BotInlineMediaResult$to_dict()}}
\item \href{#method-BotInlineMediaResult-bytes}{\code{BotInlineMediaResult$bytes()}}
\item \href{#method-BotInlineMediaResult-clone}{\code{BotInlineMediaResult$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMediaResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMediaResult-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BotInlineMediaResult with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMediaResult$new(
  id,
  type,
  send_message,
  photo = NULL,
  document = NULL,
  title = NULL,
  description = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{A string representing the ID.}

\item{\code{type}}{A string representing the type.}

\item{\code{send_message}}{A TLObject representing the send message.}

\item{\code{photo}}{A TLObject representing the photo. Optional.}

\item{\code{document}}{A TLObject representing the document. Optional.}

\item{\code{title}}{A string representing the title. Optional.}

\item{\code{description}}{A string representing the description. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMediaResult-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMediaResult-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BotInlineMediaResult object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMediaResult$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BotInlineMediaResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMediaResult-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMediaResult-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BotInlineMediaResult object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMediaResult$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BotInlineMediaResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMediaResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMediaResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMediaResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
