% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BotInlineMessageMediaGeo}
\alias{BotInlineMessageMediaGeo}
\title{BotInlineMessageMediaGeo}
\description{
Telegram API type BotInlineMessageMediaGeo
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BotInlineMessageMediaGeo}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotInlineMessageMediaGeo-new}{\code{BotInlineMessageMediaGeo$new()}}
\item \href{#method-BotInlineMessageMediaGeo-to_dict}{\code{BotInlineMessageMediaGeo$to_dict()}}
\item \href{#method-BotInlineMessageMediaGeo-bytes}{\code{BotInlineMessageMediaGeo$bytes()}}
\item \href{#method-BotInlineMessageMediaGeo-clone}{\code{BotInlineMessageMediaGeo$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaGeo-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaGeo-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BotInlineMessageMediaGeo with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaGeo$new(
  geo,
  heading = NULL,
  period = NULL,
  proximity_notification_radius = NULL,
  reply_markup = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geo}}{A TLObject representing the geo point.}

\item{\code{heading}}{An integer representing the heading. Optional.}

\item{\code{period}}{An integer representing the period. Optional.}

\item{\code{proximity_notification_radius}}{An integer representing the proximity notification radius. Optional.}

\item{\code{reply_markup}}{A TLObject representing the reply markup. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaGeo-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaGeo-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BotInlineMessageMediaGeo object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaGeo$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BotInlineMessageMediaGeo object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaGeo-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaGeo-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BotInlineMessageMediaGeo object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaGeo$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BotInlineMessageMediaGeo object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaGeo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaGeo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaGeo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
