% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BotInlineMessageMediaInvoice}
\alias{BotInlineMessageMediaInvoice}
\title{BotInlineMessageMediaInvoice}
\description{
Telegram API type BotInlineMessageMediaInvoice
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BotInlineMessageMediaInvoice}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{title}}{Field.}

\item{\code{description}}{Field.}

\item{\code{currency}}{Field.}

\item{\code{total_amount}}{Field.}

\item{\code{shipping_address_requested}}{Field.}

\item{\code{test}}{Field.}

\item{\code{photo}}{Field.}

\item{\code{reply_markup}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotInlineMessageMediaInvoice-new}{\code{BotInlineMessageMediaInvoice$new()}}
\item \href{#method-BotInlineMessageMediaInvoice-to_dict}{\code{BotInlineMessageMediaInvoice$to_dict()}}
\item \href{#method-BotInlineMessageMediaInvoice-bytes}{\code{BotInlineMessageMediaInvoice$bytes()}}
\item \href{#method-BotInlineMessageMediaInvoice-clone}{\code{BotInlineMessageMediaInvoice$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaInvoice-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaInvoice-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BotInlineMessageMediaInvoice with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaInvoice$new(
  title,
  description,
  currency,
  total_amount,
  shipping_address_requested = NULL,
  test = NULL,
  photo = NULL,
  reply_markup = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{A string representing the title.}

\item{\code{description}}{A string representing the description.}

\item{\code{currency}}{A string representing the currency.}

\item{\code{total_amount}}{An integer representing the total amount.}

\item{\code{shipping_address_requested}}{A boolean indicating if shipping address is requested. Optional.}

\item{\code{test}}{A boolean indicating if it's a test. Optional.}

\item{\code{photo}}{A TLObject representing the photo. Optional.}

\item{\code{reply_markup}}{A TLObject representing the reply markup. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaInvoice-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaInvoice-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BotInlineMessageMediaInvoice object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaInvoice$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BotInlineMessageMediaInvoice object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaInvoice-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaInvoice-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BotInlineMessageMediaInvoice object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaInvoice$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BotInlineMessageMediaInvoice object.
Reads a BotInlineMessageMediaInvoice object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaInvoice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaInvoice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaInvoice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
