% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BotInlineMessageMediaVenue}
\alias{BotInlineMessageMediaVenue}
\title{BotInlineMessageMediaVenue}
\description{
Telegram API type BotInlineMessageMediaVenue
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BotInlineMessageMediaVenue}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{geo}}{Field.}

\item{\code{title}}{Field.}

\item{\code{address}}{Field.}

\item{\code{provider}}{Field.}

\item{\code{venue_id}}{Field.}

\item{\code{venue_type}}{Field.}

\item{\code{reply_markup}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotInlineMessageMediaVenue-new}{\code{BotInlineMessageMediaVenue$new()}}
\item \href{#method-BotInlineMessageMediaVenue-to_dict}{\code{BotInlineMessageMediaVenue$to_dict()}}
\item \href{#method-BotInlineMessageMediaVenue-bytes}{\code{BotInlineMessageMediaVenue$bytes()}}
\item \href{#method-BotInlineMessageMediaVenue-clone}{\code{BotInlineMessageMediaVenue$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaVenue-new"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaVenue-new}{}}}
\subsection{Method \code{new()}}{
Initializes the BotInlineMessageMediaVenue with the given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaVenue$new(
  geo,
  title,
  address,
  provider,
  venue_id,
  venue_type,
  reply_markup = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geo}}{A TLObject representing the geo point.}

\item{\code{title}}{A string representing the title.}

\item{\code{address}}{A string representing the address.}

\item{\code{provider}}{A string representing the provider.}

\item{\code{venue_id}}{A string representing the venue ID.}

\item{\code{venue_type}}{A string representing the venue type.}

\item{\code{reply_markup}}{A TLObject representing the reply markup. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaVenue-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaVenue-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the BotInlineMessageMediaVenue object to a list (dictionary).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaVenue$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the BotInlineMessageMediaVenue object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaVenue-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaVenue-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serializes the BotInlineMessageMediaVenue object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaVenue$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized bytes of the BotInlineMessageMediaVenue object.
Reads a BotInlineMessageMediaVenue object from a BinaryReader.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotInlineMessageMediaVenue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotInlineMessageMediaVenue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotInlineMessageMediaVenue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
