% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bots.R
\name{BotMethods}
\alias{BotMethods}
\title{BotMethods Class}
\description{
A class representing methods for interacting with Telegram bots.
}
\details{
This class provides methods to interact with Telegram bots, including making inline queries.
It is designed to be used with the Telegram API and requires a valid Telegram client instance.
}
\examples{
# Assuming `client` is a valid Telegram client instance
bot_methods <- BotMethods$new(client)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BotMethods-inline_query}{\code{BotMethods$inline_query()}}
\item \href{#method-BotMethods-get_input_entity}{\code{BotMethods$get_input_entity()}}
\item \href{#method-BotMethods-invoke_function}{\code{BotMethods$invoke_function()}}
\item \href{#method-BotMethods-custom_inline_results}{\code{BotMethods$custom_inline_results()}}
\item \href{#method-BotMethods-clone}{\code{BotMethods$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotMethods-inline_query"></a>}}
\if{latex}{\out{\hypertarget{method-BotMethods-inline_query}{}}}
\subsection{Method \code{inline_query()}}{
Makes an inline query to the specified bot (e.g., "@vote New Poll").
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotMethods$inline_query(
  bot,
  query,
  entity = NULL,
  offset = NULL,
  geo_point = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bot}}{The bot entity to which the inline query should be made.}

\item{\code{query}}{The query string to send to the bot.}

\item{\code{entity}}{(Optional) The entity where the inline query is being made from.}

\item{\code{offset}}{(Optional) The string offset to use for the bot.}

\item{\code{geo_point}}{(Optional) Geo point location information for localized results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of inline results.
Placeholder for getting input entity.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotMethods-get_input_entity"></a>}}
\if{latex}{\out{\hypertarget{method-BotMethods-get_input_entity}{}}}
\subsection{Method \code{get_input_entity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotMethods$get_input_entity(entity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{The entity to process.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Processed entity.
Placeholder for invoking Telegram API functions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotMethods-invoke_function"></a>}}
\if{latex}{\out{\hypertarget{method-BotMethods-invoke_function}{}}}
\subsection{Method \code{invoke_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotMethods$invoke_function(function_name, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{function_name}}{The name of the function to invoke.}

\item{\code{params}}{The parameters for the function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response.
Placeholder for processing inline results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotMethods-custom_inline_results"></a>}}
\if{latex}{\out{\hypertarget{method-BotMethods-custom_inline_results}{}}}
\subsection{Method \code{custom_inline_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotMethods$custom_inline_results(result, peer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{The raw result from the API.}

\item{\code{peer}}{The peer entity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Processed inline results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BotMethods-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BotMethods-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BotMethods$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
