% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{cancel}
\alias{cancel}
\title{Cancel a future}
\usage{
cancel(future)
}
\arguments{
\item{future}{The future object to cancel}
}
\value{
Invisible NULL
}
\description{
Attempts to safely cancel a running future task.
}
