% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotosender.R
\name{cancel_futures}
\alias{cancel_futures}
\title{Cancel futures if they are not resolved}
\usage{
cancel_futures(log, ...)
}
\arguments{
\item{log}{Logger object for logging messages}

\item{...}{Named future objects to cancel}
}
\description{
Cancel futures if they are not resolved
}
