% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{CancelCodeRequest}
\alias{CancelCodeRequest}
\title{CancelCodeRequest R6 class}
\description{
CancelCodeRequest R6 class

CancelCodeRequest R6 class
}
\details{
Represents the TLRequest auth.CancelCodeRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0x1f040578).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0xf5b399ac).}

\item{\code{phone_number}}{character Phone number.}

\item{\code{phone_code_hash}}{character Phone code hash.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CancelCodeRequest-new}{\code{CancelCodeRequest$new()}}
\item \href{#method-CancelCodeRequest-to_list}{\code{CancelCodeRequest$to_list()}}
\item \href{#method-CancelCodeRequest-to_bytes}{\code{CancelCodeRequest$to_bytes()}}
\item \href{#method-CancelCodeRequest-from_reader}{\code{CancelCodeRequest$from_reader()}}
\item \href{#method-CancelCodeRequest-clone}{\code{CancelCodeRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CancelCodeRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-CancelCodeRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a CancelCodeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CancelCodeRequest$new(phone_number, phone_code_hash)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phone_number}}{character}

\item{\code{phone_code_hash}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CancelCodeRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CancelCodeRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CancelCodeRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CancelCodeRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-CancelCodeRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CancelCodeRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CancelCodeRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-CancelCodeRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create CancelCodeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CancelCodeRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing tgread_string()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CancelCodeRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CancelCodeRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CancelCodeRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CancelCodeRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
