% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{CdnFileTamperedError}
\alias{CdnFileTamperedError}
\title{CdnFileTamperedError}
\description{
Telegram API type CdnFileTamperedError
}
\details{
CdnFileTamperedError
}
\section{Super classes}{
\code{\link[telegramR:Exception]{telegramR::Exception}} -> \code{\link[telegramR:SecurityError]{telegramR::SecurityError}} -> \code{CdnFileTamperedError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CdnFileTamperedError-new}{\code{CdnFileTamperedError$new()}}
\item \href{#method-CdnFileTamperedError-clone}{\code{CdnFileTamperedError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdnFileTamperedError-new"></a>}}
\if{latex}{\out{\hypertarget{method-CdnFileTamperedError-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the CdnFileTamperedError class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdnFileTamperedError$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdnFileTamperedError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CdnFileTamperedError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdnFileTamperedError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
