% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{CdnRedirect}
\alias{CdnRedirect}
\title{CdnRedirect class}
\description{
Class for handling CDN redirects.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cdn_redirect}}{The CDN redirect object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CdnRedirect-new}{\code{CdnRedirect$new()}}
\item \href{#method-CdnRedirect-clone}{\code{CdnRedirect$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdnRedirect-new"></a>}}
\if{latex}{\out{\hypertarget{method-CdnRedirect-new}{}}}
\subsection{Method \code{new()}}{
Create a new CdnRedirect instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdnRedirect$new(cdn_redirect = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cdn_redirect}}{The CDN redirect object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CdnRedirect-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CdnRedirect-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CdnRedirect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
