% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ChannelAdminLogEventActionChangeProfilePeerColor}
\alias{ChannelAdminLogEventActionChangeProfilePeerColor}
\title{ChannelAdminLogEventActionChangeProfilePeerColor}
\description{
Telegram API type ChannelAdminLogEventActionChangeProfilePeerColor
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{ChannelAdminLogEventActionChangeProfilePeerColor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChannelAdminLogEventActionChangeProfilePeerColor-new}{\code{ChannelAdminLogEventActionChangeProfilePeerColor$new()}}
\item \href{#method-ChannelAdminLogEventActionChangeProfilePeerColor-to_dict}{\code{ChannelAdminLogEventActionChangeProfilePeerColor$to_dict()}}
\item \href{#method-ChannelAdminLogEventActionChangeProfilePeerColor-bytes}{\code{ChannelAdminLogEventActionChangeProfilePeerColor$bytes()}}
\item \href{#method-ChannelAdminLogEventActionChangeProfilePeerColor-clone}{\code{ChannelAdminLogEventActionChangeProfilePeerColor$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventActionChangeProfilePeerColor-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventActionChangeProfilePeerColor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventActionChangeProfilePeerColor$new(prev_value, new_value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventActionChangeProfilePeerColor-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventActionChangeProfilePeerColor-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventActionChangeProfilePeerColor$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventActionChangeProfilePeerColor-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventActionChangeProfilePeerColor-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventActionChangeProfilePeerColor$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventActionChangeProfilePeerColor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventActionChangeProfilePeerColor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventActionChangeProfilePeerColor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
