% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ChannelAdminLogEventsFilter}
\alias{ChannelAdminLogEventsFilter}
\title{ChannelAdminLogEventsFilter}
\description{
Telegram API type ChannelAdminLogEventsFilter

Telegram API type ChannelAdminLogEventsFilter
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{ChannelAdminLogEventsFilter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{join}}{Field.}

\item{\code{leave}}{Field.}

\item{\code{invite}}{Field.}

\item{\code{ban}}{Field.}

\item{\code{unban}}{Field.}

\item{\code{kick}}{Field.}

\item{\code{unkick}}{Field.}

\item{\code{promote}}{Field.}

\item{\code{demote}}{Field.}

\item{\code{info}}{Field.}

\item{\code{settings}}{Field.}

\item{\code{pinned}}{Field.}

\item{\code{edit}}{Field.}

\item{\code{delete}}{Field.}

\item{\code{group_call}}{Field.}

\item{\code{invites}}{Field.}

\item{\code{send}}{Field.}

\item{\code{forums}}{Field.}

\item{\code{sub_extend}}{Field.}

\item{\code{class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{join}}{Field.}

\item{\code{leave}}{Field.}

\item{\code{invite}}{Field.}

\item{\code{ban}}{Field.}

\item{\code{unban}}{Field.}

\item{\code{kick}}{Field.}

\item{\code{unkick}}{Field.}

\item{\code{promote}}{Field.}

\item{\code{demote}}{Field.}

\item{\code{info}}{Field.}

\item{\code{settings}}{Field.}

\item{\code{pinned}}{Field.}

\item{\code{edit}}{Field.}

\item{\code{delete}}{Field.}

\item{\code{group_call}}{Field.}

\item{\code{invites}}{Field.}

\item{\code{send}}{Field.}

\item{\code{forums}}{Field.}

\item{\code{sub_extend}}{Field.}

\item{\code{class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChannelAdminLogEventsFilter-new}{\code{ChannelAdminLogEventsFilter$new()}}
\item \href{#method-ChannelAdminLogEventsFilter-to_dict}{\code{ChannelAdminLogEventsFilter$to_dict()}}
\item \href{#method-ChannelAdminLogEventsFilter-bytes}{\code{ChannelAdminLogEventsFilter$bytes()}}
\item \href{#method-ChannelAdminLogEventsFilter-clone}{\code{ChannelAdminLogEventsFilter$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventsFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventsFilter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventsFilter$new(
  join = NULL,
  leave = NULL,
  invite = NULL,
  ban = NULL,
  unban = NULL,
  kick = NULL,
  unkick = NULL,
  promote = NULL,
  demote = NULL,
  info = NULL,
  settings = NULL,
  pinned = NULL,
  edit = NULL,
  delete = NULL,
  group_call = NULL,
  invites = NULL,
  send = NULL,
  forums = NULL,
  sub_extend = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventsFilter-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventsFilter-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventsFilter$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventsFilter-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventsFilter-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventsFilter$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventsFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventsFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventsFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{ChannelAdminLogEventsFilter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChannelAdminLogEventsFilter-new}{\code{ChannelAdminLogEventsFilter$new()}}
\item \href{#method-ChannelAdminLogEventsFilter-to_dict}{\code{ChannelAdminLogEventsFilter$to_dict()}}
\item \href{#method-ChannelAdminLogEventsFilter-bytes}{\code{ChannelAdminLogEventsFilter$bytes()}}
\item \href{#method-ChannelAdminLogEventsFilter-clone}{\code{ChannelAdminLogEventsFilter$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventsFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventsFilter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventsFilter$new(
  join = NULL,
  leave = NULL,
  invite = NULL,
  ban = NULL,
  unban = NULL,
  kick = NULL,
  unkick = NULL,
  promote = NULL,
  demote = NULL,
  info = NULL,
  settings = NULL,
  pinned = NULL,
  edit = NULL,
  delete = NULL,
  group_call = NULL,
  invites = NULL,
  send = NULL,
  forums = NULL,
  sub_extend = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventsFilter-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventsFilter-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventsFilter$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventsFilter-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventsFilter-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventsFilter$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelAdminLogEventsFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelAdminLogEventsFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelAdminLogEventsFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
