% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ChannelFull}
\alias{ChannelFull}
\title{ChannelFull}
\description{
Telegram API type ChannelFull

Telegram API type ChannelFull
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{ChannelFull}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChannelFull-new}{\code{ChannelFull$new()}}
\item \href{#method-ChannelFull-to_dict}{\code{ChannelFull$to_dict()}}
\item \href{#method-ChannelFull-clone}{\code{ChannelFull$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelFull-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelFull-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelFull$new(
  id,
  about,
  read_inbox_max_id,
  read_outbox_max_id,
  unread_count,
  chat_photo,
  notify_settings,
  bot_info,
  pts,
  can_view_participants = NULL,
  can_set_username = NULL,
  can_set_stickers = NULL,
  hidden_prehistory = NULL,
  can_set_location = NULL,
  has_scheduled = NULL,
  can_view_stats = NULL,
  blocked = NULL,
  can_delete_channel = NULL,
  antispam = NULL,
  participants_hidden = NULL,
  translations_disabled = NULL,
  stories_pinned_available = NULL,
  view_forum_as_messages = NULL,
  restricted_sponsored = NULL,
  can_view_revenue = NULL,
  paid_media_allowed = NULL,
  can_view_stars_revenue = NULL,
  paid_reactions_available = NULL,
  stargifts_available = NULL,
  paid_messages_available = NULL,
  participants_count = NULL,
  admins_count = NULL,
  kicked_count = NULL,
  banned_count = NULL,
  online_count = NULL,
  exported_invite = NULL,
  migrated_from_chat_id = NULL,
  migrated_from_max_id = NULL,
  pinned_msg_id = NULL,
  stickerset = NULL,
  available_min_id = NULL,
  folder_id = NULL,
  linked_chat_id = NULL,
  location = NULL,
  slowmode_seconds = NULL,
  slowmode_next_send_date = NULL,
  stats_dc = NULL,
  call = NULL,
  ttl_period = NULL,
  pending_suggestions = NULL,
  groupcall_default_join_as = NULL,
  theme_emoticon = NULL,
  requests_pending = NULL,
  recent_requesters = NULL,
  default_send_as = NULL,
  available_reactions = NULL,
  reactions_limit = NULL,
  stories = NULL,
  wallpaper = NULL,
  boosts_applied = NULL,
  boosts_unrestrict = NULL,
  emojiset = NULL,
  bot_verification = NULL,
  stargifts_count = NULL,
  send_paid_messages_stars = NULL,
  main_tab = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelFull-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelFull-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelFull$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelFull-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelFull-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelFull$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{ChannelFull}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChannelFull-new}{\code{ChannelFull$new()}}
\item \href{#method-ChannelFull-to_dict}{\code{ChannelFull$to_dict()}}
\item \href{#method-ChannelFull-clone}{\code{ChannelFull$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelFull-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelFull-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelFull$new(
  id,
  about,
  read_inbox_max_id,
  read_outbox_max_id,
  unread_count,
  chat_photo,
  notify_settings,
  bot_info,
  pts,
  can_view_participants = NULL,
  can_set_username = NULL,
  can_set_stickers = NULL,
  hidden_prehistory = NULL,
  can_set_location = NULL,
  has_scheduled = NULL,
  can_view_stats = NULL,
  blocked = NULL,
  can_delete_channel = NULL,
  antispam = NULL,
  participants_hidden = NULL,
  translations_disabled = NULL,
  stories_pinned_available = NULL,
  view_forum_as_messages = NULL,
  restricted_sponsored = NULL,
  can_view_revenue = NULL,
  paid_media_allowed = NULL,
  can_view_stars_revenue = NULL,
  paid_reactions_available = NULL,
  stargifts_available = NULL,
  paid_messages_available = NULL,
  participants_count = NULL,
  admins_count = NULL,
  kicked_count = NULL,
  banned_count = NULL,
  online_count = NULL,
  exported_invite = NULL,
  migrated_from_chat_id = NULL,
  migrated_from_max_id = NULL,
  pinned_msg_id = NULL,
  stickerset = NULL,
  available_min_id = NULL,
  folder_id = NULL,
  linked_chat_id = NULL,
  location = NULL,
  slowmode_seconds = NULL,
  slowmode_next_send_date = NULL,
  stats_dc = NULL,
  call = NULL,
  ttl_period = NULL,
  pending_suggestions = NULL,
  groupcall_default_join_as = NULL,
  theme_emoticon = NULL,
  requests_pending = NULL,
  recent_requesters = NULL,
  default_send_as = NULL,
  available_reactions = NULL,
  reactions_limit = NULL,
  stories = NULL,
  wallpaper = NULL,
  boosts_applied = NULL,
  boosts_unrestrict = NULL,
  emojiset = NULL,
  bot_verification = NULL,
  stargifts_count = NULL,
  send_paid_messages_stars = NULL,
  main_tab = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelFull-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelFull-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelFull$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChannelFull-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChannelFull-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChannelFull$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
