% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ChatFull}
\alias{ChatFull}
\title{ChatFull}
\description{
Telegram API type ChatFull
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{ChatFull}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{id}}{Field.}

\item{\code{about}}{Field.}

\item{\code{participants}}{Field.}

\item{\code{notify_settings}}{Field.}

\item{\code{can_set_username}}{Field.}

\item{\code{has_scheduled}}{Field.}

\item{\code{translations_disabled}}{Field.}

\item{\code{chat_photo}}{Field.}

\item{\code{exported_invite}}{Field.}

\item{\code{bot_info}}{Field.}

\item{\code{pinned_msg_id}}{Field.}

\item{\code{folder_id}}{Field.}

\item{\code{call}}{Field.}

\item{\code{ttl_period}}{Field.}

\item{\code{groupcall_default_join_as}}{Field.}

\item{\code{theme_emoticon}}{Field.}

\item{\code{requests_pending}}{Field.}

\item{\code{recent_requesters}}{Field.}

\item{\code{available_reactions}}{Field.}

\item{\code{reactions_limit}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChatFull-new}{\code{ChatFull$new()}}
\item \href{#method-ChatFull-to_list}{\code{ChatFull$to_list()}}
\item \href{#method-ChatFull-from_reader}{\code{ChatFull$from_reader()}}
\item \href{#method-ChatFull-clone}{\code{ChatFull$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatFull-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChatFull-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatFull$new(
  id,
  about,
  participants,
  notify_settings,
  can_set_username = NULL,
  has_scheduled = NULL,
  translations_disabled = NULL,
  chat_photo = NULL,
  exported_invite = NULL,
  bot_info = NULL,
  pinned_msg_id = NULL,
  folder_id = NULL,
  call = NULL,
  ttl_period = NULL,
  groupcall_default_join_as = NULL,
  theme_emoticon = NULL,
  requests_pending = NULL,
  recent_requesters = NULL,
  available_reactions = NULL,
  reactions_limit = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatFull-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ChatFull-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatFull$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatFull-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ChatFull-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatFull$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatFull-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChatFull-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatFull$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
