% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invite_link.R
\name{check_invite_link}
\alias{check_invite_link}
\title{Check a chat invite link without joining}
\usage{
check_invite_link(client, invite_link)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{invite_link}{Invite link or invite hash.}
}
\value{
The API response.
}
\description{
Check a chat invite link without joining
}
