% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R, R/utils.R
\name{check_prime_and_good}
\alias{check_prime_and_good}
\title{Check Prime and Generator}
\usage{
check_prime_and_good(prime_bytes, g)

check_prime_and_good(prime_bytes, g)
}
\arguments{
\item{prime_bytes}{A raw vector representing the prime bytes.}

\item{g}{An integer representing the generator.}
}
\value{
Raises an error if validation fails, otherwise returns nothing.
}
\description{
Checks if the prime bytes match a known good prime and validates the generator, or performs detailed checks if not matching.

Checks if the prime bytes and generator are good.
}
