% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R, R/utils.R
\name{check_prime_and_good_check}
\alias{check_prime_and_good_check}
\title{Check Prime and Generator with Detailed Checks}
\usage{
check_prime_and_good_check(prime, g)

check_prime_and_good_check(prime, g)
}
\arguments{
\item{prime}{An integer representing the prime number.}

\item{g}{An integer representing the generator.}
}
\value{
Raises an error if validation fails, otherwise returns nothing.
}
\description{
Checks if the prime and generator are good with detailed validation, including bit length, primality, and specific modular conditions.

Validates the prime number and generator g for cryptographic purposes.
}
