% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chatlists.R
\name{CheckChatlistInviteRequest}
\alias{CheckChatlistInviteRequest}
\title{CheckChatlistInviteRequest}
\description{
Telegram API type CheckChatlistInviteRequest
}
\details{
CheckChatlistInviteRequest R6 class

Request to check a chatlist invite by slug.

Serializes constructor id followed by TL-string encoded slug.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{slug}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CheckChatlistInviteRequest-new}{\code{CheckChatlistInviteRequest$new()}}
\item \href{#method-CheckChatlistInviteRequest-to_list}{\code{CheckChatlistInviteRequest$to_list()}}
\item \href{#method-CheckChatlistInviteRequest-bytes}{\code{CheckChatlistInviteRequest$bytes()}}
\item \href{#method-CheckChatlistInviteRequest-from_reader}{\code{CheckChatlistInviteRequest$from_reader()}}
\item \href{#method-CheckChatlistInviteRequest-clone}{\code{CheckChatlistInviteRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckChatlistInviteRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-CheckChatlistInviteRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize CheckChatlistInviteRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckChatlistInviteRequest$new(slug = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{slug}}{Invite slug (string)
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckChatlistInviteRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CheckChatlistInviteRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckChatlistInviteRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representation
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckChatlistInviteRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-CheckChatlistInviteRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckChatlistInviteRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector with serialized bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckChatlistInviteRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-CheckChatlistInviteRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckChatlistInviteRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckChatlistInviteRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CheckChatlistInviteRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckChatlistInviteRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
