% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{CheckGroupCallRequest}
\alias{CheckGroupCallRequest}
\title{CheckGroupCallRequest}
\description{
Telegram API type CheckGroupCallRequest
}
\details{
CheckGroupCallRequest R6 class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{Field.}

\item{\code{sources}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CheckGroupCallRequest-new}{\code{CheckGroupCallRequest$new()}}
\item \href{#method-CheckGroupCallRequest-resolve}{\code{CheckGroupCallRequest$resolve()}}
\item \href{#method-CheckGroupCallRequest-to_list}{\code{CheckGroupCallRequest$to_list()}}
\item \href{#method-CheckGroupCallRequest-bytes}{\code{CheckGroupCallRequest$bytes()}}
\item \href{#method-CheckGroupCallRequest-from_reader}{\code{CheckGroupCallRequest$from_reader()}}
\item \href{#method-CheckGroupCallRequest-clone}{\code{CheckGroupCallRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckGroupCallRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-CheckGroupCallRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a CheckGroupCallRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckGroupCallRequest$new(call, sources)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}

\item{\code{sources}}{List of integers.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckGroupCallRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-CheckGroupCallRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references

Typically used to convert high-level entities into input objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckGroupCallRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckGroupCallRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CheckGroupCallRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckGroupCallRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckGroupCallRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-CheckGroupCallRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckGroupCallRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckGroupCallRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-CheckGroupCallRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckGroupCallRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckGroupCallRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CheckGroupCallRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckGroupCallRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
