% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{CheckPasswordRequest}
\alias{CheckPasswordRequest}
\title{CheckPasswordRequest}
\description{
Telegram API type CheckPasswordRequest
}
\details{
CheckPasswordRequest R6 class

Represents the TLRequest auth.CheckPasswordRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0xd18b4d16).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0xb9e04e39).}

\item{\code{password}}{TL object TypeInputCheckPasswordSRP.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CheckPasswordRequest-new}{\code{CheckPasswordRequest$new()}}
\item \href{#method-CheckPasswordRequest-to_list}{\code{CheckPasswordRequest$to_list()}}
\item \href{#method-CheckPasswordRequest-to_bytes}{\code{CheckPasswordRequest$to_bytes()}}
\item \href{#method-CheckPasswordRequest-from_reader}{\code{CheckPasswordRequest$from_reader()}}
\item \href{#method-CheckPasswordRequest-clone}{\code{CheckPasswordRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckPasswordRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-CheckPasswordRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a CheckPasswordRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckPasswordRequest$new(password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{password}}{TL object (TypeInputCheckPasswordSRP)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckPasswordRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CheckPasswordRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckPasswordRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckPasswordRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-CheckPasswordRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckPasswordRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckPasswordRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-CheckPasswordRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create CheckPasswordRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckPasswordRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CheckPasswordRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckPasswordRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CheckPasswordRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckPasswordRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
