% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{CheckRecoveryPasswordRequest}
\alias{CheckRecoveryPasswordRequest}
\title{CheckRecoveryPasswordRequest}
\description{
Telegram API type CheckRecoveryPasswordRequest
}
\details{
CheckRecoveryPasswordRequest R6 class

Represents the TLRequest auth.CheckRecoveryPasswordRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{code}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CheckRecoveryPasswordRequest-new}{\code{CheckRecoveryPasswordRequest$new()}}
\item \href{#method-CheckRecoveryPasswordRequest-to_list}{\code{CheckRecoveryPasswordRequest$to_list()}}
\item \href{#method-CheckRecoveryPasswordRequest-to_bytes}{\code{CheckRecoveryPasswordRequest$to_bytes()}}
\item \href{#method-CheckRecoveryPasswordRequest-from_reader}{\code{CheckRecoveryPasswordRequest$from_reader()}}
\item \href{#method-CheckRecoveryPasswordRequest-clone}{\code{CheckRecoveryPasswordRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckRecoveryPasswordRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-CheckRecoveryPasswordRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a CheckRecoveryPasswordRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckRecoveryPasswordRequest$new(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckRecoveryPasswordRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CheckRecoveryPasswordRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckRecoveryPasswordRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckRecoveryPasswordRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-CheckRecoveryPasswordRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckRecoveryPasswordRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckRecoveryPasswordRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-CheckRecoveryPasswordRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create CheckRecoveryPasswordRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckRecoveryPasswordRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing tgread_string()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CheckRecoveryPasswordRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckRecoveryPasswordRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CheckRecoveryPasswordRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckRecoveryPasswordRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
