% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunks}
\alias{chunks}
\title{Chunks}
\usage{
chunks(iterable, size = 100L)
}
\arguments{
\item{iterable}{A vector or list to be chunked.}

\item{size}{An integer specifying the size of each chunk. Default is 100L.}
}
\value{
A list of vectors, each representing a chunk of the original iterable.
}
\description{
Turns the given iterable into chunks of the specified size,
which is 100 by default since that's what Telegram uses the most.
}
\examples{
\dontrun{
chunks(1:10, 3)  # Returns list(c(1,2,3), c(4,5,6), c(7,8,9), c(10))
}
}
