% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotosender.R
\name{clone_error}
\alias{clone_error}
\title{Clone an error object}
\usage{
clone_error(error)
}
\arguments{
\item{error}{The error object to clone}
}
\value{
A new error object with the same properties
}
\description{
Creates a copy of an error object preserving its class, message and other attributes
}
