% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_digest}
\alias{compute_digest}
\title{Compute Digest}
\usage{
compute_digest(algo, password)
}
\arguments{
\item{algo}{The algorithm object.}

\item{password}{A string.}
}
\value{
A raw vector.
}
\description{
Computes the digest for password.
}
