% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa.R
\name{compute_fingerprint}
\alias{compute_fingerprint}
\title{Computes the fingerprint of an RSA key.}
\usage{
compute_fingerprint(key)
}
\arguments{
\item{key}{The RSA key as a list with `n` and `e` components, a PEM string, or an openssl pubkey.}
}
\value{
The 8-byte fingerprint as a positive numeric.
}
\description{
Computes the fingerprint of an RSA key.
}
