% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_hash}
\alias{compute_hash}
\title{Compute Hash}
\usage{
compute_hash(algo, password)
}
\arguments{
\item{algo}{The algorithm object.}

\item{password}{A string.}
}
\value{
A raw vector.
}
\description{
Computes the hash for password KDF.
}
