% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{ConnectionHttp}
\alias{ConnectionHttp}
\title{HTTP Connection Class}
\description{
This class implements an HTTP connection using the HTTP packet codec.
}
\section{Super class}{
\code{\link[telegramR:Connection]{telegramR::Connection}} -> \code{ConnectionHttp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConnectionHttp-connect}{\code{ConnectionHttp$connect()}}
\item \href{#method-ConnectionHttp-clone}{\code{ConnectionHttp$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="disconnect"><a href='../../telegramR/html/Connection.html#method-Connection-disconnect'><code>telegramR::Connection$disconnect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="initialize"><a href='../../telegramR/html/Connection.html#method-Connection-initialize'><code>telegramR::Connection$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="is_connected"><a href='../../telegramR/html/Connection.html#method-Connection-is_connected'><code>telegramR::Connection$is_connected()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="recv"><a href='../../telegramR/html/Connection.html#method-Connection-recv'><code>telegramR::Connection$recv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="send"><a href='../../telegramR/html/Connection.html#method-Connection-send'><code>telegramR::Connection$send()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="to_string"><a href='../../telegramR/html/Connection.html#method-Connection-to_string'><code>telegramR::Connection$to_string()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionHttp-connect"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionHttp-connect}{}}}
\subsection{Method \code{connect()}}{
This method connects to the server using SSL if the port equals \code{SSL_PORT}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionHttp$connect(timeout = NULL, ssl = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeout}}{Optional timeout for the connection.}

\item{\code{ssl}}{Optional SSL parameter (ignored in favor of port-based selection).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A future resolving when the connection is established.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionHttp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionHttp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionHttp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
