% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpabridget.R
\name{ConnectionTcpAbridged}
\alias{ConnectionTcpAbridged}
\title{ConnectionTcpAbridged}
\description{
An R6 class that represents a TCP connection using the Abridged Packet Codec.
}
\details{
Connection Module

This module provides connection handling for Telegram API communications.
It relies on the future package for asynchronous operations and R6 for
object-oriented programming.
}
\section{Super class}{
\code{\link[telegramR:Connection]{telegramR::Connection}} -> \code{ConnectionTcpAbridged}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConnectionTcpAbridged-new}{\code{ConnectionTcpAbridged$new()}}
\item \href{#method-ConnectionTcpAbridged-clone}{\code{ConnectionTcpAbridged$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="connect"><a href='../../telegramR/html/Connection.html#method-Connection-connect'><code>telegramR::Connection$connect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="disconnect"><a href='../../telegramR/html/Connection.html#method-Connection-disconnect'><code>telegramR::Connection$disconnect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="recv"><a href='../../telegramR/html/Connection.html#method-Connection-recv'><code>telegramR::Connection$recv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="send"><a href='../../telegramR/html/Connection.html#method-Connection-send'><code>telegramR::Connection$send()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionTcpAbridged-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionTcpAbridged-new}{}}}
\subsection{Method \code{new()}}{
Initializes the connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionTcpAbridged$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters passed to the parent class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionTcpAbridged-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionTcpAbridged-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionTcpAbridged$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
