% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpfull.R
\name{ConnectionTcpFull}
\alias{ConnectionTcpFull}
\title{ConnectionTcpFull}
\description{
Default Telegram mode. Sends 12 additional bytes and calculates the CRC for each packet.
}
\section{Super class}{
\code{\link[telegramR:Connection]{telegramR::Connection}} -> \code{ConnectionTcpFull}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConnectionTcpFull-new}{\code{ConnectionTcpFull$new()}}
\item \href{#method-ConnectionTcpFull-clone}{\code{ConnectionTcpFull$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="connect"><a href='../../telegramR/html/Connection.html#method-Connection-connect'><code>telegramR::Connection$connect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="disconnect"><a href='../../telegramR/html/Connection.html#method-Connection-disconnect'><code>telegramR::Connection$disconnect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="is_connected"><a href='../../telegramR/html/Connection.html#method-Connection-is_connected'><code>telegramR::Connection$is_connected()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="recv"><a href='../../telegramR/html/Connection.html#method-Connection-recv'><code>telegramR::Connection$recv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="send"><a href='../../telegramR/html/Connection.html#method-Connection-send'><code>telegramR::Connection$send()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="to_string"><a href='../../telegramR/html/Connection.html#method-Connection-to_string'><code>telegramR::Connection$to_string()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionTcpFull-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionTcpFull-new}{}}}
\subsection{Method \code{new()}}{
Initialize the TCP full connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionTcpFull$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionTcpFull-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionTcpFull-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionTcpFull$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
