% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpintermediate.R
\name{ConnectionTcpIntermediate}
\alias{ConnectionTcpIntermediate}
\title{ConnectionTcpIntermediate}
\description{
Intermediate mode between ConnectionTcpFull and ConnectionTcpAbridged.
}
\details{
Always sends 4 extra bytes for the packet length.
}
\section{Super class}{
\code{\link[telegramR:Connection]{telegramR::Connection}} -> \code{ConnectionTcpIntermediate}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConnectionTcpIntermediate-new}{\code{ConnectionTcpIntermediate$new()}}
\item \href{#method-ConnectionTcpIntermediate-clone}{\code{ConnectionTcpIntermediate$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="connect"><a href='../../telegramR/html/Connection.html#method-Connection-connect'><code>telegramR::Connection$connect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="disconnect"><a href='../../telegramR/html/Connection.html#method-Connection-disconnect'><code>telegramR::Connection$disconnect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="recv"><a href='../../telegramR/html/Connection.html#method-Connection-recv'><code>telegramR::Connection$recv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="send"><a href='../../telegramR/html/Connection.html#method-Connection-send'><code>telegramR::Connection$send()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionTcpIntermediate-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionTcpIntermediate-new}{}}}
\subsection{Method \code{new()}}{
Initializes the intermediate TCP connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionTcpIntermediate$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters passed to the parent constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionTcpIntermediate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionTcpIntermediate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionTcpIntermediate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
