% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpmtproxy.R
\name{ConnectionTcpMTProxyIntermediate}
\alias{ConnectionTcpMTProxyIntermediate}
\title{ConnectionTcpMTProxyIntermediate}
\description{
Connects to an MTProxy server using the intermediate protocol.
}
\section{Super class}{
\code{\link[telegramR:TcpMTProxy]{telegramR::TcpMTProxy}} -> \code{ConnectionTcpMTProxyIntermediate}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{packet_codec}}{The packet codec used for this connection.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConnectionTcpMTProxyIntermediate-new}{\code{ConnectionTcpMTProxyIntermediate$new()}}
\item \href{#method-ConnectionTcpMTProxyIntermediate-clone}{\code{ConnectionTcpMTProxyIntermediate$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TcpMTProxy" data-id="connect"><a href='../../telegramR/html/TcpMTProxy.html#method-TcpMTProxy-connect'><code>telegramR::TcpMTProxy$connect()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionTcpMTProxyIntermediate-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionTcpMTProxyIntermediate-new}{}}}
\subsection{Method \code{new()}}{
Initialize connection using the intermediate protocol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionTcpMTProxyIntermediate$new(
  ip,
  port,
  dc_id,
  loggers = NULL,
  proxy = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ip}}{IP address as character.}

\item{\code{port}}{Port number as integer.}

\item{\code{dc_id}}{Data center identifier as integer.}

\item{\code{loggers}}{Optional logger.}

\item{\code{proxy}}{A list containing proxy_host, proxy_port, and secret.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionTcpMTProxyIntermediate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionTcpMTProxyIntermediate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionTcpMTProxyIntermediate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
