% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpobfuscated.R
\name{ConnectionTcpObfuscated}
\alias{ConnectionTcpObfuscated}
\title{ConnectionTcpObfuscated}
\description{
Telegram obfuscated2 connection that encrypts every message with a randomly generated key using AES-CTR.
}
\section{Super classes}{
\code{\link[telegramR:Connection]{telegramR::Connection}} -> \code{\link[telegramR:ObfuscatedConnection]{telegramR::ObfuscatedConnection}} -> \code{ConnectionTcpObfuscated}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{obfuscated_io}}{ObfuscatedIO class for handling obfuscated I/O operations.}

\item{\code{packet_codec}}{Packet codec set to AbridgedPacketCodec.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConnectionTcpObfuscated-clone}{\code{ConnectionTcpObfuscated$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="connect"><a href='../../telegramR/html/Connection.html#method-Connection-connect'><code>telegramR::Connection$connect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="disconnect"><a href='../../telegramR/html/Connection.html#method-Connection-disconnect'><code>telegramR::Connection$disconnect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="initialize"><a href='../../telegramR/html/Connection.html#method-Connection-initialize'><code>telegramR::Connection$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="is_connected"><a href='../../telegramR/html/Connection.html#method-Connection-is_connected'><code>telegramR::Connection$is_connected()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="recv"><a href='../../telegramR/html/Connection.html#method-Connection-recv'><code>telegramR::Connection$recv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="send"><a href='../../telegramR/html/Connection.html#method-Connection-send'><code>telegramR::Connection$send()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="to_string"><a href='../../telegramR/html/Connection.html#method-Connection-to_string'><code>telegramR::Connection$to_string()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionTcpObfuscated-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionTcpObfuscated-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionTcpObfuscated$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
