% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{CreateConferenceCallRequest}
\alias{CreateConferenceCallRequest}
\title{CreateConferenceCallRequest}
\description{
Telegram API type CreateConferenceCallRequest
}
\details{
CreateConferenceCallRequest R6 class

Represents CreateConferenceCallRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{muted}}{Logical or NULL. If TRUE, muted.}

\item{\code{video_stopped}}{Logical or NULL. If TRUE, video stopped.}

\item{\code{join}}{Logical or NULL. If TRUE, join.}

\item{\code{random_id}}{Integer random id.}

\item{\code{public_key}}{Integer or NULL. 256-bit public key.}

\item{\code{block}}{Raw bytes or NULL.}

\item{\code{params}}{DataJSON-like object or NULL.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CreateConferenceCallRequest-new}{\code{CreateConferenceCallRequest$new()}}
\item \href{#method-CreateConferenceCallRequest-to_list}{\code{CreateConferenceCallRequest$to_list()}}
\item \href{#method-CreateConferenceCallRequest-bytes}{\code{CreateConferenceCallRequest$bytes()}}
\item \href{#method-CreateConferenceCallRequest-from_reader}{\code{CreateConferenceCallRequest$from_reader()}}
\item \href{#method-CreateConferenceCallRequest-clone}{\code{CreateConferenceCallRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateConferenceCallRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-CreateConferenceCallRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a CreateConferenceCallRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateConferenceCallRequest$new(
  muted = NULL,
  video_stopped = NULL,
  join = NULL,
  random_id = NULL,
  public_key = NULL,
  block = NULL,
  params = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{muted}}{Logical or NULL.}

\item{\code{video_stopped}}{Logical or NULL.}

\item{\code{join}}{Logical or NULL.}

\item{\code{random_id}}{Integer or NULL (if NULL, a random 32-bit signed int is generated).}

\item{\code{public_key}}{Integer or NULL.}

\item{\code{block}}{Raw bytes or NULL.}

\item{\code{params}}{DataJSON-like object or NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateConferenceCallRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CreateConferenceCallRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateConferenceCallRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateConferenceCallRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-CreateConferenceCallRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateConferenceCallRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateConferenceCallRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-CreateConferenceCallRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateConferenceCallRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateConferenceCallRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CreateConferenceCallRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateConferenceCallRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
