% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CustomMode}
\alias{CustomMode}
\title{CustomMode}
\description{
An R6 class representing a custom parse mode with parse and unparse methods.
This class is used when a callable function is provided as the parse mode.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parse}}{A function to parse text into entities.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomMode-new}{\code{CustomMode$new()}}
\item \href{#method-CustomMode-unparse}{\code{CustomMode$unparse()}}
\item \href{#method-CustomMode-clone}{\code{CustomMode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomMode-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomMode-new}{}}}
\subsection{Method \code{new()}}{
Initialize the CustomMode object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomMode$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomMode-unparse"></a>}}
\if{latex}{\out{\hypertarget{method-CustomMode-unparse}{}}}
\subsection{Method \code{unparse()}}{
Unparse entities back to text.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomMode$unparse(text, entities)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{The text string.}

\item{\code{entities}}{The entities list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Raises NotImplementedError.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomMode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomMode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomMode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
