% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlobject.R
\name{datetime_to_timestamp}
\alias{datetime_to_timestamp}
\title{Convert a datetime object to a timestamp.}
\usage{
datetime_to_timestamp(dt)
}
\arguments{
\item{dt}{A datetime object to be converted.}
}
\value{
A numeric value representing the timestamp in seconds since the epoch.
}
\description{
This function converts a datetime object to a timestamp in seconds since the epoch.
It handles both UTC and local timezones.
}
\examples{
datetime_to_timestamp(as.POSIXct("2023-01-01 12:00:00", tz = "UTC"))
datetime_to_timestamp(as.POSIXct("2023-01-01 12:00:00"))
}
