% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{DeclineConferenceCallInviteRequest}
\alias{DeclineConferenceCallInviteRequest}
\title{DeclineConferenceCallInviteRequest}
\description{
Telegram API type DeclineConferenceCallInviteRequest
}
\details{
DeclineConferenceCallInviteRequest R6 class

Represents DeclineConferenceCallInviteRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{msg_id}}{Integer message id.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeclineConferenceCallInviteRequest-new}{\code{DeclineConferenceCallInviteRequest$new()}}
\item \href{#method-DeclineConferenceCallInviteRequest-to_list}{\code{DeclineConferenceCallInviteRequest$to_list()}}
\item \href{#method-DeclineConferenceCallInviteRequest-bytes}{\code{DeclineConferenceCallInviteRequest$bytes()}}
\item \href{#method-DeclineConferenceCallInviteRequest-from_reader}{\code{DeclineConferenceCallInviteRequest$from_reader()}}
\item \href{#method-DeclineConferenceCallInviteRequest-clone}{\code{DeclineConferenceCallInviteRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclineConferenceCallInviteRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeclineConferenceCallInviteRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a DeclineConferenceCallInviteRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclineConferenceCallInviteRequest$new(msg_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg_id}}{Integer message id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclineConferenceCallInviteRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DeclineConferenceCallInviteRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclineConferenceCallInviteRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclineConferenceCallInviteRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-DeclineConferenceCallInviteRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclineConferenceCallInviteRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclineConferenceCallInviteRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-DeclineConferenceCallInviteRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclineConferenceCallInviteRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclineConferenceCallInviteRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeclineConferenceCallInviteRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclineConferenceCallInviteRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
