% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegrambaseclient.R
\docType{data}
\name{DEFAULT_DC_ID}
\alias{DEFAULT_DC_ID}
\title{Constants}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_DC_ID
}
\description{
Constants
}
\keyword{datasets}
