% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\docType{data}
\name{DEFAULT_URL_FORMAT}
\alias{DEFAULT_URL_FORMAT}
\title{Default URL format used by unparse (deprecated in some flows).}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_URL_FORMAT
}
\description{
Default URL format used by unparse (deprecated in some flows).
}
\keyword{datasets}
