% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\docType{data}
\name{DEFAULT_URL_RE}
\alias{DEFAULT_URL_RE}
\title{Default regular expression to match inline Markdown-style URLs.}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_URL_RE
}
\description{
The pattern captures the link text in group 1 and the URL in group 2.
}
\keyword{datasets}
