% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{DeleteConferenceCallParticipantsRequest}
\alias{DeleteConferenceCallParticipantsRequest}
\title{DeleteConferenceCallParticipantsRequest R6 class}
\description{
Represents DeleteConferenceCallParticipantsRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{The input group call object (expected to implement bytes()).}

\item{\code{ids}}{List of integers.}

\item{\code{block}}{Raw bytes.}

\item{\code{only_left}}{Logical or NULL. If TRUE, only left.}

\item{\code{kick}}{Logical or NULL. If TRUE, kick.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeleteConferenceCallParticipantsRequest-new}{\code{DeleteConferenceCallParticipantsRequest$new()}}
\item \href{#method-DeleteConferenceCallParticipantsRequest-resolve}{\code{DeleteConferenceCallParticipantsRequest$resolve()}}
\item \href{#method-DeleteConferenceCallParticipantsRequest-to_list}{\code{DeleteConferenceCallParticipantsRequest$to_list()}}
\item \href{#method-DeleteConferenceCallParticipantsRequest-bytes}{\code{DeleteConferenceCallParticipantsRequest$bytes()}}
\item \href{#method-DeleteConferenceCallParticipantsRequest-from_reader}{\code{DeleteConferenceCallParticipantsRequest$from_reader()}}
\item \href{#method-DeleteConferenceCallParticipantsRequest-clone}{\code{DeleteConferenceCallParticipantsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteConferenceCallParticipantsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteConferenceCallParticipantsRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteConferenceCallParticipantsRequest$new(
  call,
  ids,
  block,
  only_left = NULL,
  kick = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}

\item{\code{ids}}{List of integers.}

\item{\code{block}}{Raw bytes.}

\item{\code{only_left}}{Logical or NULL.}

\item{\code{kick}}{Logical or NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Resolve input references

Typically used to convert high-level entities into input objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteConferenceCallParticipantsRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteConferenceCallParticipantsRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteConferenceCallParticipantsRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Convert object to a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteConferenceCallParticipantsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteConferenceCallParticipantsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteConferenceCallParticipantsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteConferenceCallParticipantsRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteConferenceCallParticipantsRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteConferenceCallParticipantsRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteConferenceCallParticipantsRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteConferenceCallParticipantsRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteConferenceCallParticipantsRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteConferenceCallParticipantsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteConferenceCallParticipantsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteConferenceCallParticipantsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
