% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chatlists.R
\name{DeleteExportedInviteRequest}
\alias{DeleteExportedInviteRequest}
\title{DeleteExportedInviteRequest}
\description{
Telegram API type DeleteExportedInviteRequest
}
\details{
DeleteExportedInviteRequest R6 class

Request to delete an exported invite for a chatlist.

Serializes constructor id, chatlist bytes, then TL-string encoded slug.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Integer constructor id.}

\item{\code{SUBCLASS_OF_ID}}{Integer subclass id.}

\item{\code{chatlist}}{TypeInputChatlist input chatlist object.}

\item{\code{slug}}{character invite slug.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeleteExportedInviteRequest-new}{\code{DeleteExportedInviteRequest$new()}}
\item \href{#method-DeleteExportedInviteRequest-to_list}{\code{DeleteExportedInviteRequest$to_list()}}
\item \href{#method-DeleteExportedInviteRequest-bytes}{\code{DeleteExportedInviteRequest$bytes()}}
\item \href{#method-DeleteExportedInviteRequest-from_reader}{\code{DeleteExportedInviteRequest$from_reader()}}
\item \href{#method-DeleteExportedInviteRequest-clone}{\code{DeleteExportedInviteRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteExportedInviteRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteExportedInviteRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize DeleteExportedInviteRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteExportedInviteRequest$new(chatlist = NULL, slug = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chatlist}}{Input chatlist object}

\item{\code{slug}}{Invite slug (string)
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteExportedInviteRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteExportedInviteRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteExportedInviteRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representation
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteExportedInviteRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteExportedInviteRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteExportedInviteRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector with serialized bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteExportedInviteRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteExportedInviteRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteExportedInviteRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteExportedInviteRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteExportedInviteRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteExportedInviteRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
