% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stickers.R
\name{DeleteStickerSetRequest}
\alias{DeleteStickerSetRequest}
\title{DeleteStickerSetRequest}
\description{
Telegram API type DeleteStickerSetRequest
}
\details{
DeleteStickerSetRequest R6 class

Represents a request to delete a sticker set.
}
\section{Methods}{

- initialize(stickerset): create instance.
- to_list(): returns a list representation suitable for JSON/inspection.
- to_bytes(): returns a raw vector representing serialized bytes (TL-like).
- from_reader(reader): reads a request from a reader object and returns a new instance.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{stickerset}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeleteStickerSetRequest-new}{\code{DeleteStickerSetRequest$new()}}
\item \href{#method-DeleteStickerSetRequest-to_list}{\code{DeleteStickerSetRequest$to_list()}}
\item \href{#method-DeleteStickerSetRequest-to_bytes}{\code{DeleteStickerSetRequest$to_bytes()}}
\item \href{#method-DeleteStickerSetRequest-from_reader}{\code{DeleteStickerSetRequest$from_reader()}}
\item \href{#method-DeleteStickerSetRequest-clone}{\code{DeleteStickerSetRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteStickerSetRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteStickerSetRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a DeleteStickerSetRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteStickerSetRequest$new(stickerset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stickerset}}{InputStickerSet (TL object or representation)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteStickerSetRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteStickerSetRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteStickerSetRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list representation of the request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteStickerSetRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteStickerSetRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to raw bytes (simple TL-like approximation)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteStickerSetRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteStickerSetRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteStickerSetRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteStickerSetRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteStickerSetRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteStickerSetRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteStickerSetRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
