% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotostate.R
\name{deque}
\alias{deque}
\title{Double-Ended Queue Implementation}
\usage{
deque(maxlen)
}
\arguments{
\item{maxlen}{Maximum length of the deque}
}
\description{
Simple implementation of a deque with max length
}
\section{Fields}{

\describe{
\item{\code{maxlen}}{Field.}
}}

