% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{DestroyAuthKeyRequest}
\alias{DestroyAuthKeyRequest}
\title{DestroyAuthKeyRequest}
\description{
R6 class representing a request to destroy the authentication key.
Contains methods for serialization to list and bytes.
Convert to list representation
Serialize to raw bytes
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DestroyAuthKeyRequest-to_list}{\code{DestroyAuthKeyRequest$to_list()}}
\item \href{#method-DestroyAuthKeyRequest-to_bytes}{\code{DestroyAuthKeyRequest$to_bytes()}}
\item \href{#method-DestroyAuthKeyRequest-clone}{\code{DestroyAuthKeyRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DestroyAuthKeyRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DestroyAuthKeyRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DestroyAuthKeyRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DestroyAuthKeyRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-DestroyAuthKeyRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DestroyAuthKeyRequest$to_bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DestroyAuthKeyRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DestroyAuthKeyRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DestroyAuthKeyRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
