% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{DestroyAuthKeyRequest_from_reader}
\alias{DestroyAuthKeyRequest_from_reader}
\title{Create DestroyAuthKeyRequest from a reader}
\usage{
DestroyAuthKeyRequest_from_reader(reader)
}
\arguments{
\item{reader}{Reader object (unused)}
}
\value{
Instance of DestroyAuthKeyRequest
}
\description{
Create DestroyAuthKeyRequest from a reader
}
