% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{DestroySessionRequest}
\alias{DestroySessionRequest}
\title{DestroySessionRequest}
\description{
R6 class representing a request to destroy a session.
Contains methods for serialization to list and bytes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor ID for the request}

\item{\code{SUBCLASS_OF_ID}}{Subclass ID for the request}

\item{\code{session_id}}{Session ID to be destroyed
Initialize the DestroySessionRequest}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DestroySessionRequest-new}{\code{DestroySessionRequest$new()}}
\item \href{#method-DestroySessionRequest-to_list}{\code{DestroySessionRequest$to_list()}}
\item \href{#method-DestroySessionRequest-to_bytes}{\code{DestroySessionRequest$to_bytes()}}
\item \href{#method-DestroySessionRequest-clone}{\code{DestroySessionRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DestroySessionRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DestroySessionRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DestroySessionRequest$new(session_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session_id}}{The ID of the session to destroy
Convert to list representation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DestroySessionRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DestroySessionRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DestroySessionRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the DestroySessionRequest
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DestroySessionRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-DestroySessionRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DestroySessionRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized DestroySessionRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DestroySessionRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DestroySessionRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DestroySessionRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
