% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{DestroySessionRequest_from_reader}
\alias{DestroySessionRequest_from_reader}
\title{Create DestroySessionRequest from a reader}
\usage{
DestroySessionRequest_from_reader(reader)
}
\arguments{
\item{reader}{Reader object}
}
\value{
Instance of DestroySessionRequest
}
\description{
Create DestroySessionRequest from a reader
}
