% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{dialog_message_key}
\alias{dialog_message_key}
\title{Helper function to get dialog message key}
\usage{
dialog_message_key(peer, message_id)
}
\arguments{
\item{peer}{The peer object}

\item{message_id}{The message ID}
}
\value{
A list containing the channel ID and message ID
}
\description{
Helper function to get dialog message key
}
