% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegram2DF.R
\name{dir2df}
\alias{dir2df}
\title{read files from a path and converts all its HTML into one tibble}
\usage{
dir2df(folder, recursive = TRUE)
}
\arguments{
\item{folder}{CHAR. The directory/folder/path to the HTML files}

\item{recursive}{BOOL. Will the . Default = TRUE}
}
\description{
read files from a path and converts all its HTML into one tibble
}
\details{
given a folder, reads all the HTML files and returns a single tibble
}
\examples{
dir2df( "~/Downloads/Telegram Desktop/ChatExport_2023-12-16 (1)" )
}
