% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{DirectDownloadIter}
\alias{DirectDownloadIter}
\title{DirectDownloadIter class}
\description{
Class for direct file downloading iteration.
}
\section{Super class}{
\code{\link[telegramR:RequestIter]{telegramR::RequestIter}} -> \code{DirectDownloadIter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{total}}{The total size of the file.}

\item{\code{stride}}{The stride size.}

\item{\code{chunk_size}}{The chunk size.}

\item{\code{last_part}}{The last part of the file.}

\item{\code{msg_data}}{Message data if applicable.}

\item{\code{timed_out}}{Whether the download timed out.}

\item{\code{exported}}{Whether the sender is exported.}

\item{\code{sender}}{The sender object.}

\item{\code{client}}{The TelegramClient instance.}

\item{\code{cdn_redirect}}{The CDN redirect object.}

\item{\code{type}}{Field.}

\item{\code{type}}{Field.}

\item{\code{type}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{Field.}

\item{\code{type}}{Field.}

\item{\code{type}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DirectDownloadIter-init}{\code{DirectDownloadIter$init()}}
\item \href{#method-DirectDownloadIter-load_next_chunk}{\code{DirectDownloadIter$load_next_chunk()}}
\item \href{#method-DirectDownloadIter-request}{\code{DirectDownloadIter$request()}}
\item \href{#method-DirectDownloadIter-close}{\code{DirectDownloadIter$close()}}
\item \href{#method-DirectDownloadIter-clone}{\code{DirectDownloadIter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id=".next"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-.next'><code>telegramR::RequestIter$.next()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="async_init"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-async_init'><code>telegramR::RequestIter$async_init()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="collect"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-collect'><code>telegramR::RequestIter$collect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="initialize"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-initialize'><code>telegramR::RequestIter$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="reset"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-reset'><code>telegramR::RequestIter$reset()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectDownloadIter-init"></a>}}
\if{latex}{\out{\hypertarget{method-DirectDownloadIter-init}{}}}
\subsection{Method \code{init()}}{
Initialize the download iterator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectDownloadIter$init(
  file,
  dc_id,
  offset,
  stride,
  chunk_size,
  request_size,
  file_size,
  msg_data,
  cdn_redirect = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{The file to download.}

\item{\code{dc_id}}{The data center ID.}

\item{\code{offset}}{The byte offset.}

\item{\code{stride}}{The stride size.}

\item{\code{chunk_size}}{The chunk size.}

\item{\code{request_size}}{The request size.}

\item{\code{file_size}}{The file size.}

\item{\code{msg_data}}{Message data if applicable.}

\item{\code{cdn_redirect}}{CDN redirect if applicable.}

\item{\code{client}}{The TelegramClient instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectDownloadIter-load_next_chunk"></a>}}
\if{latex}{\out{\hypertarget{method-DirectDownloadIter-load_next_chunk}{}}}
\subsection{Method \code{load_next_chunk()}}{
Load the next chunk of data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectDownloadIter$load_next_chunk()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectDownloadIter-request"></a>}}
\if{latex}{\out{\hypertarget{method-DirectDownloadIter-request}{}}}
\subsection{Method \code{request()}}{
Make a request to fetch data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectDownloadIter$request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectDownloadIter-close"></a>}}
\if{latex}{\out{\hypertarget{method-DirectDownloadIter-close}{}}}
\subsection{Method \code{close()}}{
Close the iterator and clean up resources.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectDownloadIter$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectDownloadIter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DirectDownloadIter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectDownloadIter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
