% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{DiscardGroupCallRequest}
\alias{DiscardGroupCallRequest}
\title{DiscardGroupCallRequest R6 class}
\description{
Represents DiscardGroupCallRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{The input group call object (expected to implement bytes()).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiscardGroupCallRequest-new}{\code{DiscardGroupCallRequest$new()}}
\item \href{#method-DiscardGroupCallRequest-resolve}{\code{DiscardGroupCallRequest$resolve()}}
\item \href{#method-DiscardGroupCallRequest-to_list}{\code{DiscardGroupCallRequest$to_list()}}
\item \href{#method-DiscardGroupCallRequest-bytes}{\code{DiscardGroupCallRequest$bytes()}}
\item \href{#method-DiscardGroupCallRequest-from_reader}{\code{DiscardGroupCallRequest$from_reader()}}
\item \href{#method-DiscardGroupCallRequest-clone}{\code{DiscardGroupCallRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscardGroupCallRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiscardGroupCallRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardGroupCallRequest$new(call)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Resolve input references

Typically used to convert high-level entities into input objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscardGroupCallRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-DiscardGroupCallRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardGroupCallRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Convert object to a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscardGroupCallRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DiscardGroupCallRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardGroupCallRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscardGroupCallRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-DiscardGroupCallRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardGroupCallRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscardGroupCallRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-DiscardGroupCallRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardGroupCallRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscardGroupCallRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiscardGroupCallRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardGroupCallRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
