% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chats.R
\name{.ProfilePhotoIter}
\alias{.ProfilePhotoIter}
\title{_ProfilePhotoIter R6 Class}
\description{
_ProfilePhotoIter R6 Class

_ProfilePhotoIter R6 Class
}
\details{
An iterator over a user's profile photos or a chat's photos.
The order is from the most recent photo to the oldest.
Inherits from RequestIter.
}
\section{Super class}{
\code{\link[telegramR:RequestIter]{telegramR::RequestIter}} -> \code{_ProfilePhotoIter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{limit}}{Field.}

\item{\code{min_date}}{Field.}

\item{\code{max_date}}{Field.}

\item{\code{offset_id}}{Field.}

\item{\code{limit}}{Field.}

\item{\code{min_id}}{Field.}

\item{\code{hash}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{limit}}{Field.}

\item{\code{min_date}}{Field.}

\item{\code{max_date}}{Field.}

\item{\code{offset_id}}{Field.}

\item{\code{limit}}{Field.}

\item{\code{min_id}}{Field.}

\item{\code{hash}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-_ProfilePhotoIter-.init}{\code{.ProfilePhotoIter$.init()}}
\item \href{#method-_ProfilePhotoIter-.load_next_chunk}{\code{.ProfilePhotoIter$.load_next_chunk()}}
\item \href{#method-_ProfilePhotoIter-clone}{\code{.ProfilePhotoIter$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id=".next"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-.next'><code>telegramR::RequestIter$.next()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="async_init"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-async_init'><code>telegramR::RequestIter$async_init()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="collect"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-collect'><code>telegramR::RequestIter$collect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="initialize"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-initialize'><code>telegramR::RequestIter$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="load_next_chunk"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-load_next_chunk'><code>telegramR::RequestIter$load_next_chunk()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="reset"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-reset'><code>telegramR::RequestIter$reset()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-_ProfilePhotoIter-.init"></a>}}
\if{latex}{\out{\hypertarget{method-_ProfilePhotoIter-.init}{}}}
\subsection{Method \code{.init()}}{
Initialize the iterator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.ProfilePhotoIter$.init(entity, offset, max_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{The entity from which to get the profile or chat photos.}

\item{\code{offset}}{How many photos should be skipped before returning the first one.}

\item{\code{max_id}}{The maximum ID allowed when fetching photos.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-_ProfilePhotoIter-.load_next_chunk"></a>}}
\if{latex}{\out{\hypertarget{method-_ProfilePhotoIter-.load_next_chunk}{}}}
\subsection{Method \code{.load_next_chunk()}}{
Load the next chunk of photos.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.ProfilePhotoIter$.load_next_chunk()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-_ProfilePhotoIter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-_ProfilePhotoIter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.ProfilePhotoIter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
