% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_downloads.R
\name{download_channel_members}
\alias{download_channel_members}
\title{Download Channel Members By Channel}
\usage{
download_channel_members(
  client,
  channel,
  limit = Inf,
  search = "",
  filter = NULL,
  include_channel = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{channel}{character or numeric. Channel username (with or without "@") or numeric id.}

\item{limit}{integer or Inf. Maximum number of members to fetch.}

\item{search}{character. Search query for participant names/usernames.}

\item{filter}{A participants filter, e.g. ChannelParticipantsAdmins.}

\item{include_channel}{logical. If TRUE, include channel fields on every row.}

\item{show_progress}{logical. If TRUE, display a progress bar.}
}
\value{
A tibble.
}
\description{
Fetches channel members (participants) by username or numeric id.
}
