% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{DropTempAuthKeysRequest}
\alias{DropTempAuthKeysRequest}
\title{DropTempAuthKeysRequest R6 class}
\description{
Represents the TLRequest auth.DropTempAuthKeysRequest.
}
\details{
Fields:
}
\section{initialize}{

 Create a new DropTempAuthKeysRequest object.
 @param except_auth_keys numeric vector (64-bit) or NULL
 @return invisible self
}

\section{to_list}{

 Return a pure R list representation of the object.
 @return list
}

\section{to_bytes}{

 Serialize the object into raw bytes for TL.
 @return raw vector
}

\section{from_reader}{

 Read fields from a reader and return a new DropTempAuthKeysRequest.
 @param reader Reader object implementing read_int(), read_long()
 @return DropTempAuthKeysRequest instance
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0x8e48a188).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0xf5b399ac).}

\item{\code{except_auth_keys}}{numeric vector 64-bit ids to drop (except list).

Methods:}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DropTempAuthKeysRequest-new}{\code{DropTempAuthKeysRequest$new()}}
\item \href{#method-DropTempAuthKeysRequest-to_list}{\code{DropTempAuthKeysRequest$to_list()}}
\item \href{#method-DropTempAuthKeysRequest-to_bytes}{\code{DropTempAuthKeysRequest$to_bytes()}}
\item \href{#method-DropTempAuthKeysRequest-from_reader}{\code{DropTempAuthKeysRequest$from_reader()}}
\item \href{#method-DropTempAuthKeysRequest-clone}{\code{DropTempAuthKeysRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DropTempAuthKeysRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DropTempAuthKeysRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DropTempAuthKeysRequest$new(except_auth_keys = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{except_auth_keys}}{numeric vector (64-bit) or NULL
Convert to an R list (similar to to_dict)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DropTempAuthKeysRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DropTempAuthKeysRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DropTempAuthKeysRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DropTempAuthKeysRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-DropTempAuthKeysRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DropTempAuthKeysRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
Read from reader and create DropTempAuthKeysRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DropTempAuthKeysRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-DropTempAuthKeysRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DropTempAuthKeysRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), read_long()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DropTempAuthKeysRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DropTempAuthKeysRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DropTempAuthKeysRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DropTempAuthKeysRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
