% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{EditGroupCallTitleRequest}
\alias{EditGroupCallTitleRequest}
\title{EditGroupCallTitleRequest}
\description{
Telegram API type EditGroupCallTitleRequest
}
\details{
EditGroupCallTitleRequest R6 class

Represents EditGroupCallTitleRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{Field.}

\item{\code{title}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EditGroupCallTitleRequest-new}{\code{EditGroupCallTitleRequest$new()}}
\item \href{#method-EditGroupCallTitleRequest-resolve}{\code{EditGroupCallTitleRequest$resolve()}}
\item \href{#method-EditGroupCallTitleRequest-to_list}{\code{EditGroupCallTitleRequest$to_list()}}
\item \href{#method-EditGroupCallTitleRequest-bytes}{\code{EditGroupCallTitleRequest$bytes()}}
\item \href{#method-EditGroupCallTitleRequest-from_reader}{\code{EditGroupCallTitleRequest$from_reader()}}
\item \href{#method-EditGroupCallTitleRequest-clone}{\code{EditGroupCallTitleRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditGroupCallTitleRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-EditGroupCallTitleRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an EditGroupCallTitleRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditGroupCallTitleRequest$new(call, title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}

\item{\code{title}}{String title.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditGroupCallTitleRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-EditGroupCallTitleRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references

Typically used to convert high-level entities into input objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditGroupCallTitleRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditGroupCallTitleRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EditGroupCallTitleRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditGroupCallTitleRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditGroupCallTitleRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-EditGroupCallTitleRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditGroupCallTitleRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditGroupCallTitleRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-EditGroupCallTitleRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditGroupCallTitleRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditGroupCallTitleRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EditGroupCallTitleRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditGroupCallTitleRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
