% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_folders.R
\name{EditPeerFoldersRequest}
\alias{EditPeerFoldersRequest}
\title{EditPeerFoldersRequest R6 class}
\value{
An R6 object of class EditPeerFoldersRequest
}
\description{
Represents the TL request EditPeerFoldersRequest.
}
\examples{
# EditPeerFoldersRequest$new(list_of_folder_peers)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (0x6847d0ab)}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (0x8af52aac)}

\item{\code{folder_peers}}{list List of InputFolderPeer objects (each should provide to_list() and to_raw())}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EditPeerFoldersRequest-new}{\code{EditPeerFoldersRequest$new()}}
\item \href{#method-EditPeerFoldersRequest-to_list}{\code{EditPeerFoldersRequest$to_list()}}
\item \href{#method-EditPeerFoldersRequest-to_raw}{\code{EditPeerFoldersRequest$to_raw()}}
\item \href{#method-EditPeerFoldersRequest-clone}{\code{EditPeerFoldersRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditPeerFoldersRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-EditPeerFoldersRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditPeerFoldersRequest$new(folder_peers = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{folder_peers}}{list List of folder peer TL objects (default: empty list)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditPeerFoldersRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EditPeerFoldersRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditPeerFoldersRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditPeerFoldersRequest-to_raw"></a>}}
\if{latex}{\out{\hypertarget{method-EditPeerFoldersRequest-to_raw}{}}}
\subsection{Method \code{to_raw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditPeerFoldersRequest$to_raw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditPeerFoldersRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EditPeerFoldersRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditPeerFoldersRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
