% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{EditPreviewMediaRequest}
\alias{EditPreviewMediaRequest}
\title{EditPreviewMediaRequest}
\description{
Telegram API type EditPreviewMediaRequest
}
\details{
EditPreviewMediaRequest R6 class

Represents the EditPreviewMediaRequest TL request.



Each method is documented inline below.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bot}}{Field.}

\item{\code{lang_code}}{Field.}

\item{\code{media}}{Field.}

\item{\code{new_media}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EditPreviewMediaRequest-new}{\code{EditPreviewMediaRequest$new()}}
\item \href{#method-EditPreviewMediaRequest-resolve}{\code{EditPreviewMediaRequest$resolve()}}
\item \href{#method-EditPreviewMediaRequest-to_list}{\code{EditPreviewMediaRequest$to_list()}}
\item \href{#method-EditPreviewMediaRequest-to_bytes}{\code{EditPreviewMediaRequest$to_bytes()}}
\item \href{#method-EditPreviewMediaRequest-clone}{\code{EditPreviewMediaRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditPreviewMediaRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-EditPreviewMediaRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize EditPreviewMediaRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditPreviewMediaRequest$new(bot, lang_code, media, new_media)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bot}}{TypeInputUser or identifier}

\item{\code{lang_code}}{character}

\item{\code{media}}{TypeInputMedia or object}

\item{\code{new_media}}{TypeInputMedia or object
Resolve references (convert entities via client/utils)

Resolves bot via client$get_input_entity and utils$get_input_user;
resolves media and new_media via utils$get_input_media.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditPreviewMediaRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-EditPreviewMediaRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditPreviewMediaRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client with get_input_entity method (synchronous)}

\item{\code{utils}}{utils with get_input_user and get_input_media}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
Convert to list (dictionary-like)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditPreviewMediaRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EditPreviewMediaRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditPreviewMediaRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditPreviewMediaRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-EditPreviewMediaRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditPreviewMediaRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditPreviewMediaRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EditPreviewMediaRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditPreviewMediaRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
