% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{EditUserInfoRequest}
\alias{EditUserInfoRequest}
\title{EditUserInfoRequest R6 class}
\description{
Edit user info (help.UserInfo). Serializes an input user, a message and a list
of message entities.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{userId}}{TLObject-like input user (environment with to_bytes()/to_list() or raw)}

\item{\code{message}}{character Message text}

\item{\code{entities}}{list List of TLObject-like message entities}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EditUserInfoRequest-new}{\code{EditUserInfoRequest$new()}}
\item \href{#method-EditUserInfoRequest-resolve}{\code{EditUserInfoRequest$resolve()}}
\item \href{#method-EditUserInfoRequest-to_list}{\code{EditUserInfoRequest$to_list()}}
\item \href{#method-EditUserInfoRequest-to_bytes}{\code{EditUserInfoRequest$to_bytes()}}
\item \href{#method-EditUserInfoRequest-clone}{\code{EditUserInfoRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditUserInfoRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-EditUserInfoRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditUserInfoRequest$new(userId = NULL, message = "", entities = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userId}}{TLObject-like or raw}

\item{\code{message}}{character}

\item{\code{entities}}{list of TLObject-like entities
Resolve userId using client and utils}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditUserInfoRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-EditUserInfoRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditUserInfoRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client object providing get_input_entity()}

\item{\code{utils}}{utils providing get_input_user()
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditUserInfoRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EditUserInfoRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditUserInfoRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `_` discriminator, user_id, message and entities as lists (if possible)
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditUserInfoRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-EditUserInfoRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditUserInfoRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EditUserInfoRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EditUserInfoRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EditUserInfoRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
