% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode_telegram_base64}
\alias{encode_telegram_base64}
\title{Encode Telegram Base64}
\usage{
encode_telegram_base64(bytes)
}
\arguments{
\item{bytes}{A raw vector of bytes to encode.}
}
\value{
A character string representing the url-safe base64-encoded data if encoding succeeds, otherwise NULL.
}
\description{
Inverse operation for `decode_telegram_base64`.
Encodes bytes into a url-safe base64 string, stripping padding.
}
