% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa.R
\name{encrypt}
\alias{encrypt}
\title{Encrypts data using the specified RSA key fingerprint.}
\usage{
encrypt(fingerprint, data, use_old = FALSE)
}
\arguments{
\item{fingerprint}{The fingerprint of the RSA key.}

\item{data}{The data to encrypt as a raw vector (or coercible).}

\item{use_old}{Logical indicating if old keys should be used.}
}
\value{
The encrypted data as a raw vector, or NULL if no matching key is found.
}
\description{
Encrypts data using the specified RSA key fingerprint.
}
