% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{EncryptedFileEmpty}
\alias{EncryptedFileEmpty}
\title{EncryptedFileEmpty}
\description{
Telegram API type EncryptedFileEmpty
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EncryptedFileEmpty-to_list}{\code{EncryptedFileEmpty$to_list()}}
\item \href{#method-EncryptedFileEmpty-clone}{\code{EncryptedFileEmpty$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncryptedFileEmpty-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EncryptedFileEmpty-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptedFileEmpty$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncryptedFileEmpty-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EncryptedFileEmpty-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptedFileEmpty$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
