% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{EncryptedMessage}
\alias{EncryptedMessage}
\title{EncryptedMessage}
\description{
Telegram API type EncryptedMessage
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{chat_id}}{Field.}

\item{\code{date}}{Field.}

\item{\code{bytes}}{Field.}

\item{\code{file}}{Field.}

\item{\code{random_id}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EncryptedMessage-new}{\code{EncryptedMessage$new()}}
\item \href{#method-EncryptedMessage-to_list}{\code{EncryptedMessage$to_list()}}
\item \href{#method-EncryptedMessage-clone}{\code{EncryptedMessage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncryptedMessage-new"></a>}}
\if{latex}{\out{\hypertarget{method-EncryptedMessage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptedMessage$new(chat_id, date = NULL, bytes, file, random_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncryptedMessage-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EncryptedMessage-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptedMessage$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncryptedMessage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EncryptedMessage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptedMessage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
