% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{EncryptedMessageService}
\alias{EncryptedMessageService}
\title{EncryptedMessageService}
\description{
Telegram API type EncryptedMessageService
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{chat_id}}{Field.}

\item{\code{date}}{Field.}

\item{\code{bytes}}{Field.}

\item{\code{random_id}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EncryptedMessageService-new}{\code{EncryptedMessageService$new()}}
\item \href{#method-EncryptedMessageService-to_list}{\code{EncryptedMessageService$to_list()}}
\item \href{#method-EncryptedMessageService-clone}{\code{EncryptedMessageService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncryptedMessageService-new"></a>}}
\if{latex}{\out{\hypertarget{method-EncryptedMessageService-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptedMessageService$new(chat_id, date = NULL, bytes, random_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncryptedMessageService-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EncryptedMessageService-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptedMessageService$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncryptedMessageService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EncryptedMessageService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptedMessageService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
