% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ensure_parent_dir_exists}
\alias{ensure_parent_dir_exists}
\title{Ensure the parent directory exists}
\usage{
ensure_parent_dir_exists(file_path)
}
\arguments{
\item{file_path}{The file path for which to ensure the parent directory exists.}
}
\value{
None. Creates the parent directory if it does not exist.
}
\description{
Ensure the parent directory exists
}
\examples{
ensure_parent_dir_exists("path/to/some/file.txt")
}
