% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{entity_type}
\alias{entity_type}
\title{Entity Type}
\usage{
entity_type(entity)
}
\arguments{
\item{entity}{Telegram entity or peer/input peer object.}
}
\value{
One of EntityType$USER, EntityType$CHAT, EntityType$CHANNEL.
}
\description{
Returns EntityType$USER, $CHAT, or $CHANNEL for known entity/peer/input types.
}
\keyword{internal}
