% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{EntityType}
\alias{EntityType}
\title{Various helpers not related to the Telegram API itself
Enum for entity types}
\format{
An object of class \code{list} of length 3.
}
\usage{
EntityType
}
\description{
Various helpers not related to the Telegram API itself
Enum for entity types
}
\keyword{datasets}
