% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{escape_html}
\alias{escape_html}
\title{Escape HTML special characters.}
\usage{
escape_html(x)
}
\arguments{
\item{x}{A character vector containing text to escape.}
}
\value{
A character vector with HTML-sensitive characters converted to their entity equivalents.
}
\description{
Replaces ampersands, angle brackets, and double quotes with their corresponding HTML entities.
}
