% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{EventBuilderDict}
\alias{EventBuilderDict}
\title{EventBuilderDict}
\description{
Handles the building of events from updates and other data.

Telegram API type EventBuilderDict
}
\details{
EventBuilderDict class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client}}{The Telegram client instance.}

\item{\code{update}}{The update data.}

\item{\code{others}}{Additional data for building events.}

\item{\code{cache}}{A cache for built events.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventBuilderDict-new}{\code{EventBuilderDict$new()}}
\item \href{#method-EventBuilderDict-get}{\code{EventBuilderDict$get()}}
\item \href{#method-EventBuilderDict-clone}{\code{EventBuilderDict$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventBuilderDict-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventBuilderDict-new}{}}}
\subsection{Method \code{new()}}{
Initializes the EventBuilderDict with the given client, update, and others.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventBuilderDict$new(client, update, others)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The Telegram client instance.}

\item{\code{update}}{The update data.}

\item{\code{others}}{Additional data for building events.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventBuilderDict-get"></a>}}
\if{latex}{\out{\hypertarget{method-EventBuilderDict-get}{}}}
\subsection{Method \code{get()}}{
Builds an event from the given builder.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventBuilderDict$get(builder)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{builder}}{The builder to use for building the event.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The built event.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventBuilderDict-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventBuilderDict-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventBuilderDict$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
