% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{Exception}
\alias{Exception}
\title{Exception Classes}
\description{
This file contains various exception classes used in the Telegram API client.
These exceptions are used to handle errors and specific conditions that may arise
during the operation of the client.

Telegram API type Exception
}
\details{
The classes defined in this file include:
- `Exception`: A base class for all exceptions.
- `ReadCancelledError`: Raised when a read operation is cancelled.
- `TypeNotFoundError`: Raised when a type is not found.
- `InvalidChecksumError`: Raised when a checksum is invalid.
- `InvalidBufferError`: Raised when a buffer is invalid.
- `AuthKeyNotFound`: Raised when an authorization key is not found.
- `SecurityError`: Raised when a security check fails.
- `CdnFileTamperedError`: Raised when a CDN file is tampered with.
- `AlreadyInConversationError`: Raised when another exclusive conversation is opened in the same chat.
- `BadMessageError`: Raised when handling a bad message notification.
- `MultiError`: A container for multiple exceptions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{message}}{\code{character} The error message.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Exception-new}{\code{Exception$new()}}
\item \href{#method-Exception-clone}{\code{Exception$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exception-new"></a>}}
\if{latex}{\out{\hypertarget{method-Exception-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the Exception class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exception$new(message = "An error occurred.")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{\code{character} The error message.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exception-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Exception-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exception$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
