% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{ExportAuthorizationRequest}
\alias{ExportAuthorizationRequest}
\title{ExportAuthorizationRequest R6 class}
\description{
Represents the TLRequest auth.ExportAuthorizationRequest.
}
\details{
Fields:
}
\section{initialize}{

 Create a new ExportAuthorizationRequest object.
 @param dc_id integer
 @return invisible self
}

\section{to_list}{

 Return a pure R list representation of the object.
 @return list
}

\section{to_bytes}{

 Serialize the object into raw bytes for TL.
 @return raw vector
}

\section{from_reader}{

 Read fields from a reader and return a new ExportAuthorizationRequest.
 @param reader Reader object implementing read_int()
 @return ExportAuthorizationRequest instance
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0xe5bfffcd).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0x5fd1ec51).}

\item{\code{dc_id}}{integer Destination DC id.

Methods:}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExportAuthorizationRequest-new}{\code{ExportAuthorizationRequest$new()}}
\item \href{#method-ExportAuthorizationRequest-to_list}{\code{ExportAuthorizationRequest$to_list()}}
\item \href{#method-ExportAuthorizationRequest-to_bytes}{\code{ExportAuthorizationRequest$to_bytes()}}
\item \href{#method-ExportAuthorizationRequest-from_reader}{\code{ExportAuthorizationRequest$from_reader()}}
\item \href{#method-ExportAuthorizationRequest-clone}{\code{ExportAuthorizationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportAuthorizationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExportAuthorizationRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportAuthorizationRequest$new(dc_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dc_id}}{integer
Convert to an R list (similar to to_dict)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportAuthorizationRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ExportAuthorizationRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportAuthorizationRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportAuthorizationRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ExportAuthorizationRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportAuthorizationRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
Read from reader and create ExportAuthorizationRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportAuthorizationRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ExportAuthorizationRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportAuthorizationRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ExportAuthorizationRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportAuthorizationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExportAuthorizationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportAuthorizationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
