% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chatlists.R
\name{ExportChatlistInviteRequest}
\alias{ExportChatlistInviteRequest}
\title{ExportChatlistInviteRequest}
\description{
Telegram API type ExportChatlistInviteRequest
}
\details{
ExportChatlistInviteRequest R6 class

Request to export an invite for a chatlist with a title and list of peers.

Serializes constructor id, chatlist bytes, title (TL-string), then peers vector.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{chatlist}}{Field.}

\item{\code{title}}{Field.}

\item{\code{peers}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExportChatlistInviteRequest-new}{\code{ExportChatlistInviteRequest$new()}}
\item \href{#method-ExportChatlistInviteRequest-resolve}{\code{ExportChatlistInviteRequest$resolve()}}
\item \href{#method-ExportChatlistInviteRequest-to_list}{\code{ExportChatlistInviteRequest$to_list()}}
\item \href{#method-ExportChatlistInviteRequest-bytes}{\code{ExportChatlistInviteRequest$bytes()}}
\item \href{#method-ExportChatlistInviteRequest-from_reader}{\code{ExportChatlistInviteRequest$from_reader()}}
\item \href{#method-ExportChatlistInviteRequest-clone}{\code{ExportChatlistInviteRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportChatlistInviteRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExportChatlistInviteRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize ExportChatlistInviteRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportChatlistInviteRequest$new(chatlist = NULL, title = NULL, peers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chatlist}}{Input chatlist object}

\item{\code{title}}{Invite title (string)}

\item{\code{peers}}{List of input peers
Resolve peer entities to input peers

Iterates over peers, resolves each via client and utils and replaces
the peers field with input_peer objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportChatlistInviteRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-ExportChatlistInviteRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportChatlistInviteRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object that provides get_input_entity}

\item{\code{utils}}{Utilities that provide get_input_peer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self
Convert to list (dictionary-like)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportChatlistInviteRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ExportChatlistInviteRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportChatlistInviteRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representation
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportChatlistInviteRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ExportChatlistInviteRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportChatlistInviteRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector with serialized bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportChatlistInviteRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ExportChatlistInviteRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportChatlistInviteRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportChatlistInviteRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExportChatlistInviteRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportChatlistInviteRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
