% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ExportedChatlistInvite}
\alias{ExportedChatlistInvite}
\title{ExportedChatlistInvite}
\description{
Telegram API type ExportedChatlistInvite
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{Field.}

\item{\code{url}}{Field.}

\item{\code{peers}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExportedChatlistInvite-new}{\code{ExportedChatlistInvite$new()}}
\item \href{#method-ExportedChatlistInvite-to_list}{\code{ExportedChatlistInvite$to_list()}}
\item \href{#method-ExportedChatlistInvite-clone}{\code{ExportedChatlistInvite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedChatlistInvite-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedChatlistInvite-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedChatlistInvite$new(title, url, peers)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedChatlistInvite-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedChatlistInvite-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedChatlistInvite$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedChatlistInvite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedChatlistInvite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedChatlistInvite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
