% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ExportedContactToken}
\alias{ExportedContactToken}
\title{ExportedContactToken}
\description{
Telegram API type ExportedContactToken
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExportedContactToken-new}{\code{ExportedContactToken$new()}}
\item \href{#method-ExportedContactToken-to_list}{\code{ExportedContactToken$to_list()}}
\item \href{#method-ExportedContactToken-clone}{\code{ExportedContactToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedContactToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedContactToken-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedContactToken$new(url, expires = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedContactToken-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedContactToken-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedContactToken$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedContactToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedContactToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedContactToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
