% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ExportedMessageLink}
\alias{ExportedMessageLink}
\title{ExportedMessageLink}
\description{
Telegram API type ExportedMessageLink
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{link}}{Field.}

\item{\code{html}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExportedMessageLink-new}{\code{ExportedMessageLink$new()}}
\item \href{#method-ExportedMessageLink-to_list}{\code{ExportedMessageLink$to_list()}}
\item \href{#method-ExportedMessageLink-clone}{\code{ExportedMessageLink$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedMessageLink-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedMessageLink-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedMessageLink$new(link, html)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedMessageLink-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedMessageLink-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedMessageLink$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedMessageLink-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedMessageLink-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedMessageLink$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
