% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ExportedStoryLink}
\alias{ExportedStoryLink}
\title{ExportedStoryLink}
\description{
Telegram API type ExportedStoryLink
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{link}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExportedStoryLink-new}{\code{ExportedStoryLink$new()}}
\item \href{#method-ExportedStoryLink-to_list}{\code{ExportedStoryLink$to_list()}}
\item \href{#method-ExportedStoryLink-clone}{\code{ExportedStoryLink$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedStoryLink-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedStoryLink-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedStoryLink$new(link)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedStoryLink-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedStoryLink-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedStoryLink$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportedStoryLink-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExportedStoryLink-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportedStoryLink$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
