% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{ExportGroupCallInviteRequest}
\alias{ExportGroupCallInviteRequest}
\title{ExportGroupCallInviteRequest}
\description{
Telegram API type ExportGroupCallInviteRequest
}
\details{
ExportGroupCallInviteRequest R6 class

Represents ExportGroupCallInviteRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{Field.}

\item{\code{can_self_unmute}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExportGroupCallInviteRequest-new}{\code{ExportGroupCallInviteRequest$new()}}
\item \href{#method-ExportGroupCallInviteRequest-resolve}{\code{ExportGroupCallInviteRequest$resolve()}}
\item \href{#method-ExportGroupCallInviteRequest-to_list}{\code{ExportGroupCallInviteRequest$to_list()}}
\item \href{#method-ExportGroupCallInviteRequest-bytes}{\code{ExportGroupCallInviteRequest$bytes()}}
\item \href{#method-ExportGroupCallInviteRequest-from_reader}{\code{ExportGroupCallInviteRequest$from_reader()}}
\item \href{#method-ExportGroupCallInviteRequest-clone}{\code{ExportGroupCallInviteRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportGroupCallInviteRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExportGroupCallInviteRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ExportGroupCallInviteRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportGroupCallInviteRequest$new(call, can_self_unmute = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}

\item{\code{can_self_unmute}}{Logical or NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportGroupCallInviteRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-ExportGroupCallInviteRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references

Typically used to convert high-level entities into input objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportGroupCallInviteRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportGroupCallInviteRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ExportGroupCallInviteRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportGroupCallInviteRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportGroupCallInviteRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ExportGroupCallInviteRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportGroupCallInviteRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportGroupCallInviteRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ExportGroupCallInviteRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportGroupCallInviteRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportGroupCallInviteRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExportGroupCallInviteRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportGroupCallInviteRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
