% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegrambaseclient.R
\name{ExportState}
\alias{ExportState}
\title{Export State Class}
\description{
This class manages the state of exported MTProto senders, including
tracking the number of borrowed senders and the time since the last
borrowed sender was returned.
}
\details{
Tracks the state of exported MTProto senders
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExportState-new}{\code{ExportState$new()}}
\item \href{#method-ExportState-add_borrow}{\code{ExportState$add_borrow()}}
\item \href{#method-ExportState-add_return}{\code{ExportState$add_return()}}
\item \href{#method-ExportState-should_disconnect}{\code{ExportState$should_disconnect()}}
\item \href{#method-ExportState-need_connect}{\code{ExportState$need_connect()}}
\item \href{#method-ExportState-mark_disconnected}{\code{ExportState$mark_disconnected()}}
\item \href{#method-ExportState-clone}{\code{ExportState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportState-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExportState-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the ExportState class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportState$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportState-add_borrow"></a>}}
\if{latex}{\out{\hypertarget{method-ExportState-add_borrow}{}}}
\subsection{Method \code{add_borrow()}}{
Add a borrowed sender
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportState$add_borrow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportState-add_return"></a>}}
\if{latex}{\out{\hypertarget{method-ExportState-add_return}{}}}
\subsection{Method \code{add_return()}}{
Return a borrowed sender
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportState$add_return()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportState-should_disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-ExportState-should_disconnect}{}}}
\subsection{Method \code{should_disconnect()}}{
Check if the sender should be disconnected
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportState$should_disconnect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if should disconnect, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportState-need_connect"></a>}}
\if{latex}{\out{\hypertarget{method-ExportState-need_connect}{}}}
\subsection{Method \code{need_connect()}}{
Check if the sender needs to connect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportState$need_connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if needs to connect, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportState-mark_disconnected"></a>}}
\if{latex}{\out{\hypertarget{method-ExportState-mark_disconnected}{}}}
\subsection{Method \code{mark_disconnected()}}{
Mark the sender as disconnected
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportState$mark_disconnected()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExportState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
