% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{F}
\alias{F}
\title{Just a wrapper around glue::glue(), For Internal usage.}
\usage{
F(
  ...,
  .sep = "",
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  .na = "NA",
  .null = character(),
  .comment = "#",
  .literal = FALSE,
  .transformer = identity_transformer,
  .trim = TRUE
)
}
\description{
Just a wrapper around glue::glue(), For Internal usage.
}
