% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{FactCheck}
\alias{FactCheck}
\title{FactCheck}
\description{
Telegram API type FactCheck
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{hash}}{Field.}

\item{\code{need_check}}{Field.}

\item{\code{country}}{Field.}

\item{\code{text}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FactCheck-new}{\code{FactCheck$new()}}
\item \href{#method-FactCheck-to_list}{\code{FactCheck$to_list()}}
\item \href{#method-FactCheck-clone}{\code{FactCheck$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactCheck-new"></a>}}
\if{latex}{\out{\hypertarget{method-FactCheck-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactCheck$new(hash, need_check = NULL, country = NULL, text = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactCheck-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-FactCheck-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactCheck$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactCheck-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FactCheck-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactCheck$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
