% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{FileHash}
\alias{FileHash}
\title{FileHash}
\description{
Telegram API type FileHash
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{offset}}{Field.}

\item{\code{limit}}{Field.}

\item{\code{hash}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileHash-new}{\code{FileHash$new()}}
\item \href{#method-FileHash-to_list}{\code{FileHash$to_list()}}
\item \href{#method-FileHash-clone}{\code{FileHash$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileHash-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileHash-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileHash$new(offset, limit, hash)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileHash-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-FileHash-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileHash$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileHash-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileHash-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileHash$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
