% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FileInfo}
\alias{FileInfo}
\title{FileInfo Class}
\description{
FileInfo Class

FileInfo Class
}
\details{
An R6 class representing file information with dc_id, location, and size.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dc_id}}{The data center ID.}

\item{\code{location}}{The file location.}

\item{\code{size}}{The file size.}

\item{\code{dc_id}}{The data center ID.}

\item{\code{location}}{The file location.}

\item{\code{size}}{The file size.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileInfo-new}{\code{FileInfo$new()}}
\item \href{#method-FileInfo-clone}{\code{FileInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize the FileInfo object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileInfo$new(dc_id, location, size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dc_id}}{The data center ID.}

\item{\code{location}}{The file location.}

\item{\code{size}}{The file size.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
