% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{FileStream}
\alias{FileStream}
\title{FileStream}
\description{
Telegram API type FileStream
}
\details{
FileStream Class
A class to handle file streams from various sources.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file}}{The original file input (path, raw vector, or connection).}

\item{\code{file_size}}{The size of the file in bytes.}

\item{\code{name}}{The name of the file (if applicable).}

\item{\code{stream}}{The underlying connection or raw connection.}

\item{\code{close_stream}}{Logical indicating if the stream should be closed on cleanup.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileStream-new}{\code{FileStream$new()}}
\item \href{#method-FileStream-read}{\code{FileStream$read()}}
\item \href{#method-FileStream-close}{\code{FileStream$close()}}
\item \href{#method-FileStream-get_file_size}{\code{FileStream$get_file_size()}}
\item \href{#method-FileStream-get_name}{\code{FileStream$get_name()}}
\item \href{#method-FileStream-clone}{\code{FileStream$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-new}{}}}
\subsection{Method \code{new()}}{
Initializes the FileStream.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$new(file, file_size = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{A file path (character), raw vector, or connection.}

\item{\code{file_size}}{Optional size of the file in bytes (if known).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new FileStream instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-read"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-read}{}}}
\subsection{Method \code{read()}}{
Reads data from the stream.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$read(n = -1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of bytes to read; defaults to -1 (read all remaining).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A raw vector containing the read data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-close"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-close}{}}}
\subsection{Method \code{close()}}{
Closes the stream if it was opened by the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$close()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-get_file_size"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-get_file_size}{}}}
\subsection{Method \code{get_file_size()}}{
Gets the size of the file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$get_file_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The size of the file in bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-get_name}{}}}
\subsection{Method \code{get_name()}}{
Gets the name of the file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$get_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The name of the file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
