% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_info.R
\name{filter_last_export}
\alias{filter_last_export}
\title{filter dataframe to show only the last exports of each chat/group}
\usage{
filter_last_export(DF)
}
\arguments{
\item{DF}{the tibble/dataframe generated with `tm_info()` function}
}
\description{
filter dataframe to show only the last exports of each chat/group
}
\examples{
tm_info("~/Downloads/Telegram Desktop/") |> filter_last_export()
}
