% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_flood}
\alias{fmt_flood}
\title{fmt_flood}
\usage{
fmt_flood(delay, request, early = FALSE, td = as.difftime)
}
\arguments{
\item{delay}{Delay time in seconds}

\item{request}{The request object}

\item{early}{Boolean indicating if the message is for an early flood wait}

\item{td}{Function to convert delay to time difference}
}
\value{
Formatted flood wait message
}
\description{
ftm_flood - Format flood wait message
}
