% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Folder}
\alias{Folder}
\title{Folder}
\description{
Telegram API type Folder
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Folder-new}{\code{Folder$new()}}
\item \href{#method-Folder-to_list}{\code{Folder$to_list()}}
\item \href{#method-Folder-clone}{\code{Folder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Folder-new"></a>}}
\if{latex}{\out{\hypertarget{method-Folder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Folder$new(
  id,
  title,
  autofill_new_broadcasts = NULL,
  autofill_public_groups = NULL,
  autofill_new_correspondents = NULL,
  photo = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Folder-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Folder-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Folder$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Folder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Folder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Folder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
