% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{FolderPeer}
\alias{FolderPeer}
\title{FolderPeer}
\description{
Telegram API type FolderPeer
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FolderPeer-new}{\code{FolderPeer$new()}}
\item \href{#method-FolderPeer-to_list}{\code{FolderPeer$to_list()}}
\item \href{#method-FolderPeer-clone}{\code{FolderPeer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FolderPeer-new"></a>}}
\if{latex}{\out{\hypertarget{method-FolderPeer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderPeer$new(peer, folder_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FolderPeer-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-FolderPeer-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderPeer$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FolderPeer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FolderPeer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderPeer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
