% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ForumTopic}
\alias{ForumTopic}
\title{ForumTopic}
\description{
Telegram API type ForumTopic
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ForumTopic-new}{\code{ForumTopic$new()}}
\item \href{#method-ForumTopic-to_list}{\code{ForumTopic$to_list()}}
\item \href{#method-ForumTopic-clone}{\code{ForumTopic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForumTopic-new"></a>}}
\if{latex}{\out{\hypertarget{method-ForumTopic-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForumTopic$new(
  id,
  date = NULL,
  title,
  icon_color,
  top_message,
  read_inbox_max_id,
  read_outbox_max_id,
  unread_count,
  unread_mentions_count,
  unread_reactions_count,
  from_id,
  notify_settings,
  my = NULL,
  closed = NULL,
  pinned = NULL,
  short = NULL,
  hidden = NULL,
  icon_emoji_id = NULL,
  draft = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForumTopic-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ForumTopic-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForumTopic$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForumTopic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ForumTopic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForumTopic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
