% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ForumTopicDeleted}
\alias{ForumTopicDeleted}
\title{ForumTopicDeleted}
\description{
Telegram API type ForumTopicDeleted
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ForumTopicDeleted-new}{\code{ForumTopicDeleted$new()}}
\item \href{#method-ForumTopicDeleted-to_list}{\code{ForumTopicDeleted$to_list()}}
\item \href{#method-ForumTopicDeleted-clone}{\code{ForumTopicDeleted$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForumTopicDeleted-new"></a>}}
\if{latex}{\out{\hypertarget{method-ForumTopicDeleted-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForumTopicDeleted$new(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForumTopicDeleted-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ForumTopicDeleted-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForumTopicDeleted$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForumTopicDeleted-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ForumTopicDeleted-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForumTopicDeleted$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
