% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{FoundStory}
\alias{FoundStory}
\title{FoundStory}
\description{
Telegram API type FoundStory
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{peer}}{Field.}

\item{\code{story}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FoundStory-new}{\code{FoundStory$new()}}
\item \href{#method-FoundStory-to_list}{\code{FoundStory$to_list()}}
\item \href{#method-FoundStory-clone}{\code{FoundStory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoundStory-new"></a>}}
\if{latex}{\out{\hypertarget{method-FoundStory-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoundStory$new(peer, story)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoundStory-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-FoundStory-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoundStory$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FoundStory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FoundStory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FoundStory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
