% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpfull.R
\name{FullPacketCodec}
\alias{FullPacketCodec}
\title{FullPacketCodec}
\description{
A codec for full TCP packets for Telegram.
}
\section{Super class}{
\code{\link[telegramR:PacketCodec]{telegramR::PacketCodec}} -> \code{FullPacketCodec}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{_send_counter}}{Counter for sent packets.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{_send_counter}}{Counter for sent packets.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FullPacketCodec-new}{\code{FullPacketCodec$new()}}
\item \href{#method-FullPacketCodec-encode_packet}{\code{FullPacketCodec$encode_packet()}}
\item \href{#method-FullPacketCodec-read_packet}{\code{FullPacketCodec$read_packet()}}
\item \href{#method-FullPacketCodec-clone}{\code{FullPacketCodec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullPacketCodec-new"></a>}}
\if{latex}{\out{\hypertarget{method-FullPacketCodec-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FullPacketCodec.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullPacketCodec$new(connection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{A connection object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullPacketCodec-encode_packet"></a>}}
\if{latex}{\out{\hypertarget{method-FullPacketCodec-encode_packet}{}}}
\subsection{Method \code{encode_packet()}}{
Encode a packet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullPacketCodec$encode_packet(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A raw vector containing the packet data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A raw vector representing the full packet.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullPacketCodec-read_packet"></a>}}
\if{latex}{\out{\hypertarget{method-FullPacketCodec-read_packet}{}}}
\subsection{Method \code{read_packet()}}{
Read a packet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullPacketCodec$read_packet(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{An object with a `readexactly` method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A raw vector representing the packet body.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullPacketCodec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FullPacketCodec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullPacketCodec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
