% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Game}
\alias{Game}
\title{Game}
\description{
Telegram API type Game
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Game-new}{\code{Game$new()}}
\item \href{#method-Game-to_list}{\code{Game$to_list()}}
\item \href{#method-Game-clone}{\code{Game$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Game-new"></a>}}
\if{latex}{\out{\hypertarget{method-Game-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$new(
  id,
  access_hash,
  short_name,
  title,
  description,
  photo,
  document = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Game-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Game-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Game-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Game-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
