% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_key_data_from_nonce}
\alias{generate_key_data_from_nonce}
\title{Generate key data from nonce}
\usage{
generate_key_data_from_nonce(server_nonce, new_nonce)
}
\arguments{
\item{server_nonce}{A raw vector representing the server nonce (16 bytes).}

\item{new_nonce}{A raw vector representing the new nonce (32 bytes).}
}
\value{
A list containing 'key' and 'iv' as raw vectors.
}
\description{
Generate key data from nonce
}
\examples{
\dontrun{
server_nonce <- as.raw(sample(0:255, 16, replace = TRUE))
new_nonce <- as.raw(sample(0:255, 32, replace = TRUE))
key_data <- generate_key_data_from_nonce(server_nonce, new_nonce)
str(key_data)
}
}
