% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_random_long}
\alias{generate_random_long}
\title{Generate a random long integer (8 bytes), optionally signed}
\usage{
generate_random_long(signed = TRUE)
}
\arguments{
\item{signed}{Logical indicating if the integer should be signed (default TRUE)}
}
\value{
A random long integer
}
\description{
Generate a random long integer (8 bytes), optionally signed
}
\examples{
\dontrun{
random_long <- generate_random_long()
print(random_long)
random_unsigned_long <- generate_random_long(signed = FALSE)
print(random_unsigned_long)
}
}
