% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{GenericDownloadIter}
\alias{GenericDownloadIter}
\title{GenericDownloadIter class}
\description{
Class for generic file downloading iteration.
}
\section{Super classes}{
\code{\link[telegramR:RequestIter]{telegramR::RequestIter}} -> \code{\link[telegramR:DirectDownloadIter]{telegramR::DirectDownloadIter}} -> \code{GenericDownloadIter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenericDownloadIter-load_next_chunk}{\code{GenericDownloadIter$load_next_chunk()}}
\item \href{#method-GenericDownloadIter-clone}{\code{GenericDownloadIter$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id=".next"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-.next'><code>telegramR::RequestIter$.next()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="async_init"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-async_init'><code>telegramR::RequestIter$async_init()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="collect"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-collect'><code>telegramR::RequestIter$collect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="initialize"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-initialize'><code>telegramR::RequestIter$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="RequestIter" data-id="reset"><a href='../../telegramR/html/RequestIter.html#method-RequestIter-reset'><code>telegramR::RequestIter$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="DirectDownloadIter" data-id="close"><a href='../../telegramR/html/DirectDownloadIter.html#method-DirectDownloadIter-close'><code>telegramR::DirectDownloadIter$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="DirectDownloadIter" data-id="init"><a href='../../telegramR/html/DirectDownloadIter.html#method-DirectDownloadIter-init'><code>telegramR::DirectDownloadIter$init()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="DirectDownloadIter" data-id="request"><a href='../../telegramR/html/DirectDownloadIter.html#method-DirectDownloadIter-request'><code>telegramR::DirectDownloadIter$request()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericDownloadIter-load_next_chunk"></a>}}
\if{latex}{\out{\hypertarget{method-GenericDownloadIter-load_next_chunk}{}}}
\subsection{Method \code{load_next_chunk()}}{
Load the next chunk for generic download.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericDownloadIter$load_next_chunk()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericDownloadIter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenericDownloadIter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericDownloadIter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
