% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{GeoPoint}
\alias{GeoPoint}
\title{GeoPoint}
\description{
Telegram API type GeoPoint
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeoPoint-new}{\code{GeoPoint$new()}}
\item \href{#method-GeoPoint-to_list}{\code{GeoPoint$to_list()}}
\item \href{#method-GeoPoint-clone}{\code{GeoPoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoPoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeoPoint-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoPoint$new(long, lat, access_hash, accuracy_radius = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoPoint-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GeoPoint-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoPoint$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoPoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeoPoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoPoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
