% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{GeoPointAddress}
\alias{GeoPointAddress}
\title{GeoPointAddress}
\description{
Telegram API type GeoPointAddress
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeoPointAddress-new}{\code{GeoPointAddress$new()}}
\item \href{#method-GeoPointAddress-to_list}{\code{GeoPointAddress$to_list()}}
\item \href{#method-GeoPointAddress-clone}{\code{GeoPointAddress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoPointAddress-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeoPointAddress-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoPointAddress$new(country_iso2, state = NULL, city = NULL, street = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoPointAddress-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GeoPointAddress-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoPointAddress$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoPointAddress-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeoPointAddress-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoPointAddress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
