% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{GeoPointEmpty}
\alias{GeoPointEmpty}
\title{GeoPointEmpty}
\description{
Telegram API type GeoPointEmpty
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeoPointEmpty-to_list}{\code{GeoPointEmpty$to_list()}}
\item \href{#method-GeoPointEmpty-clone}{\code{GeoPointEmpty$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoPointEmpty-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GeoPointEmpty-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoPointEmpty$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoPointEmpty-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeoPointEmpty-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoPointEmpty$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
