% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_appropriated_part_size}
\alias{get_appropriated_part_size}
\title{Get Appropriated Part Size}
\usage{
get_appropriated_part_size(file_size)
}
\arguments{
\item{file_size}{An integer representing the file size in bytes.}
}
\value{
An integer representing the part size.
}
\description{
Gets the appropriated part size when uploading or downloading files,
given an initial file size.
}
