% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_attributes}
\alias{get_attributes}
\title{Get Attributes for File}
\usage{
get_attributes(
  file,
  attributes = NULL,
  mime_type = NULL,
  force_document = FALSE,
  voice_note = FALSE,
  video_note = FALSE,
  supports_streaming = FALSE,
  thumb = NULL
)
}
\arguments{
\item{file}{The file object or path. Can be a string or an object with a 'name' attribute.}

\item{attributes}{A list of user-provided attributes to override defaults. Default is NULL.}

\item{mime_type}{The MIME type of the file. If NULL, it will be guessed. Default is NULL.}

\item{force_document}{Logical, whether to force the file as a document. Default is FALSE.}

\item{voice_note}{Logical, whether the file is a voice note. Default is FALSE.}

\item{video_note}{Logical, whether the file is a video note. Default is FALSE.}

\item{supports_streaming}{Logical, whether the video supports streaming. Default is FALSE.}

\item{thumb}{The thumbnail file for video attributes. Default is NULL.}
}
\value{
A list containing the attributes list and the MIME type.
}
\description{
This function retrieves a list of attributes for the given file and the MIME type as a list.
It determines attributes such as filename, audio, and video based on the file's metadata and parameters.
}
\examples{
\dontrun{
\dontrun{
# Assuming types and helper functions are defined
attrs <- get_attributes("example.mp3", voice_note = TRUE)
}
}
}
