% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa.R
\name{get_byte_array}
\alias{get_byte_array}
\title{Converts an integer to a byte array.}
\usage{
get_byte_array(integer)
}
\arguments{
\item{integer}{The integer to convert.}
}
\value{
A raw vector representing the byte array.
}
\description{
Converts an integer to a byte array.
}
