% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_info.R
\name{get_data}
\alias{get_data}
\title{returns a dataframe with general info}
\usage{
get_data(file)
}
\arguments{
\item{file}{CHAR. The folder where html exported from Telegram are located}
}
\description{
returns a dataframe with general info
}
\details{
insert on folder with html from exported chats of telegram and recieve a dataframe with info such as first date and last date of each html
}
\examples{
get_data("~/Downloads/Telegram Desktop/ChatExport_2023-12-16/messages.html")

}
