% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_display_name}
\alias{get_display_name}
\title{Get Display Name}
\usage{
get_display_name(entity)
}
\arguments{
\item{entity}{The entity object (User, Chat, or Channel).}
}
\value{
A character string representing the display name.
}
\description{
Gets the display name for the given User, Chat or Channel. Returns an empty string otherwise.
}
\examples{
\dontrun{
\dontrun{
# Assuming entity is a User object
name <- get_display_name(entity)
}
}
}
