% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_entity_pair}
\alias{get_entity_pair}
\title{Get Entity Pair}
\usage{
get_entity_pair(entity_id, entities, cache, get_input_peer = get_input_peer)
}
\arguments{
\item{entity_id}{The entity ID.}

\item{entities}{A list or environment of entities.}

\item{cache}{A cache object.}

\item{get_input_peer}{A function to get input peer, defaults to \code{get_input_peer}.}
}
\value{
A list containing entity and input_entity, or NULLs if not found.
}
\description{
Returns a list of \code{(entity, input_entity)} for the given entity ID.
}
