% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_extension}
\alias{get_extension}
\title{Get Extension}
\usage{
get_extension(media)
}
\arguments{
\item{media}{The media object to analyze.}
}
\value{
A character string representing the file extension.
}
\description{
Gets the corresponding extension for any Telegram media.
}
\examples{
\dontrun{
# Assuming media is a Document object
ext <- get_extension(media)
}
}
