% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Get File Extension}
\usage{
get_file_extension(file)
}
\arguments{
\item{file}{The file path or object.}
}
\value{
The file extension as a string.
}
\description{
Gets the extension for the given file, which can be either a
string or an open file (which has a \code{name} attribute).
}
