% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_info}
\alias{get_file_info}
\title{Get File Info}
\usage{
get_file_info(location)
}
\arguments{
\item{location}{The location object.}
}
\value{
A list with elements \code{dc_id}, \code{location}, and \code{size}.
}
\description{
Internal function to get file info.
}
