% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_profile.R
\name{get_full_user}
\alias{get_full_user}
\title{Retrieve full information about a user}
\usage{
get_full_user(client, user = "me")
}
\arguments{
\item{client}{TelegramClient instance.}

\item{user}{User (username, id, link) or "me".}
}
\value{
The API response (UserFull).
}
\description{
Retrieve full information about a user
}
