% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_channel}
\alias{get_input_channel}
\title{Get Input Channel}
\usage{
get_input_channel(entity)
}
\arguments{
\item{entity}{The entity object to convert.}
}
\value{
An InputChannel object.
}
\description{
Similar to `get_input_peer`, but for `InputChannel`'s alone.
}
\examples{
\dontrun{
# Assuming entity is a Channel object
input_channel <- get_input_channel(entity)
}
}
