% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_chat_photo}
\alias{get_input_chat_photo}
\title{Get Input Chat Photo}
\usage{
get_input_chat_photo(photo)
}
\arguments{
\item{photo}{The chat photo object to convert.}
}
\value{
An InputChatPhoto object.
}
\description{
Converts a photo-like object to an InputChatPhoto for API calls.
}
\keyword{internal}
