% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_dialog}
\alias{get_input_dialog}
\title{Get Input Dialog}
\usage{
get_input_dialog(dialog)
}
\arguments{
\item{dialog}{The dialog object to convert.}
}
\value{
An InputDialogPeer object.
}
\description{
Similar to \code{get_input_peer}, but for dialogs.
}
\examples{
\dontrun{
\dontrun{
# Assuming dialog is a dialog object
input_dialog <- get_input_dialog(dialog)
}
}
}
