% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_document}
\alias{get_input_document}
\title{Get Input Document}
\usage{
get_input_document(document)
}
\arguments{
\item{document}{The document object to convert.}
}
\value{
An InputDocument object.
}
\description{
Similar to \code{get_input_peer}, but for documents.
}
\examples{
\dontrun{
# Assuming document is a Document object
input_doc <- get_input_document(document)
}
}
