% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_geo}
\alias{get_input_geo}
\title{Get Input Geo}
\usage{
get_input_geo(geo)
}
\arguments{
\item{geo}{The geo object to convert.}
}
\value{
An InputGeoPoint object.
}
\description{
Converts a geo-like object to an InputGeoPoint for API calls.
}
\keyword{internal}
