% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_group_call}
\alias{get_input_group_call}
\title{Get Input Group Call}
\usage{
get_input_group_call(call)
}
\arguments{
\item{call}{The call object to convert.}
}
\value{
An InputGroupCall object.
}
\description{
Similar to \code{get_input_peer}, but for input calls.
}
