% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_location}
\alias{get_input_location}
\title{Get Input Location}
\usage{
get_input_location(location)
}
\arguments{
\item{location}{The location object.}
}
\value{
A list with elements \code{dc_id} and \code{location}.
}
\description{
Similar to \code{get_input_peer}, but for input messages.
}
\details{
Note that this returns a list \code{(dc_id, location)}, the
\code{dc_id} being present if known.
}
