% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_media}
\alias{get_input_media}
\title{Get Input Media}
\usage{
get_input_media(
  media,
  is_photo = FALSE,
  attributes = NULL,
  force_document = FALSE,
  voice_note = FALSE,
  video_note = FALSE,
  supports_streaming = FALSE,
  ttl = NULL
)
}
\arguments{
\item{media}{The media object to convert.}

\item{is_photo}{Logical, whether the media is a photo. Default is \code{FALSE}.}

\item{attributes}{A list of attributes for the media. Default is \code{NULL}.}

\item{force_document}{Logical, whether to force the media as a document. Default is \code{FALSE}.}

\item{voice_note}{Logical, whether the media is a voice note. Default is \code{FALSE}.}

\item{video_note}{Logical, whether the media is a video note. Default is \code{FALSE}.}

\item{supports_streaming}{Logical, whether the video supports streaming. Default is \code{FALSE}.}

\item{ttl}{The time-to-live in seconds. Default is \code{NULL}.}
}
\value{
An InputMedia object.
}
\description{
Similar to \code{get_input_peer}, but for media.
}
\details{
If the media is \code{InputFile} and \code{is_photo} is known to be \code{TRUE},
it will be treated as an \code{InputMediaUploadedPhoto}. Else, the rest
of parameters will indicate how to treat it.
}
