% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_message}
\alias{get_input_message}
\title{Get Input Message}
\usage{
get_input_message(message)
}
\arguments{
\item{message}{The message object or ID to convert.}
}
\value{
An InputMessage object.
}
\description{
Similar to \code{get_input_peer}, but for input messages.
}
