% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_photo}
\alias{get_input_photo}
\title{Get Input Photo}
\usage{
get_input_photo(photo)
}
\arguments{
\item{photo}{The photo object to convert.}
}
\value{
An InputPhoto object.
}
\description{
Converts a photo-like object to an InputPhoto for API calls.
}
\keyword{internal}
