% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_user}
\alias{get_input_user}
\title{Get Input User}
\usage{
get_input_user(entity)
}
\arguments{
\item{entity}{The entity object to convert.}
}
\value{
An InputUser object.
}
\description{
Similar to \code{get_input_peer}, but for \code{InputUser}'s alone.
}
\examples{
\dontrun{
# Assuming entity is a User object
input_user <- get_input_user(entity)
}
}
