% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentificator.R
\name{get_int}
\alias{get_int}
\title{Convert a raw byte array to an integer.}
\usage{
get_int(byte_array, signed = TRUE)
}
\arguments{
\item{byte_array}{A raw vector representing the integer.}

\item{signed}{Logical; whether the integer is signed.}
}
\value{
An integer converted from the byte array.
}
\description{
Convert a raw byte array to an integer.
}
