% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentificator.R
\name{get_int_little}
\alias{get_int_little}
\title{Convert a raw byte array in little-endian order to an integer.}
\usage{
get_int_little(byte_array, signed = TRUE)
}
\arguments{
\item{byte_array}{A raw vector in little-endian order.}

\item{signed}{Logical; whether the integer is signed.}
}
\value{
An integer converted from the byte array.
}
\description{
Convert a raw byte array in little-endian order to an integer.
}
