% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_message_id}
\alias{get_message_id}
\title{Get Message ID}
\usage{
get_message_id(message)
}
\arguments{
\item{message}{The message object or ID.}
}
\value{
The message ID as an integer, or NULL if invalid.
}
\description{
Similar to \code{get_input_peer}, but for message IDs.
}
