% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_peer_id}
\alias{get_peer_id}
\title{Get Peer ID}
\usage{
get_peer_id(peer, add_mark = TRUE)
}
\arguments{
\item{peer}{The peer object or integer to convert.}

\item{add_mark}{Logical, whether to add the mark (default TRUE).}
}
\value{
The marked or unmarked peer ID as an integer.
}
\description{
Convert the given peer into its marked ID by default.
}
\details{
This "mark" comes from the "bot api" format, and with it the peer type
can be identified back. User ID is left unmodified, chat ID is negated,
and channel ID is "prefixed" with -100:

* \code{user_id}
* \code{-chat_id}
* \code{-100channel_id}

The original ID and the peer type class can be returned with
a call to \code{resolve_id(marked_id)}.
}
\examples{
\dontrun{
\dontrun{
# Assuming peer is a PeerUser object
id <- get_peer_id(peer)
}
}
}
