% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{get_running_loop}
\alias{get_running_loop}
\title{Helper functions to simulate Python functionality}
\usage{
get_running_loop()
}
\value{
A promise object or NULL.
}
\description{
These functions are used to simulate Python's asyncio and threading behavior in R.
}
