% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetAppUpdateRequest}
\alias{GetAppUpdateRequest}
\title{GetAppUpdateRequest}
\description{
Telegram API type GetAppUpdateRequest
}
\details{
GetAppUpdateRequest R6 class

Request to get app update info.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{source}}{character Source string.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetAppUpdateRequest-new}{\code{GetAppUpdateRequest$new()}}
\item \href{#method-GetAppUpdateRequest-to_list}{\code{GetAppUpdateRequest$to_list()}}
\item \href{#method-GetAppUpdateRequest-to_bytes}{\code{GetAppUpdateRequest$to_bytes()}}
\item \href{#method-GetAppUpdateRequest-clone}{\code{GetAppUpdateRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAppUpdateRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetAppUpdateRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetAppUpdateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAppUpdateRequest$new(source = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{character
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAppUpdateRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetAppUpdateRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAppUpdateRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `_` discriminator and source
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAppUpdateRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetAppUpdateRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAppUpdateRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAppUpdateRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetAppUpdateRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAppUpdateRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
