% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthenticationKey.R
\name{getAuthenticationKey}
\alias{getAuthenticationKey}
\title{getAuthenticationKey}
\usage{
getAuthenticationKey(file_path = "~/src/telegram_credentials.yaml", bot)
}
\arguments{
\item{bot}{The yaml block which corresponds to your key.}
}
\description{
Get a local authentication key for interacting with Telegram
}
\examples{
\dontrun{
}
}
