% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_premium.R
\name{GetBoostsListRequest}
\alias{GetBoostsListRequest}
\title{GetBoostsListRequest R6 class}
\description{
GetBoostsListRequest R6 class

GetBoostsListRequest R6 class
}
\details{
Representation of the TL request "GetBoostsListRequest"
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor ID for the request}

\item{\code{SUBCLASS_OF_ID}}{Subclass ID for the request}

\item{\code{peer}}{Input peer (object or identifier)}

\item{\code{offset}}{String offset}

\item{\code{limit}}{Integer limit}

\item{\code{gifts}}{Optional logical}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetBoostsListRequest-new}{\code{GetBoostsListRequest$new()}}
\item \href{#method-GetBoostsListRequest-resolve}{\code{GetBoostsListRequest$resolve()}}
\item \href{#method-GetBoostsListRequest-to_list}{\code{GetBoostsListRequest$to_list()}}
\item \href{#method-GetBoostsListRequest-bytes}{\code{GetBoostsListRequest$bytes()}}
\item \href{#method-GetBoostsListRequest-from_reader}{\code{GetBoostsListRequest$from_reader()}}
\item \href{#method-GetBoostsListRequest-clone}{\code{GetBoostsListRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsListRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsListRequest-new}{}}}
\subsection{Method \code{new()}}{
Create a new GetBoostsListRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsListRequest$new(peer = NULL, offset = "", limit = 0L, gifts = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{input peer or identifier}

\item{\code{offset}}{string offset}

\item{\code{limit}}{integer limit}

\item{\code{gifts}}{optional logical}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsListRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsListRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve peer using client/utils helpers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsListRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client object that provides get_input_entity}

\item{\code{utils}}{utils object that provides get_input_peer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsListRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsListRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert request to list (dictionary)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsListRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsListRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsListRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize to raw bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsListRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsListRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsListRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Construct a new GetBoostsListRequest from a reader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsListRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{reader object exposing read_int / read_string / tgread_object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GetBoostsListRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsListRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsListRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsListRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
