% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_premium.R
\name{GetBoostsStatusRequest}
\alias{GetBoostsStatusRequest}
\title{GetBoostsStatusRequest R6 class}
\description{
GetBoostsStatusRequest R6 class

GetBoostsStatusRequest R6 class
}
\details{
Representation of the TL request "GetBoostsStatusRequest".
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor ID for the request}

\item{\code{SUBCLASS_OF_ID}}{Subclass ID for the request}

\item{\code{peer}}{Input peer (object or identifier)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetBoostsStatusRequest-new}{\code{GetBoostsStatusRequest$new()}}
\item \href{#method-GetBoostsStatusRequest-resolve}{\code{GetBoostsStatusRequest$resolve()}}
\item \href{#method-GetBoostsStatusRequest-to_list}{\code{GetBoostsStatusRequest$to_list()}}
\item \href{#method-GetBoostsStatusRequest-bytes}{\code{GetBoostsStatusRequest$bytes()}}
\item \href{#method-GetBoostsStatusRequest-from_reader}{\code{GetBoostsStatusRequest$from_reader()}}
\item \href{#method-GetBoostsStatusRequest-clone}{\code{GetBoostsStatusRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsStatusRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsStatusRequest-new}{}}}
\subsection{Method \code{new()}}{
Create a new GetBoostsStatusRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsStatusRequest$new(peer = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{input peer or identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsStatusRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsStatusRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve peer using client/utils helpers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsStatusRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client object that provides get_input_entity}

\item{\code{utils}}{utils object that provides get_input_peer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsStatusRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsStatusRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert request to list (dictionary)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsStatusRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsStatusRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsStatusRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize to raw bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsStatusRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsStatusRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsStatusRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Construct a new GetBoostsStatusRequest from a reader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsStatusRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{reader object exposing tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GetBoostsStatusRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBoostsStatusRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetBoostsStatusRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBoostsStatusRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
