% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{GetBotCommandsRequest}
\alias{GetBotCommandsRequest}
\title{GetBotCommandsRequest}
\description{
Telegram API type GetBotCommandsRequest
}
\details{
GetBotCommandsRequest R6 class

Represents the GetBotCommandsRequest TL request.



Each method is documented inline below.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scope}}{Field.}

\item{\code{lang_code}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetBotCommandsRequest-new}{\code{GetBotCommandsRequest$new()}}
\item \href{#method-GetBotCommandsRequest-to_list}{\code{GetBotCommandsRequest$to_list()}}
\item \href{#method-GetBotCommandsRequest-to_bytes}{\code{GetBotCommandsRequest$to_bytes()}}
\item \href{#method-GetBotCommandsRequest-clone}{\code{GetBotCommandsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBotCommandsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetBotCommandsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetBotCommandsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBotCommandsRequest$new(scope, lang_code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scope}}{TypeBotCommandScope or object}

\item{\code{lang_code}}{character
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBotCommandsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetBotCommandsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBotCommandsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBotCommandsRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetBotCommandsRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBotCommandsRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBotCommandsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetBotCommandsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBotCommandsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
