% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{GetBotRecommendationsRequest}
\alias{GetBotRecommendationsRequest}
\title{GetBotRecommendationsRequest R6 class}
\description{
Methods:
- initialize(bot)
- resolve(client, utils)
- to_list() -> list representation
- to_bytes() -> raw vector bytes

Each method is documented inline below.
}
\details{
Represents the GetBotRecommendationsRequest TL request.
}
\examples{
# o <- GetBotRecommendationsRequest$new(someBot)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bot}}{TypeInputUser input user or object implementing to_list()/to_bytes()}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetBotRecommendationsRequest-new}{\code{GetBotRecommendationsRequest$new()}}
\item \href{#method-GetBotRecommendationsRequest-resolve}{\code{GetBotRecommendationsRequest$resolve()}}
\item \href{#method-GetBotRecommendationsRequest-to_list}{\code{GetBotRecommendationsRequest$to_list()}}
\item \href{#method-GetBotRecommendationsRequest-to_bytes}{\code{GetBotRecommendationsRequest$to_bytes()}}
\item \href{#method-GetBotRecommendationsRequest-clone}{\code{GetBotRecommendationsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBotRecommendationsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetBotRecommendationsRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBotRecommendationsRequest$new(bot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bot}}{TypeInputUser or identifier
Resolve references (convert entities via client/utils)

Resolves bot via client$get_input_entity and utils$get_input_user.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBotRecommendationsRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetBotRecommendationsRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBotRecommendationsRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client with get_input_entity method}

\item{\code{utils}}{utils with get_input_user}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
Convert to list (dictionary-like)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBotRecommendationsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetBotRecommendationsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBotRecommendationsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBotRecommendationsRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetBotRecommendationsRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBotRecommendationsRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBotRecommendationsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetBotRecommendationsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetBotRecommendationsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
