% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{GetCallConfigRequest}
\alias{GetCallConfigRequest}
\title{GetCallConfigRequest R6 class}
\description{
Represents GetCallConfigRequest TLRequest.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetCallConfigRequest-to_list}{\code{GetCallConfigRequest$to_list()}}
\item \href{#method-GetCallConfigRequest-bytes}{\code{GetCallConfigRequest$bytes()}}
\item \href{#method-GetCallConfigRequest-from_reader}{\code{GetCallConfigRequest$from_reader()}}
\item \href{#method-GetCallConfigRequest-clone}{\code{GetCallConfigRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCallConfigRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetCallConfigRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCallConfigRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
Produce raw bytes for the request

Builds the TL-serialized byte representation.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCallConfigRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetCallConfigRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCallConfigRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCallConfigRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetCallConfigRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCallConfigRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCallConfigRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetCallConfigRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCallConfigRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
