% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetCdnConfigRequest}
\alias{GetCdnConfigRequest}
\title{GetCdnConfigRequest}
\description{
Telegram API type GetCdnConfigRequest
}
\details{
GetCdnConfigRequest R6 class

Request to get CDN configuration.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetCdnConfigRequest-new}{\code{GetCdnConfigRequest$new()}}
\item \href{#method-GetCdnConfigRequest-to_list}{\code{GetCdnConfigRequest$to_list()}}
\item \href{#method-GetCdnConfigRequest-to_bytes}{\code{GetCdnConfigRequest$to_bytes()}}
\item \href{#method-GetCdnConfigRequest-clone}{\code{GetCdnConfigRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnConfigRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnConfigRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetCdnConfigRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnConfigRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible NULL
Convert to list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnConfigRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnConfigRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnConfigRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `_` discriminator
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnConfigRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnConfigRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnConfigRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnConfigRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnConfigRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnConfigRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
