% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{GetCdnFileHashesRequest}
\alias{GetCdnFileHashesRequest}
\title{GetCdnFileHashesRequest R6 class}
\description{
GetCdnFileHashesRequest R6 class

GetCdnFileHashesRequest R6 class
}
\details{
Represents the TL request upload.GetCdnFileHashesRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file_token}}{raw|integer|character token bytes}

\item{\code{offset}}{numeric 64-bit offset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetCdnFileHashesRequest-new}{\code{GetCdnFileHashesRequest$new()}}
\item \href{#method-GetCdnFileHashesRequest-to_list}{\code{GetCdnFileHashesRequest$to_list()}}
\item \href{#method-GetCdnFileHashesRequest-to_bytes}{\code{GetCdnFileHashesRequest$to_bytes()}}
\item \href{#method-GetCdnFileHashesRequest-from_reader}{\code{GetCdnFileHashesRequest$from_reader()}}
\item \href{#method-GetCdnFileHashesRequest-clone}{\code{GetCdnFileHashesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnFileHashesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnFileHashesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a GetCdnFileHashesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnFileHashesRequest$new(file_token, offset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_token}}{raw|integer|character}

\item{\code{offset}}{numeric (64-bit)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnFileHashesRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnFileHashesRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnFileHashesRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnFileHashesRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnFileHashesRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to raw vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnFileHashesRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnFileHashesRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnFileHashesRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnFileHashesRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnFileHashesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnFileHashesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnFileHashesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
