% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{GetCdnFileRequest}
\alias{GetCdnFileRequest}
\title{GetCdnFileRequest}
\description{
Telegram API type GetCdnFileRequest
}
\details{
GetCdnFileRequest R6 class

Represents the TL request upload.GetCdnFileRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file_token}}{Field.}

\item{\code{offset}}{Field.}

\item{\code{limit}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetCdnFileRequest-new}{\code{GetCdnFileRequest$new()}}
\item \href{#method-GetCdnFileRequest-to_list}{\code{GetCdnFileRequest$to_list()}}
\item \href{#method-GetCdnFileRequest-to_bytes}{\code{GetCdnFileRequest$to_bytes()}}
\item \href{#method-GetCdnFileRequest-from_reader}{\code{GetCdnFileRequest$from_reader()}}
\item \href{#method-GetCdnFileRequest-clone}{\code{GetCdnFileRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnFileRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnFileRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a GetCdnFileRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnFileRequest$new(file_token, offset, limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_token}}{raw|integer|character}

\item{\code{offset}}{numeric (64-bit)}

\item{\code{limit}}{integer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnFileRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnFileRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnFileRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnFileRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnFileRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to raw vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnFileRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnFileRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnFileRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read GetCdnFileRequest from a reader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnFileRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{an object providing tgread_bytes(), read_long(), read_int()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GetCdnFileRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCdnFileRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetCdnFileRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCdnFileRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
