% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_updates.R
\name{GetChannelDifferenceRequest}
\alias{GetChannelDifferenceRequest}
\title{GetChannelDifferenceRequest R6 class}
\value{
R6 object of class GetChannelDifferenceRequest
}
\description{
Representation of updates.GetChannelDifference request.
}
\examples{
# req <- GetChannelDifferenceRequest$new(channel = some_channel, filter = some_filter, pts = 0, limit = 100)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{channel}}{TypeInputChannel Input channel (R6 object with bytes() or raw())}

\item{\code{filter}}{TypeChannelMessagesFilter Filter (R6 object with bytes() or raw())}

\item{\code{pts}}{integer PTS}

\item{\code{limit}}{integer Limit}

\item{\code{force}}{logical|NULL Optional force flag}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetChannelDifferenceRequest-new}{\code{GetChannelDifferenceRequest$new()}}
\item \href{#method-GetChannelDifferenceRequest-to_list}{\code{GetChannelDifferenceRequest$to_list()}}
\item \href{#method-GetChannelDifferenceRequest-bytes}{\code{GetChannelDifferenceRequest$bytes()}}
\item \href{#method-GetChannelDifferenceRequest-from_reader}{\code{GetChannelDifferenceRequest$from_reader()}}
\item \href{#method-GetChannelDifferenceRequest-clone}{\code{GetChannelDifferenceRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetChannelDifferenceRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetChannelDifferenceRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetChannelDifferenceRequest$new(channel, filter, pts, limit, force = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetChannelDifferenceRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetChannelDifferenceRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetChannelDifferenceRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetChannelDifferenceRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetChannelDifferenceRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetChannelDifferenceRequest$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetChannelDifferenceRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetChannelDifferenceRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetChannelDifferenceRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetChannelDifferenceRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetChannelDifferenceRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetChannelDifferenceRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
