% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chatlists.R
\name{GetChatlistUpdatesRequest}
\alias{GetChatlistUpdatesRequest}
\title{GetChatlistUpdatesRequest}
\description{
Telegram API type GetChatlistUpdatesRequest
}
\details{
GetChatlistUpdatesRequest R6 class

Request to get updates for a chatlist.

Serializes the constructor id followed by chatlist bytes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{chatlist}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetChatlistUpdatesRequest-new}{\code{GetChatlistUpdatesRequest$new()}}
\item \href{#method-GetChatlistUpdatesRequest-to_list}{\code{GetChatlistUpdatesRequest$to_list()}}
\item \href{#method-GetChatlistUpdatesRequest-bytes}{\code{GetChatlistUpdatesRequest$bytes()}}
\item \href{#method-GetChatlistUpdatesRequest-from_reader}{\code{GetChatlistUpdatesRequest$from_reader()}}
\item \href{#method-GetChatlistUpdatesRequest-clone}{\code{GetChatlistUpdatesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetChatlistUpdatesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetChatlistUpdatesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetChatlistUpdatesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetChatlistUpdatesRequest$new(chatlist = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chatlist}}{Input chatlist object
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetChatlistUpdatesRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetChatlistUpdatesRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetChatlistUpdatesRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representation
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetChatlistUpdatesRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetChatlistUpdatesRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetChatlistUpdatesRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector with serialized bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetChatlistUpdatesRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetChatlistUpdatesRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetChatlistUpdatesRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetChatlistUpdatesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetChatlistUpdatesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetChatlistUpdatesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
