% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetConfigRequest}
\alias{GetConfigRequest}
\title{GetConfigRequest R6 class}
\description{
GetConfigRequest R6 class

GetConfigRequest R6 class
}
\details{
Request to get configuration.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetConfigRequest-new}{\code{GetConfigRequest$new()}}
\item \href{#method-GetConfigRequest-to_list}{\code{GetConfigRequest$to_list()}}
\item \href{#method-GetConfigRequest-to_bytes}{\code{GetConfigRequest$to_bytes()}}
\item \href{#method-GetConfigRequest-clone}{\code{GetConfigRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetConfigRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetConfigRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetConfigRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetConfigRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible NULL
Convert to list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetConfigRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetConfigRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetConfigRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with type discriminator
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetConfigRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetConfigRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetConfigRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetConfigRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetConfigRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetConfigRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
