% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetCountriesListRequest}
\alias{GetCountriesListRequest}
\title{GetCountriesListRequest R6 class}
\description{
Request to get the list of countries (possibly not modified).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{langCode}}{character Language code.}

\item{\code{hash}}{integer Hash to check for modifications.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetCountriesListRequest-new}{\code{GetCountriesListRequest$new()}}
\item \href{#method-GetCountriesListRequest-to_list}{\code{GetCountriesListRequest$to_list()}}
\item \href{#method-GetCountriesListRequest-to_bytes}{\code{GetCountriesListRequest$to_bytes()}}
\item \href{#method-GetCountriesListRequest-clone}{\code{GetCountriesListRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCountriesListRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetCountriesListRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCountriesListRequest$new(langCode = "", hash = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{langCode}}{character}

\item{\code{hash}}{integer
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCountriesListRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetCountriesListRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCountriesListRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `_` discriminator, lang_code and hash
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCountriesListRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetCountriesListRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCountriesListRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetCountriesListRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetCountriesListRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetCountriesListRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
