% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetDeepLinkInfoRequest}
\alias{GetDeepLinkInfoRequest}
\title{GetDeepLinkInfoRequest}
\description{
Telegram API type GetDeepLinkInfoRequest
}
\details{
GetDeepLinkInfoRequest R6 class

Request to get deep link info.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{path}}{character Deep link path.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetDeepLinkInfoRequest-new}{\code{GetDeepLinkInfoRequest$new()}}
\item \href{#method-GetDeepLinkInfoRequest-to_list}{\code{GetDeepLinkInfoRequest$to_list()}}
\item \href{#method-GetDeepLinkInfoRequest-to_bytes}{\code{GetDeepLinkInfoRequest$to_bytes()}}
\item \href{#method-GetDeepLinkInfoRequest-clone}{\code{GetDeepLinkInfoRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDeepLinkInfoRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetDeepLinkInfoRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetDeepLinkInfoRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDeepLinkInfoRequest$new(path = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{character
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDeepLinkInfoRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetDeepLinkInfoRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDeepLinkInfoRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `_` discriminator and path
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDeepLinkInfoRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetDeepLinkInfoRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDeepLinkInfoRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDeepLinkInfoRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetDeepLinkInfoRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDeepLinkInfoRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
