% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_langpack.R, R/functions_updates.R
\name{GetDifferenceRequest}
\alias{GetDifferenceRequest}
\title{GetDifferenceRequest R6 class}
\description{
GetDifferenceRequest R6 class

GetDifferenceRequest R6 class

GetDifferenceRequest R6 class

GetDifferenceRequest R6 class
}
\details{
TLRequest: GetDifferenceRequest

Representation of updates.GetDifference request.
}
\examples{
# req <- GetDifferenceRequest$new("default", "en", 123)
# req <- GetDifferenceRequest$new(pts = 0, date = Sys.time(), qts = 0)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lang_pack}}{character Language pack identifier}

\item{\code{lang_code}}{character Language code}

\item{\code{from_version}}{integer Version to get difference from}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{lang_pack}}{character Language pack identifier}

\item{\code{lang_code}}{character Language code}

\item{\code{from_version}}{integer Version to get difference from}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetDifferenceRequest-new}{\code{GetDifferenceRequest$new()}}
\item \href{#method-GetDifferenceRequest-to_list}{\code{GetDifferenceRequest$to_list()}}
\item \href{#method-GetDifferenceRequest-bytes}{\code{GetDifferenceRequest$bytes()}}
\item \href{#method-GetDifferenceRequest-from_reader}{\code{GetDifferenceRequest$from_reader()}}
\item \href{#method-GetDifferenceRequest-clone}{\code{GetDifferenceRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDifferenceRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetDifferenceRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDifferenceRequest$new(
  pts,
  date = NULL,
  qts,
  pts_limit = NULL,
  pts_total_limit = NULL,
  qts_limit = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDifferenceRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetDifferenceRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDifferenceRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDifferenceRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetDifferenceRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDifferenceRequest$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDifferenceRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetDifferenceRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDifferenceRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDifferenceRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetDifferenceRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDifferenceRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pts}}{integer Current pts}

\item{\code{date}}{POSIXct|NULL Date}

\item{\code{qts}}{integer QTS}

\item{\code{pts_limit}}{integer|NULL Optional pts_limit}

\item{\code{pts_total_limit}}{integer|NULL Optional pts_total_limit}

\item{\code{qts_limit}}{integer|NULL Optional qts_limit}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetDifferenceRequest-new}{\code{GetDifferenceRequest$new()}}
\item \href{#method-GetDifferenceRequest-to_list}{\code{GetDifferenceRequest$to_list()}}
\item \href{#method-GetDifferenceRequest-bytes}{\code{GetDifferenceRequest$bytes()}}
\item \href{#method-GetDifferenceRequest-from_reader}{\code{GetDifferenceRequest$from_reader()}}
\item \href{#method-GetDifferenceRequest-clone}{\code{GetDifferenceRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDifferenceRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetDifferenceRequest-new}{}}}
\subsection{Method \code{new()}}{
Converts the request to a list representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDifferenceRequest$new(
  pts,
  date = NULL,
  qts,
  pts_limit = NULL,
  pts_total_limit = NULL,
  qts_limit = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pts}}{integer PTS}

\item{\code{date}}{POSIXct|NULL Date}

\item{\code{qts}}{integer QTS}

\item{\code{pts_limit}}{integer|NULL Optional pts_limit}

\item{\code{pts_total_limit}}{integer|NULL Optional pts_total_limit}

\item{\code{qts_limit}}{integer|NULL Optional qts_limit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDifferenceRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetDifferenceRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Initializes a new GetDifferenceRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDifferenceRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new GetDifferenceRequest instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDifferenceRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetDifferenceRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Converts the request to a list representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDifferenceRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDifferenceRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetDifferenceRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDifferenceRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetDifferenceRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetDifferenceRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetDifferenceRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
