% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chatlists.R
\name{GetExportedInvitesRequest}
\alias{GetExportedInvitesRequest}
\title{GetExportedInvitesRequest R6 class}
\description{
GetExportedInvitesRequest R6 class

GetExportedInvitesRequest R6 class
}
\details{
Request to get exported invites for a chatlist.

Serializes the constructor id followed by chatlist bytes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Integer constructor id.}

\item{\code{SUBCLASS_OF_ID}}{Integer subclass id.}

\item{\code{chatlist}}{TypeInputChatlist input chatlist object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetExportedInvitesRequest-new}{\code{GetExportedInvitesRequest$new()}}
\item \href{#method-GetExportedInvitesRequest-to_list}{\code{GetExportedInvitesRequest$to_list()}}
\item \href{#method-GetExportedInvitesRequest-bytes}{\code{GetExportedInvitesRequest$bytes()}}
\item \href{#method-GetExportedInvitesRequest-from_reader}{\code{GetExportedInvitesRequest$from_reader()}}
\item \href{#method-GetExportedInvitesRequest-clone}{\code{GetExportedInvitesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetExportedInvitesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetExportedInvitesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetExportedInvitesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetExportedInvitesRequest$new(chatlist = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chatlist}}{Input chatlist object
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetExportedInvitesRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetExportedInvitesRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetExportedInvitesRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representation
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetExportedInvitesRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetExportedInvitesRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetExportedInvitesRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector with serialized bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetExportedInvitesRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetExportedInvitesRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetExportedInvitesRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetExportedInvitesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetExportedInvitesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetExportedInvitesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
