% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{GetFileHashesRequest}
\alias{GetFileHashesRequest}
\title{GetFileHashesRequest R6 class}
\description{
GetFileHashesRequest R6 class

GetFileHashesRequest R6 class
}
\details{
Represents the TL request upload.GetFileHashesRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{location}}{TLObject-like object (input file location)}

\item{\code{offset}}{numeric 64-bit offset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetFileHashesRequest-new}{\code{GetFileHashesRequest$new()}}
\item \href{#method-GetFileHashesRequest-to_list}{\code{GetFileHashesRequest$to_list()}}
\item \href{#method-GetFileHashesRequest-to_bytes}{\code{GetFileHashesRequest$to_bytes()}}
\item \href{#method-GetFileHashesRequest-from_reader}{\code{GetFileHashesRequest$from_reader()}}
\item \href{#method-GetFileHashesRequest-clone}{\code{GetFileHashesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileHashesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileHashesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a GetFileHashesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFileHashesRequest$new(location, offset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{location}}{TLObject-like}

\item{\code{offset}}{numeric (64-bit)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileHashesRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileHashesRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFileHashesRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileHashesRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileHashesRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to raw vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFileHashesRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileHashesRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileHashesRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFileHashesRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileHashesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileHashesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFileHashesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
