% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{GetFileRequest}
\alias{GetFileRequest}
\title{GetFileRequest}
\description{
Telegram API type GetFileRequest
}
\details{
GetFileRequest R6 class

Represents the TL request upload.GetFileRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{location}}{Field.}

\item{\code{offset}}{Field.}

\item{\code{limit}}{Field.}

\item{\code{precise}}{Field.}

\item{\code{cdn_supported}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetFileRequest-new}{\code{GetFileRequest$new()}}
\item \href{#method-GetFileRequest-to_list}{\code{GetFileRequest$to_list()}}
\item \href{#method-GetFileRequest-to_bytes}{\code{GetFileRequest$to_bytes()}}
\item \href{#method-GetFileRequest-from_reader}{\code{GetFileRequest$from_reader()}}
\item \href{#method-GetFileRequest-clone}{\code{GetFileRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a GetFileRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFileRequest$new(
  location,
  offset,
  limit,
  precise = NULL,
  cdn_supported = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{location}}{TLObject-like}

\item{\code{offset}}{numeric (64-bit)}

\item{\code{limit}}{integer}

\item{\code{precise}}{logical or NULL}

\item{\code{cdn_supported}}{logical or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFileRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to raw vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFileRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read GetFileRequest from a reader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFileRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{an object providing read_int(), tgread_object(), read_long(), read_int()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GetFileRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFileRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
