% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{GetFutureSaltsRequest_from_reader}
\alias{GetFutureSaltsRequest_from_reader}
\title{Create GetFutureSaltsRequest from a reader}
\usage{
GetFutureSaltsRequest_from_reader(reader)
}
\arguments{
\item{reader}{Reader object
Subclass ID for the request}
}
\value{
Instance of GetFutureSaltsRequest
}
\description{
Create GetFutureSaltsRequest from a reader
}
