% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{GetGroupCallChainBlocksRequest}
\alias{GetGroupCallChainBlocksRequest}
\title{GetGroupCallChainBlocksRequest R6 class}
\description{
Represents GetGroupCallChainBlocksRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{The input group call object (expected to implement bytes()).}

\item{\code{sub_chain_id}}{Integer sub chain id.}

\item{\code{offset}}{Integer offset.}

\item{\code{limit}}{Integer limit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetGroupCallChainBlocksRequest-new}{\code{GetGroupCallChainBlocksRequest$new()}}
\item \href{#method-GetGroupCallChainBlocksRequest-resolve}{\code{GetGroupCallChainBlocksRequest$resolve()}}
\item \href{#method-GetGroupCallChainBlocksRequest-to_list}{\code{GetGroupCallChainBlocksRequest$to_list()}}
\item \href{#method-GetGroupCallChainBlocksRequest-bytes}{\code{GetGroupCallChainBlocksRequest$bytes()}}
\item \href{#method-GetGroupCallChainBlocksRequest-from_reader}{\code{GetGroupCallChainBlocksRequest$from_reader()}}
\item \href{#method-GetGroupCallChainBlocksRequest-clone}{\code{GetGroupCallChainBlocksRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallChainBlocksRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallChainBlocksRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallChainBlocksRequest$new(call, sub_chain_id, offset, limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}

\item{\code{sub_chain_id}}{Integer sub chain id.}

\item{\code{offset}}{Integer offset.}

\item{\code{limit}}{Integer limit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Resolve input references

Typically used to convert high-level entities into input objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallChainBlocksRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallChainBlocksRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallChainBlocksRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Convert object to a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallChainBlocksRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallChainBlocksRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallChainBlocksRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallChainBlocksRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallChainBlocksRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallChainBlocksRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallChainBlocksRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallChainBlocksRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallChainBlocksRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallChainBlocksRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallChainBlocksRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallChainBlocksRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
