% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{GetGroupCallJoinAsRequest}
\alias{GetGroupCallJoinAsRequest}
\title{GetGroupCallJoinAsRequest R6 class}
\description{
Represents GetGroupCallJoinAsRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{peer}}{The input peer object (expected to implement bytes()).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetGroupCallJoinAsRequest-new}{\code{GetGroupCallJoinAsRequest$new()}}
\item \href{#method-GetGroupCallJoinAsRequest-resolve}{\code{GetGroupCallJoinAsRequest$resolve()}}
\item \href{#method-GetGroupCallJoinAsRequest-to_list}{\code{GetGroupCallJoinAsRequest$to_list()}}
\item \href{#method-GetGroupCallJoinAsRequest-bytes}{\code{GetGroupCallJoinAsRequest$bytes()}}
\item \href{#method-GetGroupCallJoinAsRequest-from_reader}{\code{GetGroupCallJoinAsRequest$from_reader()}}
\item \href{#method-GetGroupCallJoinAsRequest-clone}{\code{GetGroupCallJoinAsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallJoinAsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallJoinAsRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallJoinAsRequest$new(peer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{Input peer object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Resolve input references

Typically used to convert high-level entities into input objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallJoinAsRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallJoinAsRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallJoinAsRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_peer method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Convert object to a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallJoinAsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallJoinAsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallJoinAsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallJoinAsRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallJoinAsRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallJoinAsRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallJoinAsRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallJoinAsRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallJoinAsRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallJoinAsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallJoinAsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallJoinAsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
