% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{GetGroupCallRequest}
\alias{GetGroupCallRequest}
\title{GetGroupCallRequest R6 class}
\description{
GetGroupCallRequest R6 class

GetGroupCallRequest R6 class
}
\details{
Represents GetGroupCallRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{The input group call object (expected to implement bytes()).}

\item{\code{limit}}{Integer limit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetGroupCallRequest-new}{\code{GetGroupCallRequest$new()}}
\item \href{#method-GetGroupCallRequest-resolve}{\code{GetGroupCallRequest$resolve()}}
\item \href{#method-GetGroupCallRequest-to_list}{\code{GetGroupCallRequest$to_list()}}
\item \href{#method-GetGroupCallRequest-bytes}{\code{GetGroupCallRequest$bytes()}}
\item \href{#method-GetGroupCallRequest-from_reader}{\code{GetGroupCallRequest$from_reader()}}
\item \href{#method-GetGroupCallRequest-clone}{\code{GetGroupCallRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a GetGroupCallRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallRequest$new(call, limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}

\item{\code{limit}}{Integer limit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references

Typically used to convert high-level entities into input objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
