% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{GetGroupCallStreamChannelsRequest}
\alias{GetGroupCallStreamChannelsRequest}
\title{GetGroupCallStreamChannelsRequest}
\description{
Telegram API type GetGroupCallStreamChannelsRequest
}
\details{
GetGroupCallStreamChannelsRequest R6 class

Represents GetGroupCallStreamChannelsRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetGroupCallStreamChannelsRequest-new}{\code{GetGroupCallStreamChannelsRequest$new()}}
\item \href{#method-GetGroupCallStreamChannelsRequest-resolve}{\code{GetGroupCallStreamChannelsRequest$resolve()}}
\item \href{#method-GetGroupCallStreamChannelsRequest-to_list}{\code{GetGroupCallStreamChannelsRequest$to_list()}}
\item \href{#method-GetGroupCallStreamChannelsRequest-bytes}{\code{GetGroupCallStreamChannelsRequest$bytes()}}
\item \href{#method-GetGroupCallStreamChannelsRequest-from_reader}{\code{GetGroupCallStreamChannelsRequest$from_reader()}}
\item \href{#method-GetGroupCallStreamChannelsRequest-clone}{\code{GetGroupCallStreamChannelsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamChannelsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamChannelsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a GetGroupCallStreamChannelsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamChannelsRequest$new(call)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamChannelsRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamChannelsRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references

Typically used to convert high-level entities into input objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamChannelsRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamChannelsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamChannelsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamChannelsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamChannelsRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamChannelsRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamChannelsRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamChannelsRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamChannelsRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamChannelsRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamChannelsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamChannelsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamChannelsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
