% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{GetGroupCallStreamRtmpUrlRequest}
\alias{GetGroupCallStreamRtmpUrlRequest}
\title{GetGroupCallStreamRtmpUrlRequest R6 class}
\description{
Represents GetGroupCallStreamRtmpUrlRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{peer}}{The input peer object (expected to implement bytes()).}

\item{\code{revoke}}{Logical indicating revoke state.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetGroupCallStreamRtmpUrlRequest-new}{\code{GetGroupCallStreamRtmpUrlRequest$new()}}
\item \href{#method-GetGroupCallStreamRtmpUrlRequest-resolve}{\code{GetGroupCallStreamRtmpUrlRequest$resolve()}}
\item \href{#method-GetGroupCallStreamRtmpUrlRequest-to_list}{\code{GetGroupCallStreamRtmpUrlRequest$to_list()}}
\item \href{#method-GetGroupCallStreamRtmpUrlRequest-bytes}{\code{GetGroupCallStreamRtmpUrlRequest$bytes()}}
\item \href{#method-GetGroupCallStreamRtmpUrlRequest-from_reader}{\code{GetGroupCallStreamRtmpUrlRequest$from_reader()}}
\item \href{#method-GetGroupCallStreamRtmpUrlRequest-clone}{\code{GetGroupCallStreamRtmpUrlRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamRtmpUrlRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamRtmpUrlRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamRtmpUrlRequest$new(peer, revoke)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{Input peer object.}

\item{\code{revoke}}{Logical.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Resolve input references

Typically used to convert high-level entities into input objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamRtmpUrlRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamRtmpUrlRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamRtmpUrlRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_peer method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Convert object to a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamRtmpUrlRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamRtmpUrlRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamRtmpUrlRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamRtmpUrlRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamRtmpUrlRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamRtmpUrlRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamRtmpUrlRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamRtmpUrlRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamRtmpUrlRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupCallStreamRtmpUrlRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupCallStreamRtmpUrlRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupCallStreamRtmpUrlRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
