% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{GetGroupParticipantsRequest}
\alias{GetGroupParticipantsRequest}
\title{GetGroupParticipantsRequest R6 class}
\description{
GetGroupParticipantsRequest R6 class

GetGroupParticipantsRequest R6 class
}
\details{
Represents GetGroupParticipantsRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{The input group call object (expected to implement bytes()).}

\item{\code{ids}}{List of input peer objects (expected to implement bytes()).}

\item{\code{sources}}{List of integers.}

\item{\code{offset}}{String offset.}

\item{\code{limit}}{Integer limit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetGroupParticipantsRequest-new}{\code{GetGroupParticipantsRequest$new()}}
\item \href{#method-GetGroupParticipantsRequest-resolve}{\code{GetGroupParticipantsRequest$resolve()}}
\item \href{#method-GetGroupParticipantsRequest-to_list}{\code{GetGroupParticipantsRequest$to_list()}}
\item \href{#method-GetGroupParticipantsRequest-bytes}{\code{GetGroupParticipantsRequest$bytes()}}
\item \href{#method-GetGroupParticipantsRequest-from_reader}{\code{GetGroupParticipantsRequest$from_reader()}}
\item \href{#method-GetGroupParticipantsRequest-clone}{\code{GetGroupParticipantsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupParticipantsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupParticipantsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a GetGroupParticipantsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupParticipantsRequest$new(call, ids, sources, offset, limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}

\item{\code{ids}}{List of input peer objects.}

\item{\code{sources}}{List of integers.}

\item{\code{offset}}{String.}

\item{\code{limit}}{Integer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupParticipantsRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupParticipantsRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references

Typically used to convert high-level entities into input objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupParticipantsRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call and get_input_peer methods.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupParticipantsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupParticipantsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupParticipantsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupParticipantsRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupParticipantsRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupParticipantsRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupParticipantsRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupParticipantsRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupParticipantsRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetGroupParticipantsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetGroupParticipantsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetGroupParticipantsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
