% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetInviteTextRequest}
\alias{GetInviteTextRequest}
\title{GetInviteTextRequest}
\description{
Telegram API type GetInviteTextRequest
}
\details{
GetInviteTextRequest R6 class

Request to get invite text.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{Field.
Serialize to bytes (raw vector)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{Field.
Serialize to bytes (raw vector)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetInviteTextRequest-new}{\code{GetInviteTextRequest$new()}}
\item \href{#method-GetInviteTextRequest-to_list}{\code{GetInviteTextRequest$to_list()}}
\item \href{#method-GetInviteTextRequest-to_bytes}{\code{GetInviteTextRequest$to_bytes()}}
\item \href{#method-GetInviteTextRequest-clone}{\code{GetInviteTextRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetInviteTextRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetInviteTextRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetInviteTextRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetInviteTextRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible NULL
Convert to list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetInviteTextRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetInviteTextRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetInviteTextRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with type discriminator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetInviteTextRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetInviteTextRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetInviteTextRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetInviteTextRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetInviteTextRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetInviteTextRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
