% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chatlists.R
\name{GetLeaveChatlistSuggestionsRequest}
\alias{GetLeaveChatlistSuggestionsRequest}
\title{GetLeaveChatlistSuggestionsRequest R6 class}
\description{
Request to get suggestions of peers to leave from a chatlist.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Integer constructor id.}

\item{\code{SUBCLASS_OF_ID}}{Integer subclass id.}

\item{\code{chatlist}}{TypeInputChatlist input chatlist object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetLeaveChatlistSuggestionsRequest-new}{\code{GetLeaveChatlistSuggestionsRequest$new()}}
\item \href{#method-GetLeaveChatlistSuggestionsRequest-to_list}{\code{GetLeaveChatlistSuggestionsRequest$to_list()}}
\item \href{#method-GetLeaveChatlistSuggestionsRequest-bytes}{\code{GetLeaveChatlistSuggestionsRequest$bytes()}}
\item \href{#method-GetLeaveChatlistSuggestionsRequest-from_reader}{\code{GetLeaveChatlistSuggestionsRequest$from_reader()}}
\item \href{#method-GetLeaveChatlistSuggestionsRequest-clone}{\code{GetLeaveChatlistSuggestionsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetLeaveChatlistSuggestionsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetLeaveChatlistSuggestionsRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetLeaveChatlistSuggestionsRequest$new(chatlist = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chatlist}}{Input chatlist object
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetLeaveChatlistSuggestionsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetLeaveChatlistSuggestionsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetLeaveChatlistSuggestionsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representation
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetLeaveChatlistSuggestionsRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetLeaveChatlistSuggestionsRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetLeaveChatlistSuggestionsRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Serializes the constructor id followed by chatlist bytes.
}

\subsection{Returns}{
raw vector with serialized bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetLeaveChatlistSuggestionsRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetLeaveChatlistSuggestionsRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetLeaveChatlistSuggestionsRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetLeaveChatlistSuggestionsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetLeaveChatlistSuggestionsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetLeaveChatlistSuggestionsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
