% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetNearestDcRequest}
\alias{GetNearestDcRequest}
\title{GetNearestDcRequest R6 class}
\description{
GetNearestDcRequest R6 class

GetNearestDcRequest R6 class
}
\details{
Request to get the nearest data-center info.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetNearestDcRequest-new}{\code{GetNearestDcRequest$new()}}
\item \href{#method-GetNearestDcRequest-to_list}{\code{GetNearestDcRequest$to_list()}}
\item \href{#method-GetNearestDcRequest-to_bytes}{\code{GetNearestDcRequest$to_bytes()}}
\item \href{#method-GetNearestDcRequest-clone}{\code{GetNearestDcRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNearestDcRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetNearestDcRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetNearestDcRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetNearestDcRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible NULL
Convert to list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNearestDcRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetNearestDcRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetNearestDcRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with type discriminator
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNearestDcRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetNearestDcRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetNearestDcRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNearestDcRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetNearestDcRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetNearestDcRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
