% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetPassportConfigRequest}
\alias{GetPassportConfigRequest}
\title{GetPassportConfigRequest}
\description{
Telegram API type GetPassportConfigRequest
}
\details{
GetPassportConfigRequest R6 class

Request to get passport configuration (possibly not modified).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{hash}}{integer Hash to check for modifications.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetPassportConfigRequest-new}{\code{GetPassportConfigRequest$new()}}
\item \href{#method-GetPassportConfigRequest-to_list}{\code{GetPassportConfigRequest$to_list()}}
\item \href{#method-GetPassportConfigRequest-to_bytes}{\code{GetPassportConfigRequest$to_bytes()}}
\item \href{#method-GetPassportConfigRequest-clone}{\code{GetPassportConfigRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPassportConfigRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetPassportConfigRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetPassportConfigRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPassportConfigRequest$new(hash = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{integer
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPassportConfigRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetPassportConfigRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPassportConfigRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with type discriminator and hash
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPassportConfigRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetPassportConfigRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPassportConfigRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPassportConfigRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetPassportConfigRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPassportConfigRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
