% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetPeerColorsRequest}
\alias{GetPeerColorsRequest}
\title{GetPeerColorsRequest}
\description{
Telegram API type GetPeerColorsRequest
}
\details{
GetPeerColorsRequest R6 class

Request to get peer colors (possibly not modified).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{hash}}{integer Hash to check for modifications.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetPeerColorsRequest-new}{\code{GetPeerColorsRequest$new()}}
\item \href{#method-GetPeerColorsRequest-to_list}{\code{GetPeerColorsRequest$to_list()}}
\item \href{#method-GetPeerColorsRequest-to_bytes}{\code{GetPeerColorsRequest$to_bytes()}}
\item \href{#method-GetPeerColorsRequest-clone}{\code{GetPeerColorsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPeerColorsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetPeerColorsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetPeerColorsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPeerColorsRequest$new(hash = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{integer
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPeerColorsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetPeerColorsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPeerColorsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `type` discriminator and hash
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPeerColorsRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetPeerColorsRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPeerColorsRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPeerColorsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetPeerColorsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPeerColorsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
