% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetPeerProfileColorsRequest}
\alias{GetPeerProfileColorsRequest}
\title{GetPeerProfileColorsRequest R6 class}
\description{
GetPeerProfileColorsRequest R6 class

GetPeerProfileColorsRequest R6 class
}
\details{
Request to get peer profile colors (possibly not modified).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{hash}}{integer Hash to check for modifications.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetPeerProfileColorsRequest-new}{\code{GetPeerProfileColorsRequest$new()}}
\item \href{#method-GetPeerProfileColorsRequest-to_list}{\code{GetPeerProfileColorsRequest$to_list()}}
\item \href{#method-GetPeerProfileColorsRequest-to_bytes}{\code{GetPeerProfileColorsRequest$to_bytes()}}
\item \href{#method-GetPeerProfileColorsRequest-clone}{\code{GetPeerProfileColorsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPeerProfileColorsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetPeerProfileColorsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetPeerProfileColorsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPeerProfileColorsRequest$new(hash = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{integer
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPeerProfileColorsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetPeerProfileColorsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPeerProfileColorsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `type` discriminator and hash
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPeerProfileColorsRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetPeerProfileColorsRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPeerProfileColorsRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPeerProfileColorsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetPeerProfileColorsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPeerProfileColorsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
