% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetPremiumPromoRequest}
\alias{GetPremiumPromoRequest}
\title{GetPremiumPromoRequest}
\description{
Telegram API type GetPremiumPromoRequest
}
\details{
GetPremiumPromoRequest R6 class

Request to get premium promotion info.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{Field.
Serialize to bytes (raw vector)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{Field.
Serialize to bytes (raw vector)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetPremiumPromoRequest-new}{\code{GetPremiumPromoRequest$new()}}
\item \href{#method-GetPremiumPromoRequest-to_list}{\code{GetPremiumPromoRequest$to_list()}}
\item \href{#method-GetPremiumPromoRequest-to_bytes}{\code{GetPremiumPromoRequest$to_bytes()}}
\item \href{#method-GetPremiumPromoRequest-clone}{\code{GetPremiumPromoRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPremiumPromoRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetPremiumPromoRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetPremiumPromoRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPremiumPromoRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible NULL
Convert to list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPremiumPromoRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetPremiumPromoRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPremiumPromoRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with type discriminator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPremiumPromoRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetPremiumPromoRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPremiumPromoRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPremiumPromoRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetPremiumPromoRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPremiumPromoRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
