% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{GetPreviewInfoRequest}
\alias{GetPreviewInfoRequest}
\title{GetPreviewInfoRequest}
\description{
Telegram API type GetPreviewInfoRequest
}
\details{
GetPreviewInfoRequest R6 class

Represents the GetPreviewInfoRequest TL request.



Each method is documented inline below.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bot}}{Field.}

\item{\code{lang_code}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetPreviewInfoRequest-new}{\code{GetPreviewInfoRequest$new()}}
\item \href{#method-GetPreviewInfoRequest-resolve}{\code{GetPreviewInfoRequest$resolve()}}
\item \href{#method-GetPreviewInfoRequest-to_list}{\code{GetPreviewInfoRequest$to_list()}}
\item \href{#method-GetPreviewInfoRequest-to_bytes}{\code{GetPreviewInfoRequest$to_bytes()}}
\item \href{#method-GetPreviewInfoRequest-clone}{\code{GetPreviewInfoRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPreviewInfoRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetPreviewInfoRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetPreviewInfoRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPreviewInfoRequest$new(bot, lang_code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bot}}{TypeInputUser or identifier}

\item{\code{lang_code}}{character
Resolve references (convert entities via client/utils)

Resolves bot using client$get_input_entity and utils$get_input_user.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPreviewInfoRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetPreviewInfoRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPreviewInfoRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client with get_input_entity method}

\item{\code{utils}}{utils with get_input_user}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
Convert to list (dictionary-like)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPreviewInfoRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetPreviewInfoRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPreviewInfoRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPreviewInfoRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetPreviewInfoRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPreviewInfoRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPreviewInfoRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetPreviewInfoRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPreviewInfoRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
