% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetPromoDataRequest}
\alias{GetPromoDataRequest}
\title{GetPromoDataRequest R6 class}
\description{
Request to get promo data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetPromoDataRequest-new}{\code{GetPromoDataRequest$new()}}
\item \href{#method-GetPromoDataRequest-to_list}{\code{GetPromoDataRequest$to_list()}}
\item \href{#method-GetPromoDataRequest-to_bytes}{\code{GetPromoDataRequest$to_bytes()}}
\item \href{#method-GetPromoDataRequest-clone}{\code{GetPromoDataRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPromoDataRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetPromoDataRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPromoDataRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible NULL
Convert to list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPromoDataRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetPromoDataRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPromoDataRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with type discriminator
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPromoDataRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetPromoDataRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPromoDataRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetPromoDataRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetPromoDataRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetPromoDataRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
