% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetRecentMeUrlsRequest}
\alias{GetRecentMeUrlsRequest}
\title{GetRecentMeUrlsRequest R6 class}
\description{
Request to get recent "me" URLs with a referer string.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{referer}}{character Referer string}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetRecentMeUrlsRequest-new}{\code{GetRecentMeUrlsRequest$new()}}
\item \href{#method-GetRecentMeUrlsRequest-to_list}{\code{GetRecentMeUrlsRequest$to_list()}}
\item \href{#method-GetRecentMeUrlsRequest-to_bytes}{\code{GetRecentMeUrlsRequest$to_bytes()}}
\item \href{#method-GetRecentMeUrlsRequest-clone}{\code{GetRecentMeUrlsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecentMeUrlsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecentMeUrlsRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecentMeUrlsRequest$new(referer = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{referer}}{character
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecentMeUrlsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecentMeUrlsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecentMeUrlsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with type discriminator and referer
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecentMeUrlsRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecentMeUrlsRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecentMeUrlsRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRecentMeUrlsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetRecentMeUrlsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetRecentMeUrlsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
