% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_users.R
\name{GetSavedMusicRequest}
\alias{GetSavedMusicRequest}
\title{GetSavedMusicRequest}
\description{
Request to get saved music for a user.
}
\details{
Fields:
- id: TLInputUser or a representation accepted by utils$get_input_user
- offset: integer
- limit: integer
- hash: integer64 (may be represented as numeric in R)

Returns: users.SavedMusic (either SavedMusicNotModified or SavedMusic)
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{GetSavedMusicRequest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetSavedMusicRequest-new}{\code{GetSavedMusicRequest$new()}}
\item \href{#method-GetSavedMusicRequest-resolve}{\code{GetSavedMusicRequest$resolve()}}
\item \href{#method-GetSavedMusicRequest-to_dict}{\code{GetSavedMusicRequest$to_dict()}}
\item \href{#method-GetSavedMusicRequest-.bytes}{\code{GetSavedMusicRequest$.bytes()}}
\item \href{#method-GetSavedMusicRequest-from_reader}{\code{GetSavedMusicRequest$from_reader()}}
\item \href{#method-GetSavedMusicRequest-clone}{\code{GetSavedMusicRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicRequest$new(id = NULL, offset = NULL, limit = NULL, hash = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicRequest-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicRequest-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicRequest$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicRequest-.bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicRequest-.bytes}{}}}
\subsection{Method \code{.bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicRequest$.bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
