% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_smsjobs.R
\name{GetSmsJobRequest}
\alias{GetSmsJobRequest}
\title{GetSmsJobRequest}
\description{
Telegram API type GetSmsJobRequest
}
\details{
GetSmsJobRequest R6 class

Representation of the GetSmsJobRequest TLRequest.

This calls a helper serialize_bytes() to encode the job_id.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{job_id}}{Field.}

\item{\code{lock_class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{lock_class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetSmsJobRequest-new}{\code{GetSmsJobRequest$new()}}
\item \href{#method-GetSmsJobRequest-to_list}{\code{GetSmsJobRequest$to_list()}}
\item \href{#method-GetSmsJobRequest-bytes}{\code{GetSmsJobRequest$bytes()}}
\item \href{#method-GetSmsJobRequest-from_reader}{\code{GetSmsJobRequest$from_reader()}}
\item \href{#method-GetSmsJobRequest-clone}{\code{GetSmsJobRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSmsJobRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetSmsJobRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a GetSmsJobRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSmsJobRequest$new(job_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job_id}}{character}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSmsJobRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetSmsJobRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSmsJobRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSmsJobRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetSmsJobRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Get raw bytes for this constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSmsJobRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSmsJobRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetSmsJobRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSmsJobRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSmsJobRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetSmsJobRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSmsJobRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
