% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_payments.R
\name{GetStarsRevenueWithdrawalUrlRequest}
\alias{GetStarsRevenueWithdrawalUrlRequest}
\title{GetStarsRevenueWithdrawalUrlRequest}
\description{
Telegram API type GetStarsRevenueWithdrawalUrlRequest
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{GetStarsRevenueWithdrawalUrlRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{peer}}{Field.}

\item{\code{password}}{Field.}

\item{\code{ton}}{Field.}

\item{\code{amount}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetStarsRevenueWithdrawalUrlRequest-new}{\code{GetStarsRevenueWithdrawalUrlRequest$new()}}
\item \href{#method-GetStarsRevenueWithdrawalUrlRequest-resolve}{\code{GetStarsRevenueWithdrawalUrlRequest$resolve()}}
\item \href{#method-GetStarsRevenueWithdrawalUrlRequest-to_dict}{\code{GetStarsRevenueWithdrawalUrlRequest$to_dict()}}
\item \href{#method-GetStarsRevenueWithdrawalUrlRequest-bytes}{\code{GetStarsRevenueWithdrawalUrlRequest$bytes()}}
\item \href{#method-GetStarsRevenueWithdrawalUrlRequest-from_reader}{\code{GetStarsRevenueWithdrawalUrlRequest$from_reader()}}
\item \href{#method-GetStarsRevenueWithdrawalUrlRequest-clone}{\code{GetStarsRevenueWithdrawalUrlRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStarsRevenueWithdrawalUrlRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetStarsRevenueWithdrawalUrlRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the GetStarsRevenueWithdrawalUrlRequest object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStarsRevenueWithdrawalUrlRequest$new(
  peer,
  password,
  ton = NULL,
  amount = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{The input peer (TypeInputPeer).}

\item{\code{password}}{The input check password SRP (TypeInputCheckPasswordSRP).}

\item{\code{ton}}{Optional boolean flag for TON.}

\item{\code{amount}}{Optional amount (integer).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStarsRevenueWithdrawalUrlRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetStarsRevenueWithdrawalUrlRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve the peer using client and utils.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStarsRevenueWithdrawalUrlRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The client object.}

\item{\code{utils}}{The utils object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStarsRevenueWithdrawalUrlRequest-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-GetStarsRevenueWithdrawalUrlRequest-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Convert the object to a dictionary-like list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStarsRevenueWithdrawalUrlRequest$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object in dictionary form.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStarsRevenueWithdrawalUrlRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetStarsRevenueWithdrawalUrlRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize the object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStarsRevenueWithdrawalUrlRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector of bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStarsRevenueWithdrawalUrlRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetStarsRevenueWithdrawalUrlRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Deserialize from a reader.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStarsRevenueWithdrawalUrlRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{The reader object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of GetStarsRevenueWithdrawalUrlRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStarsRevenueWithdrawalUrlRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetStarsRevenueWithdrawalUrlRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStarsRevenueWithdrawalUrlRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
