% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_updates.R
\name{GetStateRequest}
\alias{GetStateRequest}
\title{GetStateRequest}
\description{
Telegram API type GetStateRequest
}
\details{
GetStateRequest R6 class

Representation of updates.GetState request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetStateRequest-new}{\code{GetStateRequest$new()}}
\item \href{#method-GetStateRequest-to_list}{\code{GetStateRequest$to_list()}}
\item \href{#method-GetStateRequest-bytes}{\code{GetStateRequest$bytes()}}
\item \href{#method-GetStateRequest-from_reader}{\code{GetStateRequest$from_reader()}}
\item \href{#method-GetStateRequest-clone}{\code{GetStateRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStateRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetStateRequest-new}{}}}
\subsection{Method \code{new()}}{
Converts the request to a list representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStateRequest$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStateRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetStateRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Serializes the request to raw bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStateRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStateRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetStateRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Initializes a new GetStateRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStateRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new GetStateRequest instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStateRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetStateRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStateRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStateRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetStateRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStateRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
