% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_smsjobs.R
\name{GetStatusRequest}
\alias{GetStatusRequest}
\title{GetStatusRequest R6 class}
\description{
GetStatusRequest R6 class

GetStatusRequest R6 class
}
\details{
Representation of the GetStatusRequest TLRequest.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetStatusRequest-new}{\code{GetStatusRequest$new()}}
\item \href{#method-GetStatusRequest-to_list}{\code{GetStatusRequest$to_list()}}
\item \href{#method-GetStatusRequest-bytes}{\code{GetStatusRequest$bytes()}}
\item \href{#method-GetStatusRequest-from_reader}{\code{GetStatusRequest$from_reader()}}
\item \href{#method-GetStatusRequest-clone}{\code{GetStatusRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStatusRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetStatusRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a GetStatusRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStatusRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetStatusRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStatusRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetStatusRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Get raw bytes for this constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStatusRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetStatusRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStatusRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetStatusRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
