% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stories.R
\name{GetStoryReactionsListRequest}
\alias{GetStoryReactionsListRequest}
\title{GetStoryReactionsListRequest R6 class}
\description{
GetStoryReactionsListRequest R6 class

GetStoryReactionsListRequest R6 class
}
\details{
Request to get a list of story reactions with optional filters.
Returns stories.StoryReactionsList.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{GetStoryReactionsListRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{peer}}{Field.}

\item{\code{id}}{Field.}

\item{\code{limit}}{Field.}

\item{\code{forwards_first}}{Field.}

\item{\code{reaction}}{Field.}

\item{\code{offset}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetStoryReactionsListRequest-new}{\code{GetStoryReactionsListRequest$new()}}
\item \href{#method-GetStoryReactionsListRequest-resolve}{\code{GetStoryReactionsListRequest$resolve()}}
\item \href{#method-GetStoryReactionsListRequest-to_list}{\code{GetStoryReactionsListRequest$to_list()}}
\item \href{#method-GetStoryReactionsListRequest-to_bytes}{\code{GetStoryReactionsListRequest$to_bytes()}}
\item \href{#method-GetStoryReactionsListRequest-clone}{\code{GetStoryReactionsListRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStoryReactionsListRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetStoryReactionsListRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetStoryReactionsListRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStoryReactionsListRequest$new(
  peer,
  id,
  limit,
  forwards_first = NULL,
  reaction = NULL,
  offset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{TypeInputPeer}

\item{\code{id}}{integer story id}

\item{\code{limit}}{integer limit}

\item{\code{forwards_first}}{logical or NULL}

\item{\code{reaction}}{TypeReaction or NULL}

\item{\code{offset}}{character or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStoryReactionsListRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetStoryReactionsListRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve peer references

Convert high-level peer reference to input peer using client/utils.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStoryReactionsListRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client with get_input_entity}

\item{\code{utils}}{utils with get_input_peer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStoryReactionsListRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetStoryReactionsListRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStoryReactionsListRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStoryReactionsListRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetStoryReactionsListRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to raw TL bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStoryReactionsListRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStoryReactionsListRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetStoryReactionsListRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStoryReactionsListRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
