% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetSupportNameRequest}
\alias{GetSupportNameRequest}
\title{GetSupportNameRequest R6 class}
\description{
GetSupportNameRequest R6 class

GetSupportNameRequest R6 class
}
\details{
Request to get support name.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetSupportNameRequest-new}{\code{GetSupportNameRequest$new()}}
\item \href{#method-GetSupportNameRequest-to_list}{\code{GetSupportNameRequest$to_list()}}
\item \href{#method-GetSupportNameRequest-to_bytes}{\code{GetSupportNameRequest$to_bytes()}}
\item \href{#method-GetSupportNameRequest-clone}{\code{GetSupportNameRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSupportNameRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetSupportNameRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetSupportNameRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSupportNameRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible NULL
Convert to list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSupportNameRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetSupportNameRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSupportNameRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `type` discriminator
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSupportNameRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetSupportNameRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSupportNameRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSupportNameRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetSupportNameRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSupportNameRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
