% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetTimezonesListRequest}
\alias{GetTimezonesListRequest}
\title{GetTimezonesListRequest}
\description{
Telegram API type GetTimezonesListRequest
}
\details{
GetTimezonesListRequest R6 class

Request to get timezones list (possibly not modified).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{Field.
Serialize to bytes (raw vector)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{Field.
Serialize to bytes (raw vector)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetTimezonesListRequest-new}{\code{GetTimezonesListRequest$new()}}
\item \href{#method-GetTimezonesListRequest-to_list}{\code{GetTimezonesListRequest$to_list()}}
\item \href{#method-GetTimezonesListRequest-to_bytes}{\code{GetTimezonesListRequest$to_bytes()}}
\item \href{#method-GetTimezonesListRequest-clone}{\code{GetTimezonesListRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTimezonesListRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetTimezonesListRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetTimezonesListRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetTimezonesListRequest$new(hash = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{integer
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTimezonesListRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetTimezonesListRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetTimezonesListRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `type` discriminator and hash
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTimezonesListRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetTimezonesListRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetTimezonesListRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTimezonesListRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetTimezonesListRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetTimezonesListRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
