% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_contacts.R
\name{GetTopPeersRequest}
\alias{GetTopPeersRequest}
\title{GetTopPeersRequest}
\description{
Methods:
- initialize(offset, limit, hash, ...): create new request
- to_list(): return a list representation suitable for JSON/dumping
- to_bytes(): return raw vector of bytes for the TL request
- from_reader(reader): read from reader and return a new instance
}
\details{
R6 representation of the TL request: GetTopPeersRequest
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{GetTopPeersRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{offset}}{Field.}

\item{\code{limit}}{Field.}

\item{\code{hash}}{Field.}

\item{\code{correspondents}}{Field.}

\item{\code{bots_pm}}{Field.}

\item{\code{bots_inline}}{Field.}

\item{\code{phone_calls}}{Field.}

\item{\code{forward_users}}{Field.}

\item{\code{forward_chats}}{Field.}

\item{\code{groups}}{Field.}

\item{\code{channels}}{Field.}

\item{\code{bots_app}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetTopPeersRequest-new}{\code{GetTopPeersRequest$new()}}
\item \href{#method-GetTopPeersRequest-to_list}{\code{GetTopPeersRequest$to_list()}}
\item \href{#method-GetTopPeersRequest-to_bytes}{\code{GetTopPeersRequest$to_bytes()}}
\item \href{#method-GetTopPeersRequest-clone}{\code{GetTopPeersRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="resolve"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-resolve'><code>telegramR::TLRequest$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTopPeersRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetTopPeersRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetTopPeersRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetTopPeersRequest$new(
  offset,
  limit,
  hash,
  correspondents = NULL,
  bots_pm = NULL,
  bots_inline = NULL,
  phone_calls = NULL,
  forward_users = NULL,
  forward_chats = NULL,
  groups = NULL,
  channels = NULL,
  bots_app = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{offset}}{integer}

\item{\code{limit}}{integer}

\item{\code{hash}}{numeric/integer (64-bit)}

\item{\code{correspondents}}{logical|NULL}

\item{\code{bots_pm}}{logical|NULL}

\item{\code{bots_inline}}{logical|NULL}

\item{\code{phone_calls}}{logical|NULL}

\item{\code{forward_users}}{logical|NULL}

\item{\code{forward_chats}}{logical|NULL}

\item{\code{groups}}{logical|NULL}

\item{\code{channels}}{logical|NULL}

\item{\code{bots_app}}{logical|NULL
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTopPeersRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetTopPeersRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetTopPeersRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTopPeersRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetTopPeersRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetTopPeersRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
Convert integer to little-endian raw vector of given size (bytes)
Convert 64-bit integer (numeric) to little-endian raw vector (8 bytes)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTopPeersRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetTopPeersRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetTopPeersRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
