% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_payments.R
\name{GetUniqueStarGiftRequest}
\alias{GetUniqueStarGiftRequest}
\title{GetUniqueStarGiftRequest}
\description{
Represents a request to get a unique star gift. Inherits from TLRequest.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{GetUniqueStarGiftRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{slug}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetUniqueStarGiftRequest-new}{\code{GetUniqueStarGiftRequest$new()}}
\item \href{#method-GetUniqueStarGiftRequest-to_dict}{\code{GetUniqueStarGiftRequest$to_dict()}}
\item \href{#method-GetUniqueStarGiftRequest-bytes}{\code{GetUniqueStarGiftRequest$bytes()}}
\item \href{#method-GetUniqueStarGiftRequest-from_reader}{\code{GetUniqueStarGiftRequest$from_reader()}}
\item \href{#method-GetUniqueStarGiftRequest-clone}{\code{GetUniqueStarGiftRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="resolve"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-resolve'><code>telegramR::TLRequest$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUniqueStarGiftRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetUniqueStarGiftRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the GetUniqueStarGiftRequest object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUniqueStarGiftRequest$new(slug)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{slug}}{The slug string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUniqueStarGiftRequest-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-GetUniqueStarGiftRequest-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Convert the object to a dictionary-like list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUniqueStarGiftRequest$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object in dictionary form.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUniqueStarGiftRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetUniqueStarGiftRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize the object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUniqueStarGiftRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector of bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUniqueStarGiftRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetUniqueStarGiftRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Deserialize from a reader.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUniqueStarGiftRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{The reader object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of GetUniqueStarGiftRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUniqueStarGiftRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetUniqueStarGiftRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUniqueStarGiftRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
