% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpdates.R
\name{getUpdates}
\alias{getUpdates}
\title{getUpdates}
\usage{
getUpdates(token, lastUpdateID)
}
\arguments{
\item{token}{Your bot's API token.}

\item{last_update_id}{The last message id received. Prevents huge queries to Telegram. Only get new stuff.}
}
\description{
Gets updates from Telegram bot and parses json response.
}
\examples{
\dontrun{
results <- getUpdates(token = "XXXXXXX")
}
}
