% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_premium.R
\name{GetUserBoostsRequest}
\alias{GetUserBoostsRequest}
\title{GetUserBoostsRequest R6 class}
\description{
GetUserBoostsRequest R6 class

GetUserBoostsRequest R6 class
}
\details{
Representation of the TL request "GetUserBoostsRequest".
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor ID for the request}

\item{\code{SUBCLASS_OF_ID}}{Subclass ID for the request}

\item{\code{peer}}{Input peer (object or identifier)}

\item{\code{user_id}}{Input user (object or identifier)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetUserBoostsRequest-new}{\code{GetUserBoostsRequest$new()}}
\item \href{#method-GetUserBoostsRequest-resolve}{\code{GetUserBoostsRequest$resolve()}}
\item \href{#method-GetUserBoostsRequest-to_list}{\code{GetUserBoostsRequest$to_list()}}
\item \href{#method-GetUserBoostsRequest-bytes}{\code{GetUserBoostsRequest$bytes()}}
\item \href{#method-GetUserBoostsRequest-from_reader}{\code{GetUserBoostsRequest$from_reader()}}
\item \href{#method-GetUserBoostsRequest-clone}{\code{GetUserBoostsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserBoostsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserBoostsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetUserBoostsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserBoostsRequest$new(peer = NULL, user_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{input peer or identifier}

\item{\code{user_id}}{input user or identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserBoostsRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserBoostsRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve peer and user_id using client/utils helpers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserBoostsRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client object that provides get_input_entity}

\item{\code{utils}}{utils object that provides get_input_peer and get_input_user}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserBoostsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserBoostsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert request to list (dictionary)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserBoostsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserBoostsRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserBoostsRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize to raw bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserBoostsRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserBoostsRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserBoostsRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Construct a new GetUserBoostsRequest from a reader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserBoostsRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{reader object exposing tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GetUserBoostsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserBoostsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserBoostsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserBoostsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
