% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetUserInfoRequest}
\alias{GetUserInfoRequest}
\title{GetUserInfoRequest}
\description{
Telegram API type GetUserInfoRequest
}
\details{
GetUserInfoRequest R6 class

Request to get user info (help.UserInfo).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{userId}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetUserInfoRequest-new}{\code{GetUserInfoRequest$new()}}
\item \href{#method-GetUserInfoRequest-resolve}{\code{GetUserInfoRequest$resolve()}}
\item \href{#method-GetUserInfoRequest-to_list}{\code{GetUserInfoRequest$to_list()}}
\item \href{#method-GetUserInfoRequest-to_bytes}{\code{GetUserInfoRequest$to_bytes()}}
\item \href{#method-GetUserInfoRequest-clone}{\code{GetUserInfoRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserInfoRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserInfoRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetUserInfoRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserInfoRequest$new(userId = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userId}}{TLObject-like or raw
Resolve userId using client and utils}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserInfoRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserInfoRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserInfoRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client object providing get_input_entity()}

\item{\code{utils}}{utils providing get_input_user()
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserInfoRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserInfoRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserInfoRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `_` discriminator and user_id as list (if possible)
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserInfoRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserInfoRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserInfoRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserInfoRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserInfoRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUserInfoRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
