% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{GetWebFileRequest}
\alias{GetWebFileRequest}
\title{GetWebFileRequest R6 class}
\description{
Represents the TL request upload.GetWebFileRequest.
}
\details{
Fields:
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{location}}{TLObject-like (input web file location)}

\item{\code{offset}}{integer}

\item{\code{limit}}{integer

Methods:}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetWebFileRequest-new}{\code{GetWebFileRequest$new()}}
\item \href{#method-GetWebFileRequest-to_list}{\code{GetWebFileRequest$to_list()}}
\item \href{#method-GetWebFileRequest-to_bytes}{\code{GetWebFileRequest$to_bytes()}}
\item \href{#method-GetWebFileRequest-from_reader}{\code{GetWebFileRequest$from_reader()}}
\item \href{#method-GetWebFileRequest-clone}{\code{GetWebFileRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetWebFileRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetWebFileRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetWebFileRequest$new(location, offset, limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{location}}{TLObject-like}

\item{\code{offset}}{integer}

\item{\code{limit}}{integer
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetWebFileRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetWebFileRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetWebFileRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetWebFileRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetWebFileRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetWebFileRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetWebFileRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetWebFileRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetWebFileRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetWebFileRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetWebFileRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetWebFileRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
