% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{GlobalPrivacySettings}
\alias{GlobalPrivacySettings}
\title{GlobalPrivacySettings}
\description{
Telegram API type GlobalPrivacySettings
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobalPrivacySettings-new}{\code{GlobalPrivacySettings$new()}}
\item \href{#method-GlobalPrivacySettings-to_list}{\code{GlobalPrivacySettings$to_list()}}
\item \href{#method-GlobalPrivacySettings-clone}{\code{GlobalPrivacySettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalPrivacySettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalPrivacySettings-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalPrivacySettings$new(
  archive_and_mute_new_noncontactpeers = NULL,
  keep_archived_unmuted = NULL,
  keep_archived_folders = NULL,
  hide_read_marks = NULL,
  new_noncontactpeers_require_premium = NULL,
  display_gifts_button = NULL,
  noncontactpeers_paid_stars = NULL,
  disallowed_gifts = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalPrivacySettings-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalPrivacySettings-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalPrivacySettings$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalPrivacySettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalPrivacySettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalPrivacySettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
