% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\docType{data}
\name{good_prime}
\alias{good_prime}
\title{Known good prime constant for Telegram cryptographic operations}
\format{
An object of class \code{raw} of length 256.
}
\usage{
good_prime
}
\description{
Known good prime constant for Telegram cryptographic operations
}
\keyword{internal}
