% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R, R/telegrambaseclient.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing helper}
\usage{
x \%||\% y

x \%||\% y
}
\arguments{
\item{x}{The first value}

\item{y}{The second value}

\item{a}{The primary value to be inspected for `NULL`.}

\item{b}{The fallback value to use if `a` is `NULL`.}
}
\value{
The first non-`NULL` argument among `a` and `b`.

The first value if not NULL, otherwise the second value
}
\description{
Returns the first argument if it is not `NULL`; otherwise, it returns the
provided fallback value.
}
\examples{
1 \%||\% 2
NULL \%||\% 2
}
