% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{GroupCallStreamChannel}
\alias{GroupCallStreamChannel}
\title{GroupCallStreamChannel}
\description{
Telegram API type GroupCallStreamChannel
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupCallStreamChannel-new}{\code{GroupCallStreamChannel$new()}}
\item \href{#method-GroupCallStreamChannel-to_list}{\code{GroupCallStreamChannel$to_list()}}
\item \href{#method-GroupCallStreamChannel-bytes}{\code{GroupCallStreamChannel$bytes()}}
\item \href{#method-GroupCallStreamChannel-clone}{\code{GroupCallStreamChannel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupCallStreamChannel-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupCallStreamChannel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupCallStreamChannel$new(channel, scale, lastTimestampMs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupCallStreamChannel-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GroupCallStreamChannel-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupCallStreamChannel$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupCallStreamChannel-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GroupCallStreamChannel-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupCallStreamChannel$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupCallStreamChannel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupCallStreamChannel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupCallStreamChannel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
