% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gsub2}
\alias{gsub2}
\title{wrapper around gsub(). For Internal usage.}
\usage{
gsub2(text, arg1, subs2, ic = T)
}
\arguments{
\item{text}{CHAR. the text to be substituted}

\item{arg1}{CHAR. the argument to be substituted}

\item{ic}{BOOLEAN. ignore case. Default = TRUE}

\item{arg2}{CHAR. the argument with the substitution}
}
\description{
wrapper around gsub(). For Internal usage.
}
\details{
A wrapper around gsub, in order to make it usable with native pipe `|>`
}
\examples{
"foo foo FOO bar bar" |> gsub2("foo", "fool")
}
