% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_extension}
\alias{guess_extension}
\title{Guess Extension}
\usage{
guess_extension(mime_type)
}
\arguments{
\item{mime_type}{Character string of the MIME type.}
}
\value{
A character string with the file extension (including dot), or NULL.
}
\description{
Guesses a file extension from a MIME type.
}
\keyword{internal}
