% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chatlists.R
\name{HideChatlistUpdatesRequest}
\alias{HideChatlistUpdatesRequest}
\title{HideChatlistUpdatesRequest R6 class}
\description{
HideChatlistUpdatesRequest R6 class

HideChatlistUpdatesRequest R6 class
}
\details{
Request to hide updates for a chatlist.

Serializes the constructor id followed by chatlist bytes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Integer constructor id.}

\item{\code{SUBCLASS_OF_ID}}{Integer subclass id.}

\item{\code{chatlist}}{TypeInputChatlist input chatlist object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HideChatlistUpdatesRequest-new}{\code{HideChatlistUpdatesRequest$new()}}
\item \href{#method-HideChatlistUpdatesRequest-to_list}{\code{HideChatlistUpdatesRequest$to_list()}}
\item \href{#method-HideChatlistUpdatesRequest-bytes}{\code{HideChatlistUpdatesRequest$bytes()}}
\item \href{#method-HideChatlistUpdatesRequest-from_reader}{\code{HideChatlistUpdatesRequest$from_reader()}}
\item \href{#method-HideChatlistUpdatesRequest-clone}{\code{HideChatlistUpdatesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HideChatlistUpdatesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-HideChatlistUpdatesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize HideChatlistUpdatesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HideChatlistUpdatesRequest$new(chatlist = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chatlist}}{Input chatlist object
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HideChatlistUpdatesRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-HideChatlistUpdatesRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HideChatlistUpdatesRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representation
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HideChatlistUpdatesRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-HideChatlistUpdatesRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HideChatlistUpdatesRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector with serialized bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HideChatlistUpdatesRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-HideChatlistUpdatesRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HideChatlistUpdatesRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HideChatlistUpdatesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HideChatlistUpdatesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HideChatlistUpdatesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
