% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{HidePromoDataRequest}
\alias{HidePromoDataRequest}
\title{HidePromoDataRequest R6 class}
\description{
HidePromoDataRequest R6 class

HidePromoDataRequest R6 class
}
\details{
Request to hide promo data for a peer.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{peer}}{TLObject-like input peer (environment with to_bytes()/to_list() or raw)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HidePromoDataRequest-new}{\code{HidePromoDataRequest$new()}}
\item \href{#method-HidePromoDataRequest-resolve}{\code{HidePromoDataRequest$resolve()}}
\item \href{#method-HidePromoDataRequest-to_list}{\code{HidePromoDataRequest$to_list()}}
\item \href{#method-HidePromoDataRequest-to_bytes}{\code{HidePromoDataRequest$to_bytes()}}
\item \href{#method-HidePromoDataRequest-clone}{\code{HidePromoDataRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HidePromoDataRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-HidePromoDataRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize HidePromoDataRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HidePromoDataRequest$new(peer = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{TLObject-like or raw
Resolve peer using client and utils}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HidePromoDataRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-HidePromoDataRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HidePromoDataRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client object providing get_input_entity()}

\item{\code{utils}}{utils providing get_input_peer()
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HidePromoDataRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-HidePromoDataRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HidePromoDataRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `_` discriminator and peer as list (if possible)
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HidePromoDataRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-HidePromoDataRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HidePromoDataRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HidePromoDataRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HidePromoDataRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HidePromoDataRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
