% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{HighScore}
\alias{HighScore}
\title{HighScore}
\description{
Telegram API type HighScore
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pos}}{Field.}

\item{\code{userId}}{Field.}

\item{\code{score}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HighScore-new}{\code{HighScore$new()}}
\item \href{#method-HighScore-to_list}{\code{HighScore$to_list()}}
\item \href{#method-HighScore-bytes}{\code{HighScore$bytes()}}
\item \href{#method-HighScore-clone}{\code{HighScore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HighScore-new"></a>}}
\if{latex}{\out{\hypertarget{method-HighScore-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HighScore$new(pos, userId, score)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HighScore-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-HighScore-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HighScore$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HighScore-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-HighScore-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HighScore$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HighScore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HighScore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HighScore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
