% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegram2DF.R
\name{html2df}
\alias{html2df}
\title{convert the the HTML into a tibble/dataframe}
\usage{
html2df(html_file)
}
\arguments{
\item{html_file}{the HTML file path}
}
\description{
convert the the HTML into a tibble/dataframe
}
\details{
from html file of the exported chats of telegram, get a dataframe with chat name, message id, date time, user name, text and link to audio/video`
}
\examples{
html2df("~/Downloads/Telegram Desktop/ChatExport_2023-12-16/messages.html")

}
