% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{HttpWait}
\alias{HttpWait}
\title{HttpWait}
\description{
Telegram API type HttpWait
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HttpWait-new}{\code{HttpWait$new()}}
\item \href{#method-HttpWait-to_list}{\code{HttpWait$to_list()}}
\item \href{#method-HttpWait-bytes}{\code{HttpWait$bytes()}}
\item \href{#method-HttpWait-clone}{\code{HttpWait$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpWait-new"></a>}}
\if{latex}{\out{\hypertarget{method-HttpWait-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpWait$new(max_delay, wait_after, max_wait)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpWait-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-HttpWait-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpWait$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpWait-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-HttpWait-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpWait$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpWait-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HttpWait-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpWait$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
