% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{ImportBotAuthorizationRequest}
\alias{ImportBotAuthorizationRequest}
\title{ImportBotAuthorizationRequest}
\description{
Telegram API type ImportBotAuthorizationRequest
}
\details{
ImportBotAuthorizationRequest R6 class

Represents the TLRequest auth.ImportBotAuthorizationRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{flags}}{Field.}

\item{\code{api_id}}{Field.}

\item{\code{api_hash}}{Field.}

\item{\code{bot_auth_token}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImportBotAuthorizationRequest-new}{\code{ImportBotAuthorizationRequest$new()}}
\item \href{#method-ImportBotAuthorizationRequest-to_list}{\code{ImportBotAuthorizationRequest$to_list()}}
\item \href{#method-ImportBotAuthorizationRequest-to_bytes}{\code{ImportBotAuthorizationRequest$to_bytes()}}
\item \href{#method-ImportBotAuthorizationRequest-from_reader}{\code{ImportBotAuthorizationRequest$from_reader()}}
\item \href{#method-ImportBotAuthorizationRequest-clone}{\code{ImportBotAuthorizationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportBotAuthorizationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImportBotAuthorizationRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ImportBotAuthorizationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportBotAuthorizationRequest$new(flags, api_id, api_hash, bot_auth_token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flags}}{integer}

\item{\code{api_id}}{integer}

\item{\code{api_hash}}{character}

\item{\code{bot_auth_token}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportBotAuthorizationRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ImportBotAuthorizationRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportBotAuthorizationRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportBotAuthorizationRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ImportBotAuthorizationRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportBotAuthorizationRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportBotAuthorizationRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ImportBotAuthorizationRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create ImportBotAuthorizationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportBotAuthorizationRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ImportBotAuthorizationRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportBotAuthorizationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImportBotAuthorizationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportBotAuthorizationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
