% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ImportedContact}
\alias{ImportedContact}
\title{ImportedContact}
\description{
Telegram API type ImportedContact
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImportedContact-new}{\code{ImportedContact$new()}}
\item \href{#method-ImportedContact-to_list}{\code{ImportedContact$to_list()}}
\item \href{#method-ImportedContact-bytes}{\code{ImportedContact$bytes()}}
\item \href{#method-ImportedContact-clone}{\code{ImportedContact$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportedContact-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImportedContact-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportedContact$new(user_id, client_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportedContact-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ImportedContact-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportedContact$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportedContact-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ImportedContact-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportedContact$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportedContact-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImportedContact-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportedContact$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
