% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{ImportLoginTokenRequest}
\alias{ImportLoginTokenRequest}
\title{ImportLoginTokenRequest}
\description{
Telegram API type ImportLoginTokenRequest
}
\details{
ImportLoginTokenRequest R6 class

Represents the TLRequest auth.ImportLoginTokenRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{token}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImportLoginTokenRequest-new}{\code{ImportLoginTokenRequest$new()}}
\item \href{#method-ImportLoginTokenRequest-to_list}{\code{ImportLoginTokenRequest$to_list()}}
\item \href{#method-ImportLoginTokenRequest-to_bytes}{\code{ImportLoginTokenRequest$to_bytes()}}
\item \href{#method-ImportLoginTokenRequest-from_reader}{\code{ImportLoginTokenRequest$from_reader()}}
\item \href{#method-ImportLoginTokenRequest-clone}{\code{ImportLoginTokenRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLoginTokenRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLoginTokenRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ImportLoginTokenRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLoginTokenRequest$new(token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{raw vector or string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLoginTokenRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLoginTokenRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLoginTokenRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLoginTokenRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLoginTokenRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLoginTokenRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLoginTokenRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLoginTokenRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create ImportLoginTokenRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLoginTokenRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_bytes(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ImportLoginTokenRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLoginTokenRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLoginTokenRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLoginTokenRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
