% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{ImportWebTokenAuthorizationRequest}
\alias{ImportWebTokenAuthorizationRequest}
\title{ImportWebTokenAuthorizationRequest}
\description{
Telegram API type ImportWebTokenAuthorizationRequest
}
\details{
ImportWebTokenAuthorizationRequest R6 class

Represents the TLRequest auth.ImportWebTokenAuthorizationRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0x2db873a9).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0xb9e04e39).}

\item{\code{api_id}}{integer API id.}

\item{\code{api_hash}}{character API hash.}

\item{\code{web_auth_token}}{character Web auth token.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImportWebTokenAuthorizationRequest-new}{\code{ImportWebTokenAuthorizationRequest$new()}}
\item \href{#method-ImportWebTokenAuthorizationRequest-to_list}{\code{ImportWebTokenAuthorizationRequest$to_list()}}
\item \href{#method-ImportWebTokenAuthorizationRequest-to_bytes}{\code{ImportWebTokenAuthorizationRequest$to_bytes()}}
\item \href{#method-ImportWebTokenAuthorizationRequest-from_reader}{\code{ImportWebTokenAuthorizationRequest$from_reader()}}
\item \href{#method-ImportWebTokenAuthorizationRequest-clone}{\code{ImportWebTokenAuthorizationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportWebTokenAuthorizationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImportWebTokenAuthorizationRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ImportWebTokenAuthorizationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportWebTokenAuthorizationRequest$new(api_id, api_hash, web_auth_token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_id}}{integer}

\item{\code{api_hash}}{character}

\item{\code{web_auth_token}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportWebTokenAuthorizationRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ImportWebTokenAuthorizationRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportWebTokenAuthorizationRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportWebTokenAuthorizationRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ImportWebTokenAuthorizationRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportWebTokenAuthorizationRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportWebTokenAuthorizationRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ImportWebTokenAuthorizationRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create ImportWebTokenAuthorizationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportWebTokenAuthorizationRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_bytes(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ImportWebTokenAuthorizationRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportWebTokenAuthorizationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImportWebTokenAuthorizationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportWebTokenAuthorizationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
