% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_view_count.R
\name{increase_view_count}
\alias{increase_view_count}
\title{Increase view count for channel messages}
\usage{
increase_view_count(client, channel, message_ids)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{channel}{Channel (username, link, or entity).}

\item{message_ids}{Numeric vector of message IDs.}
}
\value{
The API response (views).
}
\description{
Increase view count for channel messages
}
\examples{
\dontrun{
client <- TelegramClient$new("my_session", api_id = 123, api_hash = "abc")
client$connect()

increase_view_count(client, "channelname", message_ids = 1:10)
}
}
