% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InlineBotSwitchPM}
\alias{InlineBotSwitchPM}
\title{InlineBotSwitchPM}
\description{
Telegram API type InlineBotSwitchPM
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{text}}{Field.}

\item{\code{start_param}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineBotSwitchPM-new}{\code{InlineBotSwitchPM$new()}}
\item \href{#method-InlineBotSwitchPM-to_list}{\code{InlineBotSwitchPM$to_list()}}
\item \href{#method-InlineBotSwitchPM-bytes}{\code{InlineBotSwitchPM$bytes()}}
\item \href{#method-InlineBotSwitchPM-serialize_bytes}{\code{InlineBotSwitchPM$serialize_bytes()}}
\item \href{#method-InlineBotSwitchPM-clone}{\code{InlineBotSwitchPM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineBotSwitchPM-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineBotSwitchPM-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBotSwitchPM$new(text, start_param)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineBotSwitchPM-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InlineBotSwitchPM-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBotSwitchPM$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineBotSwitchPM-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InlineBotSwitchPM-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBotSwitchPM$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineBotSwitchPM-serialize_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InlineBotSwitchPM-serialize_bytes}{}}}
\subsection{Method \code{serialize_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBotSwitchPM$serialize_bytes(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineBotSwitchPM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineBotSwitchPM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBotSwitchPM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
