% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InlineBotWebView}
\alias{InlineBotWebView}
\title{InlineBotWebView}
\description{
Telegram API type InlineBotWebView
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineBotWebView-new}{\code{InlineBotWebView$new()}}
\item \href{#method-InlineBotWebView-to_list}{\code{InlineBotWebView$to_list()}}
\item \href{#method-InlineBotWebView-bytes}{\code{InlineBotWebView$bytes()}}
\item \href{#method-InlineBotWebView-serialize_bytes}{\code{InlineBotWebView$serialize_bytes()}}
\item \href{#method-InlineBotWebView-clone}{\code{InlineBotWebView$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineBotWebView-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineBotWebView-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBotWebView$new(text, url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineBotWebView-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InlineBotWebView-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBotWebView$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineBotWebView-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InlineBotWebView-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBotWebView$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineBotWebView-serialize_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InlineBotWebView-serialize_bytes}{}}}
\subsection{Method \code{serialize_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBotWebView$serialize_bytes(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineBotWebView-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineBotWebView-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineBotWebView$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
