% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputAppEvent}
\alias{InputAppEvent}
\title{InputAppEvent}
\description{
Telegram API type InputAppEvent
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time}}{Field.}

\item{\code{type}}{Field.}

\item{\code{peer}}{Field.}

\item{\code{data}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputAppEvent-new}{\code{InputAppEvent$new()}}
\item \href{#method-InputAppEvent-to_list}{\code{InputAppEvent$to_list()}}
\item \href{#method-InputAppEvent-bytes}{\code{InputAppEvent$bytes()}}
\item \href{#method-InputAppEvent-clone}{\code{InputAppEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputAppEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputAppEvent-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputAppEvent$new(time, type, peer, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputAppEvent-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputAppEvent-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputAppEvent$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputAppEvent-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputAppEvent-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputAppEvent$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputAppEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputAppEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputAppEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
