% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputBotAppID}
\alias{InputBotAppID}
\title{InputBotAppID}
\description{
Telegram API type InputBotAppID
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputBotAppID-new}{\code{InputBotAppID$new()}}
\item \href{#method-InputBotAppID-to_list}{\code{InputBotAppID$to_list()}}
\item \href{#method-InputBotAppID-bytes}{\code{InputBotAppID$bytes()}}
\item \href{#method-InputBotAppID-clone}{\code{InputBotAppID$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBotAppID-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputBotAppID-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBotAppID$new(id, access_hash)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBotAppID-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputBotAppID-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBotAppID$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBotAppID-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputBotAppID-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBotAppID$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBotAppID-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputBotAppID-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBotAppID$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
