% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputBusinessAwayMessage}
\alias{InputBusinessAwayMessage}
\title{InputBusinessAwayMessage}
\description{
Telegram API type InputBusinessAwayMessage
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputBusinessAwayMessage-new}{\code{InputBusinessAwayMessage$new()}}
\item \href{#method-InputBusinessAwayMessage-to_list}{\code{InputBusinessAwayMessage$to_list()}}
\item \href{#method-InputBusinessAwayMessage-bytes}{\code{InputBusinessAwayMessage$bytes()}}
\item \href{#method-InputBusinessAwayMessage-clone}{\code{InputBusinessAwayMessage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBusinessAwayMessage-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputBusinessAwayMessage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBusinessAwayMessage$new(
  shortcut_id,
  schedule,
  recipients,
  offline_only = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBusinessAwayMessage-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputBusinessAwayMessage-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBusinessAwayMessage$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBusinessAwayMessage-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputBusinessAwayMessage-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBusinessAwayMessage$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBusinessAwayMessage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputBusinessAwayMessage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBusinessAwayMessage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
