% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputBusinessBotRecipients}
\alias{InputBusinessBotRecipients}
\title{InputBusinessBotRecipients}
\description{
Telegram API type InputBusinessBotRecipients
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputBusinessBotRecipients-new}{\code{InputBusinessBotRecipients$new()}}
\item \href{#method-InputBusinessBotRecipients-to_list}{\code{InputBusinessBotRecipients$to_list()}}
\item \href{#method-InputBusinessBotRecipients-bytes}{\code{InputBusinessBotRecipients$bytes()}}
\item \href{#method-InputBusinessBotRecipients-clone}{\code{InputBusinessBotRecipients$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBusinessBotRecipients-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputBusinessBotRecipients-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBusinessBotRecipients$new(
  existing_chats = NULL,
  new_chats = NULL,
  contacts = NULL,
  non_contacts = NULL,
  exclude_selected = NULL,
  users = NULL,
  exclude_users = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBusinessBotRecipients-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputBusinessBotRecipients-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBusinessBotRecipients$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBusinessBotRecipients-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputBusinessBotRecipients-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBusinessBotRecipients$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputBusinessBotRecipients-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputBusinessBotRecipients-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputBusinessBotRecipients$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
