% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputEncryptedFileBigUploaded}
\alias{InputEncryptedFileBigUploaded}
\title{InputEncryptedFileBigUploaded}
\description{
Telegram API type InputEncryptedFileBigUploaded
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputEncryptedFileBigUploaded-new}{\code{InputEncryptedFileBigUploaded$new()}}
\item \href{#method-InputEncryptedFileBigUploaded-to_list}{\code{InputEncryptedFileBigUploaded$to_list()}}
\item \href{#method-InputEncryptedFileBigUploaded-bytes}{\code{InputEncryptedFileBigUploaded$bytes()}}
\item \href{#method-InputEncryptedFileBigUploaded-clone}{\code{InputEncryptedFileBigUploaded$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileBigUploaded-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileBigUploaded-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileBigUploaded$new(id, parts, key_fingerprint)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileBigUploaded-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileBigUploaded-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileBigUploaded$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileBigUploaded-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileBigUploaded-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileBigUploaded$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileBigUploaded-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileBigUploaded-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileBigUploaded$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
