% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputEncryptedFileEmpty}
\alias{InputEncryptedFileEmpty}
\title{InputEncryptedFileEmpty}
\description{
Telegram API type InputEncryptedFileEmpty
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputEncryptedFileEmpty-to_list}{\code{InputEncryptedFileEmpty$to_list()}}
\item \href{#method-InputEncryptedFileEmpty-bytes}{\code{InputEncryptedFileEmpty$bytes()}}
\item \href{#method-InputEncryptedFileEmpty-clone}{\code{InputEncryptedFileEmpty$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileEmpty-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileEmpty-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileEmpty$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileEmpty-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileEmpty-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileEmpty$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileEmpty-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileEmpty-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileEmpty$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
