% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputEncryptedFileLocation}
\alias{InputEncryptedFileLocation}
\title{InputEncryptedFileLocation}
\description{
Telegram API type InputEncryptedFileLocation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputEncryptedFileLocation-new}{\code{InputEncryptedFileLocation$new()}}
\item \href{#method-InputEncryptedFileLocation-to_list}{\code{InputEncryptedFileLocation$to_list()}}
\item \href{#method-InputEncryptedFileLocation-bytes}{\code{InputEncryptedFileLocation$bytes()}}
\item \href{#method-InputEncryptedFileLocation-clone}{\code{InputEncryptedFileLocation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileLocation-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileLocation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileLocation$new(id, access_hash)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileLocation-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileLocation-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileLocation$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileLocation-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileLocation-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileLocation$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileLocation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileLocation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileLocation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
