% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputEncryptedFileUploaded}
\alias{InputEncryptedFileUploaded}
\title{InputEncryptedFileUploaded}
\description{
Telegram API type InputEncryptedFileUploaded
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputEncryptedFileUploaded-new}{\code{InputEncryptedFileUploaded$new()}}
\item \href{#method-InputEncryptedFileUploaded-to_list}{\code{InputEncryptedFileUploaded$to_list()}}
\item \href{#method-InputEncryptedFileUploaded-bytes}{\code{InputEncryptedFileUploaded$bytes()}}
\item \href{#method-InputEncryptedFileUploaded-clone}{\code{InputEncryptedFileUploaded$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileUploaded-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileUploaded-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileUploaded$new(id, parts, md5_checksum, key_fingerprint)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileUploaded-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileUploaded-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileUploaded$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileUploaded-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileUploaded-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileUploaded$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputEncryptedFileUploaded-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputEncryptedFileUploaded-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputEncryptedFileUploaded$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
