% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputFile}
\alias{InputFile}
\title{InputFile}
\description{
Telegram API type InputFile
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Field.}

\item{\code{parts}}{Field.}

\item{\code{name}}{Field.}

\item{\code{md5_checksum}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputFile-new}{\code{InputFile$new()}}
\item \href{#method-InputFile-to_list}{\code{InputFile$to_list()}}
\item \href{#method-InputFile-bytes}{\code{InputFile$bytes()}}
\item \href{#method-InputFile-clone}{\code{InputFile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFile-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputFile-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFile$new(id, parts, name, md5_checksum)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFile-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputFile-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFile$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFile-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputFile-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFile$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
