% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputFileBig}
\alias{InputFileBig}
\title{InputFileBig}
\description{
Telegram API type InputFileBig
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Field.}

\item{\code{parts}}{Field.}

\item{\code{name}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputFileBig-new}{\code{InputFileBig$new()}}
\item \href{#method-InputFileBig-to_list}{\code{InputFileBig$to_list()}}
\item \href{#method-InputFileBig-bytes}{\code{InputFileBig$bytes()}}
\item \href{#method-InputFileBig-clone}{\code{InputFileBig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileBig-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileBig-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileBig$new(id, parts, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileBig-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileBig-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileBig$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileBig-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileBig-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileBig$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileBig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileBig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileBig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
