% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputFileLocation}
\alias{InputFileLocation}
\title{InputFileLocation}
\description{
Telegram API type InputFileLocation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{volume_id}}{Field.}

\item{\code{local_id}}{Field.}

\item{\code{secret}}{Field.}

\item{\code{file_reference}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputFileLocation-new}{\code{InputFileLocation$new()}}
\item \href{#method-InputFileLocation-to_list}{\code{InputFileLocation$to_list()}}
\item \href{#method-InputFileLocation-bytes}{\code{InputFileLocation$bytes()}}
\item \href{#method-InputFileLocation-clone}{\code{InputFileLocation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileLocation-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileLocation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileLocation$new(volume_id, local_id, secret, file_reference)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileLocation-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileLocation-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileLocation$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileLocation-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileLocation-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileLocation$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileLocation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileLocation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileLocation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
