% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputFileStoryDocument}
\alias{InputFileStoryDocument}
\title{InputFileStoryDocument}
\description{
Telegram API type InputFileStoryDocument
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputFileStoryDocument-new}{\code{InputFileStoryDocument$new()}}
\item \href{#method-InputFileStoryDocument-to_list}{\code{InputFileStoryDocument$to_list()}}
\item \href{#method-InputFileStoryDocument-bytes}{\code{InputFileStoryDocument$bytes()}}
\item \href{#method-InputFileStoryDocument-clone}{\code{InputFileStoryDocument$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileStoryDocument-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileStoryDocument-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileStoryDocument$new(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileStoryDocument-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileStoryDocument-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileStoryDocument$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileStoryDocument-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileStoryDocument-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileStoryDocument$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFileStoryDocument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputFileStoryDocument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFileStoryDocument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
