% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputFolderPeer}
\alias{InputFolderPeer}
\title{InputFolderPeer}
\description{
Telegram API type InputFolderPeer
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{peer}}{Field.}

\item{\code{folder_id}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputFolderPeer-new}{\code{InputFolderPeer$new()}}
\item \href{#method-InputFolderPeer-to_list}{\code{InputFolderPeer$to_list()}}
\item \href{#method-InputFolderPeer-bytes}{\code{InputFolderPeer$bytes()}}
\item \href{#method-InputFolderPeer-clone}{\code{InputFolderPeer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFolderPeer-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputFolderPeer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFolderPeer$new(peer, folder_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFolderPeer-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputFolderPeer-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFolderPeer$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFolderPeer-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputFolderPeer-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFolderPeer$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputFolderPeer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputFolderPeer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputFolderPeer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
