% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputGameID}
\alias{InputGameID}
\title{InputGameID}
\description{
Telegram API type InputGameID
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputGameID-new}{\code{InputGameID$new()}}
\item \href{#method-InputGameID-to_list}{\code{InputGameID$to_list()}}
\item \href{#method-InputGameID-bytes}{\code{InputGameID$bytes()}}
\item \href{#method-InputGameID-clone}{\code{InputGameID$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGameID-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputGameID-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGameID$new(id, access_hash)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGameID-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputGameID-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGameID$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGameID-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputGameID-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGameID$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGameID-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputGameID-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGameID$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
