% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputGameShortName}
\alias{InputGameShortName}
\title{InputGameShortName}
\description{
Telegram API type InputGameShortName
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bot_id}}{Field.}

\item{\code{short_name}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputGameShortName-new}{\code{InputGameShortName$new()}}
\item \href{#method-InputGameShortName-to_list}{\code{InputGameShortName$to_list()}}
\item \href{#method-InputGameShortName-bytes}{\code{InputGameShortName$bytes()}}
\item \href{#method-InputGameShortName-clone}{\code{InputGameShortName$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGameShortName-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputGameShortName-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGameShortName$new(bot_id, short_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGameShortName-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputGameShortName-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGameShortName$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGameShortName-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputGameShortName-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGameShortName$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGameShortName-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputGameShortName-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGameShortName$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
