% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputGeoPoint}
\alias{InputGeoPoint}
\title{InputGeoPoint}
\description{
Telegram API type InputGeoPoint
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputGeoPoint-new}{\code{InputGeoPoint$new()}}
\item \href{#method-InputGeoPoint-to_dict}{\code{InputGeoPoint$to_dict()}}
\item \href{#method-InputGeoPoint-bytes}{\code{InputGeoPoint$bytes()}}
\item \href{#method-InputGeoPoint-clone}{\code{InputGeoPoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGeoPoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputGeoPoint-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGeoPoint$new(lat, long, accuracy_radius = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGeoPoint-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-InputGeoPoint-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGeoPoint$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGeoPoint-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputGeoPoint-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGeoPoint$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGeoPoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputGeoPoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGeoPoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
