% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputGeoPointEmpty}
\alias{InputGeoPointEmpty}
\title{InputGeoPointEmpty}
\description{
Telegram API type InputGeoPointEmpty
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputGeoPointEmpty-to_dict}{\code{InputGeoPointEmpty$to_dict()}}
\item \href{#method-InputGeoPointEmpty-bytes}{\code{InputGeoPointEmpty$bytes()}}
\item \href{#method-InputGeoPointEmpty-clone}{\code{InputGeoPointEmpty$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGeoPointEmpty-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-InputGeoPointEmpty-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGeoPointEmpty$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGeoPointEmpty-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputGeoPointEmpty-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGeoPointEmpty$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGeoPointEmpty-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputGeoPointEmpty-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGeoPointEmpty$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
