% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputGroupCallStream}
\alias{InputGroupCallStream}
\title{InputGroupCallStream}
\description{
Telegram API type InputGroupCallStream
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{Field.}

\item{\code{time_ms}}{Field.}

\item{\code{scale}}{Field.}

\item{\code{video_channel}}{Field.}

\item{\code{video_quality}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputGroupCallStream-new}{\code{InputGroupCallStream$new()}}
\item \href{#method-InputGroupCallStream-to_list}{\code{InputGroupCallStream$to_list()}}
\item \href{#method-InputGroupCallStream-bytes}{\code{InputGroupCallStream$bytes()}}
\item \href{#method-InputGroupCallStream-clone}{\code{InputGroupCallStream$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGroupCallStream-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputGroupCallStream-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGroupCallStream$new(
  call,
  time_ms,
  scale,
  video_channel = NULL,
  video_quality = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGroupCallStream-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputGroupCallStream-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGroupCallStream$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGroupCallStream-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputGroupCallStream-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGroupCallStream$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputGroupCallStream-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputGroupCallStream-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputGroupCallStream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
