% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputInvoiceBusinessBotTransferStars}
\alias{InputInvoiceBusinessBotTransferStars}
\title{InputInvoiceBusinessBotTransferStars}
\description{
Telegram API type InputInvoiceBusinessBotTransferStars
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputInvoiceBusinessBotTransferStars-new}{\code{InputInvoiceBusinessBotTransferStars$new()}}
\item \href{#method-InputInvoiceBusinessBotTransferStars-to_list}{\code{InputInvoiceBusinessBotTransferStars$to_list()}}
\item \href{#method-InputInvoiceBusinessBotTransferStars-bytes}{\code{InputInvoiceBusinessBotTransferStars$bytes()}}
\item \href{#method-InputInvoiceBusinessBotTransferStars-clone}{\code{InputInvoiceBusinessBotTransferStars$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceBusinessBotTransferStars-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceBusinessBotTransferStars-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceBusinessBotTransferStars$new(bot, stars)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceBusinessBotTransferStars-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceBusinessBotTransferStars-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceBusinessBotTransferStars$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceBusinessBotTransferStars-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceBusinessBotTransferStars-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceBusinessBotTransferStars$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceBusinessBotTransferStars-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceBusinessBotTransferStars-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceBusinessBotTransferStars$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
