% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputInvoicePremiumGiftStars}
\alias{InputInvoicePremiumGiftStars}
\title{InputInvoicePremiumGiftStars}
\description{
Telegram API type InputInvoicePremiumGiftStars
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputInvoicePremiumGiftStars-new}{\code{InputInvoicePremiumGiftStars$new()}}
\item \href{#method-InputInvoicePremiumGiftStars-to_dict}{\code{InputInvoicePremiumGiftStars$to_dict()}}
\item \href{#method-InputInvoicePremiumGiftStars-bytes}{\code{InputInvoicePremiumGiftStars$bytes()}}
\item \href{#method-InputInvoicePremiumGiftStars-clone}{\code{InputInvoicePremiumGiftStars$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoicePremiumGiftStars-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoicePremiumGiftStars-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoicePremiumGiftStars$new(user_id, months, message = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoicePremiumGiftStars-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoicePremiumGiftStars-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoicePremiumGiftStars$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoicePremiumGiftStars-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoicePremiumGiftStars-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoicePremiumGiftStars$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoicePremiumGiftStars-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoicePremiumGiftStars-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoicePremiumGiftStars$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
