% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputInvoiceSlug}
\alias{InputInvoiceSlug}
\title{InputInvoiceSlug}
\description{
Telegram API type InputInvoiceSlug
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{slug}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputInvoiceSlug-new}{\code{InputInvoiceSlug$new()}}
\item \href{#method-InputInvoiceSlug-to_dict}{\code{InputInvoiceSlug$to_dict()}}
\item \href{#method-InputInvoiceSlug-bytes}{\code{InputInvoiceSlug$bytes()}}
\item \href{#method-InputInvoiceSlug-clone}{\code{InputInvoiceSlug$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceSlug-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceSlug-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceSlug$new(slug)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceSlug-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceSlug-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceSlug$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceSlug-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceSlug-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceSlug$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceSlug-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceSlug-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceSlug$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
