% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputInvoiceStarGift}
\alias{InputInvoiceStarGift}
\title{InputInvoiceStarGift}
\description{
Telegram API type InputInvoiceStarGift
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputInvoiceStarGift-new}{\code{InputInvoiceStarGift$new()}}
\item \href{#method-InputInvoiceStarGift-to_list}{\code{InputInvoiceStarGift$to_list()}}
\item \href{#method-InputInvoiceStarGift-bytes}{\code{InputInvoiceStarGift$bytes()}}
\item \href{#method-InputInvoiceStarGift-clone}{\code{InputInvoiceStarGift$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceStarGift-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceStarGift-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceStarGift$new(
  peer,
  gift_id,
  hide_name = NULL,
  include_upgrade = NULL,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceStarGift-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceStarGift-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceStarGift$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceStarGift-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceStarGift-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceStarGift$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceStarGift-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceStarGift-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceStarGift$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
