% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputInvoiceStarGiftPrepaidUpgrade}
\alias{InputInvoiceStarGiftPrepaidUpgrade}
\title{InputInvoiceStarGiftPrepaidUpgrade}
\description{
Telegram API type InputInvoiceStarGiftPrepaidUpgrade
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputInvoiceStarGiftPrepaidUpgrade-new}{\code{InputInvoiceStarGiftPrepaidUpgrade$new()}}
\item \href{#method-InputInvoiceStarGiftPrepaidUpgrade-to_list}{\code{InputInvoiceStarGiftPrepaidUpgrade$to_list()}}
\item \href{#method-InputInvoiceStarGiftPrepaidUpgrade-bytes}{\code{InputInvoiceStarGiftPrepaidUpgrade$bytes()}}
\item \href{#method-InputInvoiceStarGiftPrepaidUpgrade-clone}{\code{InputInvoiceStarGiftPrepaidUpgrade$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceStarGiftPrepaidUpgrade-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceStarGiftPrepaidUpgrade-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceStarGiftPrepaidUpgrade$new(peer, hash)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceStarGiftPrepaidUpgrade-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceStarGiftPrepaidUpgrade-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceStarGiftPrepaidUpgrade$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceStarGiftPrepaidUpgrade-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceStarGiftPrepaidUpgrade-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceStarGiftPrepaidUpgrade$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputInvoiceStarGiftPrepaidUpgrade-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputInvoiceStarGiftPrepaidUpgrade-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputInvoiceStarGiftPrepaidUpgrade$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
