% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputStorePaymentStarsGiveaway}
\alias{InputStorePaymentStarsGiveaway}
\title{InputStorePaymentStarsGiveaway}
\description{
Telegram API type InputStorePaymentStarsGiveaway
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{InputStorePaymentStarsGiveaway}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputStorePaymentStarsGiveaway-new}{\code{InputStorePaymentStarsGiveaway$new()}}
\item \href{#method-InputStorePaymentStarsGiveaway-to_dict}{\code{InputStorePaymentStarsGiveaway$to_dict()}}
\item \href{#method-InputStorePaymentStarsGiveaway-bytes}{\code{InputStorePaymentStarsGiveaway$bytes()}}
\item \href{#method-InputStorePaymentStarsGiveaway-clone}{\code{InputStorePaymentStarsGiveaway$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputStorePaymentStarsGiveaway-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputStorePaymentStarsGiveaway-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputStorePaymentStarsGiveaway$new(
  stars,
  boost_peer,
  until_date,
  currency,
  amount,
  users,
  only_new_subscribers = NULL,
  winners_are_visible = NULL,
  additional_peers = NULL,
  countries_iso2 = NULL,
  prize_description = NULL,
  random_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputStorePaymentStarsGiveaway-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-InputStorePaymentStarsGiveaway-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputStorePaymentStarsGiveaway$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputStorePaymentStarsGiveaway-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InputStorePaymentStarsGiveaway-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputStorePaymentStarsGiveaway$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputStorePaymentStarsGiveaway-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputStorePaymentStarsGiveaway-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputStorePaymentStarsGiveaway$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
