% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputTheme}
\alias{InputTheme}
\title{InputTheme}
\description{
Telegram API type InputTheme
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Field.}

\item{\code{access_hash}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputTheme-new}{\code{InputTheme$new()}}
\item \href{#method-InputTheme-to_dict}{\code{InputTheme$to_dict()}}
\item \href{#method-InputTheme-clone}{\code{InputTheme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputTheme-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputTheme-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputTheme$new(id, access_hash)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputTheme-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-InputTheme-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputTheme$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputTheme-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputTheme-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputTheme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
