% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{InputThemeSettings}
\alias{InputThemeSettings}
\title{InputThemeSettings}
\description{
Telegram API type InputThemeSettings
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputThemeSettings-new}{\code{InputThemeSettings$new()}}
\item \href{#method-InputThemeSettings-to_dict}{\code{InputThemeSettings$to_dict()}}
\item \href{#method-InputThemeSettings-clone}{\code{InputThemeSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputThemeSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputThemeSettings-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputThemeSettings$new(
  base_theme,
  accent_color,
  message_colors_animated = NULL,
  outbox_accent_color = NULL,
  message_colors = NULL,
  wallpaper = NULL,
  wallpaper_settings = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputThemeSettings-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-InputThemeSettings-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputThemeSettings$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputThemeSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputThemeSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputThemeSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
