% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int_from_bytes}
\alias{int_from_bytes}
\title{Convert Bytes to Bigz Integer}
\usage{
int_from_bytes(data, endian = "big")
}
\arguments{
\item{data}{A raw vector.}

\item{endian}{Either "big" or "little".}
}
\value{
A bigz integer.
}
\description{
Convert Bytes to Bigz Integer
}
