% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int_to_bytes}
\alias{int_to_bytes}
\title{Convert Bigz Integer/Numeric to Bytes}
\usage{
int_to_bytes(val, length, endian = "big")
}
\arguments{
\item{val}{A bigz integer or numeric.}

\item{length}{The length of the resulting raw vector.}

\item{endian}{Either "big" or "little".}
}
\value{
A raw vector.
}
\description{
Convert Bigz Integer/Numeric to Bytes
}
