% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int_to_raw_le}
\alias{int_to_raw_le}
\title{Helper to convert integer to little-endian raw vector of given width (bytes)}
\usage{
int_to_raw_le(x, width = 4L)
}
\arguments{
\item{x}{integer or numeric}

\item{width}{number of bytes (default 4)}
}
\value{
raw vector
}
\description{
Helper to convert integer to little-endian raw vector of given width (bytes)
}
