% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{InvalidBufferError}
\alias{InvalidBufferError}
\title{InvalidBufferError}
\description{
Telegram API type InvalidBufferError
}
\details{
InvalidBufferError
}
\section{Super class}{
\code{\link[telegramR:Exception]{telegramR::Exception}} -> \code{InvalidBufferError}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{payload}}{\code{raw} The invalid buffer payload.}

\item{\code{code}}{\code{numeric} The HTTP error code.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvalidBufferError-new}{\code{InvalidBufferError$new()}}
\item \href{#method-InvalidBufferError-clone}{\code{InvalidBufferError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidBufferError-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidBufferError-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the InvalidBufferError class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidBufferError$new(payload)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{payload}}{\code{raw} The invalid buffer payload.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidBufferError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidBufferError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidBufferError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
