% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{InvalidChecksumError}
\alias{InvalidChecksumError}
\title{InvalidChecksumError}
\description{
Occurs when using the TCP full mode and the checksum of a received packet
doesn't match the expected checksum.
}
\section{Super class}{
\code{\link[telegramR:Exception]{telegramR::Exception}} -> \code{InvalidChecksumError}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{checksum}}{\code{numeric} The received checksum.}

\item{\code{valid_checksum}}{\code{numeric} The expected checksum.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvalidChecksumError-new}{\code{InvalidChecksumError$new()}}
\item \href{#method-InvalidChecksumError-clone}{\code{InvalidChecksumError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidChecksumError-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidChecksumError-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the InvalidChecksumError class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidChecksumError$new(checksum, valid_checksum)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{checksum}}{\code{numeric} The received checksum.}

\item{\code{valid_checksum}}{\code{numeric} The expected checksum.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidChecksumError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidChecksumError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidChecksumError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
