% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{InvokeWebViewCustomMethodRequest}
\alias{InvokeWebViewCustomMethodRequest}
\title{InvokeWebViewCustomMethodRequest R6 class}
\description{
Methods:
- initialize(bot, custom_method, params)
- resolve(client, utils)
- to_list() -> list representation
- to_bytes() -> raw vector bytes

Each method is documented inline below.
}
\details{
Represents the InvokeWebViewCustomMethodRequest TL request.
}
\examples{
# o <- InvokeWebViewCustomMethodRequest$new(bot, "methodName", paramsObj)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bot}}{TypeInputUser input user or object implementing to_list()/to_bytes()}

\item{\code{custom_method}}{character custom method name}

\item{\code{params}}{TypeDataJSON object or representation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvokeWebViewCustomMethodRequest-new}{\code{InvokeWebViewCustomMethodRequest$new()}}
\item \href{#method-InvokeWebViewCustomMethodRequest-resolve}{\code{InvokeWebViewCustomMethodRequest$resolve()}}
\item \href{#method-InvokeWebViewCustomMethodRequest-to_list}{\code{InvokeWebViewCustomMethodRequest$to_list()}}
\item \href{#method-InvokeWebViewCustomMethodRequest-to_bytes}{\code{InvokeWebViewCustomMethodRequest$to_bytes()}}
\item \href{#method-InvokeWebViewCustomMethodRequest-clone}{\code{InvokeWebViewCustomMethodRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWebViewCustomMethodRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWebViewCustomMethodRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWebViewCustomMethodRequest$new(bot, custom_method, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bot}}{TypeInputUser or identifier}

\item{\code{custom_method}}{character}

\item{\code{params}}{TypeDataJSON or object
Resolve references (convert entities via client/utils)

Resolves bot via client$get_input_entity and utils$get_input_user.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWebViewCustomMethodRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWebViewCustomMethodRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWebViewCustomMethodRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client with get_input_entity method}

\item{\code{utils}}{utils with get_input_user}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
Convert to list (dictionary-like)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWebViewCustomMethodRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWebViewCustomMethodRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWebViewCustomMethodRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWebViewCustomMethodRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWebViewCustomMethodRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWebViewCustomMethodRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWebViewCustomMethodRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWebViewCustomMethodRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWebViewCustomMethodRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
