% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/requests.R
\name{InvokeWithoutUpdatesRequest}
\alias{InvokeWithoutUpdatesRequest}
\title{InvokeWithoutUpdatesRequest R6 class}
\value{
An R6 object of class \code{InvokeWithoutUpdatesRequest}.
}
\description{
Represents the TL request `InvokeWithoutUpdatesRequest`.

R6 class representing a request that invokes a nested `query` without producing
any updates. The nested `query` object is expected to implement `to_list()`
and `to_bytes()` for proper serialization.
}
\details{
Fields:
- query: TypeX (an object representing a TL query). May be an R6 TL object with to_raw()/to_list().

Methods:
- new(query): create new instance.
- to_list(): return an R list representation.
- to_raw(): serialize to raw vector (bytes) in little endian as used in TL.

Note: to_raw() expects the query to provide a to_raw() method returning a raw vector,
or to already be a raw vector. If neither is true, the query is written via as.raw
of its serialized form where appropriate.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{query}}{Field.
Initialize an InvokeWithoutUpdatesRequest}

\item{\code{class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvokeWithoutUpdatesRequest-new}{\code{InvokeWithoutUpdatesRequest$new()}}
\item \href{#method-InvokeWithoutUpdatesRequest-to_list}{\code{InvokeWithoutUpdatesRequest$to_list()}}
\item \href{#method-InvokeWithoutUpdatesRequest-to_bytes}{\code{InvokeWithoutUpdatesRequest$to_bytes()}}
\item \href{#method-InvokeWithoutUpdatesRequest-clone}{\code{InvokeWithoutUpdatesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWithoutUpdatesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWithoutUpdatesRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWithoutUpdatesRequest$new(query)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWithoutUpdatesRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWithoutUpdatesRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWithoutUpdatesRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWithoutUpdatesRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWithoutUpdatesRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWithoutUpdatesRequest$to_bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWithoutUpdatesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWithoutUpdatesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWithoutUpdatesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{query}}{Field.
Initialize a new InvokeWithoutUpdatesRequest}

\item{\code{class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvokeWithoutUpdatesRequest-new}{\code{InvokeWithoutUpdatesRequest$new()}}
\item \href{#method-InvokeWithoutUpdatesRequest-to_list}{\code{InvokeWithoutUpdatesRequest$to_list()}}
\item \href{#method-InvokeWithoutUpdatesRequest-to_bytes}{\code{InvokeWithoutUpdatesRequest$to_bytes()}}
\item \href{#method-InvokeWithoutUpdatesRequest-clone}{\code{InvokeWithoutUpdatesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWithoutUpdatesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWithoutUpdatesRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWithoutUpdatesRequest$new(query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{Nested query object that responds to \code{to_list} and \code{to_bytes}.
Convert the request to a list

Produces a plain R list suitable for debugging or JSON serialization.
If the nested \code{query} implements \code{to_list()}, that representation
will be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWithoutUpdatesRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWithoutUpdatesRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWithoutUpdatesRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with keys: \code{"_"} and \code{query}.
Serialize the request to raw bytes

The serialization consists of the constructor id followed by the nested
query's serialized bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWithoutUpdatesRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWithoutUpdatesRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWithoutUpdatesRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw vector containing the serialized request bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvokeWithoutUpdatesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvokeWithoutUpdatesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvokeWithoutUpdatesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
