% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_audio}
\alias{is_audio}
\title{Check if File is Audio}
\usage{
is_audio(file)
}
\arguments{
\item{file}{The file path or object to check.}
}
\value{
A logical value indicating whether the file is audio.
}
\description{
Returns `TRUE` if the file has an audio mime type, either from metadata
or guessed from the file extension.
}
