% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_gif}
\alias{is_gif}
\title{Check if File is a GIF}
\usage{
is_gif(file)
}
\arguments{
\item{file}{The file path or object to check.}
}
\value{
A logical value indicating whether the file is a GIF.
}
\description{
Returns `TRUE` if the file extension looks like a GIF file to Telegram.
}
