% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_good_large}
\alias{is_good_large}
\title{Is Good Large}
\usage{
is_good_large(number, p)
}
\arguments{
\item{number}{An integer.}

\item{p}{An integer.}
}
\value{
A logical value.
}
\description{
Checks if a number is good and large relative to p.
}
