% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_good_mod_exp_first}
\alias{is_good_mod_exp_first}
\title{Is Good Mod Exp First}
\usage{
is_good_mod_exp_first(modexp, prime)
}
\arguments{
\item{modexp}{An integer.}

\item{prime}{An integer.}
}
\value{
A logical value.
}
\description{
Checks if modexp is good for modular exponentiation.
}
