% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotosender.R
\name{is_list_like}
\alias{is_list_like}
\title{Check if an object is list-like (list, vector, etc.)}
\usage{
is_list_like(obj, allow_data_frames = FALSE)
}
\arguments{
\item{obj}{The object to check}

\item{allow_data_frames}{Whether to consider data frames as list-like (default: FALSE)}
}
\value{
TRUE if the object is list-like, FALSE otherwise
}
\description{
Determines whether an object can be treated like a list or collection
in the context of MTProto requests.
}
