% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_video}
\alias{is_video}
\title{Check if File is Video}
\usage{
is_video(file)
}
\arguments{
\item{file}{The file path or object to check.}
}
\value{
A logical value indicating whether the file is video.
}
\description{
Returns `TRUE` if the file has a video mime type, either from metadata
or guessed from the file extension.
}
