% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_smsjobs.R
\name{IsEligibleToJoinRequest}
\alias{IsEligibleToJoinRequest}
\title{IsEligibleToJoinRequest R6 class}
\description{
IsEligibleToJoinRequest R6 class

IsEligibleToJoinRequest R6 class
}
\details{
Representation of the IsEligibleToJoinRequest TLRequest.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IsEligibleToJoinRequest-new}{\code{IsEligibleToJoinRequest$new()}}
\item \href{#method-IsEligibleToJoinRequest-to_list}{\code{IsEligibleToJoinRequest$to_list()}}
\item \href{#method-IsEligibleToJoinRequest-bytes}{\code{IsEligibleToJoinRequest$bytes()}}
\item \href{#method-IsEligibleToJoinRequest-from_reader}{\code{IsEligibleToJoinRequest$from_reader()}}
\item \href{#method-IsEligibleToJoinRequest-clone}{\code{IsEligibleToJoinRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsEligibleToJoinRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-IsEligibleToJoinRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an IsEligibleToJoinRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsEligibleToJoinRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsEligibleToJoinRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-IsEligibleToJoinRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsEligibleToJoinRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsEligibleToJoinRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-IsEligibleToJoinRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Get raw bytes for this constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsEligibleToJoinRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsEligibleToJoinRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-IsEligibleToJoinRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsEligibleToJoinRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IsEligibleToJoinRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IsEligibleToJoinRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IsEligibleToJoinRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
