% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_channel.R
\name{join_channel}
\alias{join_channel}
\title{Join a chat or channel by username or invite link}
\usage{
join_channel(client, channel)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{channel}{Username/link or invite link.}
}
\value{
The API response.
}
\description{
Join a chat or channel by username or invite link
}
