% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_channel.R
\name{join_public_channel}
\alias{join_public_channel}
\title{Join a public channel by username}
\usage{
join_public_channel(client, channel)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{channel}{Public channel username or link (non-invite).}
}
\value{
The API response.
}
\description{
Join a public channel by username
}
\examples{
\dontrun{
client <- TelegramClient$new("my_session", api_id = 123, api_hash = "abc")
client$connect()

join_public_channel(client, "channelname")
}
}
