% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chatlists.R
\name{JoinChatlistUpdatesRequest}
\alias{JoinChatlistUpdatesRequest}
\title{JoinChatlistUpdatesRequest R6 class}
\description{
JoinChatlistUpdatesRequest R6 class

JoinChatlistUpdatesRequest R6 class
}
\details{
Request to join a chatlist with specific peers.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Integer constructor id.}

\item{\code{SUBCLASS_OF_ID}}{Integer subclass id.}

\item{\code{chatlist}}{TypeInputChatlist input chatlist object.}

\item{\code{peers}}{List of TypeInputPeer input peers.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JoinChatlistUpdatesRequest-new}{\code{JoinChatlistUpdatesRequest$new()}}
\item \href{#method-JoinChatlistUpdatesRequest-resolve}{\code{JoinChatlistUpdatesRequest$resolve()}}
\item \href{#method-JoinChatlistUpdatesRequest-to_list}{\code{JoinChatlistUpdatesRequest$to_list()}}
\item \href{#method-JoinChatlistUpdatesRequest-bytes}{\code{JoinChatlistUpdatesRequest$bytes()}}
\item \href{#method-JoinChatlistUpdatesRequest-from_reader}{\code{JoinChatlistUpdatesRequest$from_reader()}}
\item \href{#method-JoinChatlistUpdatesRequest-clone}{\code{JoinChatlistUpdatesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinChatlistUpdatesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-JoinChatlistUpdatesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize JoinChatlistUpdatesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinChatlistUpdatesRequest$new(chatlist = NULL, peers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chatlist}}{Input chatlist object}

\item{\code{peers}}{List of input peers
Resolve peer entities to input peers

Iterates over peers, resolves each via client and utils and replaces
the peers field with input_peer objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinChatlistUpdatesRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-JoinChatlistUpdatesRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinChatlistUpdatesRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object that provides get_input_entity}

\item{\code{utils}}{Utilities that provide get_input_peer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self
Convert to list (dictionary-like)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinChatlistUpdatesRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-JoinChatlistUpdatesRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinChatlistUpdatesRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representation
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinChatlistUpdatesRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-JoinChatlistUpdatesRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinChatlistUpdatesRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector with serialized bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinChatlistUpdatesRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-JoinChatlistUpdatesRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinChatlistUpdatesRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinChatlistUpdatesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JoinChatlistUpdatesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinChatlistUpdatesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
