% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{JoinGroupCallPresentationRequest}
\alias{JoinGroupCallPresentationRequest}
\title{JoinGroupCallPresentationRequest R6 class}
\description{
Represents JoinGroupCallPresentationRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{The input group call object (expected to implement bytes()).}

\item{\code{params}}{DataJSON-like object (expected to implement bytes()).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JoinGroupCallPresentationRequest-new}{\code{JoinGroupCallPresentationRequest$new()}}
\item \href{#method-JoinGroupCallPresentationRequest-resolve}{\code{JoinGroupCallPresentationRequest$resolve()}}
\item \href{#method-JoinGroupCallPresentationRequest-to_list}{\code{JoinGroupCallPresentationRequest$to_list()}}
\item \href{#method-JoinGroupCallPresentationRequest-bytes}{\code{JoinGroupCallPresentationRequest$bytes()}}
\item \href{#method-JoinGroupCallPresentationRequest-from_reader}{\code{JoinGroupCallPresentationRequest$from_reader()}}
\item \href{#method-JoinGroupCallPresentationRequest-clone}{\code{JoinGroupCallPresentationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinGroupCallPresentationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-JoinGroupCallPresentationRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinGroupCallPresentationRequest$new(call, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}

\item{\code{params}}{DataJSON-like object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Resolve input references

Typically used to convert high-level entities into input objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinGroupCallPresentationRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-JoinGroupCallPresentationRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinGroupCallPresentationRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
Convert object to a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinGroupCallPresentationRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-JoinGroupCallPresentationRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinGroupCallPresentationRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinGroupCallPresentationRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-JoinGroupCallPresentationRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinGroupCallPresentationRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinGroupCallPresentationRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-JoinGroupCallPresentationRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinGroupCallPresentationRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinGroupCallPresentationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JoinGroupCallPresentationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinGroupCallPresentationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
