% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_smsjobs.R
\name{JoinRequest}
\alias{JoinRequest}
\title{JoinRequest}
\description{
Telegram API type JoinRequest
}
\details{
JoinRequest R6 class

Representation of the JoinRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lock_class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{lock_class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JoinRequest-new}{\code{JoinRequest$new()}}
\item \href{#method-JoinRequest-to_list}{\code{JoinRequest$to_list()}}
\item \href{#method-JoinRequest-bytes}{\code{JoinRequest$bytes()}}
\item \href{#method-JoinRequest-from_reader}{\code{JoinRequest$from_reader()}}
\item \href{#method-JoinRequest-clone}{\code{JoinRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-JoinRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a JoinRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-JoinRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-JoinRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Get raw bytes for this constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-JoinRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JoinRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JoinRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JoinRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
