% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlobject.R
\name{json_default}
\alias{json_default}
\title{Default JSON serialization function.}
\usage{
json_default(value)
}
\arguments{
\item{value}{The value to be serialized.}
}
\value{
A character string representing the serialized value.
}
\description{
This function is used to serialize various data types to JSON format.
It handles raw data, POSIXt objects, and other types.
}
\examples{
json_default(charToRaw("test"))
json_default(as.POSIXct("2023-01-01 12:00:00", tz = "UTC"))
json_default(123)
}
