% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chatlists.R
\name{LeaveChatlistRequest}
\alias{LeaveChatlistRequest}
\title{LeaveChatlistRequest R6 class}
\description{
Request to leave a chatlist for specific peers.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Integer constructor id.}

\item{\code{SUBCLASS_OF_ID}}{Integer subclass id.}

\item{\code{chatlist}}{TypeInputChatlist input chatlist object.}

\item{\code{peers}}{List of TypeInputPeer input peers.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LeaveChatlistRequest-new}{\code{LeaveChatlistRequest$new()}}
\item \href{#method-LeaveChatlistRequest-resolve}{\code{LeaveChatlistRequest$resolve()}}
\item \href{#method-LeaveChatlistRequest-to_list}{\code{LeaveChatlistRequest$to_list()}}
\item \href{#method-LeaveChatlistRequest-bytes}{\code{LeaveChatlistRequest$bytes()}}
\item \href{#method-LeaveChatlistRequest-from_reader}{\code{LeaveChatlistRequest$from_reader()}}
\item \href{#method-LeaveChatlistRequest-clone}{\code{LeaveChatlistRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveChatlistRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveChatlistRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveChatlistRequest$new(chatlist = NULL, peers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chatlist}}{Input chatlist object}

\item{\code{peers}}{List of input peers
Resolve peer entities to input peers}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveChatlistRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveChatlistRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveChatlistRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object that provides get_input_entity}

\item{\code{utils}}{Utilities that provide get_input_peer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self
Convert to list (dictionary-like)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveChatlistRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveChatlistRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveChatlistRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representation
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveChatlistRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveChatlistRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveChatlistRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector with serialized bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveChatlistRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveChatlistRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveChatlistRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveChatlistRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveChatlistRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveChatlistRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
