% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{LeaveGroupCallPresentationRequest}
\alias{LeaveGroupCallPresentationRequest}
\title{LeaveGroupCallPresentationRequest}
\description{
Telegram API type LeaveGroupCallPresentationRequest
}
\details{
LeaveGroupCallPresentationRequest R6 class

Represents LeaveGroupCallPresentationRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{The input group call object (expected to implement bytes()).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LeaveGroupCallPresentationRequest-new}{\code{LeaveGroupCallPresentationRequest$new()}}
\item \href{#method-LeaveGroupCallPresentationRequest-resolve}{\code{LeaveGroupCallPresentationRequest$resolve()}}
\item \href{#method-LeaveGroupCallPresentationRequest-to_list}{\code{LeaveGroupCallPresentationRequest$to_list()}}
\item \href{#method-LeaveGroupCallPresentationRequest-bytes}{\code{LeaveGroupCallPresentationRequest$bytes()}}
\item \href{#method-LeaveGroupCallPresentationRequest-from_reader}{\code{LeaveGroupCallPresentationRequest$from_reader()}}
\item \href{#method-LeaveGroupCallPresentationRequest-clone}{\code{LeaveGroupCallPresentationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveGroupCallPresentationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveGroupCallPresentationRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a LeaveGroupCallPresentationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveGroupCallPresentationRequest$new(call)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveGroupCallPresentationRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveGroupCallPresentationRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references

Typically used to convert high-level entities into input objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveGroupCallPresentationRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveGroupCallPresentationRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveGroupCallPresentationRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveGroupCallPresentationRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveGroupCallPresentationRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveGroupCallPresentationRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes `call$bytes()` exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveGroupCallPresentationRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveGroupCallPresentationRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveGroupCallPresentationRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveGroupCallPresentationRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveGroupCallPresentationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveGroupCallPresentationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveGroupCallPresentationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
