% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_smsjobs.R
\name{LeaveRequest}
\alias{LeaveRequest}
\title{LeaveRequest R6 class}
\description{
Representation of the LeaveRequest TLRequest.
}
\details{
Methods:
- new(): create instance
- to_list(): return a list representation
- bytes(): return raw bytes for the TL constructor
- from_reader(reader): class method to construct from a reader
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LeaveRequest-new}{\code{LeaveRequest$new()}}
\item \href{#method-LeaveRequest-to_list}{\code{LeaveRequest$to_list()}}
\item \href{#method-LeaveRequest-bytes}{\code{LeaveRequest$bytes()}}
\item \href{#method-LeaveRequest-from_reader}{\code{LeaveRequest$from_reader()}}
\item \href{#method-LeaveRequest-clone}{\code{LeaveRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
Convert to list (dictionary-like)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Get raw bytes for this constructor
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeaveRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LeaveRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeaveRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
