% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listenTextMessage.R
\name{listenTextMessage}
\alias{listenTextMessage}
\title{listenTextMessage}
\usage{
listenTextMessage(token, chat_id, text_message)
}
\arguments{
\item{token}{Your bot's API token.}
}
\description{
Send a message to a chatId
}
\examples{
\dontrun{
postTextReply(token = token, chat_id = "475093487", text_message = "Hi there!")
}

}
