% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logg}
\alias{logg}
\title{Logging setup
A simple logging function that prints formatted messages to the console.}
\usage{
logg(...)
}
\arguments{
\item{...}{Arguments passed to sprintf for formatting the log message.}
}
\value{
None. Prints the formatted message to the console.
}
\description{
Logging setup
A simple logging function that prints formatted messages to the console.
}
\examples{
logg("This is a log message with a number: \%d", 42)
}
