% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{LogOutRequest}
\alias{LogOutRequest}
\title{LogOutRequest R6 class}
\description{
Represents the TLRequest auth.LogOutRequest.
}
\details{
Fields:
}
\section{initialize}{

 Create a new LogOutRequest object.
 @return invisible self
}

\section{to_list}{

 Return a pure R list representation of the object.
 @return list
}

\section{to_bytes}{

 Serialize the object into raw bytes for TL. Uses serialize_bytes helper
 from TL base environment.
 @return raw vector
}

\section{from_reader}{

 Factory that reads fields from a reader and returns a new LogOutRequest instance.
 @param reader Reader object implementing read_int(), tgread_string(), tgread_object()
 @return LogOutRequest instance
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0x3e72ba19).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0x0a804315).

Methods:}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogOutRequest-new}{\code{LogOutRequest$new()}}
\item \href{#method-LogOutRequest-to_list}{\code{LogOutRequest$to_list()}}
\item \href{#method-LogOutRequest-to_bytes}{\code{LogOutRequest$to_bytes()}}
\item \href{#method-LogOutRequest-from_reader}{\code{LogOutRequest$from_reader()}}
\item \href{#method-LogOutRequest-clone}{\code{LogOutRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogOutRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-LogOutRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogOutRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible self
Convert to an R list (similar to to_dict)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogOutRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-LogOutRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogOutRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogOutRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-LogOutRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogOutRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
Read from reader and create LogOutRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogOutRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-LogOutRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogOutRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
LogOutRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogOutRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogOutRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogOutRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
