% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotostate.R
\docType{data}
\name{MAX_RECENT_MSG_IDS}
\alias{MAX_RECENT_MSG_IDS}
\title{Constants}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_RECENT_MSG_IDS
}
\description{
Constants
}
\keyword{datasets}
