% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Message}
\alias{Message}
\title{Message}
\description{
Telegram API type Message
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{Message}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Message-new}{\code{Message$new()}}
\item \href{#method-Message-to_dict}{\code{Message$to_dict()}}
\item \href{#method-Message-bytes}{\code{Message$bytes()}}
\item \href{#method-Message-clone}{\code{Message$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-new"></a>}}
\if{latex}{\out{\hypertarget{method-Message-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$new(
  id,
  peer_id,
  date,
  message,
  out = NULL,
  mentioned = NULL,
  media_unread = NULL,
  silent = NULL,
  post = NULL,
  from_scheduled = NULL,
  legacy = NULL,
  edit_hide = NULL,
  pinned = NULL,
  noforwards = NULL,
  invert_media = NULL,
  offline = NULL,
  video_processing_pending = NULL,
  paid_suggested_post_stars = NULL,
  paid_suggested_post_ton = NULL,
  from_id = NULL,
  from_boosts_applied = NULL,
  saved_peer_id = NULL,
  fwd_from = NULL,
  via_bot_id = NULL,
  via_business_bot_id = NULL,
  reply_to = NULL,
  media = NULL,
  reply_markup = NULL,
  entities = NULL,
  views = NULL,
  forwards = NULL,
  replies = NULL,
  edit_date = NULL,
  post_author = NULL,
  grouped_id = NULL,
  reactions = NULL,
  restriction_reason = NULL,
  ttl_period = NULL,
  quick_reply_shortcut_id = NULL,
  effect = NULL,
  factcheck = NULL,
  report_delivery_until_date = NULL,
  paid_message_stars = NULL,
  suggested_post = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-Message-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-Message-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Message-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
