% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{MessageActionStarGiftUnique}
\alias{MessageActionStarGiftUnique}
\title{MessageActionStarGiftUnique}
\description{
Telegram API type MessageActionStarGiftUnique
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{MessageActionStarGiftUnique}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gift}}{Field.}

\item{\code{upgrade}}{Field.}

\item{\code{transferred}}{Field.}

\item{\code{saved}}{Field.}

\item{\code{refunded}}{Field.}

\item{\code{prepaid_upgrade}}{Field.}

\item{\code{can_export_at}}{Field.}

\item{\code{transfer_stars}}{Field.}

\item{\code{from_id}}{Field.}

\item{\code{peer}}{Field.}

\item{\code{saved_id}}{Field.}

\item{\code{resale_amount}}{Field.}

\item{\code{can_transfer_at}}{Field.}

\item{\code{can_resell_at}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MessageActionStarGiftUnique-new}{\code{MessageActionStarGiftUnique$new()}}
\item \href{#method-MessageActionStarGiftUnique-to_dict}{\code{MessageActionStarGiftUnique$to_dict()}}
\item \href{#method-MessageActionStarGiftUnique-bytes}{\code{MessageActionStarGiftUnique$bytes()}}
\item \href{#method-MessageActionStarGiftUnique-clone}{\code{MessageActionStarGiftUnique$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageActionStarGiftUnique-new"></a>}}
\if{latex}{\out{\hypertarget{method-MessageActionStarGiftUnique-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageActionStarGiftUnique$new(
  gift,
  upgrade = NULL,
  transferred = NULL,
  saved = NULL,
  refunded = NULL,
  prepaid_upgrade = NULL,
  can_export_at = NULL,
  transfer_stars = NULL,
  from_id = NULL,
  peer = NULL,
  saved_id = NULL,
  resale_amount = NULL,
  can_transfer_at = NULL,
  can_resell_at = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageActionStarGiftUnique-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-MessageActionStarGiftUnique-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageActionStarGiftUnique$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageActionStarGiftUnique-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-MessageActionStarGiftUnique-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageActionStarGiftUnique$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageActionStarGiftUnique-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MessageActionStarGiftUnique-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageActionStarGiftUnique$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
