% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{MessageFwdHeader}
\alias{MessageFwdHeader}
\title{MessageFwdHeader}
\description{
Telegram API type MessageFwdHeader
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{MessageFwdHeader}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{date}}{Field.}

\item{\code{imported}}{Field.}

\item{\code{saved_out}}{Field.}

\item{\code{from_id}}{Field.}

\item{\code{from_name}}{Field.}

\item{\code{channel_post}}{Field.}

\item{\code{post_author}}{Field.}

\item{\code{saved_frompeer}}{Field.}

\item{\code{saved_from_msg_id}}{Field.}

\item{\code{saved_from_id}}{Field.}

\item{\code{saved_from_name}}{Field.}

\item{\code{saved_date}}{Field.}

\item{\code{psa_type}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MessageFwdHeader-new}{\code{MessageFwdHeader$new()}}
\item \href{#method-MessageFwdHeader-to_dict}{\code{MessageFwdHeader$to_dict()}}
\item \href{#method-MessageFwdHeader-bytes}{\code{MessageFwdHeader$bytes()}}
\item \href{#method-MessageFwdHeader-clone}{\code{MessageFwdHeader$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageFwdHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-MessageFwdHeader-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageFwdHeader$new(
  date = NULL,
  imported = NULL,
  saved_out = NULL,
  from_id = NULL,
  from_name = NULL,
  channel_post = NULL,
  post_author = NULL,
  saved_frompeer = NULL,
  saved_from_msg_id = NULL,
  saved_from_id = NULL,
  saved_from_name = NULL,
  saved_date = NULL,
  psa_type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageFwdHeader-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-MessageFwdHeader-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageFwdHeader$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageFwdHeader-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-MessageFwdHeader-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageFwdHeader$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageFwdHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MessageFwdHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageFwdHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
