% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{MessageReplyHeader}
\alias{MessageReplyHeader}
\title{MessageReplyHeader}
\description{
Telegram API type MessageReplyHeader
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{MessageReplyHeader}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{reply_to_scheduled}}{Field.}

\item{\code{forum_topic}}{Field.}

\item{\code{quote}}{Field.}

\item{\code{reply_to_msg_id}}{Field.}

\item{\code{reply_topeer_id}}{Field.}

\item{\code{reply_from}}{Field.}

\item{\code{reply_media}}{Field.}

\item{\code{reply_to_top_id}}{Field.}

\item{\code{quote_text}}{Field.}

\item{\code{quote_entities}}{Field.}

\item{\code{quote_offset}}{Field.}

\item{\code{todo_item_id}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MessageReplyHeader-new}{\code{MessageReplyHeader$new()}}
\item \href{#method-MessageReplyHeader-to_dict}{\code{MessageReplyHeader$to_dict()}}
\item \href{#method-MessageReplyHeader-bytes}{\code{MessageReplyHeader$bytes()}}
\item \href{#method-MessageReplyHeader-clone}{\code{MessageReplyHeader$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageReplyHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-MessageReplyHeader-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageReplyHeader$new(
  reply_to_scheduled = NULL,
  forum_topic = NULL,
  quote = NULL,
  reply_to_msg_id = NULL,
  reply_topeer_id = NULL,
  reply_from = NULL,
  reply_media = NULL,
  reply_to_top_id = NULL,
  quote_text = NULL,
  quote_entities = NULL,
  quote_offset = NULL,
  todo_item_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageReplyHeader-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-MessageReplyHeader-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageReplyHeader$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageReplyHeader-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-MessageReplyHeader-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageReplyHeader$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageReplyHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MessageReplyHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageReplyHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
