% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\docType{data}
\name{MIN_CHUNK_SIZE}
\alias{MIN_CHUNK_SIZE}
\title{Constants for file operations}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_CHUNK_SIZE
}
\description{
Constants for file operations
}
\keyword{datasets}
