% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentificator.R
\name{modexp}
\alias{modexp}
\title{Modular exponentiation.}
\usage{
modexp(base, exp, mod)
}
\arguments{
\item{base}{The base value.}

\item{exp}{The exponent.}

\item{mod}{The modulus.}
}
\value{
The result of \code{base^exp \% mod \%}.
}
\description{
Modular exponentiation.
}
