% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotoplainsender.R
\name{MTProtoPlainSender}
\alias{MTProtoPlainSender}
\title{MTProtoPlainSender}
\description{
Telegram API type MTProtoPlainSender
}
\details{
This module contains the class used to communicate with Telegram's servers
in plain text, when no authorization key has been created yet.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MTProtoPlainSender-new}{\code{MTProtoPlainSender$new()}}
\item \href{#method-MTProtoPlainSender-send}{\code{MTProtoPlainSender$send()}}
\item \href{#method-MTProtoPlainSender-clone}{\code{MTProtoPlainSender$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTProtoPlainSender-new"></a>}}
\if{latex}{\out{\hypertarget{method-MTProtoPlainSender-new}{}}}
\subsection{Method \code{new()}}{
Initializes the MTProto plain sender.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTProtoPlainSender$new(connection, loggers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{the Connection to be used.}

\item{\code{loggers}}{Optional loggers map.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTProtoPlainSender-send"></a>}}
\if{latex}{\out{\hypertarget{method-MTProtoPlainSender-send}{}}}
\subsection{Method \code{send()}}{
Sends and receives the result for the given request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTProtoPlainSender$send(request)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request}}{The request to send.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The response object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTProtoPlainSender-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MTProtoPlainSender-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTProtoPlainSender$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
