% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpmtproxy.R
\name{MTProxyIO}
\alias{MTProxyIO}
\title{MTProxyIO}
\description{
Handles MTProxy header initialization and I/O encryption/decryption.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{header}}{The MTProxy header.}

\item{\code{encryptor}}{The AESModeCTR encryptor.}

\item{\code{decryptor}}{The AESModeCTR decryptor.}

\item{\code{reader}}{The reader object.}

\item{\code{writer}}{The writer object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MTProxyIO-new}{\code{MTProxyIO$new()}}
\item \href{#method-MTProxyIO-readexactly}{\code{MTProxyIO$readexactly()}}
\item \href{#method-MTProxyIO-write}{\code{MTProxyIO$write()}}
\item \href{#method-MTProxyIO-clone}{\code{MTProxyIO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTProxyIO-new"></a>}}
\if{latex}{\out{\hypertarget{method-MTProxyIO-new}{}}}
\subsection{Method \code{new()}}{
Initialize MTProxyIO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTProxyIO$new(connection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{A connection object with fields: reader, writer, secret, dc_id, and packet_codec.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTProxyIO-readexactly"></a>}}
\if{latex}{\out{\hypertarget{method-MTProxyIO-readexactly}{}}}
\subsection{Method \code{readexactly()}}{
Asynchronously read exactly n bytes and decrypt them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTProxyIO$readexactly(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of bytes to read.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A raw vector with decrypted data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTProxyIO-write"></a>}}
\if{latex}{\out{\hypertarget{method-MTProxyIO-write}{}}}
\subsection{Method \code{write()}}{
Encrypt and write data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTProxyIO$write(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A raw vector to write.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTProxyIO-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MTProxyIO-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTProxyIO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
