% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{MultiError}
\alias{MultiError}
\title{MultiError}
\description{
Exception container for multiple `TLRequest`s.
}
\section{Super class}{
\code{\link[telegramR:Exception]{telegramR::Exception}} -> \code{MultiError}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{exceptions}}{\code{list} A list of exceptions.}

\item{\code{results}}{\code{list} A list of results.}

\item{\code{requests}}{\code{list} A list of requests.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiError-new}{\code{MultiError$new()}}
\item \href{#method-MultiError-clone}{\code{MultiError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiError-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiError-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the MultiError class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiError$new(exceptions, results, requests)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exceptions}}{\code{list} A list of exceptions.}

\item{\code{results}}{\code{list} A list of results.}

\item{\code{requests}}{\code{list} A list of requests.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
