% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newEventRouter.R
\name{newEventRouter}
\alias{newEventRouter}
\title{newEventRouter}
\usage{
newEventRouter(chat_id, message_text)
}
\arguments{
\item{chat_id}{A chat id.}

\item{token}{Your bot's API token.}

\item{text_message}{The text message you want to send.}
}
\description{
Route a new message to the appropriate service
}
\examples{
\dontrun{
newEventRouter(chat_id, message_text)
}

}
