% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegrambaseclient.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{The first value}

\item{y}{The second value}
}
\value{
The first value if not NULL, otherwise the second value
}
\description{
Null coalescing operator
}
