% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_bytes_for_hash}
\alias{num_bytes_for_hash}
\title{Num Bytes for Hash}
\usage{
num_bytes_for_hash(number)
}
\arguments{
\item{number}{A raw vector.}
}
\value{
A raw vector.
}
\description{
Num Bytes for Hash
}
