% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{ObfuscatedConnection}
\alias{ObfuscatedConnection}
\title{ObfuscatedConnection Class}
\description{
Base class for obfuscated connections (e.g., obfuscated2, mtproto proxy).
}
\section{Super class}{
\code{\link[telegramR:Connection]{telegramR::Connection}} -> \code{ObfuscatedConnection}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{obfuscated_io}}{The obfuscation class to use.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObfuscatedConnection-clone}{\code{ObfuscatedConnection$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="connect"><a href='../../telegramR/html/Connection.html#method-Connection-connect'><code>telegramR::Connection$connect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="disconnect"><a href='../../telegramR/html/Connection.html#method-Connection-disconnect'><code>telegramR::Connection$disconnect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="initialize"><a href='../../telegramR/html/Connection.html#method-Connection-initialize'><code>telegramR::Connection$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="recv"><a href='../../telegramR/html/Connection.html#method-Connection-recv'><code>telegramR::Connection$recv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="Connection" data-id="send"><a href='../../telegramR/html/Connection.html#method-Connection-send'><code>telegramR::Connection$send()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObfuscatedConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObfuscatedConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObfuscatedConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
