% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa.R
\name{odcstring}
\alias{odcstring}
\title{Converts a raw vector to an odd-length string.}
\usage{
odcstring(raw_data)
}
\arguments{
\item{raw_data}{The raw vector to convert.}
}
\value{
A string with odd-length characters.
}
\description{
Converts a raw vector to an odd-length string.
}
