% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotostate.R
\name{OpaqueRequest}
\alias{OpaqueRequest}
\title{Opaque Request Class}
\description{
Opaque Request Class

Opaque Request Class
}
\details{
Wraps a serialized request into a type that can be serialized again.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Raw vector of serialized request data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OpaqueRequest-new}{\code{OpaqueRequest$new()}}
\item \href{#method-OpaqueRequest-to_bytes}{\code{OpaqueRequest$to_bytes()}}
\item \href{#method-OpaqueRequest-clone}{\code{OpaqueRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpaqueRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OpaqueRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize with serialized request data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpaqueRequest$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw vector of serialized request data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpaqueRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-OpaqueRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Convert to bytes representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpaqueRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw vector of the wrapped data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpaqueRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OpaqueRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpaqueRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
