% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pack_bot_file_id}
\alias{pack_bot_file_id}
\title{Pack Bot File ID}
\usage{
pack_bot_file_id(file)
}
\arguments{
\item{file}{A Document or Photo object (or MessageMediaDocument/MessageMediaPhoto wrapper).}
}
\value{
A base64-encoded string representing the file ID, or NULL if invalid.
}
\description{
Inverse operation for `resolve_bot_file_id`. This function takes a Document or Photo
object and returns a variable-length base64-encoded `file_id` string. If an invalid
parameter is given, it returns NULL.
}
