% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{PacketCodec}
\alias{PacketCodec}
\title{PacketCodec Class}
\description{
Abstract base class for packet codecs.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PacketCodec-new}{\code{PacketCodec$new()}}
\item \href{#method-PacketCodec-encode_packet}{\code{PacketCodec$encode_packet()}}
\item \href{#method-PacketCodec-read_packet}{\code{PacketCodec$read_packet()}}
\item \href{#method-PacketCodec-clone}{\code{PacketCodec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PacketCodec-new"></a>}}
\if{latex}{\out{\hypertarget{method-PacketCodec-new}{}}}
\subsection{Method \code{new()}}{
Initialize a PacketCodec instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PacketCodec$new(connection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{The associated Connection object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PacketCodec-encode_packet"></a>}}
\if{latex}{\out{\hypertarget{method-PacketCodec-encode_packet}{}}}
\subsection{Method \code{encode_packet()}}{
Encode a packet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PacketCodec$encode_packet(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to encode.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Encoded bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PacketCodec-read_packet"></a>}}
\if{latex}{\out{\hypertarget{method-PacketCodec-read_packet}{}}}
\subsection{Method \code{read_packet()}}{
Read a packet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PacketCodec$read_packet(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{An object with a readexactly method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A promise resolving with the packet.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PacketCodec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PacketCodec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PacketCodec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
