% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotoplainsender.R
\name{packInt32}
\alias{packInt32}
\title{Pack a 32-bit integer into a raw vector (little-endian)}
\usage{
packInt32(value)
}
\arguments{
\item{value}{The integer value to pack}
}
\value{
A raw vector representation of the integer
}
\description{
Pack a 32-bit integer into a raw vector (little-endian)
}
