% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotostate.R
\name{packRandomBytes}
\alias{packRandomBytes}
\title{Pack random bytes into a raw vector}
\usage{
packRandomBytes(n)
}
\arguments{
\item{n}{Number of random bytes to generate}
}
\value{
Raw vector of random bytes
}
\description{
Pack random bytes into a raw vector
}
