% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotostate.R
\name{packRandomLong}
\alias{packRandomLong}
\title{Generate random 64-bit value bytes (little-endian)}
\usage{
packRandomLong()
}
\value{
Raw(8) random bytes
}
\description{
Generate random 64-bit value bytes (little-endian)
}
