% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chat_history.R
\name{parse_chat_history}
\alias{parse_chat_history}
\title{Parse Telegram chat history files (WIP)}
\usage{
parse_chat_history(chat_dir)
}
\arguments{
\item{chat_dir}{Directory where the Chat History was saved}
}
\value{
lines of messages
}
\description{
Parse Telegram chat history files (WIP)
}
\examples{
parse_chat_history("chat_history")
}
