% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{parse_html_to_telegram}
\alias{parse_html_to_telegram}
\title{Parse HTML into plain text and telegram-style entities}
\usage{
parse_html_to_telegram(html)
}
\arguments{
\item{html}{character(1) HTML fragment}
}
\value{
named list with elements \code{text} and \code{entities}
}
\description{
Parse HTML into plain text and telegram-style entities
}
