% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_phone}
\alias{parse_phone}
\title{Parse Phone Number}
\usage{
parse_phone(phone)
}
\arguments{
\item{phone}{The phone number to parse, as an integer or string.}
}
\value{
A character string of the parsed phone number, or `NULL` if invalid.
}
\description{
Parses the given phone number, removing non-digit characters, and returns it as a string
if it consists only of digits. Returns `NULL` if invalid.
}
\examples{
\dontrun{
parse_phone(1234567890) # "1234567890"
parse_phone("+1 (234) 567-890") # "1234567890"
parse_phone("invalid") # NULL
}
}
