% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pbkdf2sha512}
\alias{pbkdf2sha512}
\title{PBKDF2 SHA512}
\usage{
pbkdf2sha512(password, salt, iterations)
}
\arguments{
\item{password}{A raw vector.}

\item{salt}{A raw vector.}

\item{iterations}{An integer.}
}
\value{
A raw vector.
}
\description{
Computes PBKDF2 with SHA512.
}
