% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{PeerSettings}
\alias{PeerSettings}
\title{PeerSettings}
\description{
Telegram API type PeerSettings
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{PeerSettings}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{report_spam}}{Field.}

\item{\code{add_contact}}{Field.}

\item{\code{block_contact}}{Field.}

\item{\code{share_contact}}{Field.}

\item{\code{need_contacts_exception}}{Field.}

\item{\code{report_geo}}{Field.}

\item{\code{autoarchived}}{Field.}

\item{\code{invite_members}}{Field.}

\item{\code{request_chat_broadcast}}{Field.}

\item{\code{business_bot_paused}}{Field.}

\item{\code{business_bot_can_reply}}{Field.}

\item{\code{geo_distance}}{Field.}

\item{\code{request_chat_title}}{Field.}

\item{\code{request_chat_date}}{Field.}

\item{\code{business_bot_id}}{Field.}

\item{\code{business_bot_manage_url}}{Field.}

\item{\code{charge_paid_message_stars}}{Field.}

\item{\code{registration_month}}{Field.}

\item{\code{phone_country}}{Field.}

\item{\code{name_change_date}}{Field.}

\item{\code{photo_change_date}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PeerSettings-new}{\code{PeerSettings$new()}}
\item \href{#method-PeerSettings-to_dict}{\code{PeerSettings$to_dict()}}
\item \href{#method-PeerSettings-bytes}{\code{PeerSettings$bytes()}}
\item \href{#method-PeerSettings-clone}{\code{PeerSettings$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeerSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-PeerSettings-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeerSettings$new(
  report_spam = NULL,
  add_contact = NULL,
  block_contact = NULL,
  share_contact = NULL,
  need_contacts_exception = NULL,
  report_geo = NULL,
  autoarchived = NULL,
  invite_members = NULL,
  request_chat_broadcast = NULL,
  business_bot_paused = NULL,
  business_bot_can_reply = NULL,
  geo_distance = NULL,
  request_chat_title = NULL,
  request_chat_date = NULL,
  business_bot_id = NULL,
  business_bot_manage_url = NULL,
  charge_paid_message_stars = NULL,
  registration_month = NULL,
  phone_country = NULL,
  name_change_date = NULL,
  photo_change_date = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeerSettings-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-PeerSettings-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeerSettings$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeerSettings-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-PeerSettings-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeerSettings$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeerSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PeerSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeerSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
