% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pow}
\alias{pow}
\title{Modular Exponentiation (Alias)}
\usage{
pow(base, exp, mod)
}
\arguments{
\item{base}{The base.}

\item{exp}{The exponent.}

\item{mod}{The modulus.}
}
\value{
The result of (base^exp) %% mod as bigz.
}
\description{
Modular Exponentiation (Alias)
}
