% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{PremiumSubscriptionOption}
\alias{PremiumSubscriptionOption}
\title{PremiumSubscriptionOption}
\description{
Telegram API type PremiumSubscriptionOption
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{PremiumSubscriptionOption}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{months}}{Field.}

\item{\code{currency}}{Field.}

\item{\code{amount}}{Field.}

\item{\code{bot_url}}{Field.}

\item{\code{current}}{Field.}

\item{\code{can_purchase_upgrade}}{Field.}

\item{\code{transaction}}{Field.}

\item{\code{store_product}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PremiumSubscriptionOption-new}{\code{PremiumSubscriptionOption$new()}}
\item \href{#method-PremiumSubscriptionOption-to_dict}{\code{PremiumSubscriptionOption$to_dict()}}
\item \href{#method-PremiumSubscriptionOption-bytes}{\code{PremiumSubscriptionOption$bytes()}}
\item \href{#method-PremiumSubscriptionOption-clone}{\code{PremiumSubscriptionOption$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PremiumSubscriptionOption-new"></a>}}
\if{latex}{\out{\hypertarget{method-PremiumSubscriptionOption-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PremiumSubscriptionOption$new(
  months = NULL,
  currency = NULL,
  amount = NULL,
  bot_url = NULL,
  current = NULL,
  can_purchase_upgrade = NULL,
  transaction = NULL,
  store_product = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PremiumSubscriptionOption-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-PremiumSubscriptionOption-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PremiumSubscriptionOption$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PremiumSubscriptionOption-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-PremiumSubscriptionOption-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PremiumSubscriptionOption$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PremiumSubscriptionOption-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PremiumSubscriptionOption-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PremiumSubscriptionOption$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
