% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raise_cast_fail}
\alias{raise_cast_fail}
\title{Raise Cast Fail}
\usage{
raise_cast_fail(entity, target)
}
\arguments{
\item{entity}{The entity object that failed to cast.}

\item{target}{A character string representing the target type.}
}
\description{
Raises an error indicating that the entity cannot be cast to the target type.
}
\examples{
\dontrun{
\dontrun{
raise_cast_fail(some_entity, "InputPeer")
}
}
}
