% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{ReadCancelledError}
\alias{ReadCancelledError}
\title{ReadCancelledError}
\description{
Telegram API type ReadCancelledError
}
\details{
ReadCancelledError
}
\section{Super class}{
\code{\link[telegramR:Exception]{telegramR::Exception}} -> \code{ReadCancelledError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReadCancelledError-new}{\code{ReadCancelledError$new()}}
\item \href{#method-ReadCancelledError-clone}{\code{ReadCancelledError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadCancelledError-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReadCancelledError-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the ReadCancelledError class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadCancelledError$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{\code{character} The error message.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadCancelledError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReadCancelledError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadCancelledError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
