% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{Reader}
\alias{Reader}
\title{Reader Class}
\description{
Implements a simple reader.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{socket}}{Field.}

\item{\code{timeout}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Reader-readexactly}{\code{Reader$readexactly()}}
\item \href{#method-Reader-clone}{\code{Reader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reader-readexactly"></a>}}
\if{latex}{\out{\hypertarget{method-Reader-readexactly}{}}}
\subsection{Method \code{readexactly()}}{
Read a fixed number of bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reader$readexactly(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of bytes to read.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A promise resolving with raw data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Reader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
