% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_payments.R
\name{RefundStarsChargeRequest}
\alias{RefundStarsChargeRequest}
\title{RefundStarsChargeRequest}
\description{
Represents a request to refund a stars charge. Inherits from TLRequest.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{RefundStarsChargeRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{The constructor ID for this request.}

\item{\code{SUBCLASS_OF_ID}}{The subclass ID.}

\item{\code{user_id}}{The input user.}

\item{\code{charge_id}}{The charge ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RefundStarsChargeRequest-new}{\code{RefundStarsChargeRequest$new()}}
\item \href{#method-RefundStarsChargeRequest-resolve}{\code{RefundStarsChargeRequest$resolve()}}
\item \href{#method-RefundStarsChargeRequest-to_dict}{\code{RefundStarsChargeRequest$to_dict()}}
\item \href{#method-RefundStarsChargeRequest-bytes}{\code{RefundStarsChargeRequest$bytes()}}
\item \href{#method-RefundStarsChargeRequest-from_reader}{\code{RefundStarsChargeRequest$from_reader()}}
\item \href{#method-RefundStarsChargeRequest-clone}{\code{RefundStarsChargeRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefundStarsChargeRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-RefundStarsChargeRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the RefundStarsChargeRequest object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefundStarsChargeRequest$new(user_id, charge_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{The input user (TypeInputUser).}

\item{\code{charge_id}}{The charge ID (string).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefundStarsChargeRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-RefundStarsChargeRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve the user_id using client and utils.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefundStarsChargeRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The client object.}

\item{\code{utils}}{The utils object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefundStarsChargeRequest-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-RefundStarsChargeRequest-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Convert the object to a dictionary-like list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefundStarsChargeRequest$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object in dictionary form.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefundStarsChargeRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-RefundStarsChargeRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize the object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefundStarsChargeRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector of bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefundStarsChargeRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-RefundStarsChargeRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Deserialize from a reader.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefundStarsChargeRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{The reader object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of RefundStarsChargeRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefundStarsChargeRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RefundStarsChargeRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefundStarsChargeRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
