% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{ReorderPreviewMediasRequest}
\alias{ReorderPreviewMediasRequest}
\title{ReorderPreviewMediasRequest}
\description{
Methods:
- initialize(bot, lang_code, order)
- resolve(client, utils)
- to_list() -> list representation
- to_bytes() -> raw vector bytes

Each method is documented inline below.

Telegram API type ReorderPreviewMediasRequest
}
\details{
ReorderPreviewMediasRequest R6 class

Represents the ReorderPreviewMediasRequest TL request.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bot}}{TypeInputUser input user or object implementing to_list()/to_bytes()}

\item{\code{lang_code}}{character language code}

\item{\code{order}}{list of TypeInputMedia objects or representations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReorderPreviewMediasRequest-new}{\code{ReorderPreviewMediasRequest$new()}}
\item \href{#method-ReorderPreviewMediasRequest-resolve}{\code{ReorderPreviewMediasRequest$resolve()}}
\item \href{#method-ReorderPreviewMediasRequest-to_list}{\code{ReorderPreviewMediasRequest$to_list()}}
\item \href{#method-ReorderPreviewMediasRequest-to_bytes}{\code{ReorderPreviewMediasRequest$to_bytes()}}
\item \href{#method-ReorderPreviewMediasRequest-clone}{\code{ReorderPreviewMediasRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReorderPreviewMediasRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReorderPreviewMediasRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize ReorderPreviewMediasRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReorderPreviewMediasRequest$new(bot, lang_code, order)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bot}}{TypeInputUser or identifier}

\item{\code{lang_code}}{character}

\item{\code{order}}{list of TypeInputMedia or objects
Resolve references (convert entities via client/utils)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReorderPreviewMediasRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-ReorderPreviewMediasRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReorderPreviewMediasRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client with get_input_entity method}

\item{\code{utils}}{utils with get_input_user / get_input_media}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
Convert to list (dictionary-like)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReorderPreviewMediasRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ReorderPreviewMediasRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReorderPreviewMediasRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReorderPreviewMediasRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ReorderPreviewMediasRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReorderPreviewMediasRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReorderPreviewMediasRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReorderPreviewMediasRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReorderPreviewMediasRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
