% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{ReqPqMultiRequest_from_reader}
\alias{ReqPqMultiRequest_from_reader}
\title{Create ReqPqMultiRequest from a binary reader}
\usage{
ReqPqMultiRequest_from_reader(reader)
}
\arguments{
\item{reader}{Reader object providing `read_large_int(bits = 128)`.}
}
\value{
A new instance of `ReqPqMultiRequest`.
}
\description{
Reads a 128-bit `nonce` from the provided `reader` and constructs a
`ReqPqMultiRequest` instance.
}
