% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{RequestPasswordRecoveryRequest}
\alias{RequestPasswordRecoveryRequest}
\title{RequestPasswordRecoveryRequest}
\description{
Telegram API type RequestPasswordRecoveryRequest
}
\details{
RequestPasswordRecoveryRequest R6 class

Represents the TLRequest auth.RequestPasswordRecoveryRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0xd897bc66).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0xfa72d43a).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RequestPasswordRecoveryRequest-new}{\code{RequestPasswordRecoveryRequest$new()}}
\item \href{#method-RequestPasswordRecoveryRequest-to_list}{\code{RequestPasswordRecoveryRequest$to_list()}}
\item \href{#method-RequestPasswordRecoveryRequest-to_bytes}{\code{RequestPasswordRecoveryRequest$to_bytes()}}
\item \href{#method-RequestPasswordRecoveryRequest-from_reader}{\code{RequestPasswordRecoveryRequest$from_reader()}}
\item \href{#method-RequestPasswordRecoveryRequest-clone}{\code{RequestPasswordRecoveryRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestPasswordRecoveryRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-RequestPasswordRecoveryRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a RequestPasswordRecoveryRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestPasswordRecoveryRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestPasswordRecoveryRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-RequestPasswordRecoveryRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestPasswordRecoveryRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestPasswordRecoveryRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-RequestPasswordRecoveryRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestPasswordRecoveryRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestPasswordRecoveryRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-RequestPasswordRecoveryRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create RequestPasswordRecoveryRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestPasswordRecoveryRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RequestPasswordRecoveryRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestPasswordRecoveryRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RequestPasswordRecoveryRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestPasswordRecoveryRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
