% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_messages.R
\name{RequestSimpleWebViewRequest}
\alias{RequestSimpleWebViewRequest}
\title{RequestSimpleWebViewRequest}
\description{
Represents a request to request a simple web view for a bot. This class inherits from TLRequest.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{RequestSimpleWebViewRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{The constructor ID.}

\item{\code{SUBCLASS_OF_ID}}{The subclass ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RequestSimpleWebViewRequest-new}{\code{RequestSimpleWebViewRequest$new()}}
\item \href{#method-RequestSimpleWebViewRequest-resolve}{\code{RequestSimpleWebViewRequest$resolve()}}
\item \href{#method-RequestSimpleWebViewRequest-to_dict}{\code{RequestSimpleWebViewRequest$to_dict()}}
\item \href{#method-RequestSimpleWebViewRequest-bytes}{\code{RequestSimpleWebViewRequest$bytes()}}
\item \href{#method-RequestSimpleWebViewRequest-clone}{\code{RequestSimpleWebViewRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestSimpleWebViewRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-RequestSimpleWebViewRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the RequestSimpleWebViewRequest object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestSimpleWebViewRequest$new(
  bot,
  platform,
  from_switch_webview = NULL,
  from_side_menu = NULL,
  compact = NULL,
  fullscreen = NULL,
  url = NULL,
  start_param = NULL,
  theme_params = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bot}}{The input user (bot).}

\item{\code{platform}}{The platform string.}

\item{\code{from_switch_webview}}{Optional from switch webview flag.}

\item{\code{from_side_menu}}{Optional from side menu flag.}

\item{\code{compact}}{Optional compact flag.}

\item{\code{fullscreen}}{Optional fullscreen flag.}

\item{\code{url}}{Optional URL.}

\item{\code{start_param}}{Optional start parameter.}

\item{\code{theme_params}}{Optional theme parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestSimpleWebViewRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-RequestSimpleWebViewRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve the request using client and utils.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestSimpleWebViewRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The client object.}

\item{\code{utils}}{The utils object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestSimpleWebViewRequest-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-RequestSimpleWebViewRequest-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Convert the object to a dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestSimpleWebViewRequest$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestSimpleWebViewRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-RequestSimpleWebViewRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize the object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestSimpleWebViewRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector of bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestSimpleWebViewRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RequestSimpleWebViewRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestSimpleWebViewRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
