% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requeststate.R
\name{RequestState}
\alias{RequestState}
\title{RequestState}
\description{
Holds information relevant to sent messages, including the message ID assigned to the request,
the container ID to which it belongs, the request itself, the request as raw bytes, and a future result.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RequestState-new}{\code{RequestState$new()}}
\item \href{#method-RequestState-clone}{\code{RequestState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestState-new"></a>}}
\if{latex}{\out{\hypertarget{method-RequestState-new}{}}}
\subsection{Method \code{new()}}{
Creates a new RequestState instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestState$new(request, after = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request}}{The request object.}

\item{\code{after}}{Optional additional parameter. Defaults to NULL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RequestState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
