% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{ResetLoginEmailRequest}
\alias{ResetLoginEmailRequest}
\title{ResetLoginEmailRequest R6 class}
\description{
Represents the TLRequest auth.ResetLoginEmailRequest.
}
\details{
Fields:
}
\section{initialize}{

 Create a new ResetLoginEmailRequest object.
 @param phone_number character
 @param phone_code_hash character
}

\section{to_list}{

 Return a pure R list representation of the object.
 @return list
}

\section{to_bytes}{

 Serialize the object into raw bytes for TL. Uses serialize_bytes helper
 from TL base environment.
 @return raw vector
}

\section{from_reader}{

 Factory that reads fields from a reader and returns a new ResetLoginEmailRequest instance.
 @param reader Reader object implementing read_int(), tgread_string(), tgread_object()
 @return ResetLoginEmailRequest instance
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0x7e960193).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0x6ce87081).}

\item{\code{phone_number}}{character Phone number.}

\item{\code{phone_code_hash}}{character Phone code hash.

Methods:}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResetLoginEmailRequest-new}{\code{ResetLoginEmailRequest$new()}}
\item \href{#method-ResetLoginEmailRequest-to_list}{\code{ResetLoginEmailRequest$to_list()}}
\item \href{#method-ResetLoginEmailRequest-to_bytes}{\code{ResetLoginEmailRequest$to_bytes()}}
\item \href{#method-ResetLoginEmailRequest-from_reader}{\code{ResetLoginEmailRequest$from_reader()}}
\item \href{#method-ResetLoginEmailRequest-clone}{\code{ResetLoginEmailRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetLoginEmailRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResetLoginEmailRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResetLoginEmailRequest$new(phone_number, phone_code_hash)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phone_number}}{character}

\item{\code{phone_code_hash}}{character
Convert to an R list (similar to to_dict)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetLoginEmailRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ResetLoginEmailRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResetLoginEmailRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetLoginEmailRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ResetLoginEmailRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResetLoginEmailRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
Read from reader and create ResetLoginEmailRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetLoginEmailRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ResetLoginEmailRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResetLoginEmailRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ResetLoginEmailRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetLoginEmailRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResetLoginEmailRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResetLoginEmailRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
