% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_bot_file_id}
\alias{resolve_bot_file_id}
\title{Resolve Bot File ID}
\usage{
resolve_bot_file_id(file_id)
}
\arguments{
\item{file_id}{A character string representing the Bot API-style file ID.}
}
\value{
A `Document` or `Photo` object if valid, otherwise `NULL`.
}
\description{
Given a Bot API-style `file_id`, returns the media it represents.
If the `file_id` is not valid, `NULL` is returned instead.
}
\details{
Note that the `file_id` does not have information such as image dimensions
or file size, so these will be zero if present.

For thumbnails, the photo ID and hash will always be zero.
}
