% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_id}
\alias{resolve_id}
\title{Resolve Marked ID}
\usage{
resolve_id(marked_id)
}
\arguments{
\item{marked_id}{An integer representing the marked ID.}
}
\value{
A list containing the original ID and the Peer type (e.g., types$PeerUser, types$PeerChat, or types$PeerChannel).
}
\description{
Given a marked ID, returns the original ID and its Peer type.
}
