% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_inline_message_id}
\alias{resolve_inline_message_id}
\title{Resolve Inline Message ID}
\usage{
resolve_inline_message_id(inline_msg_id)
}
\arguments{
\item{inline_msg_id}{A character string representing the inline message ID.}
}
\value{
A list containing message_id, peer, dc_id, access_hash, or NULLs if invalid.
}
\description{
Resolves an inline message ID. Returns a list of
\code{(message_id, peer, dc_id, access_hash)}.
}
\details{
The \code{peer} may either be a \code{PeerUser} referencing
the user who sent the message via the bot in a private
conversation or small group chat, or a \code{PeerChannel}
if the message was sent in a channel.

The \code{access_hash} does not have any use yet.
}
