% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_invite_link}
\alias{resolve_invite_link}
\title{Resolve Invite Link}
\usage{
resolve_invite_link(link)
}
\arguments{
\item{link}{A character string representing the invite link.}
}
\value{
A list containing link creator user id, global chat id, random int,
  or a list of three NULLs if invalid.
}
\description{
Resolves the given invite link. Returns a list of
\code{(link creator user id, global chat id, random int)}.
}
\details{
Note that for broadcast channels or with the newest link format, the link
creator user ID will be zero to protect their identity. Normal chats and
megagroup channels will have such ID.

Note that the chat ID may not be accurate for chats with a link that were
upgraded to megagroup, since the link can remain the same, but the chat
ID will be correct once a new link is generated.
}
