% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_channels.R
\name{RestrictSponsoredMessagesRequest}
\alias{RestrictSponsoredMessagesRequest}
\title{RestrictSponsoredMessagesRequest}
\description{
Represents a request to restrict sponsored messages in a channel.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{RestrictSponsoredMessagesRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{The constructor ID for this request.}

\item{\code{SUBCLASS_OF_ID}}{The subclass ID for this request.}

\item{\code{channel}}{The input channel.}

\item{\code{restricted}}{Whether sponsored messages are restricted.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RestrictSponsoredMessagesRequest-new}{\code{RestrictSponsoredMessagesRequest$new()}}
\item \href{#method-RestrictSponsoredMessagesRequest-resolve}{\code{RestrictSponsoredMessagesRequest$resolve()}}
\item \href{#method-RestrictSponsoredMessagesRequest-to_dict}{\code{RestrictSponsoredMessagesRequest$to_dict()}}
\item \href{#method-RestrictSponsoredMessagesRequest-bytes}{\code{RestrictSponsoredMessagesRequest$bytes()}}
\item \href{#method-RestrictSponsoredMessagesRequest-clone}{\code{RestrictSponsoredMessagesRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RestrictSponsoredMessagesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-RestrictSponsoredMessagesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the RestrictSponsoredMessagesRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RestrictSponsoredMessagesRequest$new(channel, restricted)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channel}}{The input channel.}

\item{\code{restricted}}{Whether sponsored messages are restricted.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RestrictSponsoredMessagesRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-RestrictSponsoredMessagesRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve the channel entity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RestrictSponsoredMessagesRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The client object.}

\item{\code{utils}}{The utilities object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RestrictSponsoredMessagesRequest-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-RestrictSponsoredMessagesRequest-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Convert the object to a dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RestrictSponsoredMessagesRequest$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RestrictSponsoredMessagesRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-RestrictSponsoredMessagesRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize the object to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RestrictSponsoredMessagesRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector of bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RestrictSponsoredMessagesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RestrictSponsoredMessagesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RestrictSponsoredMessagesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
