% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotosender.R
\name{retry_range}
\alias{retry_range}
\title{Generate a sequence for retry attempts}
\usage{
retry_range(n, force_retry = FALSE)
}
\arguments{
\item{n}{Number of retries to perform}

\item{force_retry}{Whether to force at least one retry attempt even if n is 0 (default: FALSE)}
}
\value{
An integer sequence to iterate over for retry attempts
}
\description{
Creates a sequence from 1 to n for retry attempts with optional forcing of at least one retry
}
