% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{ReuploadCdnFileRequest}
\alias{ReuploadCdnFileRequest}
\title{ReuploadCdnFileRequest}
\description{
Telegram API type ReuploadCdnFileRequest
}
\details{
ReuploadCdnFileRequest R6 class

Represents the TL request upload.ReuploadCdnFileRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file_token}}{Field.}

\item{\code{request_token}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReuploadCdnFileRequest-new}{\code{ReuploadCdnFileRequest$new()}}
\item \href{#method-ReuploadCdnFileRequest-to_list}{\code{ReuploadCdnFileRequest$to_list()}}
\item \href{#method-ReuploadCdnFileRequest-to_bytes}{\code{ReuploadCdnFileRequest$to_bytes()}}
\item \href{#method-ReuploadCdnFileRequest-from_reader}{\code{ReuploadCdnFileRequest$from_reader()}}
\item \href{#method-ReuploadCdnFileRequest-clone}{\code{ReuploadCdnFileRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReuploadCdnFileRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReuploadCdnFileRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ReuploadCdnFileRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReuploadCdnFileRequest$new(file_token, request_token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_token}}{raw|integer|character}

\item{\code{request_token}}{raw|integer|character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReuploadCdnFileRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ReuploadCdnFileRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReuploadCdnFileRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReuploadCdnFileRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ReuploadCdnFileRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to raw vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReuploadCdnFileRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReuploadCdnFileRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ReuploadCdnFileRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read ReuploadCdnFileRequest from a reader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReuploadCdnFileRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{an object providing tgread_bytes()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ReuploadCdnFileRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReuploadCdnFileRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReuploadCdnFileRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReuploadCdnFileRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
