% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rle_decode}
\alias{rle_decode}
\title{Decode Run-Length-Encoded Data}
\usage{
rle_decode(data)
}
\arguments{
\item{data}{A raw vector representing the encoded data.}
}
\value{
A raw vector of the decoded data.
}
\description{
Decodes run-length-encoded data.
}
