% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rle_encode}
\alias{rle_encode}
\title{Encode Data with Run-Length Encoding}
\usage{
rle_encode(string)
}
\arguments{
\item{string}{A raw vector to encode.}
}
\value{
A raw vector of the encoded data.
}
\description{
Encodes a raw vector using run-length encoding.
}
