% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{rpc_message_to_error}
\alias{rpc_message_to_error}
\title{Convert an RPC error to an R condition}
\usage{
rpc_message_to_error(rpc_error, request)
}
\arguments{
\item{rpc_error}{A list or R6 object with error_code and error_message fields}

\item{request}{The request that caused the error}
}
\value{
A condition object
}
\description{
Convert an RPC error to an R condition
}
\keyword{internal}
