% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{RpcDropAnswerRequest}
\alias{RpcDropAnswerRequest}
\title{RpcDropAnswerRequest R6 class}
\value{
An R6 object of class \code{RpcDropAnswerRequest}.
}
\description{
R6 class representing a request to drop an RPC answer for a previously sent
request message. The server will respond with one of the \code{RpcAnswer*}
response variants (e.g. \code{RpcAnswerUnknown}, \code{RpcAnswerDroppedRunning},
\code{RpcAnswerDropped}).
}
\examples{
# req <- RpcDropAnswerRequest$new(req_msg_id = 1234567890123456789)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Integer constructor identifier.}

\item{\code{SUBCLASS_OF_ID}}{Integer subclass identifier.}

\item{\code{req_msg_id}}{Integer64 request message identifier whose answer should be dropped.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RpcDropAnswerRequest-new}{\code{RpcDropAnswerRequest$new()}}
\item \href{#method-RpcDropAnswerRequest-to_list}{\code{RpcDropAnswerRequest$to_list()}}
\item \href{#method-RpcDropAnswerRequest-to_bytes}{\code{RpcDropAnswerRequest$to_bytes()}}
\item \href{#method-RpcDropAnswerRequest-clone}{\code{RpcDropAnswerRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$new(req_msg_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$to_bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
