% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{RpcDropAnswerRequest_from_reader}
\alias{RpcDropAnswerRequest_from_reader}
\title{Create RpcDropAnswerRequest from a binary reader}
\usage{
RpcDropAnswerRequest_from_reader(reader)
}
\arguments{
\item{reader}{Reader object providing \code{read_long()} to read 64-bit integers.}
}
\value{
Instance of \code{RpcDropAnswerRequest}.
}
\description{
Reads a 64-bit \code{req_msg_id} from the provided \code{reader} and
constructs an instance of \code{RpcDropAnswerRequest}.
}
