% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{SaveDefaultGroupCallJoinAsRequest}
\alias{SaveDefaultGroupCallJoinAsRequest}
\title{SaveDefaultGroupCallJoinAsRequest R6 class}
\description{
SaveDefaultGroupCallJoinAsRequest R6 class

SaveDefaultGroupCallJoinAsRequest R6 class
}
\details{
Represents SaveDefaultGroupCallJoinAsRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{peer}}{Input peer object (must provide bytes()).}

\item{\code{join_as}}{Input peer to join as (must provide bytes()).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SaveDefaultGroupCallJoinAsRequest-new}{\code{SaveDefaultGroupCallJoinAsRequest$new()}}
\item \href{#method-SaveDefaultGroupCallJoinAsRequest-resolve}{\code{SaveDefaultGroupCallJoinAsRequest$resolve()}}
\item \href{#method-SaveDefaultGroupCallJoinAsRequest-to_list}{\code{SaveDefaultGroupCallJoinAsRequest$to_list()}}
\item \href{#method-SaveDefaultGroupCallJoinAsRequest-bytes}{\code{SaveDefaultGroupCallJoinAsRequest$bytes()}}
\item \href{#method-SaveDefaultGroupCallJoinAsRequest-from_reader}{\code{SaveDefaultGroupCallJoinAsRequest$from_reader()}}
\item \href{#method-SaveDefaultGroupCallJoinAsRequest-clone}{\code{SaveDefaultGroupCallJoinAsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveDefaultGroupCallJoinAsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SaveDefaultGroupCallJoinAsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a SaveDefaultGroupCallJoinAsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveDefaultGroupCallJoinAsRequest$new(peer, join_as)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{Input peer object.}

\item{\code{join_as}}{Input peer object to join as.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveDefaultGroupCallJoinAsRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-SaveDefaultGroupCallJoinAsRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references

Typically used to convert high-level entities into input objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveDefaultGroupCallJoinAsRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_peer method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveDefaultGroupCallJoinAsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SaveDefaultGroupCallJoinAsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveDefaultGroupCallJoinAsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveDefaultGroupCallJoinAsRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SaveDefaultGroupCallJoinAsRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes `peer$bytes()` and
`join_as$bytes()` exist.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveDefaultGroupCallJoinAsRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(),
        a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveDefaultGroupCallJoinAsRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-SaveDefaultGroupCallJoinAsRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveDefaultGroupCallJoinAsRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveDefaultGroupCallJoinAsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SaveDefaultGroupCallJoinAsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveDefaultGroupCallJoinAsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
