% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{SaveFilePartRequest}
\alias{SaveFilePartRequest}
\title{SaveFilePartRequest R6 class}
\description{
Represents the TL request upload.SaveFilePartRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file_id}}{numeric 64-bit file id (stored as numeric)}

\item{\code{file_part}}{integer part index}

\item{\code{bytes_data}}{raw|integer|character payload bytes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SaveFilePartRequest-new}{\code{SaveFilePartRequest$new()}}
\item \href{#method-SaveFilePartRequest-to_list}{\code{SaveFilePartRequest$to_list()}}
\item \href{#method-SaveFilePartRequest-to_bytes}{\code{SaveFilePartRequest$to_bytes()}}
\item \href{#method-SaveFilePartRequest-from_reader}{\code{SaveFilePartRequest$from_reader()}}
\item \href{#method-SaveFilePartRequest-clone}{\code{SaveFilePartRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveFilePartRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SaveFilePartRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveFilePartRequest$new(file_id, file_part, bytes_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_id}}{numeric (64-bit)}

\item{\code{file_part}}{integer}

\item{\code{bytes_data}}{raw|integer|character
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveFilePartRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SaveFilePartRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveFilePartRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveFilePartRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SaveFilePartRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveFilePartRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveFilePartRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-SaveFilePartRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveFilePartRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveFilePartRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SaveFilePartRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveFilePartRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
