% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{SecurityError}
\alias{SecurityError}
\title{SecurityError}
\description{
Generic security error, mostly used when generating a new AuthKey.
}
\section{Super class}{
\code{\link[telegramR:Exception]{telegramR::Exception}} -> \code{SecurityError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecurityError-new}{\code{SecurityError$new()}}
\item \href{#method-SecurityError-clone}{\code{SecurityError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityError-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityError-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the SecurityError class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityError$new(message = "A security check failed.")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{\code{character} The error message.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
