% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{SendCodeRequest}
\alias{SendCodeRequest}
\title{SendCodeRequest R6 class}
\description{
SendCodeRequest R6 class

SendCodeRequest R6 class
}
\details{
Represents the TLRequest auth.SendCodeRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0xa677244f).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0x6ce87081).}

\item{\code{phone_number}}{character Phone number.}

\item{\code{api_id}}{integer API id.}

\item{\code{api_hash}}{character API hash.}

\item{\code{settings}}{TL object TypeCodeSettings.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SendCodeRequest-new}{\code{SendCodeRequest$new()}}
\item \href{#method-SendCodeRequest-to_list}{\code{SendCodeRequest$to_list()}}
\item \href{#method-SendCodeRequest-to_bytes}{\code{SendCodeRequest$to_bytes()}}
\item \href{#method-SendCodeRequest-from_reader}{\code{SendCodeRequest$from_reader()}}
\item \href{#method-SendCodeRequest-clone}{\code{SendCodeRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SendCodeRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SendCodeRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a SendCodeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SendCodeRequest$new(phone_number, api_id, api_hash, settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phone_number}}{character}

\item{\code{api_id}}{integer}

\item{\code{api_hash}}{character}

\item{\code{settings}}{TL object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SendCodeRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SendCodeRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SendCodeRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SendCodeRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SendCodeRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SendCodeRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SendCodeRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-SendCodeRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create SendCodeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SendCodeRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SendCodeRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SendCodeRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SendCodeRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SendCodeRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
