% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{SendSignalingDataRequest}
\alias{SendSignalingDataRequest}
\title{SendSignalingDataRequest}
\description{
Telegram API type SendSignalingDataRequest
}
\details{
SendSignalingDataRequest R6 class

Represents SendSignalingDataRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{peer}}{Field.}

\item{\code{data}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SendSignalingDataRequest-new}{\code{SendSignalingDataRequest$new()}}
\item \href{#method-SendSignalingDataRequest-to_list}{\code{SendSignalingDataRequest$to_list()}}
\item \href{#method-SendSignalingDataRequest-bytes}{\code{SendSignalingDataRequest$bytes()}}
\item \href{#method-SendSignalingDataRequest-from_reader}{\code{SendSignalingDataRequest$from_reader()}}
\item \href{#method-SendSignalingDataRequest-clone}{\code{SendSignalingDataRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SendSignalingDataRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SendSignalingDataRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a SendSignalingDataRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SendSignalingDataRequest$new(peer, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{Input phone call object.}

\item{\code{data}}{Raw vector containing bytes to send.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SendSignalingDataRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SendSignalingDataRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SendSignalingDataRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SendSignalingDataRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SendSignalingDataRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes `peer$bytes()` exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SendSignalingDataRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SendSignalingDataRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-SendSignalingDataRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SendSignalingDataRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SendSignalingDataRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SendSignalingDataRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SendSignalingDataRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
