% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendTextMessage.R
\name{sendTextMessage}
\alias{sendTextMessage}
\title{sendTextMessage}
\usage{
sendTextMessage(token, chat_id, text_message)
}
\arguments{
\item{token}{Your bot's API token.}

\item{chat_id}{A chat id.}

\item{text_message}{The text message you want to send.}
}
\description{
Send a text message to a chatId.
}
\examples{
\dontrun{
sendTextMessage(token = token, chat_id = "475093487", text_message = "Hi there!")
}

}
