% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlobject.R
\name{serialize_bytes}
\alias{serialize_bytes}
\title{Serialize data to bytes.}
\usage{
serialize_bytes(data)
}
\arguments{
\item{data}{The data to be serialized.}
}
\value{
A raw vector representing the serialized data.
}
\description{
This function serializes data to a byte array. It handles both raw data and character strings.
}
\examples{
\dontrun{
serialize_bytes(charToRaw("test"))
serialize_bytes("test")
}
}
