% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlobject.R
\name{serialize_datetime}
\alias{serialize_datetime}
\title{Serialize a datetime object to bytes.}
\usage{
serialize_datetime(dt)
}
\arguments{
\item{dt}{The datetime object to be serialized.}
}
\value{
A raw vector representing the serialized datetime.
}
\description{
This function serializes a datetime object to a byte array. It handles various datetime formats,
including POSIXct, Date, numeric, and difftime.
}
\examples{
serialize_datetime(as.POSIXct("2023-01-01 12:00:00", tz = "UTC"))
serialize_datetime(as.Date("2023-01-01"))
serialize_datetime(as.numeric(1234567890))
serialize_datetime(as.difftime(3600, units = "secs"))
}
