% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serialize_int}
\alias{serialize_int}
\title{Serialize an integer to 4 bytes little-endian.}
\usage{
serialize_int(x)
}
\arguments{
\item{x}{integer}
}
\value{
raw vector
}
\description{
Serialize an integer to 4 bytes little-endian.
}
