% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serialize_string}
\alias{serialize_string}
\title{Serialize a string to Telegram bytes format.}
\usage{
serialize_string(s)
}
\arguments{
\item{s}{string}
}
\value{
raw vector
}
\description{
Serialize a string to Telegram bytes format.
}
