% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{SetBotMenuButtonRequest}
\alias{SetBotMenuButtonRequest}
\title{SetBotMenuButtonRequest R6 class}
\description{
Methods:
- initialize(user_id, button)
- resolve(client, utils)
- to_list() -> list representation
- to_bytes() -> raw vector bytes

Each method is documented inline below.
}
\details{
Represents the SetBotMenuButtonRequest TL request.
}
\examples{
# o <- SetBotMenuButtonRequest$new(user_id, button)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_id}}{TypeInputUser input user or object implementing to_list()/to_bytes()}

\item{\code{button}}{TypeBotMenuButton object or representation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SetBotMenuButtonRequest-new}{\code{SetBotMenuButtonRequest$new()}}
\item \href{#method-SetBotMenuButtonRequest-resolve}{\code{SetBotMenuButtonRequest$resolve()}}
\item \href{#method-SetBotMenuButtonRequest-to_list}{\code{SetBotMenuButtonRequest$to_list()}}
\item \href{#method-SetBotMenuButtonRequest-to_bytes}{\code{SetBotMenuButtonRequest$to_bytes()}}
\item \href{#method-SetBotMenuButtonRequest-clone}{\code{SetBotMenuButtonRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotMenuButtonRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotMenuButtonRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize SetBotMenuButtonRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotMenuButtonRequest$new(user_id, button)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{TypeInputUser or identifier}

\item{\code{button}}{TypeBotMenuButton or object
Resolve references (convert entities via client/utils)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotMenuButtonRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotMenuButtonRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotMenuButtonRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client with get_input_entity method}

\item{\code{utils}}{utils with get_input_user
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotMenuButtonRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotMenuButtonRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotMenuButtonRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotMenuButtonRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotMenuButtonRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotMenuButtonRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotMenuButtonRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotMenuButtonRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotMenuButtonRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
