% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{SetBotUpdatesStatusRequest}
\alias{SetBotUpdatesStatusRequest}
\title{SetBotUpdatesStatusRequest}
\description{
Telegram API type SetBotUpdatesStatusRequest
}
\details{
SetBotUpdatesStatusRequest R6 class

Notify server of bot updates status.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{pendingUpdatesCount}}{integer Count of pending updates.}

\item{\code{message}}{character Additional message.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SetBotUpdatesStatusRequest-new}{\code{SetBotUpdatesStatusRequest$new()}}
\item \href{#method-SetBotUpdatesStatusRequest-to_list}{\code{SetBotUpdatesStatusRequest$to_list()}}
\item \href{#method-SetBotUpdatesStatusRequest-to_bytes}{\code{SetBotUpdatesStatusRequest$to_bytes()}}
\item \href{#method-SetBotUpdatesStatusRequest-clone}{\code{SetBotUpdatesStatusRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotUpdatesStatusRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotUpdatesStatusRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize SetBotUpdatesStatusRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotUpdatesStatusRequest$new(pendingUpdatesCount = 0L, message = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pendingUpdatesCount}}{integer}

\item{\code{message}}{character
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotUpdatesStatusRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotUpdatesStatusRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotUpdatesStatusRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list representation
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotUpdatesStatusRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotUpdatesStatusRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotUpdatesStatusRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotUpdatesStatusRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotUpdatesStatusRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotUpdatesStatusRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
