% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{SetCustomVerificationRequest}
\alias{SetCustomVerificationRequest}
\title{SetCustomVerificationRequest}
\description{
Methods:
- initialize(peer, enabled = NULL, bot = NULL, custom_description = NULL)
- resolve(client, utils)
- to_list() -> list representation
- to_bytes() -> raw vector bytes

Telegram API type SetCustomVerificationRequest
}
\details{
SetCustomVerificationRequest R6 class

Represents the SetCustomVerificationRequest TL request.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{peer}}{TypeInputPeer input peer or object implementing to_list()/to_bytes()}

\item{\code{enabled}}{logical|null optional}

\item{\code{bot}}{TypeInputUser|null optional}

\item{\code{custom_description}}{character|null optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SetCustomVerificationRequest-new}{\code{SetCustomVerificationRequest$new()}}
\item \href{#method-SetCustomVerificationRequest-resolve}{\code{SetCustomVerificationRequest$resolve()}}
\item \href{#method-SetCustomVerificationRequest-to_list}{\code{SetCustomVerificationRequest$to_list()}}
\item \href{#method-SetCustomVerificationRequest-to_bytes}{\code{SetCustomVerificationRequest$to_bytes()}}
\item \href{#method-SetCustomVerificationRequest-clone}{\code{SetCustomVerificationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetCustomVerificationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SetCustomVerificationRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize SetCustomVerificationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetCustomVerificationRequest$new(
  peer,
  enabled = NULL,
  bot = NULL,
  custom_description = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{TypeInputPeer or identifier}

\item{\code{enabled}}{logical|null}

\item{\code{bot}}{TypeInputUser or identifier|null}

\item{\code{custom_description}}{character|null
Resolve references (convert entities via client/utils)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetCustomVerificationRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-SetCustomVerificationRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetCustomVerificationRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client with get_input_entity method}

\item{\code{utils}}{utils with get_input_user / get_input_peer
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetCustomVerificationRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SetCustomVerificationRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetCustomVerificationRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetCustomVerificationRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SetCustomVerificationRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetCustomVerificationRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetCustomVerificationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SetCustomVerificationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetCustomVerificationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
