% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{SignInRequest}
\alias{SignInRequest}
\title{SignInRequest}
\description{
Telegram API type SignInRequest
}
\details{
SignInRequest R6 class

Represents the TLRequest auth.SignInRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{phone_number}}{Field.}

\item{\code{phone_code_hash}}{Field.}

\item{\code{phone_code}}{Field.}

\item{\code{email_verification}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SignInRequest-new}{\code{SignInRequest$new()}}
\item \href{#method-SignInRequest-to_list}{\code{SignInRequest$to_list()}}
\item \href{#method-SignInRequest-to_bytes}{\code{SignInRequest$to_bytes()}}
\item \href{#method-SignInRequest-from_reader}{\code{SignInRequest$from_reader()}}
\item \href{#method-SignInRequest-clone}{\code{SignInRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignInRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SignInRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a SignInRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignInRequest$new(
  phone_number,
  phone_code_hash,
  phone_code = NULL,
  email_verification = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phone_number}}{character}

\item{\code{phone_code_hash}}{character}

\item{\code{phone_code}}{character or NULL}

\item{\code{email_verification}}{TL object or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignInRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SignInRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignInRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignInRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SignInRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignInRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignInRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-SignInRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create SignInRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignInRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SignInRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignInRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SignInRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignInRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
