% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{StartScheduledGroupCallRequest}
\alias{StartScheduledGroupCallRequest}
\title{StartScheduledGroupCallRequest}
\description{
Telegram API type StartScheduledGroupCallRequest
}
\details{
StartScheduledGroupCallRequest R6 class

Represents StartScheduledGroupCallRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{The input group call object (expected to implement bytes()).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StartScheduledGroupCallRequest-new}{\code{StartScheduledGroupCallRequest$new()}}
\item \href{#method-StartScheduledGroupCallRequest-resolve}{\code{StartScheduledGroupCallRequest$resolve()}}
\item \href{#method-StartScheduledGroupCallRequest-to_list}{\code{StartScheduledGroupCallRequest$to_list()}}
\item \href{#method-StartScheduledGroupCallRequest-bytes}{\code{StartScheduledGroupCallRequest$bytes()}}
\item \href{#method-StartScheduledGroupCallRequest-from_reader}{\code{StartScheduledGroupCallRequest$from_reader()}}
\item \href{#method-StartScheduledGroupCallRequest-clone}{\code{StartScheduledGroupCallRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartScheduledGroupCallRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-StartScheduledGroupCallRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a StartScheduledGroupCallRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StartScheduledGroupCallRequest$new(call)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call (object).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartScheduledGroupCallRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-StartScheduledGroupCallRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references

Typically used to convert high-level entities into input objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StartScheduledGroupCallRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartScheduledGroupCallRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-StartScheduledGroupCallRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StartScheduledGroupCallRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartScheduledGroupCallRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-StartScheduledGroupCallRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes `self$call$bytes()` exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StartScheduledGroupCallRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(),
        a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartScheduledGroupCallRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-StartScheduledGroupCallRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StartScheduledGroupCallRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartScheduledGroupCallRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StartScheduledGroupCallRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StartScheduledGroupCallRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
