% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stickers.R
\name{SuggestShortNameRequest}
\alias{SuggestShortNameRequest}
\title{SuggestShortNameRequest}
\description{
Telegram API type SuggestShortNameRequest
}
\details{
SuggestShortNameRequest R6 class

Represents a request to suggest a short name for a sticker set.
}
\section{Methods}{

- initialize(title): create instance.
- to_list(): returns a list representation suitable for JSON/inspection.
- to_bytes(): returns a raw vector representing serialized bytes (TL-like).
- from_reader(reader): reads a request from a reader object and returns a new instance.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{title}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SuggestShortNameRequest-new}{\code{SuggestShortNameRequest$new()}}
\item \href{#method-SuggestShortNameRequest-to_list}{\code{SuggestShortNameRequest$to_list()}}
\item \href{#method-SuggestShortNameRequest-to_bytes}{\code{SuggestShortNameRequest$to_bytes()}}
\item \href{#method-SuggestShortNameRequest-from_reader}{\code{SuggestShortNameRequest$from_reader()}}
\item \href{#method-SuggestShortNameRequest-clone}{\code{SuggestShortNameRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SuggestShortNameRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SuggestShortNameRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a SuggestShortNameRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuggestShortNameRequest$new(title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{character Title string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SuggestShortNameRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SuggestShortNameRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuggestShortNameRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list representation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SuggestShortNameRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SuggestShortNameRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to raw bytes (simple TL-like approximation)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuggestShortNameRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SuggestShortNameRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-SuggestShortNameRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuggestShortNameRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SuggestShortNameRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SuggestShortNameRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuggestShortNameRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
