% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{TakeoutClient}
\alias{TakeoutClient}
\title{TakeoutClient}
\description{
Telegram API type TakeoutClient
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{finalize}}{A logical indicating whether to finalize the takeout session.}

\item{\code{client}}{An instance of the client to use for the takeout session.}

\item{\code{request}}{An optional request object for the takeout session.}

\item{\code{success}}{A logical indicating whether the takeout session was successful.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TakeoutClient-new}{\code{TakeoutClient$new()}}
\item \href{#method-TakeoutClient-get_success}{\code{TakeoutClient$get_success()}}
\item \href{#method-TakeoutClient-set_success}{\code{TakeoutClient$set_success()}}
\item \href{#method-TakeoutClient-aenter}{\code{TakeoutClient$aenter()}}
\item \href{#method-TakeoutClient-aexit}{\code{TakeoutClient$aexit()}}
\item \href{#method-TakeoutClient-call}{\code{TakeoutClient$call()}}
\item \href{#method-TakeoutClient-get_attribute}{\code{TakeoutClient$get_attribute()}}
\item \href{#method-TakeoutClient-get_attr}{\code{TakeoutClient$get_attr()}}
\item \href{#method-TakeoutClient-set_attr}{\code{TakeoutClient$set_attr()}}
\item \href{#method-TakeoutClient-clone}{\code{TakeoutClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TakeoutClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-TakeoutClient-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new `TakeoutClient` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakeoutClient$new(finalize, client, request)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{finalize}}{A logical indicating whether to finalize the takeout session.}

\item{\code{client}}{An instance of the client to use for the takeout session.}

\item{\code{request}}{An optional request object for the takeout session.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TakeoutClient` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TakeoutClient-get_success"></a>}}
\if{latex}{\out{\hypertarget{method-TakeoutClient-get_success}{}}}
\subsection{Method \code{get_success()}}{
Get the success status of the takeout session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakeoutClient$get_success()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A logical indicating whether the takeout session was successful.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TakeoutClient-set_success"></a>}}
\if{latex}{\out{\hypertarget{method-TakeoutClient-set_success}{}}}
\subsection{Method \code{set_success()}}{
Set the success status of the takeout session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakeoutClient$set_success(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A logical indicating whether the takeout session was successful.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TakeoutClient-aenter"></a>}}
\if{latex}{\out{\hypertarget{method-TakeoutClient-aenter}{}}}
\subsection{Method \code{aenter()}}{
Enter the takeout session context.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakeoutClient$aenter()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current instance of the `TakeoutClient`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TakeoutClient-aexit"></a>}}
\if{latex}{\out{\hypertarget{method-TakeoutClient-aexit}{}}}
\subsection{Method \code{aexit()}}{
Exit the takeout session context.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakeoutClient$aexit(exc_type, exc_value, traceback)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exc_type}}{The type of exception raised, if any.}

\item{\code{exc_value}}{The value of the exception raised, if any.}

\item{\code{traceback}}{The traceback of the exception raised, if any.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TakeoutClient-call"></a>}}
\if{latex}{\out{\hypertarget{method-TakeoutClient-call}{}}}
\subsection{Method \code{call()}}{
Make a request to the Telegram API using the takeout session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakeoutClient$call(request, ordered = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request}}{The request object to send.}

\item{\code{ordered}}{A logical indicating whether the request should be ordered.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TakeoutClient-get_attribute"></a>}}
\if{latex}{\out{\hypertarget{method-TakeoutClient-get_attribute}{}}}
\subsection{Method \code{get_attribute()}}{
Get the result of the request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakeoutClient$get_attribute(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the attribute to get.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TakeoutClient-get_attr"></a>}}
\if{latex}{\out{\hypertarget{method-TakeoutClient-get_attr}{}}}
\subsection{Method \code{get_attr()}}{
Get the attribute of the request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakeoutClient$get_attr(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the attribute to get.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The value of the attribute.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TakeoutClient-set_attr"></a>}}
\if{latex}{\out{\hypertarget{method-TakeoutClient-set_attr}{}}}
\subsection{Method \code{set_attr()}}{
Set the attribute of the request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakeoutClient$set_attr(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the attribute to set.}

\item{\code{value}}{The value to set for the attribute.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TakeoutClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TakeoutClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakeoutClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
