% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tele_info.R
\name{tele_info}
\alias{tele_info}
\title{get credentials for the chat bot}
\usage{
tele_info(token)
}
\arguments{
\item{token}{the HTTP API token provided by the Telegram Botfather}
}
\value{
returns a list of credentials to identify the user
}
\description{
get credentials for the chat bot
}

