% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tele_send.R
\name{tele_send}
\alias{tele_send}
\title{Sending messages to the Telegram}
\usage{
tele_send(myinfo, message)
}
\arguments{
\item{myinfo}{a list returned by tele_info() containting API Token and the Chat ID}

\item{message}{a string containing the message to be sent by the Bot}
}
\description{
Sending messages to the Telegram
}

