% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegramclient.R
\name{TelegramClient}
\alias{TelegramClient}
\title{TelegramClient}
\value{
An R6 generator object of class \code{TelegramClient}.
}
\description{
High-level 'Telegram' 'MTProto' client. Create one with
\code{TelegramClient$new(session, api_id, api_hash)}, then \code{$start()}
(interactive login) or \code{$connect()} together with
\code{$send_code_request()} and \code{$sign_in()} to authenticate. Once
connected it exposes the messaging, channel, media and download methods used
throughout the package documentation.
}
\details{
This is an R6 class. Typical usage:
\preformatted{
client <- TelegramClient$new("my_session", api_id = 123, api_hash = "...")
client$connect()
}
}
\examples{
\dontrun{
client <- TelegramClient$new("my_session", api_id = 123, api_hash = "...")
client$start()
}
}
\section{Super class}{
\code{telegramR::TelegramBaseClient} -> \code{TelegramClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TelegramClient-takeout}{\code{TelegramClient$takeout()}}
\item \href{#method-TelegramClient-end_takeout}{\code{TelegramClient$end_takeout()}}
\item \href{#method-TelegramClient-start}{\code{TelegramClient$start()}}
\item \href{#method-TelegramClient-sign_in}{\code{TelegramClient$sign_in()}}
\item \href{#method-TelegramClient-sign_up}{\code{TelegramClient$sign_up()}}
\item \href{#method-TelegramClient-send_code_request}{\code{TelegramClient$send_code_request()}}
\item \href{#method-TelegramClient-qr_login}{\code{TelegramClient$qr_login()}}
\item \href{#method-TelegramClient-log_out}{\code{TelegramClient$log_out()}}
\item \href{#method-TelegramClient-edit_2fa}{\code{TelegramClient$edit_2fa()}}
\item \href{#method-TelegramClient-parse_phone_and_hash}{\code{TelegramClient$parse_phone_and_hash()}}
\item \href{#method-TelegramClient-on_login}{\code{TelegramClient$on_login()}}
\item \href{#method-TelegramClient-start_impl}{\code{TelegramClient$start_impl()}}
\item \href{#method-TelegramClient-parse_phone}{\code{TelegramClient$parse_phone()}}
\item \href{#method-TelegramClient-compute_check}{\code{TelegramClient$compute_check()}}
\item \href{#method-TelegramClient-compute_digest}{\code{TelegramClient$compute_digest()}}
\item \href{#method-TelegramClient-get_display_name}{\code{TelegramClient$get_display_name()}}
\item \href{#method-TelegramClient-download_profile_photo}{\code{TelegramClient$download_profile_photo()}}
\item \href{#method-TelegramClient-download_media}{\code{TelegramClient$download_media()}}
\item \href{#method-TelegramClient-download_file}{\code{TelegramClient$download_file()}}
\item \href{#method-TelegramClient-.download_file}{\code{TelegramClient$.download_file()}}
\item \href{#method-TelegramClient-iter_download}{\code{TelegramClient$iter_download()}}
\item \href{#method-TelegramClient-.iter_download}{\code{TelegramClient$.iter_download()}}
\item \href{#method-TelegramClient-get_thumb}{\code{TelegramClient$get_thumb()}}
\item \href{#method-TelegramClient-download_cached_photo_size}{\code{TelegramClient$download_cached_photo_size()}}
\item \href{#method-TelegramClient-download_photo}{\code{TelegramClient$download_photo()}}
\item \href{#method-TelegramClient-get_kind_and_names}{\code{TelegramClient$get_kind_and_names()}}
\item \href{#method-TelegramClient-download_document}{\code{TelegramClient$download_document()}}
\item \href{#method-TelegramClient-download_contact}{\code{TelegramClient$download_contact()}}
\item \href{#method-TelegramClient-download_web_document}{\code{TelegramClient$download_web_document()}}
\item \href{#method-TelegramClient-get_proper_filename}{\code{TelegramClient$get_proper_filename()}}
\item \href{#method-TelegramClient-iter_dialogs}{\code{TelegramClient$iter_dialogs()}}
\item \href{#method-TelegramClient-get_dialogs}{\code{TelegramClient$get_dialogs()}}
\item \href{#method-TelegramClient-iter_drafts}{\code{TelegramClient$iter_drafts()}}
\item \href{#method-TelegramClient-get_drafts}{\code{TelegramClient$get_drafts()}}
\item \href{#method-TelegramClient-edit_folder}{\code{TelegramClient$edit_folder()}}
\item \href{#method-TelegramClient-delete_dialog}{\code{TelegramClient$delete_dialog()}}
\item \href{#method-TelegramClient-conversation}{\code{TelegramClient$conversation()}}
\item \href{#method-TelegramClient-iter_participants}{\code{TelegramClient$iter_participants()}}
\item \href{#method-TelegramClient-get_participants}{\code{TelegramClient$get_participants()}}
\item \href{#method-TelegramClient-iter_admin_log}{\code{TelegramClient$iter_admin_log()}}
\item \href{#method-TelegramClient-get_admin_log}{\code{TelegramClient$get_admin_log()}}
\item \href{#method-TelegramClient-iter_profile_photos}{\code{TelegramClient$iter_profile_photos()}}
\item \href{#method-TelegramClient-get_profile_photos}{\code{TelegramClient$get_profile_photos()}}
\item \href{#method-TelegramClient-action}{\code{TelegramClient$action()}}
\item \href{#method-TelegramClient-edit_admin}{\code{TelegramClient$edit_admin()}}
\item \href{#method-TelegramClient-edit_permissions}{\code{TelegramClient$edit_permissions()}}
\item \href{#method-TelegramClient-kick_participant}{\code{TelegramClient$kick_participant()}}
\item \href{#method-TelegramClient-get_permissions}{\code{TelegramClient$get_permissions()}}
\item \href{#method-TelegramClient-get_stats}{\code{TelegramClient$get_stats()}}
\item \href{#method-TelegramClient-inline_query}{\code{TelegramClient$inline_query()}}
\item \href{#method-TelegramClient-invoke_function}{\code{TelegramClient$invoke_function()}}
\item \href{#method-TelegramClient-custom_inline_results}{\code{TelegramClient$custom_inline_results()}}
\item \href{#method-TelegramClient-iter_messages}{\code{TelegramClient$iter_messages()}}
\item \href{#method-TelegramClient-iter_messages_page}{\code{TelegramClient$iter_messages_page()}}
\item \href{#method-TelegramClient-get_messages}{\code{TelegramClient$get_messages()}}
\item \href{#method-TelegramClient-channels_get_messages}{\code{TelegramClient$channels_get_messages()}}
\item \href{#method-TelegramClient-collect}{\code{TelegramClient$collect()}}
\item \href{#method-TelegramClient-collect_one}{\code{TelegramClient$collect_one()}}
\item \href{#method-TelegramClient-send_message}{\code{TelegramClient$send_message()}}
\item \href{#method-TelegramClient-forward_messages}{\code{TelegramClient$forward_messages()}}
\item \href{#method-TelegramClient-edit_message}{\code{TelegramClient$edit_message()}}
\item \href{#method-TelegramClient-delete_messages}{\code{TelegramClient$delete_messages()}}
\item \href{#method-TelegramClient-send_read_acknowledge}{\code{TelegramClient$send_read_acknowledge()}}
\item \href{#method-TelegramClient-pin_message}{\code{TelegramClient$pin_message()}}
\item \href{#method-TelegramClient-unpin_message}{\code{TelegramClient$unpin_message()}}
\item \href{#method-TelegramClient-get_comment_data}{\code{TelegramClient$get_comment_data()}}
\item \href{#method-TelegramClient-pin_internal}{\code{TelegramClient$pin_internal()}}
\item \href{#method-TelegramClient-send_file}{\code{TelegramClient$send_file()}}
\item \href{#method-TelegramClient-upload_file}{\code{TelegramClient$upload_file()}}
\item \href{#method-TelegramClient-file_to_media}{\code{TelegramClient$file_to_media()}}
\item \href{#method-TelegramClient-resize_photo_if_needed}{\code{TelegramClient$resize_photo_if_needed()}}
\item \href{#method-TelegramClient-is_image}{\code{TelegramClient$is_image()}}
\item \href{#method-TelegramClient-get_appropriated_part_size}{\code{TelegramClient$get_appropriated_part_size()}}
\item \href{#method-TelegramClient-invoke}{\code{TelegramClient$invoke()}}
\item \href{#method-TelegramClient-build_reply_markup}{\code{TelegramClient$build_reply_markup()}}
\item \href{#method-TelegramClient-run_until_disconnected}{\code{TelegramClient$run_until_disconnected()}}
\item \href{#method-TelegramClient-set_receive_updates}{\code{TelegramClient$set_receive_updates()}}
\item \href{#method-TelegramClient-on}{\code{TelegramClient$on()}}
\item \href{#method-TelegramClient-add_event_handler}{\code{TelegramClient$add_event_handler()}}
\item \href{#method-TelegramClient-remove_event_handler}{\code{TelegramClient$remove_event_handler()}}
\item \href{#method-TelegramClient-list_event_handlers}{\code{TelegramClient$list_event_handlers()}}
\item \href{#method-TelegramClient-catch_up}{\code{TelegramClient$catch_up()}}
\item \href{#method-TelegramClient-update_loop}{\code{TelegramClient$update_loop()}}
\item \href{#method-TelegramClient-preprocess_updates}{\code{TelegramClient$preprocess_updates()}}
\item \href{#method-TelegramClient-keepalive_loop}{\code{TelegramClient$keepalive_loop()}}
\item \href{#method-TelegramClient-dispatch_update}{\code{TelegramClient$dispatch_update()}}
\item \href{#method-TelegramClient-handle_auto_reconnect}{\code{TelegramClient$handle_auto_reconnect()}}
\item \href{#method-TelegramClient-set_parse_mode}{\code{TelegramClient$set_parse_mode()}}
\item \href{#method-TelegramClient-get_parse_mode}{\code{TelegramClient$get_parse_mode()}}
\item \href{#method-TelegramClient-sanitize_parse_mode}{\code{TelegramClient$sanitize_parse_mode()}}
\item \href{#method-TelegramClient-parse_message_text}{\code{TelegramClient$parse_message_text()}}
\item \href{#method-TelegramClient-replace_with_mention}{\code{TelegramClient$replace_with_mention()}}
\item \href{#method-TelegramClient-get_response_message}{\code{TelegramClient$get_response_message()}}
\item \href{#method-TelegramClient-call}{\code{TelegramClient$call()}}
\item \href{#method-TelegramClient-call_internal}{\code{TelegramClient$call_internal()}}
\item \href{#method-TelegramClient-get_me}{\code{TelegramClient$get_me()}}
\item \href{#method-TelegramClient-self_id}{\code{TelegramClient$self_id()}}
\item \href{#method-TelegramClient-is_bot}{\code{TelegramClient$is_bot()}}
\item \href{#method-TelegramClient-is_user_authorized}{\code{TelegramClient$is_user_authorized()}}
\item \href{#method-TelegramClient-get_entity}{\code{TelegramClient$get_entity()}}
\item \href{#method-TelegramClient-get_input_entity}{\code{TelegramClient$get_input_entity()}}
\item \href{#method-TelegramClient-getInputEntity}{\code{TelegramClient$getInputEntity()}}
\item \href{#method-TelegramClient-getInputPeer}{\code{TelegramClient$getInputPeer()}}
\item \href{#method-TelegramClient-getInputUser}{\code{TelegramClient$getInputUser()}}
\item \href{#method-TelegramClient-getInputChannel}{\code{TelegramClient$getInputChannel()}}
\item \href{#method-TelegramClient-getInputMessage}{\code{TelegramClient$getInputMessage()}}
\item \href{#method-TelegramClient-getInputMedia}{\code{TelegramClient$getInputMedia()}}
\item \href{#method-TelegramClient-getInputDocument}{\code{TelegramClient$getInputDocument()}}
\item \href{#method-TelegramClient-getInputPhoto}{\code{TelegramClient$getInputPhoto()}}
\item \href{#method-TelegramClient-getInputChatPhoto}{\code{TelegramClient$getInputChatPhoto()}}
\item \href{#method-TelegramClient-getInputGroupCall}{\code{TelegramClient$getInputGroupCall()}}
\item \href{#method-TelegramClient-get_peer}{\code{TelegramClient$get_peer()}}
\item \href{#method-TelegramClient-get_peer_id}{\code{TelegramClient$get_peer_id()}}
\item \href{#method-TelegramClient-get_entity_from_string}{\code{TelegramClient$get_entity_from_string()}}
\item \href{#method-TelegramClient-get_input_dialog}{\code{TelegramClient$get_input_dialog()}}
\item \href{#method-TelegramClient-get_input_notify}{\code{TelegramClient$get_input_notify()}}
\item \href{#method-TelegramClient-download_files_parallel}{\code{TelegramClient$download_files_parallel()}}
\item \href{#method-TelegramClient-new}{\code{TelegramClient$new()}}
\item \href{#method-TelegramClient-clone}{\code{TelegramClient$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="clear_session_auth_key"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-clear_session_auth_key'><code>telegramR::TelegramBaseClient$clear_session_auth_key()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="connect"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-connect'><code>telegramR::TelegramBaseClient$connect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="disconnect"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-disconnect'><code>telegramR::TelegramBaseClient$disconnect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="get_flood_sleep_threshold"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-get_flood_sleep_threshold'><code>telegramR::TelegramBaseClient$get_flood_sleep_threshold()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="get_proxy"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-get_proxy'><code>telegramR::TelegramBaseClient$get_proxy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="get_session_path"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-get_session_path'><code>telegramR::TelegramBaseClient$get_session_path()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="get_version"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-get_version'><code>telegramR::TelegramBaseClient$get_version()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="is_connected"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-is_connected'><code>telegramR::TelegramBaseClient$is_connected()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="set_dc"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-set_dc'><code>telegramR::TelegramBaseClient$set_dc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="set_flood_sleep_threshold"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-set_flood_sleep_threshold'><code>telegramR::TelegramBaseClient$set_flood_sleep_threshold()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="set_proxy"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-set_proxy'><code>telegramR::TelegramBaseClient$set_proxy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="switch_dc"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-switch_dc'><code>telegramR::TelegramBaseClient$switch_dc()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-takeout"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-takeout}{}}}
\subsection{Method \code{takeout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$takeout(
  finalize = TRUE,
  contacts = NULL,
  users = NULL,
  chats = NULL,
  megagroups = NULL,
  channels = NULL,
  files = NULL,
  max_file_size = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-end_takeout"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-end_takeout}{}}}
\subsection{Method \code{end_takeout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$end_takeout(success)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-start"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-start}{}}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$start(
  phone = NULL,
  password = NULL,
  bot_token = NULL,
  force_sms = FALSE,
  code_callback = NULL,
  first_name = "New User",
  last_name = "",
  max_attempts = 3
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-sign_in"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-sign_in}{}}}
\subsection{Method \code{sign_in()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$sign_in(
  phone = NULL,
  code = NULL,
  password = NULL,
  bot_token = NULL,
  phone_code_hash = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-sign_up"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-sign_up}{}}}
\subsection{Method \code{sign_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$sign_up(
  code,
  first_name,
  last_name = "",
  phone = NULL,
  phone_code_hash = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-send_code_request"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-send_code_request}{}}}
\subsection{Method \code{send_code_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$send_code_request(phone, force_sms = FALSE, retry_count = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-qr_login"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-qr_login}{}}}
\subsection{Method \code{qr_login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$qr_login(ignored_ids = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-log_out"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-log_out}{}}}
\subsection{Method \code{log_out()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$log_out()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-edit_2fa"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-edit_2fa}{}}}
\subsection{Method \code{edit_2fa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$edit_2fa(
  current_password = NULL,
  new_password = NULL,
  hint = "",
  email = NULL,
  email_code_callback = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-parse_phone_and_hash"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-parse_phone_and_hash}{}}}
\subsection{Method \code{parse_phone_and_hash()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$parse_phone_and_hash(phone, phone_hash)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-on_login"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-on_login}{}}}
\subsection{Method \code{on_login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$on_login(user)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-start_impl"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-start_impl}{}}}
\subsection{Method \code{start_impl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$start_impl(
  phone,
  password,
  bot_token,
  force_sms,
  code_callback,
  first_name,
  last_name,
  max_attempts
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-parse_phone"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-parse_phone}{}}}
\subsection{Method \code{parse_phone()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$parse_phone(phone)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-compute_check"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-compute_check}{}}}
\subsection{Method \code{compute_check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$compute_check(request, password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-compute_digest"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-compute_digest}{}}}
\subsection{Method \code{compute_digest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$compute_digest(algo, password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_display_name"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_display_name}{}}}
\subsection{Method \code{get_display_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_display_name(user)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-download_profile_photo"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-download_profile_photo}{}}}
\subsection{Method \code{download_profile_photo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$download_profile_photo(entity, file = NULL, download_big = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-download_media"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-download_media}{}}}
\subsection{Method \code{download_media()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$download_media(
  message,
  file = NULL,
  thumb = NULL,
  progress_callback = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-download_file"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-download_file}{}}}
\subsection{Method \code{download_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$download_file(
  input_location,
  file = NULL,
  part_size_kb = NULL,
  file_size = NULL,
  progress_callback = NULL,
  dc_id = NULL,
  key = NULL,
  iv = NULL,
  msg_data = NULL,
  cdn_redirect = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-.download_file"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-.download_file}{}}}
\subsection{Method \code{.download_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$.download_file(
  input_location,
  file = NULL,
  part_size_kb = NULL,
  file_size = NULL,
  progress_callback = NULL,
  dc_id = NULL,
  key = NULL,
  iv = NULL,
  msg_data = NULL,
  cdn_redirect = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-iter_download"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-iter_download}{}}}
\subsection{Method \code{iter_download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$iter_download(
  file,
  offset = 0,
  stride = NULL,
  limit = NULL,
  chunk_size = NULL,
  request_size = MAX_CHUNK_SIZE,
  file_size = NULL,
  dc_id = NULL,
  msg_data = NULL,
  cdn_redirect = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-.iter_download"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-.iter_download}{}}}
\subsection{Method \code{.iter_download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$.iter_download(
  file,
  offset = 0,
  stride = NULL,
  limit = NULL,
  chunk_size = NULL,
  request_size = MAX_CHUNK_SIZE,
  file_size = NULL,
  dc_id = NULL,
  msg_data = NULL,
  cdn_redirect = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_thumb"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_thumb}{}}}
\subsection{Method \code{get_thumb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_thumb(thumbs, thumb)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-download_cached_photo_size"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-download_cached_photo_size}{}}}
\subsection{Method \code{download_cached_photo_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$download_cached_photo_size(size, file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-download_photo"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-download_photo}{}}}
\subsection{Method \code{download_photo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$download_photo(photo, file, date, thumb, progress_callback)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_kind_and_names"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_kind_and_names}{}}}
\subsection{Method \code{get_kind_and_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_kind_and_names(attributes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-download_document"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-download_document}{}}}
\subsection{Method \code{download_document()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$download_document(
  document,
  file,
  date,
  thumb,
  progress_callback,
  msg_data
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-download_contact"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-download_contact}{}}}
\subsection{Method \code{download_contact()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$download_contact(mm_contact, file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-download_web_document"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-download_web_document}{}}}
\subsection{Method \code{download_web_document()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$download_web_document(web, file, progress_callback)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_proper_filename"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_proper_filename}{}}}
\subsection{Method \code{get_proper_filename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_proper_filename(
  file,
  kind,
  extension,
  date = NULL,
  possible_names = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-iter_dialogs"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-iter_dialogs}{}}}
\subsection{Method \code{iter_dialogs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$iter_dialogs(
  limit = NULL,
  offset_date = NULL,
  offset_id = 0,
  offset_peer = InputPeerEmpty$new(),
  ignore_pinned = FALSE,
  ignore_migrated = FALSE,
  folder = NULL,
  archived = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_dialogs"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_dialogs}{}}}
\subsection{Method \code{get_dialogs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_dialogs(
  limit = NULL,
  offset_date = NULL,
  offset_id = 0,
  offset_peer = InputPeerEmpty$new(),
  ignore_pinned = FALSE,
  ignore_migrated = FALSE,
  folder = NULL,
  archived = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-iter_drafts"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-iter_drafts}{}}}
\subsection{Method \code{iter_drafts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$iter_drafts(entity = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_drafts"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_drafts}{}}}
\subsection{Method \code{get_drafts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_drafts(entity = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-edit_folder"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-edit_folder}{}}}
\subsection{Method \code{edit_folder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$edit_folder(entity = NULL, folder = NULL, unpack = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-delete_dialog"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-delete_dialog}{}}}
\subsection{Method \code{delete_dialog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$delete_dialog(entity, revoke = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-conversation"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-conversation}{}}}
\subsection{Method \code{conversation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$conversation(
  entity,
  timeout = 60,
  total_timeout = NULL,
  max_messages = 100,
  exclusive = TRUE,
  replies_are_responses = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-iter_participants"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-iter_participants}{}}}
\subsection{Method \code{iter_participants()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$iter_participants(
  entity,
  limit = NULL,
  search = "",
  filter = NULL,
  aggressive = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_participants"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_participants}{}}}
\subsection{Method \code{get_participants()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_participants(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-iter_admin_log"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-iter_admin_log}{}}}
\subsection{Method \code{iter_admin_log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$iter_admin_log(
  entity,
  limit = NULL,
  max_id = 0,
  min_id = 0,
  search = NULL,
  admins = NULL,
  join = NULL,
  leave = NULL,
  invite = NULL,
  restrict = NULL,
  unrestrict = NULL,
  ban = NULL,
  unban = NULL,
  promote = NULL,
  demote = NULL,
  info = NULL,
  settings = NULL,
  pinned = NULL,
  edit = NULL,
  delete = NULL,
  group_call = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_admin_log"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_admin_log}{}}}
\subsection{Method \code{get_admin_log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_admin_log(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-iter_profile_photos"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-iter_profile_photos}{}}}
\subsection{Method \code{iter_profile_photos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$iter_profile_photos(
  entity,
  limit = NULL,
  offset = 0,
  max_id = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_profile_photos"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_profile_photos}{}}}
\subsection{Method \code{get_profile_photos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_profile_photos(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-action"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-action}{}}}
\subsection{Method \code{action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$action(entity, action, delay = 4, auto_cancel = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-edit_admin"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-edit_admin}{}}}
\subsection{Method \code{edit_admin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$edit_admin(
  entity,
  user,
  change_info = NULL,
  post_messages = NULL,
  edit_messages = NULL,
  delete_messages = NULL,
  ban_users = NULL,
  invite_users = NULL,
  pin_messages = NULL,
  add_admins = NULL,
  manage_call = NULL,
  anonymous = NULL,
  is_admin = NULL,
  title = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-edit_permissions"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-edit_permissions}{}}}
\subsection{Method \code{edit_permissions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$edit_permissions(
  entity,
  user = NULL,
  until_date = NULL,
  view_messages = TRUE,
  send_messages = TRUE,
  send_media = TRUE,
  send_stickers = TRUE,
  send_gifs = TRUE,
  send_games = TRUE,
  send_inline = TRUE,
  embed_link_previews = TRUE,
  send_polls = TRUE,
  change_info = TRUE,
  invite_users = TRUE,
  pin_messages = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-kick_participant"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-kick_participant}{}}}
\subsection{Method \code{kick_participant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$kick_participant(entity, user)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_permissions"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_permissions}{}}}
\subsection{Method \code{get_permissions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_permissions(entity, user = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_stats"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_stats}{}}}
\subsection{Method \code{get_stats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_stats(entity, message = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-inline_query"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-inline_query}{}}}
\subsection{Method \code{inline_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$inline_query(
  bot,
  query,
  entity = NULL,
  offset = NULL,
  geo_point = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-invoke_function"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-invoke_function}{}}}
\subsection{Method \code{invoke_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$invoke_function(function_name, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-custom_inline_results"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-custom_inline_results}{}}}
\subsection{Method \code{custom_inline_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$custom_inline_results(result, peer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-iter_messages"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-iter_messages}{}}}
\subsection{Method \code{iter_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$iter_messages(
  entity,
  limit = NULL,
  offset_date = NULL,
  offset_id = 0L,
  max_id = 0L,
  min_id = 0L,
  add_offset = 0L,
  search = NULL,
  filter = NULL,
  from_user = NULL,
  wait_time = NULL,
  ids = NULL,
  reverse = FALSE,
  reply_to = NULL,
  scheduled = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-iter_messages_page"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-iter_messages_page}{}}}
\subsection{Method \code{iter_messages_page()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$iter_messages_page(
  entity,
  limit = NULL,
  offset_date = NULL,
  offset_id = 0L,
  max_id = 0L,
  min_id = 0L,
  add_offset = 0L,
  search = NULL,
  filter = NULL,
  from_user = NULL,
  wait_time = NULL,
  reverse = FALSE,
  reply_to = NULL,
  scheduled = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_messages"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_messages}{}}}
\subsection{Method \code{get_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_messages(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-channels_get_messages"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-channels_get_messages}{}}}
\subsection{Method \code{channels_get_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$channels_get_messages(channel, ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-collect"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-collect}{}}}
\subsection{Method \code{collect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$collect(it)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-collect_one"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-collect_one}{}}}
\subsection{Method \code{collect_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$collect_one(it)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-send_message"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-send_message}{}}}
\subsection{Method \code{send_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$send_message(
  entity,
  message = "",
  reply_to = NULL,
  attributes = NULL,
  parse_mode = NULL,
  formatting_entities = NULL,
  link_preview = TRUE,
  file = NULL,
  thumb = NULL,
  force_document = FALSE,
  clear_draft = FALSE,
  buttons = NULL,
  silent = NULL,
  background = NULL,
  supports_streaming = FALSE,
  schedule = NULL,
  comment_to = NULL,
  nosound_video = NULL,
  send_as = NULL,
  message_effect_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-forward_messages"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-forward_messages}{}}}
\subsection{Method \code{forward_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$forward_messages(
  entity,
  messages,
  from_peer = NULL,
  background = NULL,
  with_my_score = NULL,
  silent = NULL,
  as_album = NULL,
  schedule = NULL,
  drop_author = NULL,
  drop_media_captions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-edit_message"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-edit_message}{}}}
\subsection{Method \code{edit_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$edit_message(
  entity,
  message = NULL,
  text = NULL,
  parse_mode = NULL,
  attributes = NULL,
  formatting_entities = NULL,
  link_preview = TRUE,
  file = NULL,
  thumb = NULL,
  force_document = FALSE,
  buttons = NULL,
  supports_streaming = FALSE,
  schedule = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-delete_messages"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-delete_messages}{}}}
\subsection{Method \code{delete_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$delete_messages(entity, message_ids, revoke = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-send_read_acknowledge"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-send_read_acknowledge}{}}}
\subsection{Method \code{send_read_acknowledge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$send_read_acknowledge(
  entity,
  message = NULL,
  max_id = NULL,
  clear_mentions = FALSE,
  clear_reactions = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-pin_message"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-pin_message}{}}}
\subsection{Method \code{pin_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$pin_message(entity, message, notify = FALSE, pm_oneside = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-unpin_message"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-unpin_message}{}}}
\subsection{Method \code{unpin_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$unpin_message(entity, message = NULL, notify = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_comment_data"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_comment_data}{}}}
\subsection{Method \code{get_comment_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_comment_data(entity, message)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-pin_internal"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-pin_internal}{}}}
\subsection{Method \code{pin_internal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$pin_internal(
  entity,
  message,
  unpin,
  notify = FALSE,
  pm_oneside = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-send_file"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-send_file}{}}}
\subsection{Method \code{send_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$send_file(
  entity,
  file,
  caption = NULL,
  force_document = FALSE,
  file_size = NULL,
  progress_callback = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-upload_file"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-upload_file}{}}}
\subsection{Method \code{upload_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$upload_file(
  file,
  part_size_kb = NULL,
  file_size = NULL,
  progress_callback = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-file_to_media"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-file_to_media}{}}}
\subsection{Method \code{file_to_media()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$file_to_media(
  file,
  force_document = FALSE,
  file_size = NULL,
  progress_callback = NULL,
  attributes = NULL,
  thumb = NULL,
  allow_cache = TRUE,
  voice_note = FALSE,
  video_note = FALSE,
  supports_streaming = FALSE,
  mime_type = NULL,
  as_image = NULL,
  ttl = NULL,
  nosound_video = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-resize_photo_if_needed"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-resize_photo_if_needed}{}}}
\subsection{Method \code{resize_photo_if_needed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$resize_photo_if_needed(
  file,
  is_image,
  width = 2560,
  height = 2560
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-is_image"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-is_image}{}}}
\subsection{Method \code{is_image()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$is_image(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_appropriated_part_size"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_appropriated_part_size}{}}}
\subsection{Method \code{get_appropriated_part_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_appropriated_part_size(file_size)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-invoke"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-invoke}{}}}
\subsection{Method \code{invoke()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$invoke(request)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-build_reply_markup"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-build_reply_markup}{}}}
\subsection{Method \code{build_reply_markup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$build_reply_markup(buttons = NULL, inline_only = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-run_until_disconnected"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-run_until_disconnected}{}}}
\subsection{Method \code{run_until_disconnected()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$run_until_disconnected()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-set_receive_updates"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-set_receive_updates}{}}}
\subsection{Method \code{set_receive_updates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$set_receive_updates(receive_updates)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-on"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-on}{}}}
\subsection{Method \code{on()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$on(event)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-add_event_handler"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-add_event_handler}{}}}
\subsection{Method \code{add_event_handler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$add_event_handler(callback, event = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-remove_event_handler"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-remove_event_handler}{}}}
\subsection{Method \code{remove_event_handler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$remove_event_handler(callback, event = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-list_event_handlers"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-list_event_handlers}{}}}
\subsection{Method \code{list_event_handlers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$list_event_handlers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-catch_up"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-catch_up}{}}}
\subsection{Method \code{catch_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$catch_up()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-update_loop"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-update_loop}{}}}
\subsection{Method \code{update_loop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$update_loop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-preprocess_updates"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-preprocess_updates}{}}}
\subsection{Method \code{preprocess_updates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$preprocess_updates(updates, users, chats)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-keepalive_loop"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-keepalive_loop}{}}}
\subsection{Method \code{keepalive_loop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$keepalive_loop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-dispatch_update"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-dispatch_update}{}}}
\subsection{Method \code{dispatch_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$dispatch_update(update)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-handle_auto_reconnect"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-handle_auto_reconnect}{}}}
\subsection{Method \code{handle_auto_reconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$handle_auto_reconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-set_parse_mode"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-set_parse_mode}{}}}
\subsection{Method \code{set_parse_mode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$set_parse_mode(mode)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_parse_mode"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_parse_mode}{}}}
\subsection{Method \code{get_parse_mode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_parse_mode()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-sanitize_parse_mode"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-sanitize_parse_mode}{}}}
\subsection{Method \code{sanitize_parse_mode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$sanitize_parse_mode(mode)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-parse_message_text"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-parse_message_text}{}}}
\subsection{Method \code{parse_message_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$parse_message_text(message, parse_mode = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-replace_with_mention"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-replace_with_mention}{}}}
\subsection{Method \code{replace_with_mention()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$replace_with_mention(entities, i, user)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_response_message"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_response_message}{}}}
\subsection{Method \code{get_response_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_response_message(request, result, input_chat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-call"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-call}{}}}
\subsection{Method \code{call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$call(request, ordered = FALSE, flood_sleep_threshold = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-call_internal"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-call_internal}{}}}
\subsection{Method \code{call_internal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$call_internal(
  sender,
  request,
  ordered = FALSE,
  flood_sleep_threshold = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_me"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_me}{}}}
\subsection{Method \code{get_me()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_me(input_peer = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-self_id"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-self_id}{}}}
\subsection{Method \code{self_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$self_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-is_bot"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-is_bot}{}}}
\subsection{Method \code{is_bot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$is_bot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-is_user_authorized"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-is_user_authorized}{}}}
\subsection{Method \code{is_user_authorized()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$is_user_authorized()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_entity"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_entity}{}}}
\subsection{Method \code{get_entity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_entity(entity)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_input_entity"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_input_entity}{}}}
\subsection{Method \code{get_input_entity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_input_entity(peer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-getInputEntity"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-getInputEntity}{}}}
\subsection{Method \code{getInputEntity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$getInputEntity(peer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-getInputPeer"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-getInputPeer}{}}}
\subsection{Method \code{getInputPeer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$getInputPeer(peer, allow_self = TRUE, check_hash = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-getInputUser"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-getInputUser}{}}}
\subsection{Method \code{getInputUser()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$getInputUser(entity)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-getInputChannel"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-getInputChannel}{}}}
\subsection{Method \code{getInputChannel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$getInputChannel(entity)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-getInputMessage"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-getInputMessage}{}}}
\subsection{Method \code{getInputMessage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$getInputMessage(message)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-getInputMedia"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-getInputMedia}{}}}
\subsection{Method \code{getInputMedia()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$getInputMedia(
  media,
  is_photo = FALSE,
  attributes = NULL,
  force_document = FALSE,
  file_size = NULL,
  progress_callback = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-getInputDocument"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-getInputDocument}{}}}
\subsection{Method \code{getInputDocument()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$getInputDocument(document)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-getInputPhoto"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-getInputPhoto}{}}}
\subsection{Method \code{getInputPhoto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$getInputPhoto(photo)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-getInputChatPhoto"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-getInputChatPhoto}{}}}
\subsection{Method \code{getInputChatPhoto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$getInputChatPhoto(photo)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-getInputGroupCall"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-getInputGroupCall}{}}}
\subsection{Method \code{getInputGroupCall()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$getInputGroupCall(call)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_peer"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_peer}{}}}
\subsection{Method \code{get_peer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_peer(peer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_peer_id"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_peer_id}{}}}
\subsection{Method \code{get_peer_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_peer_id(peer, add_mark = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_entity_from_string"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_entity_from_string}{}}}
\subsection{Method \code{get_entity_from_string()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_entity_from_string(string)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_input_dialog"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_input_dialog}{}}}
\subsection{Method \code{get_input_dialog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_input_dialog(dialog)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-get_input_notify"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-get_input_notify}{}}}
\subsection{Method \code{get_input_notify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$get_input_notify(notify)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-download_files_parallel"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-download_files_parallel}{}}}
\subsection{Method \code{download_files_parallel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$download_files_parallel(
  items,
  output_dir = NULL,
  workers = 4L,
  stagger_ms = 500L
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
