% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegramR-package.R
\docType{package}
\name{telegramR-package}
\alias{telegramR}
\alias{telegramR-package}
\title{telegramR: deal with exported chats/channels from Telegram}
\description{
The telegramR is a package to deal with exported chats from Telegram messenger (https://www.telegram.org/). This package allows the transformation of HTML from exported chats/channels in tibble/dataframe, as well as have some functions to a brief summary about the exported chats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://soaresalisson.github.io/telegramR/}
}

}
\author{
\strong{Maintainer}: Alisson Soares \email{alissonmsoares@gmail.com} (\href{https://orcid.org/0009-0004-5497-5572}{ORCID})

}
\keyword{internal}
