% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{TG_JOIN_RE}
\alias{TG_JOIN_RE}
\title{Telegram Join Regular Expression}
\format{
An object of class \code{character} of length 1.
}
\usage{
TG_JOIN_RE
}
\description{
Regular expression pattern for matching Telegram join links.
}
\keyword{datasets}
