% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_info.R
\name{tm_info}
\alias{tm_info}
\title{return general information}
\usage{
tm_info(folders)
}
\arguments{
\item{folders}{CHAR. a folder or vector of folders with telegram exported HTML files}
}
\description{
return general information
}
