% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_account.R
\name{ToggleConnectedBotPausedRequest}
\alias{ToggleConnectedBotPausedRequest}
\title{ToggleConnectedBotPausedRequest}
\description{
R6 class representing a ToggleConnectedBotPausedRequest.
}
\details{
This class handles toggling the paused state of a connected bot for a peer.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{ToggleConnectedBotPausedRequest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ToggleConnectedBotPausedRequest-new}{\code{ToggleConnectedBotPausedRequest$new()}}
\item \href{#method-ToggleConnectedBotPausedRequest-resolve}{\code{ToggleConnectedBotPausedRequest$resolve()}}
\item \href{#method-ToggleConnectedBotPausedRequest-toDict}{\code{ToggleConnectedBotPausedRequest$toDict()}}
\item \href{#method-ToggleConnectedBotPausedRequest-bytes}{\code{ToggleConnectedBotPausedRequest$bytes()}}
\item \href{#method-ToggleConnectedBotPausedRequest-fromReader}{\code{ToggleConnectedBotPausedRequest$fromReader()}}
\item \href{#method-ToggleConnectedBotPausedRequest-clone}{\code{ToggleConnectedBotPausedRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleConnectedBotPausedRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleConnectedBotPausedRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the ToggleConnectedBotPausedRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleConnectedBotPausedRequest$new(peer, paused)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{The input peer.}

\item{\code{paused}}{Boolean indicating if the bot is paused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleConnectedBotPausedRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleConnectedBotPausedRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve the peer using client and utils.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleConnectedBotPausedRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The client object.}

\item{\code{utils}}{The utils object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleConnectedBotPausedRequest-toDict"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleConnectedBotPausedRequest-toDict}{}}}
\subsection{Method \code{toDict()}}{
Convert to dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleConnectedBotPausedRequest$toDict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleConnectedBotPausedRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleConnectedBotPausedRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleConnectedBotPausedRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleConnectedBotPausedRequest-fromReader"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleConnectedBotPausedRequest-fromReader}{}}}
\subsection{Method \code{fromReader()}}{
Create from reader.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleConnectedBotPausedRequest$fromReader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{The reader object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of ToggleConnectedBotPausedRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleConnectedBotPausedRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleConnectedBotPausedRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleConnectedBotPausedRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
