% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{ToggleGroupCallRecordRequest}
\alias{ToggleGroupCallRecordRequest}
\title{ToggleGroupCallRecordRequest}
\description{
Telegram API type ToggleGroupCallRecordRequest
}
\details{
ToggleGroupCallRecordRequest R6 class

Represents ToggleGroupCallRecordRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{Field.}

\item{\code{start}}{Field.}

\item{\code{video}}{Field.}

\item{\code{title}}{Field.}

\item{\code{video_portrait}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ToggleGroupCallRecordRequest-new}{\code{ToggleGroupCallRecordRequest$new()}}
\item \href{#method-ToggleGroupCallRecordRequest-resolve}{\code{ToggleGroupCallRecordRequest$resolve()}}
\item \href{#method-ToggleGroupCallRecordRequest-to_list}{\code{ToggleGroupCallRecordRequest$to_list()}}
\item \href{#method-ToggleGroupCallRecordRequest-bytes}{\code{ToggleGroupCallRecordRequest$bytes()}}
\item \href{#method-ToggleGroupCallRecordRequest-from_reader}{\code{ToggleGroupCallRecordRequest$from_reader()}}
\item \href{#method-ToggleGroupCallRecordRequest-clone}{\code{ToggleGroupCallRecordRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallRecordRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallRecordRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ToggleGroupCallRecordRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallRecordRequest$new(
  call,
  start = NULL,
  video = NULL,
  title = NULL,
  video_portrait = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call (object).}

\item{\code{start}}{Logical or NULL.}

\item{\code{video}}{Logical or NULL.}

\item{\code{title}}{Optional string or NULL.}

\item{\code{video_portrait}}{Logical or NULL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallRecordRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallRecordRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallRecordRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallRecordRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallRecordRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallRecordRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallRecordRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallRecordRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

This constructs the TL-serialized bytes. Assumes `call$bytes()` exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallRecordRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallRecordRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallRecordRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallRecordRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallRecordRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallRecordRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallRecordRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
