% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{ToggleGroupCallStartSubscriptionRequest}
\alias{ToggleGroupCallStartSubscriptionRequest}
\title{ToggleGroupCallStartSubscriptionRequest}
\description{
Telegram API type ToggleGroupCallStartSubscriptionRequest
}
\details{
ToggleGroupCallStartSubscriptionRequest R6 class

Represents ToggleGroupCallStartSubscriptionRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{call}}{The input group call object (expected to implement bytes()).}

\item{\code{subscribed}}{Logical indicating subscription state.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ToggleGroupCallStartSubscriptionRequest-new}{\code{ToggleGroupCallStartSubscriptionRequest$new()}}
\item \href{#method-ToggleGroupCallStartSubscriptionRequest-resolve}{\code{ToggleGroupCallStartSubscriptionRequest$resolve()}}
\item \href{#method-ToggleGroupCallStartSubscriptionRequest-to_list}{\code{ToggleGroupCallStartSubscriptionRequest$to_list()}}
\item \href{#method-ToggleGroupCallStartSubscriptionRequest-bytes}{\code{ToggleGroupCallStartSubscriptionRequest$bytes()}}
\item \href{#method-ToggleGroupCallStartSubscriptionRequest-from_reader}{\code{ToggleGroupCallStartSubscriptionRequest$from_reader()}}
\item \href{#method-ToggleGroupCallStartSubscriptionRequest-clone}{\code{ToggleGroupCallStartSubscriptionRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallStartSubscriptionRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallStartSubscriptionRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ToggleGroupCallStartSubscriptionRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallStartSubscriptionRequest$new(call, subscribed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{Input group call (object).}

\item{\code{subscribed}}{Logical (TRUE/FALSE).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallStartSubscriptionRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallStartSubscriptionRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve input references
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallStartSubscriptionRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{Client object (passed to utils functions).}

\item{\code{utils}}{Utility object with get_input_group_call method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallStartSubscriptionRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallStartSubscriptionRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallStartSubscriptionRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallStartSubscriptionRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallStartSubscriptionRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

This constructs the TL-serialized bytes. Assumes `call$bytes()` exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallStartSubscriptionRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallStartSubscriptionRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallStartSubscriptionRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallStartSubscriptionRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleGroupCallStartSubscriptionRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleGroupCallStartSubscriptionRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ToggleGroupCallStartSubscriptionRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
