% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{TypeNotFoundError}
\alias{TypeNotFoundError}
\title{TypeNotFoundError}
\description{
Telegram API type TypeNotFoundError
}
\details{
TypeNotFoundError
}
\section{Super class}{
\code{\link[telegramR:Exception]{telegramR::Exception}} -> \code{TypeNotFoundError}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{invalid_constructor_id}}{\code{numeric} The invalid constructor ID.}

\item{\code{remaining}}{\code{character} The remaining bytes in the buffer.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TypeNotFoundError-new}{\code{TypeNotFoundError$new()}}
\item \href{#method-TypeNotFoundError-clone}{\code{TypeNotFoundError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TypeNotFoundError-new"></a>}}
\if{latex}{\out{\hypertarget{method-TypeNotFoundError-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the TypeNotFoundError class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeNotFoundError$new(invalid_constructor_id, remaining)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{invalid_constructor_id}}{\code{numeric} The invalid constructor ID.}

\item{\code{remaining}}{\code{character} The remaining bytes in the buffer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TypeNotFoundError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TypeNotFoundError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TypeNotFoundError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
