% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_contacts.R
\name{UnblockRequest}
\alias{UnblockRequest}
\title{UnblockRequest}
\description{
Methods:
- initialize(id, my_stories_from = NULL): create new request
- resolve(client, utils): resolve id into input peer using client/utils
- to_list(): return a list representation
- to_bytes(): return raw vector of bytes for the TL request (little-endian integer packing used)
- from_reader(reader): class method to read from a reader and return a new instance
}
\details{
R6 representation of the TL request: UnblockRequest
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{UnblockRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex)}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex)}

\item{\code{id}}{TLObject|list Input peer object (will be resolved with utils)}

\item{\code{my_stories_from}}{logical|NULL Optional flag}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnblockRequest-new}{\code{UnblockRequest$new()}}
\item \href{#method-UnblockRequest-resolve}{\code{UnblockRequest$resolve()}}
\item \href{#method-UnblockRequest-to_list}{\code{UnblockRequest$to_list()}}
\item \href{#method-UnblockRequest-to_bytes}{\code{UnblockRequest$to_bytes()}}
\item \href{#method-UnblockRequest-clone}{\code{UnblockRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnblockRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnblockRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize UnblockRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnblockRequest$new(id, my_stories_from = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{input peer (object or identifier)}

\item{\code{my_stories_from}}{logical or NULL
Resolve entities using client and utils

This will replace `id` with an input_peer object obtained via utils$get_input_peer(...)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnblockRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-UnblockRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnblockRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client object with method get_input_entity()}

\item{\code{utils}}{utils object with method get_input_peer()
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnblockRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-UnblockRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnblockRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)

Packs flags as uint32 little-endian and appends id bytes.
Assumes id has method to_bytes() which returns raw vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnblockRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-UnblockRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnblockRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
Convert integer to little-endian raw vector of given size (bytes)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnblockRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnblockRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnblockRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
