% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unpack}
\alias{unpack}
\title{Unpack data from binary format}
\usage{
unpack(format, data)
}
\arguments{
\item{format}{A character string specifying the format.}

\item{data}{A raw vector of data to unpack.}
}
\value{
A list of unpacked values.
}
\description{
Mimics Python's struct.unpack.
}
