% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{unparse_telegram_to_html}
\alias{unparse_telegram_to_html}
\title{Unparse plain text and entities back to HTML}
\usage{
unparse_telegram_to_html(text, entities)
}
\arguments{
\item{text}{character(1) plain text}

\item{entities}{list of entity lists as produced by \code{parse_html_to_telegram}}
}
\value{
character(1) HTML string
}
\description{
Unparse plain text and entities back to HTML
}
