% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_profile.R
\name{update_profile}
\alias{update_profile}
\title{Update your name and/or bio}
\usage{
update_profile(client, first_name = NULL, last_name = NULL, bio = NULL)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{first_name}{First name (optional).}

\item{last_name}{Last name (optional).}

\item{bio}{About/bio text (optional).}
}
\value{
The API response.
}
\description{
Update your name and/or bio
}
