% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_profile.R
\name{update_profile_photo}
\alias{update_profile_photo}
\title{Update your profile photo}
\usage{
update_profile_photo(client, file, progress_callback = NULL)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{file}{Path to image file.}

\item{progress_callback}{Optional callback (pos, total).}
}
\value{
The API response.
}
\description{
Update your profile photo
}
