% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_profile.R
\name{update_username}
\alias{update_username}
\title{Update your username}
\usage{
update_username(client, username)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{username}{New username (without @).}
}
\value{
The API response.
}
\description{
Update your username
}
\examples{
\dontrun{
client <- TelegramClient$new("my_session", api_id = 123, api_hash = "abc")
client$connect()

update_username(client, "new_username")
}
}
