% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_account.R
\name{UpdateBusinessGreetingMessageRequest}
\alias{UpdateBusinessGreetingMessageRequest}
\title{UpdateBusinessGreetingMessageRequest}
\description{
R6 class representing an UpdateBusinessGreetingMessageRequest.
}
\details{
This class handles updating the business greeting message with an optional input business greeting message.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{UpdateBusinessGreetingMessageRequest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UpdateBusinessGreetingMessageRequest-new}{\code{UpdateBusinessGreetingMessageRequest$new()}}
\item \href{#method-UpdateBusinessGreetingMessageRequest-toDict}{\code{UpdateBusinessGreetingMessageRequest$toDict()}}
\item \href{#method-UpdateBusinessGreetingMessageRequest-bytes}{\code{UpdateBusinessGreetingMessageRequest$bytes()}}
\item \href{#method-UpdateBusinessGreetingMessageRequest-fromReader}{\code{UpdateBusinessGreetingMessageRequest$fromReader()}}
\item \href{#method-UpdateBusinessGreetingMessageRequest-clone}{\code{UpdateBusinessGreetingMessageRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="resolve"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-resolve'><code>telegramR::TLRequest$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateBusinessGreetingMessageRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateBusinessGreetingMessageRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the UpdateBusinessGreetingMessageRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateBusinessGreetingMessageRequest$new(message = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{Optional input business greeting message.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateBusinessGreetingMessageRequest-toDict"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateBusinessGreetingMessageRequest-toDict}{}}}
\subsection{Method \code{toDict()}}{
Convert to dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateBusinessGreetingMessageRequest$toDict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateBusinessGreetingMessageRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateBusinessGreetingMessageRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateBusinessGreetingMessageRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateBusinessGreetingMessageRequest-fromReader"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateBusinessGreetingMessageRequest-fromReader}{}}}
\subsection{Method \code{fromReader()}}{
Create from reader.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateBusinessGreetingMessageRequest$fromReader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{The reader object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of UpdateBusinessGreetingMessageRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateBusinessGreetingMessageRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateBusinessGreetingMessageRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateBusinessGreetingMessageRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
