% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_account.R
\name{UpdateNotifySettingsRequest}
\alias{UpdateNotifySettingsRequest}
\title{UpdateNotifySettingsRequest}
\description{
R6 class representing an UpdateNotifySettingsRequest.
}
\details{
This class handles updating notification settings for a peer.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{UpdateNotifySettingsRequest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UpdateNotifySettingsRequest-new}{\code{UpdateNotifySettingsRequest$new()}}
\item \href{#method-UpdateNotifySettingsRequest-resolve}{\code{UpdateNotifySettingsRequest$resolve()}}
\item \href{#method-UpdateNotifySettingsRequest-toDict}{\code{UpdateNotifySettingsRequest$toDict()}}
\item \href{#method-UpdateNotifySettingsRequest-bytes}{\code{UpdateNotifySettingsRequest$bytes()}}
\item \href{#method-UpdateNotifySettingsRequest-fromReader}{\code{UpdateNotifySettingsRequest$fromReader()}}
\item \href{#method-UpdateNotifySettingsRequest-clone}{\code{UpdateNotifySettingsRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateNotifySettingsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateNotifySettingsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the UpdateNotifySettingsRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateNotifySettingsRequest$new(peer, settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{The input notify peer.}

\item{\code{settings}}{The input peer notify settings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateNotifySettingsRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateNotifySettingsRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve the peer using client and utils.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateNotifySettingsRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The client object.}

\item{\code{utils}}{The utils object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateNotifySettingsRequest-toDict"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateNotifySettingsRequest-toDict}{}}}
\subsection{Method \code{toDict()}}{
Convert to dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateNotifySettingsRequest$toDict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateNotifySettingsRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateNotifySettingsRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateNotifySettingsRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateNotifySettingsRequest-fromReader"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateNotifySettingsRequest-fromReader}{}}}
\subsection{Method \code{fromReader()}}{
Create from reader.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateNotifySettingsRequest$fromReader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{The reader object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of UpdateNotifySettingsRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateNotifySettingsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateNotifySettingsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateNotifySettingsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
