% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_account.R
\name{UpdatePersonalChannelRequest}
\alias{UpdatePersonalChannelRequest}
\title{UpdatePersonalChannelRequest}
\description{
R6 class representing an UpdatePersonalChannelRequest.
}
\details{
This class handles updating the personal channel.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{UpdatePersonalChannelRequest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UpdatePersonalChannelRequest-new}{\code{UpdatePersonalChannelRequest$new()}}
\item \href{#method-UpdatePersonalChannelRequest-resolve}{\code{UpdatePersonalChannelRequest$resolve()}}
\item \href{#method-UpdatePersonalChannelRequest-toDict}{\code{UpdatePersonalChannelRequest$toDict()}}
\item \href{#method-UpdatePersonalChannelRequest-bytes}{\code{UpdatePersonalChannelRequest$bytes()}}
\item \href{#method-UpdatePersonalChannelRequest-fromReader}{\code{UpdatePersonalChannelRequest$fromReader()}}
\item \href{#method-UpdatePersonalChannelRequest-clone}{\code{UpdatePersonalChannelRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePersonalChannelRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePersonalChannelRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the UpdatePersonalChannelRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePersonalChannelRequest$new(channel)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channel}}{The input channel.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePersonalChannelRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePersonalChannelRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve the channel using client and utils.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePersonalChannelRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{The client object.}

\item{\code{utils}}{The utils object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePersonalChannelRequest-toDict"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePersonalChannelRequest-toDict}{}}}
\subsection{Method \code{toDict()}}{
Convert to dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePersonalChannelRequest$toDict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePersonalChannelRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePersonalChannelRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePersonalChannelRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePersonalChannelRequest-fromReader"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePersonalChannelRequest-fromReader}{}}}
\subsection{Method \code{fromReader()}}{
Create from reader.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePersonalChannelRequest$fromReader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{The reader object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of UpdatePersonalChannelRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePersonalChannelRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePersonalChannelRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePersonalChannelRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
