% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{UpdateUserEmojiStatusRequest}
\alias{UpdateUserEmojiStatusRequest}
\title{UpdateUserEmojiStatusRequest}
\description{
Methods:
- initialize(user_id, emoji_status)
- resolve(client, utils)
- to_list() -> list representation
- to_bytes() -> raw vector bytes

Telegram API type UpdateUserEmojiStatusRequest
}
\details{
UpdateUserEmojiStatusRequest R6 class

Represents the UpdateUserEmojiStatusRequest TL request.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_id}}{TypeInputUser}

\item{\code{emoji_status}}{TypeEmojiStatus}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UpdateUserEmojiStatusRequest-new}{\code{UpdateUserEmojiStatusRequest$new()}}
\item \href{#method-UpdateUserEmojiStatusRequest-resolve}{\code{UpdateUserEmojiStatusRequest$resolve()}}
\item \href{#method-UpdateUserEmojiStatusRequest-to_list}{\code{UpdateUserEmojiStatusRequest$to_list()}}
\item \href{#method-UpdateUserEmojiStatusRequest-to_bytes}{\code{UpdateUserEmojiStatusRequest$to_bytes()}}
\item \href{#method-UpdateUserEmojiStatusRequest-clone}{\code{UpdateUserEmojiStatusRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateUserEmojiStatusRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateUserEmojiStatusRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize UpdateUserEmojiStatusRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateUserEmojiStatusRequest$new(user_id, emoji_status)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{TypeInputUser or identifier}

\item{\code{emoji_status}}{TypeEmojiStatus object
Resolve references (convert entities via client/utils)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateUserEmojiStatusRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateUserEmojiStatusRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateUserEmojiStatusRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client with get_input_entity method}

\item{\code{utils}}{utils with get_input_user / get_input_media etc
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateUserEmojiStatusRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateUserEmojiStatusRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateUserEmojiStatusRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateUserEmojiStatusRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateUserEmojiStatusRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateUserEmojiStatusRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateUserEmojiStatusRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateUserEmojiStatusRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateUserEmojiStatusRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
