% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{User}
\alias{User}
\title{User}
\description{
Telegram API type User
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{User}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-User-new}{\code{User$new()}}
\item \href{#method-User-to_dict}{\code{User$to_dict()}}
\item \href{#method-User-bytes}{\code{User$bytes()}}
\item \href{#method-User-clone}{\code{User$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-new"></a>}}
\if{latex}{\out{\hypertarget{method-User-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(
  id,
  is_self = NULL,
  contact = NULL,
  mutual_contact = NULL,
  deleted = NULL,
  bot = NULL,
  bot_chat_history = NULL,
  bot_nochats = NULL,
  verified = NULL,
  restricted = NULL,
  min = NULL,
  bot_inline_geo = NULL,
  support = NULL,
  scam = NULL,
  apply_min_photo = NULL,
  fake = NULL,
  bot_attach_menu = NULL,
  premium = NULL,
  attach_menu_enabled = NULL,
  bot_can_edit = NULL,
  close_friend = NULL,
  stories_hidden = NULL,
  stories_unavailable = NULL,
  contact_require_premium = NULL,
  bot_business = NULL,
  bot_has_main_app = NULL,
  bot_forum_view = NULL,
  access_hash = NULL,
  first_name = NULL,
  last_name = NULL,
  username = NULL,
  phone = NULL,
  photo = NULL,
  status = NULL,
  bot_info_version = NULL,
  restriction_reason = NULL,
  bot_inline_placeholder = NULL,
  lang_code = NULL,
  emoji_status = NULL,
  usernames = NULL,
  stories_max_id = NULL,
  color = NULL,
  profile_color = NULL,
  bot_active_users = NULL,
  bot_verification_icon = NULL,
  send_paid_messages_stars = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-User-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-User-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-clone"></a>}}
\if{latex}{\out{\hypertarget{method-User-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
