% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{UserFull}
\alias{UserFull}
\title{UserFull}
\description{
Telegram API type UserFull
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{UserFull}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserFull-new}{\code{UserFull$new()}}
\item \href{#method-UserFull-to_dict}{\code{UserFull$to_dict()}}
\item \href{#method-UserFull-bytes}{\code{UserFull$bytes()}}
\item \href{#method-UserFull-clone}{\code{UserFull$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserFull-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserFull-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserFull$new(
  id,
  settings,
  notify_settings,
  common_chats_count,
  blocked = NULL,
  phone_calls_available = NULL,
  phone_calls_private = NULL,
  can_pin_message = NULL,
  has_scheduled = NULL,
  video_calls_available = NULL,
  voice_messages_forbidden = NULL,
  translations_disabled = NULL,
  stories_pinned_available = NULL,
  blocked_my_stories_from = NULL,
  wallpaper_overridden = NULL,
  contact_require_premium = NULL,
  read_dates_private = NULL,
  sponsored_enabled = NULL,
  can_view_revenue = NULL,
  bot_can_manage_emoji_status = NULL,
  display_gifts_button = NULL,
  about = NULL,
  personal_photo = NULL,
  profile_photo = NULL,
  fallback_photo = NULL,
  bot_info = NULL,
  pinned_msg_id = NULL,
  folder_id = NULL,
  ttl_period = NULL,
  theme = NULL,
  private_forward_name = NULL,
  bot_group_admin_rights = NULL,
  bot_broadcast_admin_rights = NULL,
  wallpaper = NULL,
  stories = NULL,
  business_work_hours = NULL,
  business_location = NULL,
  business_greeting_message = NULL,
  business_away_message = NULL,
  business_intro = NULL,
  birthday = NULL,
  personal_channel_id = NULL,
  personal_channel_message = NULL,
  stargifts_count = NULL,
  starref_program = NULL,
  bot_verification = NULL,
  send_paid_messages_stars = NULL,
  disallowed_gifts = NULL,
  stars_rating = NULL,
  stars_my_pending_rating = NULL,
  stars_my_pending_rating_date = NULL,
  main_tab = NULL,
  saved_music = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserFull-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-UserFull-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserFull$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserFull-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-UserFull-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserFull$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserFull-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserFull-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserFull$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
