% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{USERNAME_RE}
\alias{USERNAME_RE}
\title{Username Regular Expression}
\format{
An object of class \code{character} of length 1.
}
\usage{
USERNAME_RE
}
\description{
Regular expression pattern for matching usernames in Telegram links or mentions.
}
\keyword{datasets}
