% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{VALID_USERNAME_RE}
\alias{VALID_USERNAME_RE}
\title{Valid Username Regular Expression}
\format{
An object of class \code{character} of length 1.
}
\usage{
VALID_USERNAME_RE
}
\description{
Regular expression pattern for validating Telegram usernames.
}
\keyword{datasets}
