% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_account.R
\name{VerifyPhoneRequest}
\alias{VerifyPhoneRequest}
\title{VerifyPhoneRequest}
\description{
R6 class representing a VerifyPhoneRequest.
}
\details{
This class handles phone verification requests.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{VerifyPhoneRequest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VerifyPhoneRequest-new}{\code{VerifyPhoneRequest$new()}}
\item \href{#method-VerifyPhoneRequest-toDict}{\code{VerifyPhoneRequest$toDict()}}
\item \href{#method-VerifyPhoneRequest-bytes}{\code{VerifyPhoneRequest$bytes()}}
\item \href{#method-VerifyPhoneRequest-clone}{\code{VerifyPhoneRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="resolve"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-resolve'><code>telegramR::TLRequest$resolve()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VerifyPhoneRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-VerifyPhoneRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the VerifyPhoneRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VerifyPhoneRequest$new(phoneNumber, phoneCodeHash, phoneCode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phoneNumber}}{The phone number.}

\item{\code{phoneCodeHash}}{The phone code hash.}

\item{\code{phoneCode}}{The phone code.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VerifyPhoneRequest-toDict"></a>}}
\if{latex}{\out{\hypertarget{method-VerifyPhoneRequest-toDict}{}}}
\subsection{Method \code{toDict()}}{
Convert to dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VerifyPhoneRequest$toDict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VerifyPhoneRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-VerifyPhoneRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize to bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VerifyPhoneRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VerifyPhoneRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VerifyPhoneRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VerifyPhoneRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
