% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{Writer}
\alias{Writer}
\title{Writer Class}
\description{
Implements a simple writer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Writer-write}{\code{Writer$write()}}
\item \href{#method-Writer-drain}{\code{Writer$drain()}}
\item \href{#method-Writer-close}{\code{Writer$close()}}
\item \href{#method-Writer-clone}{\code{Writer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Writer-write"></a>}}
\if{latex}{\out{\hypertarget{method-Writer-write}{}}}
\subsection{Method \code{write()}}{
Write data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Writer$write(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data to write.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Writer-drain"></a>}}
\if{latex}{\out{\hypertarget{method-Writer-drain}{}}}
\subsection{Method \code{drain()}}{
Drain the write buffer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Writer$drain()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A promise that resolves when drained.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Writer-close"></a>}}
\if{latex}{\out{\hypertarget{method-Writer-close}{}}}
\subsection{Method \code{close()}}{
Close the writer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Writer$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Writer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Writer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Writer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
