% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{AcceptLoginTokenRequest}
\alias{AcceptLoginTokenRequest}
\title{AcceptLoginTokenRequest R6 class}
\description{
AcceptLoginTokenRequest R6 class

AcceptLoginTokenRequest R6 class
}
\details{
Represents the TLRequest auth.AcceptLoginTokenRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0xe894ad4d).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0xc913c01a).}

\item{\code{token}}{raw|character Token bytes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AcceptLoginTokenRequest-new}{\code{AcceptLoginTokenRequest$new()}}
\item \href{#method-AcceptLoginTokenRequest-to_list}{\code{AcceptLoginTokenRequest$to_list()}}
\item \href{#method-AcceptLoginTokenRequest-to_bytes}{\code{AcceptLoginTokenRequest$to_bytes()}}
\item \href{#method-AcceptLoginTokenRequest-from_reader}{\code{AcceptLoginTokenRequest$from_reader()}}
\item \href{#method-AcceptLoginTokenRequest-clone}{\code{AcceptLoginTokenRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptLoginTokenRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptLoginTokenRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an AcceptLoginTokenRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptLoginTokenRequest$new(token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{raw vector or string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptLoginTokenRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptLoginTokenRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptLoginTokenRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptLoginTokenRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptLoginTokenRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptLoginTokenRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptLoginTokenRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptLoginTokenRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create AcceptLoginTokenRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptLoginTokenRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing tgread_bytes()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AcceptLoginTokenRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcceptLoginTokenRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AcceptLoginTokenRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcceptLoginTokenRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
