% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{AccountMethods}
\alias{AccountMethods}
\title{AccountMethods class}
\description{
Provides methods for managing account-related operations, including takeout sessions.
}
\details{
The `AccountMethods` class is used to manage account-related operations in the Telegram API.
It provides methods for creating and managing takeout sessions, as well as handling the success status of those sessions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AccountMethods-takeout}{\code{AccountMethods$takeout()}}
\item \href{#method-AccountMethods-end_takeout}{\code{AccountMethods$end_takeout()}}
\item \href{#method-AccountMethods-clone}{\code{AccountMethods$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccountMethods-takeout"></a>}}
\if{latex}{\out{\hypertarget{method-AccountMethods-takeout}{}}}
\subsection{Method \code{takeout()}}{
Initialize a new `AccountMethods` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccountMethods$takeout(
  finalize = TRUE,
  contacts = NULL,
  users = NULL,
  chats = NULL,
  megagroups = NULL,
  channels = NULL,
  files = NULL,
  max_file_size = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{finalize}}{A logical indicating whether to finalize the takeout session.}

\item{\code{contacts}}{A logical indicating whether to include contacts in the takeout session.}

\item{\code{users}}{A logical indicating whether to include users in the takeout session.}

\item{\code{chats}}{A logical indicating whether to include chats in the takeout session.}

\item{\code{megagroups}}{A logical indicating whether to include megagroups in the takeout session.}

\item{\code{channels}}{A logical indicating whether to include channels in the takeout session.}

\item{\code{files}}{A logical indicating whether to include files in the takeout session.}

\item{\code{max_file_size}}{The maximum file size to include in the takeout session.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `AccountMethods` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccountMethods-end_takeout"></a>}}
\if{latex}{\out{\hypertarget{method-AccountMethods-end_takeout}{}}}
\subsection{Method \code{end_takeout()}}{
Finalize the takeout session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccountMethods$end_takeout(success)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{success}}{A logical indicating whether the takeout session was successful.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical indicating whether the takeout session was finalized successfully.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccountMethods-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AccountMethods-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccountMethods$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
