% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_premium.R
\name{ApplyBoostRequest}
\alias{ApplyBoostRequest}
\title{ApplyBoostRequest}
\description{
Telegram API type ApplyBoostRequest
}
\details{
ApplyBoostRequest R6 class

Representation of the TL request "ApplyBoostRequest".
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor ID for the request}

\item{\code{SUBCLASS_OF_ID}}{Subclass ID for the request}

\item{\code{peer}}{Input peer (object or identifier)}

\item{\code{slots}}{Integer vector of slots or NULL}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApplyBoostRequest-new}{\code{ApplyBoostRequest$new()}}
\item \href{#method-ApplyBoostRequest-resolve}{\code{ApplyBoostRequest$resolve()}}
\item \href{#method-ApplyBoostRequest-to_list}{\code{ApplyBoostRequest$to_list()}}
\item \href{#method-ApplyBoostRequest-bytes}{\code{ApplyBoostRequest$bytes()}}
\item \href{#method-ApplyBoostRequest-from_reader}{\code{ApplyBoostRequest$from_reader()}}
\item \href{#method-ApplyBoostRequest-clone}{\code{ApplyBoostRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplyBoostRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApplyBoostRequest-new}{}}}
\subsection{Method \code{new()}}{
Create a new ApplyBoostRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplyBoostRequest$new(peer = NULL, slots = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{input peer or identifier}

\item{\code{slots}}{integer vector of slots or NULL}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplyBoostRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-ApplyBoostRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolve peer using client/utils helpers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplyBoostRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{client object that provides get_input_entity}

\item{\code{utils}}{utils object that provides get_input_peer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplyBoostRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ApplyBoostRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert request to list (dictionary)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplyBoostRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplyBoostRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ApplyBoostRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize to raw bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplyBoostRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplyBoostRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ApplyBoostRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Construct a new ApplyBoostRequest from a reader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplyBoostRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{reader object exposing read_int / tgread_object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ApplyBoostRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplyBoostRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApplyBoostRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplyBoostRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
