% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{BindTempAuthKeyRequest}
\alias{BindTempAuthKeyRequest}
\title{BindTempAuthKeyRequest}
\description{
Telegram API type BindTempAuthKeyRequest
}
\details{
BindTempAuthKeyRequest R6 class

Represents the TLRequest auth.BindTempAuthKeyRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{perm_auth_key_id}}{Field.}

\item{\code{nonce}}{Field.}

\item{\code{expires_at}}{Field.}

\item{\code{encrypted_message}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BindTempAuthKeyRequest-new}{\code{BindTempAuthKeyRequest$new()}}
\item \href{#method-BindTempAuthKeyRequest-to_list}{\code{BindTempAuthKeyRequest$to_list()}}
\item \href{#method-BindTempAuthKeyRequest-to_bytes}{\code{BindTempAuthKeyRequest$to_bytes()}}
\item \href{#method-BindTempAuthKeyRequest-from_reader}{\code{BindTempAuthKeyRequest$from_reader()}}
\item \href{#method-BindTempAuthKeyRequest-clone}{\code{BindTempAuthKeyRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BindTempAuthKeyRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-BindTempAuthKeyRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a BindTempAuthKeyRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BindTempAuthKeyRequest$new(
  perm_auth_key_id,
  nonce,
  expires_at = NULL,
  encrypted_message
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{perm_auth_key_id}}{numeric or integer (64-bit)}

\item{\code{nonce}}{numeric or integer (64-bit)}

\item{\code{expires_at}}{POSIXct|Date or NULL}

\item{\code{encrypted_message}}{raw vector or string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BindTempAuthKeyRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-BindTempAuthKeyRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BindTempAuthKeyRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BindTempAuthKeyRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-BindTempAuthKeyRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BindTempAuthKeyRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BindTempAuthKeyRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-BindTempAuthKeyRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create BindTempAuthKeyRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BindTempAuthKeyRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_long(), tgread_date(), tgread_bytes()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
BindTempAuthKeyRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BindTempAuthKeyRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BindTempAuthKeyRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BindTempAuthKeyRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
