% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BusinessBotRights}
\alias{BusinessBotRights}
\title{BusinessBotRights}
\description{
Telegram API type BusinessBotRights
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{BusinessBotRights}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{reply}}{Field.}

\item{\code{read_messages}}{Field.}

\item{\code{delete_sent_messages}}{Field.}

\item{\code{delete_received_messages}}{Field.}

\item{\code{edit_name}}{Field.}

\item{\code{edit_bio}}{Field.}

\item{\code{edit_profile_photo}}{Field.}

\item{\code{edit_username}}{Field.}

\item{\code{view_gifts}}{Field.}

\item{\code{sell_gifts}}{Field.}

\item{\code{change_gift_settings}}{Field.}

\item{\code{transfer_and_upgrade_gifts}}{Field.}

\item{\code{transfer_stars}}{Field.}

\item{\code{manage_stories}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BusinessBotRights-new}{\code{BusinessBotRights$new()}}
\item \href{#method-BusinessBotRights-to_dict}{\code{BusinessBotRights$to_dict()}}
\item \href{#method-BusinessBotRights-clone}{\code{BusinessBotRights$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessBotRights-new"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessBotRights-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessBotRights$new(
  reply = NULL,
  read_messages = NULL,
  delete_sent_messages = NULL,
  delete_received_messages = NULL,
  edit_name = NULL,
  edit_bio = NULL,
  edit_profile_photo = NULL,
  edit_username = NULL,
  view_gifts = NULL,
  sell_gifts = NULL,
  change_gift_settings = NULL,
  transfer_and_upgrade_gifts = NULL,
  transfer_stars = NULL,
  manage_stories = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessBotRights-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessBotRights-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessBotRights$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessBotRights-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessBotRights-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessBotRights$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
