% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Channel}
\alias{Channel}
\title{Channel}
\description{
Telegram API type Channel
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{Channel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Channel-new}{\code{Channel$new()}}
\item \href{#method-Channel-to_dict}{\code{Channel$to_dict()}}
\item \href{#method-Channel-bytes}{\code{Channel$bytes()}}
\item \href{#method-Channel-clone}{\code{Channel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Channel-new"></a>}}
\if{latex}{\out{\hypertarget{method-Channel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Channel$new(
  id,
  title,
  photo,
  date,
  creator = NULL,
  left = NULL,
  broadcast = NULL,
  verified = NULL,
  megagroup = NULL,
  restricted = NULL,
  signatures = NULL,
  min = NULL,
  scam = NULL,
  has_link = NULL,
  has_geo = NULL,
  slowmode_enabled = NULL,
  call_active = NULL,
  call_not_empty = NULL,
  fake = NULL,
  gigagroup = NULL,
  noforwards = NULL,
  join_to_send = NULL,
  join_request = NULL,
  forum = NULL,
  stories_hidden = NULL,
  stories_hidden_min = NULL,
  stories_unavailable = NULL,
  signature_profiles = NULL,
  autotranslation = NULL,
  broadcast_messages_allowed = NULL,
  monoforum = NULL,
  forum_tabs = NULL,
  access_hash = NULL,
  username = NULL,
  restriction_reason = NULL,
  admin_rights = NULL,
  banned_rights = NULL,
  default_banned_rights = NULL,
  participants_count = NULL,
  usernames = NULL,
  stories_max_id = NULL,
  color = NULL,
  profile_color = NULL,
  emoji_status = NULL,
  level = NULL,
  subscription_until_date = NULL,
  bot_verification_icon = NULL,
  send_paid_messages_stars = NULL,
  linked_monoforum_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Channel-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-Channel-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Channel$to_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Channel-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-Channel-bytes}{}}}
\subsection{Method \code{bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Channel$bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Channel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Channel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Channel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
