% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stickers.R
\name{CheckShortNameRequest}
\alias{CheckShortNameRequest}
\title{CheckShortNameRequest}
\description{
Telegram API type CheckShortNameRequest
}
\details{
CheckShortNameRequest R6 class

Represents a request to check whether a short name is available/valid.
}
\section{Methods}{

- initialize(short_name): create instance.
- to_list(): returns a list representation suitable for JSON/inspection.
- to_bytes(): returns a raw vector representing serialized bytes (TL-like).
- from_reader(reader): reads a request from a reader object and returns a new instance.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{short_name}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CheckShortNameRequest-new}{\code{CheckShortNameRequest$new()}}
\item \href{#method-CheckShortNameRequest-to_list}{\code{CheckShortNameRequest$to_list()}}
\item \href{#method-CheckShortNameRequest-to_bytes}{\code{CheckShortNameRequest$to_bytes()}}
\item \href{#method-CheckShortNameRequest-from_reader}{\code{CheckShortNameRequest$from_reader()}}
\item \href{#method-CheckShortNameRequest-clone}{\code{CheckShortNameRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckShortNameRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-CheckShortNameRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a CheckShortNameRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckShortNameRequest$new(short_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{short_name}}{character short name string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckShortNameRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CheckShortNameRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckShortNameRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list representation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckShortNameRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-CheckShortNameRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to raw bytes (TL-like approximation)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckShortNameRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckShortNameRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-CheckShortNameRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckShortNameRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckShortNameRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CheckShortNameRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckShortNameRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
