% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_check}
\alias{compute_check}
\title{Compute Check}
\usage{
compute_check(request, password)
}
\arguments{
\item{request}{The password request object.}

\item{password}{A string.}
}
\value{
An InputCheckPasswordSRP object.
}
\description{
Computes the check for SRP password.
}
