% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{compute_check}
\alias{compute_check}
\title{Compute check}
\usage{
compute_check(request, password)
}
\arguments{
\item{request}{An object of type ReqPqMulti or ReqPq.}

\item{password}{A string representing the password.}
}
\value{
An object of type InputCheckPasswordSRP.
}
\description{
Compute check
}
