% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Config}
\alias{Config}
\title{Config}
\description{
Telegram API type Config
}
\section{Super class}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{Config}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{date}}{Field.}

\item{\code{expires}}{Field.}

\item{\code{test_mode}}{Field.}

\item{\code{this_dc}}{Field.}

\item{\code{dc_options}}{Field.}

\item{\code{dc_txt_domain_name}}{Field.}

\item{\code{chat_size_max}}{Field.}

\item{\code{megagroup_size_max}}{Field.}

\item{\code{forwarded_count_max}}{Field.}

\item{\code{online_update_period_ms}}{Field.}

\item{\code{offline_blur_timeout_ms}}{Field.}

\item{\code{offline_idle_timeout_ms}}{Field.}

\item{\code{online_cloud_timeout_ms}}{Field.}

\item{\code{notify_cloud_delay_ms}}{Field.}

\item{\code{notify_default_delay_ms}}{Field.}

\item{\code{push_chat_period_ms}}{Field.}

\item{\code{push_chat_limit}}{Field.}

\item{\code{edit_time_limit}}{Field.}

\item{\code{revoke_time_limit}}{Field.}

\item{\code{revoke_pm_time_limit}}{Field.}

\item{\code{rating_e_decay}}{Field.}

\item{\code{stickers_recent_limit}}{Field.}

\item{\code{channels_read_media_period}}{Field.}

\item{\code{call_receive_timeout_ms}}{Field.}

\item{\code{call_ring_timeout_ms}}{Field.}

\item{\code{call_connect_timeout_ms}}{Field.}

\item{\code{call_packet_timeout_ms}}{Field.}

\item{\code{me_url_prefix}}{Field.}

\item{\code{caption_length_max}}{Field.}

\item{\code{message_length_max}}{Field.}

\item{\code{webfile_dc_id}}{Field.}

\item{\code{default_p2p_contacts}}{Field.}

\item{\code{preload_featured_stickers}}{Field.}

\item{\code{revoke_pm_inbox}}{Field.}

\item{\code{blocked_mode}}{Field.}

\item{\code{force_try_ipv6}}{Field.}

\item{\code{tmp_sessions}}{Field.}

\item{\code{autoupdate_url_prefix}}{Field.}

\item{\code{gif_search_username}}{Field.}

\item{\code{venue_search_username}}{Field.}

\item{\code{img_search_username}}{Field.}

\item{\code{static_maps_provider}}{Field.}

\item{\code{suggested_lang_code}}{Field.}

\item{\code{lang_pack_version}}{Field.}

\item{\code{base_lang_pack_version}}{Field.}

\item{\code{reactions_default}}{Field.}

\item{\code{autologin_token}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Config-new}{\code{Config$new()}}
\item \href{#method-Config-to_list}{\code{Config$to_list()}}
\item \href{#method-Config-clone}{\code{Config$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-.bytes'><code>telegramR::TLObject$.bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="from_reader"><a href='../../telegramR/html/TLObject.html#method-TLObject-from_reader'><code>telegramR::TLObject$from_reader()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_dict"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_dict'><code>telegramR::TLObject$to_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Config-new"></a>}}
\if{latex}{\out{\hypertarget{method-Config-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$new(
  date,
  expires,
  test_mode,
  this_dc,
  dc_options,
  dc_txt_domain_name,
  chat_size_max,
  megagroup_size_max,
  forwarded_count_max,
  online_update_period_ms,
  offline_blur_timeout_ms,
  offline_idle_timeout_ms,
  online_cloud_timeout_ms,
  notify_cloud_delay_ms,
  notify_default_delay_ms,
  push_chat_period_ms,
  push_chat_limit,
  edit_time_limit,
  revoke_time_limit,
  revoke_pm_time_limit,
  rating_e_decay,
  stickers_recent_limit,
  channels_read_media_period,
  call_receive_timeout_ms,
  call_ring_timeout_ms,
  call_connect_timeout_ms,
  call_packet_timeout_ms,
  me_url_prefix,
  caption_length_max,
  message_length_max,
  webfile_dc_id,
  default_p2p_contacts = NULL,
  preload_featured_stickers = NULL,
  revoke_pm_inbox = NULL,
  blocked_mode = NULL,
  force_try_ipv6 = NULL,
  tmp_sessions = NULL,
  autoupdate_url_prefix = NULL,
  gif_search_username = NULL,
  venue_search_username = NULL,
  img_search_username = NULL,
  static_maps_provider = NULL,
  suggested_lang_code = NULL,
  lang_pack_version = NULL,
  base_lang_pack_version = NULL,
  reactions_default = NULL,
  autologin_token = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Config-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Config-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Config-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Config-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
