% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phone.R
\name{ConfirmCallRequest}
\alias{ConfirmCallRequest}
\title{ConfirmCallRequest}
\description{
Telegram API type ConfirmCallRequest
}
\details{
ConfirmCallRequest R6 class

Represents ConfirmCallRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{peer}}{Field.}

\item{\code{g_a}}{Field.}

\item{\code{key_fingerprint}}{Field.}

\item{\code{protocol}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConfirmCallRequest-new}{\code{ConfirmCallRequest$new()}}
\item \href{#method-ConfirmCallRequest-to_list}{\code{ConfirmCallRequest$to_list()}}
\item \href{#method-ConfirmCallRequest-bytes}{\code{ConfirmCallRequest$bytes()}}
\item \href{#method-ConfirmCallRequest-from_reader}{\code{ConfirmCallRequest$from_reader()}}
\item \href{#method-ConfirmCallRequest-clone}{\code{ConfirmCallRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfirmCallRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConfirmCallRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ConfirmCallRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfirmCallRequest$new(peer, g_a, key_fingerprint, protocol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peer}}{Input phone call object.}

\item{\code{g_a}}{Raw bytes.}

\item{\code{key_fingerprint}}{Integer key fingerprint.}

\item{\code{protocol}}{Protocol object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfirmCallRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ConfirmCallRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfirmCallRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list representing the request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfirmCallRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ConfirmCallRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Produce raw bytes for the request

Builds the TL-serialized byte representation. Assumes dependent objects provide bytes().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfirmCallRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector (bytes). If dependent objects don't provide bytes(), a warning is emitted and raw(0) is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfirmCallRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ConfirmCallRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfirmCallRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfirmCallRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConfirmCallRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfirmCallRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
