% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DestroyAuthKeyRequest}
\alias{DestroyAuthKeyRequest}
\title{DestroyAuthKeyRequest R6 class}
\description{
Represents the TL request `DestroyAuthKeyRequest`.
}
\details{
Fields:
- (no fields)

Methods:
- new(): create new instance.
- to_list(): return an R list representation.
- to_raw(): serialize to raw vector (bytes) in little endian as used in TL.

Writes constructor id 0xd1435160 as little-endian bytes: 0x60 0x51 0x43 0xD1
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DestroyAuthKeyRequest-to_list}{\code{DestroyAuthKeyRequest$to_list()}}
\item \href{#method-DestroyAuthKeyRequest-to_bytes}{\code{DestroyAuthKeyRequest$to_bytes()}}
\item \href{#method-DestroyAuthKeyRequest-clone}{\code{DestroyAuthKeyRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DestroyAuthKeyRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DestroyAuthKeyRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DestroyAuthKeyRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DestroyAuthKeyRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-DestroyAuthKeyRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DestroyAuthKeyRequest$to_bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DestroyAuthKeyRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DestroyAuthKeyRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DestroyAuthKeyRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
