% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{DropTempAuthKeysRequest}
\alias{DropTempAuthKeysRequest}
\title{DropTempAuthKeysRequest}
\description{
Telegram API type DropTempAuthKeysRequest
}
\details{
DropTempAuthKeysRequest R6 class

Represents the TLRequest auth.DropTempAuthKeysRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0x8e48a188).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0xf5b399ac).}

\item{\code{except_auth_keys}}{numeric vector 64-bit ids to drop (except list).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DropTempAuthKeysRequest-new}{\code{DropTempAuthKeysRequest$new()}}
\item \href{#method-DropTempAuthKeysRequest-to_list}{\code{DropTempAuthKeysRequest$to_list()}}
\item \href{#method-DropTempAuthKeysRequest-to_bytes}{\code{DropTempAuthKeysRequest$to_bytes()}}
\item \href{#method-DropTempAuthKeysRequest-from_reader}{\code{DropTempAuthKeysRequest$from_reader()}}
\item \href{#method-DropTempAuthKeysRequest-clone}{\code{DropTempAuthKeysRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DropTempAuthKeysRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DropTempAuthKeysRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a DropTempAuthKeysRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DropTempAuthKeysRequest$new(except_auth_keys = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{except_auth_keys}}{numeric vector (64-bit) or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DropTempAuthKeysRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DropTempAuthKeysRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DropTempAuthKeysRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DropTempAuthKeysRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-DropTempAuthKeysRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DropTempAuthKeysRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DropTempAuthKeysRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-DropTempAuthKeysRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create DropTempAuthKeysRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DropTempAuthKeysRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), read_long()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DropTempAuthKeysRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DropTempAuthKeysRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DropTempAuthKeysRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DropTempAuthKeysRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
