% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/uploads.R
\name{FileStream}
\alias{FileStream}
\title{FileStream Class
A class to handle file streams from various sources.}
\description{
This class can handle file paths, raw vectors, and connections. It provides
methods to read data, close the stream, and get file size and name.

A helper class to handle file streams for uploading files.
}
\examples{
# fs <- FileStream$new("path/to/file.txt")
# data <- fs$read(100)
# fs$close()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file}}{The original file input (path, raw vector, or connection).}

\item{\code{file_size}}{The size of the file in bytes.}

\item{\code{name}}{The name of the file (if applicable).}

\item{\code{stream}}{The underlying connection or raw connection.}

\item{\code{close_stream}}{Logical indicating if the stream should be closed on cleanup.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{stream}}{The underlying connection or raw connection.}

\item{\code{close_stream}}{Logical indicating if the stream should be closed on cleanup.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileStream-new}{\code{FileStream$new()}}
\item \href{#method-FileStream-read}{\code{FileStream$read()}}
\item \href{#method-FileStream-clone}{\code{FileStream$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$new(file, file_size = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-read"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$read(size)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file}}{The file to be streamed (can be a path, raw bytes, etc.).}

\item{\code{file_size}}{The size of the file in bytes.}

\item{\code{name}}{The name of the file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileStream-new}{\code{FileStream$new()}}
\item \href{#method-FileStream-read}{\code{FileStream$read()}}
\item \href{#method-FileStream-clone}{\code{FileStream$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-new}{}}}
\subsection{Method \code{new()}}{
Initializes the FileStream object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$new(file, file_size = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{The file to be streamed (path, raw bytes, etc.).}

\item{\code{file_size}}{Optional size of the file in bytes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-read"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-read}{}}}
\subsection{Method \code{read()}}{
Reads a chunk of the file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$read(size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{The size of the chunk to read in bytes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A raw vector containing the file chunk.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileStream-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileStream-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileStream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
