% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libssl.R
\name{find_ssl_lib}
\alias{find_ssl_lib}
\title{Finds and loads the system's SSL library.}
\usage{
find_ssl_lib()
}
\value{
A reference (path) to the loaded SSL library.
}
\description{
This function attempts to locate the system's SSL library and load it using R's
`dyn.load` function. On macOS, it handles versioning issues for compatibility.
}
