% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_smsjobs.R
\name{FinishJobRequest}
\alias{FinishJobRequest}
\title{FinishJobRequest}
\description{
Telegram API type FinishJobRequest
}
\details{
FinishJobRequest R6 class

Representation of the FinishJobRequest TLRequest.

Encodes constructor id, flags (bit 0 for error), job_id and optional error.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{job_id}}{character}

\item{\code{error}}{character or NULL}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FinishJobRequest-new}{\code{FinishJobRequest$new()}}
\item \href{#method-FinishJobRequest-to_list}{\code{FinishJobRequest$to_list()}}
\item \href{#method-FinishJobRequest-bytes}{\code{FinishJobRequest$bytes()}}
\item \href{#method-FinishJobRequest-from_reader}{\code{FinishJobRequest$from_reader()}}
\item \href{#method-FinishJobRequest-clone}{\code{FinishJobRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FinishJobRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-FinishJobRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a FinishJobRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FinishJobRequest$new(job_id, error = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job_id}}{character}

\item{\code{error}}{character or NULL}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FinishJobRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-FinishJobRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FinishJobRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FinishJobRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-FinishJobRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Get raw bytes for this constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FinishJobRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FinishJobRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-FinishJobRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FinishJobRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FinishJobRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FinishJobRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FinishJobRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
