% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{GetAdminedBotsRequest}
\alias{GetAdminedBotsRequest}
\title{GetAdminedBotsRequest R6 class}
\description{
Methods:
- initialize()
- to_list() -> list representation
- to_bytes() -> raw vector bytes
}
\details{
Represents the GetAdminedBotsRequest TL request.

No fields.
}
\examples{
# o <- GetAdminedBotsRequest$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetAdminedBotsRequest-new}{\code{GetAdminedBotsRequest$new()}}
\item \href{#method-GetAdminedBotsRequest-to_list}{\code{GetAdminedBotsRequest$to_list()}}
\item \href{#method-GetAdminedBotsRequest-to_bytes}{\code{GetAdminedBotsRequest$to_bytes()}}
\item \href{#method-GetAdminedBotsRequest-clone}{\code{GetAdminedBotsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAdminedBotsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetAdminedBotsRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAdminedBotsRequest$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAdminedBotsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetAdminedBotsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAdminedBotsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAdminedBotsRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetAdminedBotsRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAdminedBotsRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAdminedBotsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetAdminedBotsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAdminedBotsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
