% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetAppConfigRequest}
\alias{GetAppConfigRequest}
\title{GetAppConfigRequest R6 class}
\description{
GetAppConfigRequest R6 class

GetAppConfigRequest R6 class
}
\details{
Request to get app configuration (possibly not modified).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex).}

\item{\code{hash}}{integer Hash to check for modifications.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetAppConfigRequest-new}{\code{GetAppConfigRequest$new()}}
\item \href{#method-GetAppConfigRequest-to_list}{\code{GetAppConfigRequest$to_list()}}
\item \href{#method-GetAppConfigRequest-to_bytes}{\code{GetAppConfigRequest$to_bytes()}}
\item \href{#method-GetAppConfigRequest-clone}{\code{GetAppConfigRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAppConfigRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetAppConfigRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetAppConfigRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAppConfigRequest$new(hash = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{integer
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAppConfigRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetAppConfigRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAppConfigRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with `_` discriminator and hash
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAppConfigRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetAppConfigRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAppConfigRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAppConfigRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetAppConfigRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetAppConfigRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
