% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{GetFutureSaltsRequest}
\alias{GetFutureSaltsRequest}
\title{GetFutureSaltsRequest}
\description{
R6 class representing a request to get future salts.
Contains methods for serialization to list and bytes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor ID for the request}

\item{\code{SUBCLASS_OF_ID}}{Subclass ID for the request}

\item{\code{num}}{Number of future salts to request
Initialize the GetFutureSaltsRequest}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetFutureSaltsRequest-new}{\code{GetFutureSaltsRequest$new()}}
\item \href{#method-GetFutureSaltsRequest-to_list}{\code{GetFutureSaltsRequest$to_list()}}
\item \href{#method-GetFutureSaltsRequest-to_bytes}{\code{GetFutureSaltsRequest$to_bytes()}}
\item \href{#method-GetFutureSaltsRequest-clone}{\code{GetFutureSaltsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFutureSaltsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetFutureSaltsRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFutureSaltsRequest$new(num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num}}{The number of future salts to request
Convert to list representation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFutureSaltsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetFutureSaltsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFutureSaltsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the GetFutureSaltsRequest
Serialize to raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFutureSaltsRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetFutureSaltsRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFutureSaltsRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the serialized GetFutureSaltsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFutureSaltsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetFutureSaltsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetFutureSaltsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
