% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_premium.R
\name{GetMyBoostsRequest}
\alias{GetMyBoostsRequest}
\title{GetMyBoostsRequest}
\description{
Telegram API type GetMyBoostsRequest
}
\details{
GetMyBoostsRequest R6 class

Representation of the TL request "GetMyBoostsRequest".
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor ID for the request}

\item{\code{SUBCLASS_OF_ID}}{Subclass ID for the request}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetMyBoostsRequest-new}{\code{GetMyBoostsRequest$new()}}
\item \href{#method-GetMyBoostsRequest-to_list}{\code{GetMyBoostsRequest$to_list()}}
\item \href{#method-GetMyBoostsRequest-bytes}{\code{GetMyBoostsRequest$bytes()}}
\item \href{#method-GetMyBoostsRequest-from_reader}{\code{GetMyBoostsRequest$from_reader()}}
\item \href{#method-GetMyBoostsRequest-clone}{\code{GetMyBoostsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetMyBoostsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetMyBoostsRequest-new}{}}}
\subsection{Method \code{new()}}{
Create a new GetMyBoostsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetMyBoostsRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetMyBoostsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetMyBoostsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert request to a list (dictionary)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetMyBoostsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetMyBoostsRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetMyBoostsRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Serialize to raw bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetMyBoostsRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
Read object from reader and return new instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetMyBoostsRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetMyBoostsRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetMyBoostsRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{an object exposing tgread_object / read_int / read_string as in the surrounding codebase}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GetMyBoostsRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetMyBoostsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetMyBoostsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetMyBoostsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
