% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_users.R
\name{GetSavedMusicByIDRequest}
\alias{GetSavedMusicByIDRequest}
\title{GetSavedMusicByIDRequest}
\description{
GetSavedMusicByIDRequest

GetSavedMusicByIDRequest
}
\details{
Request to get saved music by document IDs for a user.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{GetSavedMusicByIDRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor ID for the GetSavedMusicByIDRequest class.}

\item{\code{SUBCLASS_OF_ID}}{Subclass ID for the GetSavedMusicByIDRequest class.}

\item{\code{id}}{TLInputUser or a representation accepted by get_input_user}

\item{\code{documents}}{List of TLInputDocument or representations accepted by get_input_document}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetSavedMusicByIDRequest-new}{\code{GetSavedMusicByIDRequest$new()}}
\item \href{#method-GetSavedMusicByIDRequest-resolve}{\code{GetSavedMusicByIDRequest$resolve()}}
\item \href{#method-GetSavedMusicByIDRequest-to_dict}{\code{GetSavedMusicByIDRequest$to_dict()}}
\item \href{#method-GetSavedMusicByIDRequest-.bytes}{\code{GetSavedMusicByIDRequest$.bytes()}}
\item \href{#method-GetSavedMusicByIDRequest-from_reader}{\code{GetSavedMusicByIDRequest$from_reader()}}
\item \href{#method-GetSavedMusicByIDRequest-clone}{\code{GetSavedMusicByIDRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicByIDRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicByIDRequest-new}{}}}
\subsection{Method \code{new()}}{
Request to get saved music by document IDs for a user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicByIDRequest$new(id = NULL, documents = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{TLInputUser or a representation accepted by get_input_user}

\item{\code{documents}}{List of TLInputDocument or representations accepted by get_input_document}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicByIDRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicByIDRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Converts the GetSavedMusicByIDRequest to a dictionary (list).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicByIDRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{A client object to resolve user references.}

\item{\code{utils}}{A utils object with methods to get input user and document representations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicByIDRequest-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicByIDRequest-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the GetSavedMusicByIDRequest to raw bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicByIDRequest$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the GetSavedMusicByIDRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicByIDRequest-.bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicByIDRequest-.bytes}{}}}
\subsection{Method \code{.bytes()}}{
Initializes the GetSavedMusicByIDRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicByIDRequest$.bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw vector representing the GetSavedMusicByIDRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicByIDRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicByIDRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Request to get saved music by document IDs for a user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicByIDRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{A reader object to read the response.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of GetSavedMusicByIDRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSavedMusicByIDRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetSavedMusicByIDRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSavedMusicByIDRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
