% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{GetSupportRequest}
\alias{GetSupportRequest}
\title{GetSupportRequest}
\description{
Telegram API type GetSupportRequest
}
\details{
GetSupportRequest R6 class

Request to get support information.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{Field.
Serialize to bytes (raw vector)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{Field.
Serialize to bytes (raw vector)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetSupportRequest-new}{\code{GetSupportRequest$new()}}
\item \href{#method-GetSupportRequest-to_list}{\code{GetSupportRequest$to_list()}}
\item \href{#method-GetSupportRequest-to_bytes}{\code{GetSupportRequest$to_bytes()}}
\item \href{#method-GetSupportRequest-clone}{\code{GetSupportRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSupportRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetSupportRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize GetSupportRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSupportRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible NULL
Convert to list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSupportRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GetSupportRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSupportRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with type discriminator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSupportRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetSupportRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSupportRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSupportRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetSupportRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSupportRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
