% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_users.R
\name{GetUsersRequest}
\alias{GetUsersRequest}
\title{GetUsersRequest}
\description{
Request to get a vector of users by input user references.
}
\section{Super classes}{
\code{\link[telegramR:TLObject]{telegramR::TLObject}} -> \code{\link[telegramR:TLRequest]{telegramR::TLRequest}} -> \code{GetUsersRequest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor ID for the GetUsersRequest class.}

\item{\code{SUBCLASS_OF_ID}}{Subclass ID for the GetUsersRequest class.}

\item{\code{id}}{TLInputUser or a representation accepted by get_input_user}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetUsersRequest-new}{\code{GetUsersRequest$new()}}
\item \href{#method-GetUsersRequest-resolve}{\code{GetUsersRequest$resolve()}}
\item \href{#method-GetUsersRequest-to_dict}{\code{GetUsersRequest$to_dict()}}
\item \href{#method-GetUsersRequest-.bytes}{\code{GetUsersRequest$.bytes()}}
\item \href{#method-GetUsersRequest-from_reader}{\code{GetUsersRequest$from_reader()}}
\item \href{#method-GetUsersRequest-clone}{\code{GetUsersRequest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".eq"><a href='../../telegramR/html/TLObject.html#method-TLObject-.eq'><code>telegramR::TLObject$.eq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".ne"><a href='../../telegramR/html/TLObject.html#method-TLObject-.ne'><code>telegramR::TLObject$.ne()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id=".str"><a href='../../telegramR/html/TLObject.html#method-TLObject-.str'><code>telegramR::TLObject$.str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="pretty_format"><a href='../../telegramR/html/TLObject.html#method-TLObject-pretty_format'><code>telegramR::TLObject$pretty_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_bytes"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_bytes'><code>telegramR::TLObject$serialize_bytes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="serialize_datetime"><a href='../../telegramR/html/TLObject.html#method-TLObject-serialize_datetime'><code>telegramR::TLObject$serialize_datetime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="stringify"><a href='../../telegramR/html/TLObject.html#method-TLObject-stringify'><code>telegramR::TLObject$stringify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLObject" data-id="to_json"><a href='../../telegramR/html/TLObject.html#method-TLObject-to_json'><code>telegramR::TLObject$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TLRequest" data-id="read_result"><a href='../../telegramR/html/TLRequest.html#method-TLRequest-read_result'><code>telegramR::TLRequest$read_result()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUsersRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetUsersRequest-new}{}}}
\subsection{Method \code{new()}}{
Initializes the GetUsersRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUsersRequest$new(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{List of TLInputUser or representations accepted by get_input_user}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUsersRequest-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-GetUsersRequest-resolve}{}}}
\subsection{Method \code{resolve()}}{
Resolves the id field using the provided client and utils.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUsersRequest$resolve(client, utils)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{A client object to resolve user references.}

\item{\code{utils}}{A utils object with methods to get input user representations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUsersRequest-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-GetUsersRequest-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Request to get a vector of users by input user references.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUsersRequest$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of resolved id objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUsersRequest-.bytes"></a>}}
\if{latex}{\out{\hypertarget{method-GetUsersRequest-.bytes}{}}}
\subsection{Method \code{.bytes()}}{
Initializes the GetUsersRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUsersRequest$.bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of resolved id objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUsersRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-GetUsersRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Converts the GetUsersRequest to raw bytes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUsersRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{A reader object to read the response.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A raw vector representing the GetUsersRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUsersRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetUsersRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetUsersRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
