% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegrambaseclient.R
\name{\%||\%}
\alias{\%||\%}
\title{Helper function for NULL handling (similar to %||% in R)}
\usage{
x \%||\% y
}
\arguments{
\item{x}{The first value}

\item{y}{The second value}
}
\value{
The first value if not NULL, otherwise the second value
}
\description{
Helper function for NULL handling (similar to %||% in R)
}
