% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{ImportAuthorizationRequest}
\alias{ImportAuthorizationRequest}
\title{ImportAuthorizationRequest}
\description{
Telegram API type ImportAuthorizationRequest
}
\details{
ImportAuthorizationRequest R6 class

Represents the TLRequest auth.ImportAuthorizationRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0xa57a7dad).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0xb9e04e39).}

\item{\code{id}}{numeric 64-bit id.}

\item{\code{bytes}}{raw|character Token bytes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImportAuthorizationRequest-new}{\code{ImportAuthorizationRequest$new()}}
\item \href{#method-ImportAuthorizationRequest-to_list}{\code{ImportAuthorizationRequest$to_list()}}
\item \href{#method-ImportAuthorizationRequest-to_bytes}{\code{ImportAuthorizationRequest$to_bytes()}}
\item \href{#method-ImportAuthorizationRequest-from_reader}{\code{ImportAuthorizationRequest$from_reader()}}
\item \href{#method-ImportAuthorizationRequest-clone}{\code{ImportAuthorizationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportAuthorizationRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImportAuthorizationRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ImportAuthorizationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportAuthorizationRequest$new(id, bytes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{numeric or integer (64-bit)}

\item{\code{bytes}}{raw vector or string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportAuthorizationRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ImportAuthorizationRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportAuthorizationRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportAuthorizationRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ImportAuthorizationRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportAuthorizationRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportAuthorizationRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ImportAuthorizationRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create ImportAuthorizationRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportAuthorizationRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_long(), tgread_bytes()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ImportAuthorizationRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportAuthorizationRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImportAuthorizationRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportAuthorizationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
