% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_good_mod_exp_first}
\alias{is_good_mod_exp_first}
\title{Is Good Mod Exp First}
\usage{
is_good_mod_exp_first(modexp, prime)
}
\arguments{
\item{modexp}{A bigz integer.}

\item{prime}{A bigz integer.}
}
\value{
A logical value.

A logical value.
}
\description{
Checks if a modular exponentiation result is good for the first check.
}
\details{
Is Good Mod Exp First

Checks if modexp is good for modular exponentiation.
}
