% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_channel.R
\name{join_private_chat}
\alias{join_private_chat}
\title{Join a private chat or channel by invite link}
\usage{
join_private_chat(client, invite_link)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{invite_link}{Invite link (e.g. https://t.me/joinchat/... or https://t.me/+...).}
}
\value{
The API response.
}
\description{
Join a private chat or channel by invite link
}
