% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotostate.R
\docType{data}
\name{MAX_RECENT_MSG_IDS}
\alias{MAX_RECENT_MSG_IDS}
\title{MTProtoState Class}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_RECENT_MSG_IDS
}
\description{
This class manages the state needed by MTProtoSender for encrypting and
decrypting incoming/outgoing messages and generating message IDs.
}
\details{
MTProtoState handles authentication key management, message sequencing,
and encryption/decryption required by Telegram's MTProto protocol.
It is not concerned with storing information to disk, as users may create
multiple senders for different data centers or CDNs.
}
\keyword{datasets}
