% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{mimetypes}
\alias{mimetypes}
\title{Extend the mimetypes character vector with additional common types.}
\format{
An object of class \code{character} of length 1565.
}
\usage{
mimetypes
}
\description{
Extend the mimetypes character vector with additional common types.
}
\keyword{datasets}
