% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtprotostate.R
\name{packRandomLong}
\alias{packRandomLong}
\title{Pack a random 64-bit integer and return as raw vector}
\usage{
packRandomLong()
}
\value{
Raw vector representation of a random long integer
}
\description{
Pack a random 64-bit integer and return as raw vector
}
