% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{+.raw_bytes}
\alias{+.raw_bytes}
\title{Addition operator for raw_bytes}
\usage{
\method{+}{raw_bytes}(e1, e2)
}
\arguments{
\item{e1}{Left operand.}

\item{e2}{Right operand.}
}
\value{
A new raw_bytes object.
}
\description{
Allows using '+' to concatenate raw_bytes objects, mimicking Python's byte concatenation.
}
