% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{RecoverPasswordRequest}
\alias{RecoverPasswordRequest}
\title{RecoverPasswordRequest}
\description{
Telegram API type RecoverPasswordRequest
}
\details{
RecoverPasswordRequest R6 class

Represents the TLRequest auth.RecoverPasswordRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{code}}{Field.}

\item{\code{new_settings}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RecoverPasswordRequest-new}{\code{RecoverPasswordRequest$new()}}
\item \href{#method-RecoverPasswordRequest-to_list}{\code{RecoverPasswordRequest$to_list()}}
\item \href{#method-RecoverPasswordRequest-to_bytes}{\code{RecoverPasswordRequest$to_bytes()}}
\item \href{#method-RecoverPasswordRequest-from_reader}{\code{RecoverPasswordRequest$from_reader()}}
\item \href{#method-RecoverPasswordRequest-clone}{\code{RecoverPasswordRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecoverPasswordRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-RecoverPasswordRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a RecoverPasswordRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecoverPasswordRequest$new(code, new_settings = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{character}

\item{\code{new_settings}}{TL object or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecoverPasswordRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-RecoverPasswordRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecoverPasswordRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecoverPasswordRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-RecoverPasswordRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecoverPasswordRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecoverPasswordRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-RecoverPasswordRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create RecoverPasswordRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecoverPasswordRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RecoverPasswordRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecoverPasswordRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RecoverPasswordRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecoverPasswordRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
