% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{ReportMissingCodeRequest}
\alias{ReportMissingCodeRequest}
\title{ReportMissingCodeRequest R6 class}
\description{
Represents the TLRequest auth.ReportMissingCodeRequest.
}
\details{
Fields:
}
\section{initialize}{

 Create a new ReportMissingCodeRequest object.
 @param phone_number character
 @param phone_code_hash character
 @param mnc character
}

\section{to_list}{

 Return a pure R list representation of the object.
 @return list
}

\section{to_bytes}{

 Serialize the object into raw bytes for TL. Uses serialize_bytes helper
 from TL base environment.
 @return raw vector
}

\section{from_reader}{

 Factory that reads fields from a reader and returns a new ReportMissingCodeRequest instance.
 @param reader Reader object implementing read_int(), tgread_string(), tgread_object()
 @return ReportMissingCodeRequest instance
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0xcb9deff6).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0xf5b399ac).}

\item{\code{phone_number}}{character Phone number.}

\item{\code{phone_code_hash}}{character Phone code hash.}

\item{\code{mnc}}{character Mobile network code.

Methods:}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReportMissingCodeRequest-new}{\code{ReportMissingCodeRequest$new()}}
\item \href{#method-ReportMissingCodeRequest-to_list}{\code{ReportMissingCodeRequest$to_list()}}
\item \href{#method-ReportMissingCodeRequest-to_bytes}{\code{ReportMissingCodeRequest$to_bytes()}}
\item \href{#method-ReportMissingCodeRequest-from_reader}{\code{ReportMissingCodeRequest$from_reader()}}
\item \href{#method-ReportMissingCodeRequest-clone}{\code{ReportMissingCodeRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportMissingCodeRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReportMissingCodeRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportMissingCodeRequest$new(phone_number, phone_code_hash, mnc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phone_number}}{character}

\item{\code{phone_code_hash}}{character}

\item{\code{mnc}}{character
Convert to an R list (similar to to_dict)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportMissingCodeRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ReportMissingCodeRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportMissingCodeRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportMissingCodeRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ReportMissingCodeRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportMissingCodeRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
Read from reader and create ReportMissingCodeRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportMissingCodeRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ReportMissingCodeRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportMissingCodeRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ReportMissingCodeRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportMissingCodeRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReportMissingCodeRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportMissingCodeRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
