% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{ResendCodeRequest}
\alias{ResendCodeRequest}
\title{ResendCodeRequest}
\description{
Telegram API type ResendCodeRequest
}
\details{
ResendCodeRequest R6 class

Represents the TLRequest auth.ResendCodeRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}

\item{\code{phone_number}}{Field.}

\item{\code{phone_code_hash}}{Field.}

\item{\code{reason}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResendCodeRequest-new}{\code{ResendCodeRequest$new()}}
\item \href{#method-ResendCodeRequest-to_list}{\code{ResendCodeRequest$to_list()}}
\item \href{#method-ResendCodeRequest-to_bytes}{\code{ResendCodeRequest$to_bytes()}}
\item \href{#method-ResendCodeRequest-from_reader}{\code{ResendCodeRequest$from_reader()}}
\item \href{#method-ResendCodeRequest-clone}{\code{ResendCodeRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResendCodeRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResendCodeRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ResendCodeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResendCodeRequest$new(phone_number, phone_code_hash, reason = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phone_number}}{character}

\item{\code{phone_code_hash}}{character}

\item{\code{reason}}{character or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResendCodeRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ResendCodeRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResendCodeRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResendCodeRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ResendCodeRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResendCodeRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResendCodeRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ResendCodeRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create ResendCodeRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResendCodeRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ResendCodeRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResendCodeRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResendCodeRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResendCodeRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
