% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{ResetAuthorizationsRequest}
\alias{ResetAuthorizationsRequest}
\title{ResetAuthorizationsRequest}
\description{
Telegram API type ResetAuthorizationsRequest
}
\details{
ResetAuthorizationsRequest R6 class

Represents the TLRequest auth.ResetAuthorizationsRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Constructor identifier for this TL object.}

\item{\code{SUBCLASS_OF_ID}}{Subclass identifier for this TL object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResetAuthorizationsRequest-new}{\code{ResetAuthorizationsRequest$new()}}
\item \href{#method-ResetAuthorizationsRequest-to_list}{\code{ResetAuthorizationsRequest$to_list()}}
\item \href{#method-ResetAuthorizationsRequest-to_bytes}{\code{ResetAuthorizationsRequest$to_bytes()}}
\item \href{#method-ResetAuthorizationsRequest-from_reader}{\code{ResetAuthorizationsRequest$from_reader()}}
\item \href{#method-ResetAuthorizationsRequest-clone}{\code{ResetAuthorizationsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetAuthorizationsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResetAuthorizationsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ResetAuthorizationsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResetAuthorizationsRequest$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetAuthorizationsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ResetAuthorizationsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResetAuthorizationsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetAuthorizationsRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-ResetAuthorizationsRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResetAuthorizationsRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetAuthorizationsRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-ResetAuthorizationsRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create ResetAuthorizationsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResetAuthorizationsRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int(), tgread_string(), tgread_object()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ResetAuthorizationsRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetAuthorizationsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResetAuthorizationsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResetAuthorizationsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
