% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/requests.R
\name{RpcDropAnswerRequest}
\alias{RpcDropAnswerRequest}
\title{RpcDropAnswerRequest R6 class}
\value{
An R6 object of class \code{RpcDropAnswerRequest}.
}
\description{
Methods:
- new(req_msg_id): create new instance.
- to_list(): return an R list representation.
- to_raw(): serialize to raw vector (bytes) in little endian as used in TL.
- from_reader(reader): class method, construct instance from a `reader` object
  that exposes `read_long()` (returns numeric/integer) and similar methods.

Note: to_raw() uses writeBin on numeric for 8-byte values. If exact 64-bit
two's-complement preservation is required, replace with a dedicated 64-bit
integer serialization utility.

Initialize a RpcDropAnswerRequest

R6 class representing a request to drop an RPC answer for a previously sent
request message. The server will respond with one of the \code{RpcAnswer*}
response variants (e.g. \code{RpcAnswerUnknown}, \code{RpcAnswerDroppedRunning},
\code{RpcAnswerDropped}).
}
\details{
Represents the TL request `RpcDropAnswerRequest`.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{req_msg_id}}{integer|numeric Request message id (64-bit).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RpcDropAnswerRequest-new}{\code{RpcDropAnswerRequest$new()}}
\item \href{#method-RpcDropAnswerRequest-to_list}{\code{RpcDropAnswerRequest$to_list()}}
\item \href{#method-RpcDropAnswerRequest-to_bytes}{\code{RpcDropAnswerRequest$to_bytes()}}
\item \href{#method-RpcDropAnswerRequest-clone}{\code{RpcDropAnswerRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$new(req_msg_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$to_bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{Integer constructor identifier.}

\item{\code{SUBCLASS_OF_ID}}{Integer subclass identifier.}

\item{\code{req_msg_id}}{Integer64 request message identifier whose answer should be dropped.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RpcDropAnswerRequest-new}{\code{RpcDropAnswerRequest$new()}}
\item \href{#method-RpcDropAnswerRequest-to_list}{\code{RpcDropAnswerRequest$to_list()}}
\item \href{#method-RpcDropAnswerRequest-to_bytes}{\code{RpcDropAnswerRequest$to_bytes()}}
\item \href{#method-RpcDropAnswerRequest-clone}{\code{RpcDropAnswerRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$new(req_msg_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$to_bytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RpcDropAnswerRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RpcDropAnswerRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RpcDropAnswerRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
