% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{SaveAppLogRequest}
\alias{SaveAppLogRequest}
\title{SaveAppLogRequest}
\description{
Telegram API type SaveAppLogRequest
}
\details{
SaveAppLogRequest R6 class

Request to save application log events.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{events}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SaveAppLogRequest-new}{\code{SaveAppLogRequest$new()}}
\item \href{#method-SaveAppLogRequest-to_list}{\code{SaveAppLogRequest$to_list()}}
\item \href{#method-SaveAppLogRequest-to_bytes}{\code{SaveAppLogRequest$to_bytes()}}
\item \href{#method-SaveAppLogRequest-clone}{\code{SaveAppLogRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveAppLogRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SaveAppLogRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize SaveAppLogRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveAppLogRequest$new(events = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{events}}{list of TLObject-like events
Convert to list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveAppLogRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SaveAppLogRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveAppLogRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with a `_` discriminator and events as lists (if possible)
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveAppLogRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SaveAppLogRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveAppLogRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveAppLogRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SaveAppLogRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveAppLogRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
