% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{SaveBigFilePartRequest}
\alias{SaveBigFilePartRequest}
\title{SaveBigFilePartRequest R6 class}
\description{
Represents the TL request upload.SaveBigFilePartRequest.
}
\details{
Fields:
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file_id}}{numeric 64-bit file id (stored as numeric)}

\item{\code{file_part}}{integer part index}

\item{\code{file_total_parts}}{integer total number of parts}

\item{\code{bytes_data}}{raw|integer|character payload bytes

Methods:}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SaveBigFilePartRequest-new}{\code{SaveBigFilePartRequest$new()}}
\item \href{#method-SaveBigFilePartRequest-to_list}{\code{SaveBigFilePartRequest$to_list()}}
\item \href{#method-SaveBigFilePartRequest-to_bytes}{\code{SaveBigFilePartRequest$to_bytes()}}
\item \href{#method-SaveBigFilePartRequest-from_reader}{\code{SaveBigFilePartRequest$from_reader()}}
\item \href{#method-SaveBigFilePartRequest-clone}{\code{SaveBigFilePartRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveBigFilePartRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SaveBigFilePartRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveBigFilePartRequest$new(file_id, file_part, file_total_parts, bytes_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_id}}{numeric (64-bit)}

\item{\code{file_part}}{integer}

\item{\code{file_total_parts}}{integer}

\item{\code{bytes_data}}{raw|integer|character
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveBigFilePartRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SaveBigFilePartRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveBigFilePartRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveBigFilePartRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SaveBigFilePartRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveBigFilePartRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveBigFilePartRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-SaveBigFilePartRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveBigFilePartRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveBigFilePartRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SaveBigFilePartRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveBigFilePartRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
