% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bots.R
\name{SetBotGroupDefaultAdminRightsRequest}
\alias{SetBotGroupDefaultAdminRightsRequest}
\title{SetBotGroupDefaultAdminRightsRequest R6 class}
\description{
Methods:
- initialize(admin_rights)
- to_list() -> list representation
- to_bytes() -> raw vector bytes

Each method is documented inline below.
}
\details{
Represents the SetBotGroupDefaultAdminRightsRequest TL request.
}
\examples{
# o <- SetBotGroupDefaultAdminRightsRequest$new(admin_rights)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{admin_rights}}{TypeChatAdminRights object or representation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SetBotGroupDefaultAdminRightsRequest-new}{\code{SetBotGroupDefaultAdminRightsRequest$new()}}
\item \href{#method-SetBotGroupDefaultAdminRightsRequest-to_list}{\code{SetBotGroupDefaultAdminRightsRequest$to_list()}}
\item \href{#method-SetBotGroupDefaultAdminRightsRequest-to_bytes}{\code{SetBotGroupDefaultAdminRightsRequest$to_bytes()}}
\item \href{#method-SetBotGroupDefaultAdminRightsRequest-clone}{\code{SetBotGroupDefaultAdminRightsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotGroupDefaultAdminRightsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotGroupDefaultAdminRightsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize SetBotGroupDefaultAdminRightsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotGroupDefaultAdminRightsRequest$new(admin_rights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{admin_rights}}{TypeChatAdminRights or object
Convert to list (dictionary-like)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotGroupDefaultAdminRightsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotGroupDefaultAdminRightsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotGroupDefaultAdminRightsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Serialize to bytes (raw vector)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotGroupDefaultAdminRightsRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotGroupDefaultAdminRightsRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotGroupDefaultAdminRightsRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetBotGroupDefaultAdminRightsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SetBotGroupDefaultAdminRightsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SetBotGroupDefaultAdminRightsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
