% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auth.R
\name{SignUpRequest}
\alias{SignUpRequest}
\title{SignUpRequest R6 class}
\description{
SignUpRequest R6 class

SignUpRequest R6 class
}
\details{
Represents the TLRequest auth.SignUpRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CONSTRUCTOR_ID}}{integer Constructor id (hex 0xaac7b717).}

\item{\code{SUBCLASS_OF_ID}}{integer Subclass id (hex 0xb9e04e39).}

\item{\code{phone_number}}{character Phone number.}

\item{\code{phone_code_hash}}{character Phone code hash.}

\item{\code{first_name}}{character First name.}

\item{\code{last_name}}{character Last name.}

\item{\code{no_joined_notifications}}{logical Optional flag.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SignUpRequest-new}{\code{SignUpRequest$new()}}
\item \href{#method-SignUpRequest-to_list}{\code{SignUpRequest$to_list()}}
\item \href{#method-SignUpRequest-to_bytes}{\code{SignUpRequest$to_bytes()}}
\item \href{#method-SignUpRequest-from_reader}{\code{SignUpRequest$from_reader()}}
\item \href{#method-SignUpRequest-clone}{\code{SignUpRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignUpRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-SignUpRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a SignUpRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignUpRequest$new(
  phone_number,
  phone_code_hash,
  first_name,
  last_name,
  no_joined_notifications = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phone_number}}{character}

\item{\code{phone_code_hash}}{character}

\item{\code{first_name}}{character}

\item{\code{last_name}}{character}

\item{\code{no_joined_notifications}}{logical or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignUpRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SignUpRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to an R list (similar to to_dict)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignUpRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignUpRequest-to_bytes"></a>}}
\if{latex}{\out{\hypertarget{method-SignUpRequest-to_bytes}{}}}
\subsection{Method \code{to_bytes()}}{
Serialize to bytes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignUpRequest$to_bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignUpRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-SignUpRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
Read from reader and create SignUpRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignUpRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reader}}{Reader object implementing read_int() and tgread_string()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
SignUpRequest instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignUpRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SignUpRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignUpRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
