% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpmtproxy.R
\name{TcpMTProxy}
\alias{TcpMTProxy}
\title{TcpMTProxy}
\description{
Connector for MTProxy, handling proxy connections and secret normalization.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TcpMTProxy-new}{\code{TcpMTProxy$new()}}
\item \href{#method-TcpMTProxy-connect}{\code{TcpMTProxy$connect()}}
\item \href{#method-TcpMTProxy-clone}{\code{TcpMTProxy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TcpMTProxy-new"></a>}}
\if{latex}{\out{\hypertarget{method-TcpMTProxy-new}{}}}
\subsection{Method \code{new()}}{
Initialize TcpMTProxy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TcpMTProxy$new(ip, port, dc_id, loggers = NULL, proxy = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ip}}{IP address as character.}

\item{\code{port}}{Port number as integer.}

\item{\code{dc_id}}{Data center identifier as integer.}

\item{\code{loggers}}{Optional logger.}

\item{\code{proxy}}{A list containing proxy_host, proxy_port, and secret.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TcpMTProxy-connect"></a>}}
\if{latex}{\out{\hypertarget{method-TcpMTProxy-connect}{}}}
\subsection{Method \code{connect()}}{
Asynchronously connect to the proxy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TcpMTProxy$connect(timeout = NULL, ssl = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeout}}{Timeout value in seconds.}

\item{\code{ssl}}{Logical indicating whether to use SSL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TcpMTProxy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TcpMTProxy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TcpMTProxy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
