% Generated manually
\name{TelegramBaseClient}
\alias{TelegramBaseClient}
\title{TelegramBaseClient}
\description{
Low-level Telegram MTProto client. Most users should use \code{TelegramClient}.
}
\details{
This is an R6 class providing lower-level transport and protocol operations.
}
\usage{TelegramBaseClient}
\seealso{\code{\link{TelegramClient}}}
