% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegramclient.R
\name{TelegramClient}
\alias{TelegramClient}
\title{TelegramClient Class}
\description{
TelegramClient Class

TelegramClient Class
}
\details{
An R6 class that combines functionality from multiple method classes to interact with Telegram API
}
\section{Super class}{
\code{\link[telegramR:TelegramBaseClient]{telegramR::TelegramBaseClient}} -> \code{TelegramClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TelegramClient-new}{\code{TelegramClient$new()}}
\item \href{#method-TelegramClient-clone}{\code{TelegramClient$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="connect"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-connect'><code>telegramR::TelegramBaseClient$connect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="disconnect"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-disconnect'><code>telegramR::TelegramBaseClient$disconnect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="get_flood_sleep_threshold"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-get_flood_sleep_threshold'><code>telegramR::TelegramBaseClient$get_flood_sleep_threshold()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="get_version"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-get_version'><code>telegramR::TelegramBaseClient$get_version()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="is_connected"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-is_connected'><code>telegramR::TelegramBaseClient$is_connected()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="set_flood_sleep_threshold"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-set_flood_sleep_threshold'><code>telegramR::TelegramBaseClient$set_flood_sleep_threshold()</code></a></span></li>
<li><span class="pkg-link" data-pkg="telegramR" data-topic="TelegramBaseClient" data-id="set_proxy"><a href='../../telegramR/html/TelegramBaseClient.html#method-TelegramBaseClient-set_proxy'><code>telegramR::TelegramBaseClient$set_proxy()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-new}{}}}
\subsection{Method \code{new()}}{
Create a new TelegramClient instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to the parent class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
