% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegramclient.R
\name{TelegramClient}
\alias{TelegramClient}
\title{TelegramClient Class}
\description{
TelegramClient Class

TelegramClient Class
}
\details{
An R6 class that combines functionality from multiple method classes to interact with Telegram API
}
\section{Super class}{
\code{base::NA} -> \code{TelegramClient}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{account}}{An instance of AccountMethods for account-related operations}

\item{\code{auth}}{An instance of AuthMethods for authentication-related operations}

\item{\code{download}}{An instance of DownloadMethods for downloading media}

\item{\code{dialog}}{An instance of DialogMethods for dialog-related operations}

\item{\code{chat}}{An instance of ChatMethods for chat-related operations}

\item{\code{bot}}{An instance of BotMethods for bot-related operations}

\item{\code{message}}{An instance of MessageMethods for message-related operations}

\item{\code{upload}}{An instance of UploadMethods for uploading media}

\item{\code{button}}{An instance of ButtonMethods for button-related operations}

\item{\code{update}}{An instance of UpdateMethods for update-related operations}

\item{\code{parser}}{An instance of MessageParseMethods for parsing messages}

\item{\code{user}}{An instance of UserMethods for user-related operations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TelegramClient-new}{\code{TelegramClient$new()}}
\item \href{#method-TelegramClient-clone}{\code{TelegramClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-new}{}}}
\subsection{Method \code{new()}}{
Create a new TelegramClient instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to the parent class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
