% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_profile.R
\name{update_username}
\alias{update_username}
\title{Update your username}
\usage{
update_username(client, username)
}
\arguments{
\item{client}{TelegramClient instance.}

\item{username}{New username (without @).}
}
\value{
The API response.
}
\description{
Update your username
}
