% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_smsjobs.R
\name{UpdateSettingsRequest}
\alias{UpdateSettingsRequest}
\title{UpdateSettingsRequest}
\description{
Telegram API type UpdateSettingsRequest
}
\details{
UpdateSettingsRequest R6 class

Representation of the UpdateSettingsRequest TLRequest.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{allow_international}}{Field.}

\item{\code{lock_class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{lock_class}}{Field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UpdateSettingsRequest-new}{\code{UpdateSettingsRequest$new()}}
\item \href{#method-UpdateSettingsRequest-to_list}{\code{UpdateSettingsRequest$to_list()}}
\item \href{#method-UpdateSettingsRequest-bytes}{\code{UpdateSettingsRequest$bytes()}}
\item \href{#method-UpdateSettingsRequest-from_reader}{\code{UpdateSettingsRequest$from_reader()}}
\item \href{#method-UpdateSettingsRequest-clone}{\code{UpdateSettingsRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateSettingsRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateSettingsRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize UpdateSettingsRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateSettingsRequest$new(allow_international = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_international}}{logical or NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateSettingsRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateSettingsRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to list (dictionary-like)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateSettingsRequest$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateSettingsRequest-bytes"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateSettingsRequest-bytes}{}}}
\subsection{Method \code{bytes()}}{
Get raw bytes for this constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateSettingsRequest$bytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateSettingsRequest-from_reader"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateSettingsRequest-from_reader}{}}}
\subsection{Method \code{from_reader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateSettingsRequest$from_reader(reader)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateSettingsRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateSettingsRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateSettingsRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
